Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- RAND_UNIFORM$instantiation()

# Define a closed form expression
cf <- CLOSED_FORM$instantiation(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1]  8.736533 10.404916  9.273406  8.887594  9.474172

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  8.594188 10.744404  9.218781  8.693085  8.608401
#> 
#> $rhs
#> $rhs$b
#> [1] 0.1464474 0.8631860 0.3546452 0.1794133 0.1511853
#> 
#> 

d <- RAND_NORMAL$instantiation()

# Define a closed form expression with another closed form expression
cf2 <- CLOSED_FORM$instantiation(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1] 12.126557 10.611815  7.795485 11.810744 14.877491

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 11.846011  6.610608 10.870915  8.420365 10.410281
#> 
#> $rhs
#> $rhs$b
#> [1] 0.8511970 0.9619494 0.8479353 0.2814064 0.7022692
#> 
#> $rhs$cf
#> [1] 10.708436 11.040694 10.698651  8.999065 10.261653
#> 
#> $rhs$d
#> [1]  0.37919142 -1.47669521  0.05742110 -0.19289972  0.04954269
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- CLOSED_FORM$instantiation(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  0.7785321 -2.1004735 -0.4816856 -2.6520733 -1.8795318
#> 
#> $rhs
#> $rhs$d
#> [1]  0.3892660 -1.0502368 -0.2408428 -1.3260367 -0.9397659
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- CLOSED_FORM$instantiation(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -3.615897 -2.655734  1.379416 -1.554384  8.810995
#> 
#> $rhs
#> $rhs$d
#> [1] -0.9039743 -0.6639336  0.3448540 -0.3885960  2.2027487
#> 
#> $rhs$cf3
#> [1] -1.8079486 -1.3278672  0.6897079 -0.7771921  4.4054974
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- CLOSED_FORM$instantiation(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 0.18380657 4.83682055 0.01295284 9.41592559 5.13511621
#> 
#> $rhs
#> $rhs$d
#> [1] 0.030634428 0.806136758 0.002158806 1.569320931 0.855852701
#> 
#> $rhs$cf3
#> [1] 0.061268855 1.612273515 0.004317612 3.138641862 1.711705402
#> 
#> $rhs$cf4
#> [1] 0.122537710 3.224547031 0.008635223 6.277283724 3.423410804
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>