Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>            y          x           e      .resid  .fitted test_name   statistic
#> 1  2.9923469 0.90402308  0.27106612  0.11875133 2.873596    F-test  0.83195425
#> 2  0.9852594 0.02907062 -0.04465635 -0.46188929 1.447149    F-test  0.83195425
#> 3  1.2572615 0.66452484 -0.84885660 -1.22587684 2.483138    F-test  0.83195425
#> 4  3.0140831 0.93682869  0.19960640  0.08700405 2.927079    F-test  0.83195425
#> 5  3.4864424 0.61963672  1.48285604  1.07648579 2.409957    F-test  0.83195425
#> 6  3.4828152 0.60122062  1.52012838  1.10288262 2.379933    F-test  0.83195425
#> 7  1.3970630 0.73158671 -0.86974279 -1.19540720 2.592470    F-test  0.83195425
#> 8  2.6603870 0.67335731  0.53361961  0.16284894 2.497538    F-test  0.83195425
#> 9  3.6685588 0.72412868  1.42006781  1.08824753 2.580311    F-test  0.83195425
#> 10 2.1926107 0.94822440 -0.65474321 -0.75304691 2.945658    F-test  0.83195425
#> 11 2.3972523 0.90402308  0.27106612 -0.47634330 2.873596    F-test  0.02596382
#> 12 1.3369557 0.02907062 -0.04465635 -0.11019299 1.447149    F-test  0.02596382
#> 13 3.1905874 0.66452484 -0.84885660  0.70744907 2.483138    F-test  0.02596382
#> 14 3.3115622 0.93682869  0.19960640  0.38448319 2.927079    F-test  0.02596382
#> 15 0.6535535 0.61963672  1.48285604 -1.75640309 2.409957    F-test  0.02596382
#> 16 4.1695997 0.60122062  1.52012838  1.78966708 2.379933    F-test  0.02596382
#> 17 2.3122257 0.73158671 -0.86974279 -0.28024450 2.592470    F-test  0.02596382
#> 18 2.2421326 0.67335731  0.53361961 -0.25540541 2.497538    F-test  0.02596382
#> 19 2.5994107 0.72412868  1.42006781  0.01909944 2.580311    F-test  0.02596382
#> 20 2.9235481 0.94822440 -0.65474321 -0.02210948 2.945658    F-test  0.02596382
#> 21 1.9995186 0.90402308  0.27106612 -0.87407702 2.873596    F-test 23.14803186
#> 22 0.3888705 0.02907062 -0.04465635 -1.05827817 1.447149    F-test 23.14803186
#> 23 2.8736313 0.66452484 -0.84885660  0.39049291 2.483138    F-test 23.14803186
#> 24 2.5031392 0.93682869  0.19960640 -0.42393987 2.927079    F-test 23.14803186
#> 25 3.4815501 0.61963672  1.48285604  1.07159346 2.409957    F-test 23.14803186
#> 26 2.5210636 0.60122062  1.52012838  0.14113098 2.379933    F-test 23.14803186
#> 27 3.6637241 0.73158671 -0.86974279  1.07125383 2.592470    F-test 23.14803186
#> 28 2.8838551 0.67335731  0.53361961  0.38631707 2.497538    F-test 23.14803186
#> 29 3.3213293 0.72412868  1.42006781  0.74101802 2.580311    F-test 23.14803186
#> 30 1.5001464 0.94822440 -0.65474321 -1.44551122 2.945658    F-test 23.14803186
#>        p_value k  null
#> 1  0.392037056 1 FALSE
#> 2  0.392037056 1 FALSE
#> 3  0.392037056 1 FALSE
#> 4  0.392037056 1 FALSE
#> 5  0.392037056 1 FALSE
#> 6  0.392037056 1 FALSE
#> 7  0.392037056 1 FALSE
#> 8  0.392037056 1 FALSE
#> 9  0.392037056 1 FALSE
#> 10 0.392037056 1 FALSE
#> 11 0.876540588 2  TRUE
#> 12 0.876540588 2  TRUE
#> 13 0.876540588 2  TRUE
#> 14 0.876540588 2  TRUE
#> 15 0.876540588 2  TRUE
#> 16 0.876540588 2  TRUE
#> 17 0.876540588 2  TRUE
#> 18 0.876540588 2  TRUE
#> 19 0.876540588 2  TRUE
#> 20 0.876540588 2  TRUE
#> 21 0.001940952 3  TRUE
#> 22 0.001940952 3  TRUE
#> 23 0.001940952 3  TRUE
#> 24 0.001940952 3  TRUE
#> 25 0.001940952 3  TRUE
#> 26 0.001940952 3  TRUE
#> 27 0.001940952 3  TRUE
#> 28 0.001940952 3  TRUE
#> 29 0.001940952 3  TRUE
#> 30 0.001940952 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e         x     .resid  .fitted test_name  statistic   p_value k
#> 1   2.558938  1 0.3994098 -4.5421654 7.101103    F-test 0.09915187 0.7620284 3
#> 2   3.814079  2 0.5315420 -3.5829418 7.397021    F-test 0.09915187 0.7620284 3
#> 3   4.148767  3 0.1314798 -2.3522926 6.501059    F-test 0.09915187 0.7620284 3
#> 4   5.343061  4 0.2701046 -1.4684561 6.811517    F-test 0.09915187 0.7620284 3
#> 5   7.726967  5 0.9060467 -0.5087773 8.235745    F-test 0.09915187 0.7620284 3
#> 6   7.494598  6 0.3629008  0.4752583 7.019340    F-test 0.09915187 0.7620284 3
#> 7   8.430156  7 0.3247158  1.4963341 6.933822    F-test 0.09915187 0.7620284 3
#> 8  10.424981  8 0.7942104  2.4396995 7.985281    F-test 0.09915187 0.7620284 3
#> 9  10.542037  9 0.3899645  3.4620867 7.079950    F-test 0.09915187 0.7620284 3
#> 10 11.247124 10 0.2050705  4.5812546 6.665870    F-test 0.09915187 0.7620284 3
#> 11  8.076742  1 0.3994098  0.9756385 7.101103    F-test 0.04408842 0.8396698 1
#> 12  5.675946  2 0.5315420 -1.7210744 7.397021    F-test 0.04408842 0.8396698 1
#> 13  3.963732  3 0.1314798 -2.5373271 6.501059    F-test 0.04408842 0.8396698 1
#> 14  9.169233  4 0.2701046  2.3577155 6.811517    F-test 0.04408842 0.8396698 1
#> 15 10.897881  5 0.9060467  2.6621367 8.235745    F-test 0.04408842 0.8396698 1
#> 16  6.338115  6 0.3629008 -0.6812245 7.019340    F-test 0.04408842 0.8396698 1
#> 17 12.717060  7 0.3247158  5.7832385 6.933822    F-test 0.04408842 0.8396698 1
#> 18  5.219949  8 0.7942104 -2.7653320 7.985281    F-test 0.04408842 0.8396698 1
#> 19  2.811964  9 0.3899645 -4.2679865 7.079950    F-test 0.04408842 0.8396698 1
#> 20  6.860085 10 0.2050705  0.1942153 6.665870    F-test 0.04408842 0.8396698 1
#> 21 10.446511  1 0.3994098  3.3454076 7.101103    F-test 1.33559279 0.2857304 2
#> 22  9.590827  2 0.5315420  2.1938060 7.397021    F-test 1.33559279 0.2857304 2
#> 23  5.378271  3 0.1314798 -1.1227886 6.501059    F-test 1.33559279 0.2857304 2
#> 24  3.923386  4 0.2701046 -2.8881313 6.811517    F-test 1.33559279 0.2857304 2
#> 25  9.019373  5 0.9060467  0.7836284 8.235745    F-test 1.33559279 0.2857304 2
#> 26 10.802531  6 0.3629008  3.7831913 7.019340    F-test 1.33559279 0.2857304 2
#> 27 10.301708  7 0.3247158  3.3678862 6.933822    F-test 1.33559279 0.2857304 2
#> 28  4.013439  8 0.7942104 -3.9718420 7.985281    F-test 1.33559279 0.2857304 2
#> 29  4.534993  9 0.3899645 -2.5449572 7.079950    F-test 1.33559279 0.2857304 2
#> 30  3.719669 10 0.2050705 -2.9462004 6.665870    F-test 1.33559279 0.2857304 2
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE