Calculate p-value for multiple lineups.

This function calculates p-value for multiple lineups by using function calc_p_value(). If comb = TRUE, then function calc_p_value_comb() will be used instead.

calc_p_value_multi(
  dat,
  lineup_id = "lineup_id",
  detected = "detected",
  n_sel = "n_sel",
  comb = FALSE,
  n_eval = 1,
  replace_0 = TRUE,
  replace_full = TRUE,
  n_plot = 20,
  n_sim = 50000,
  cache_env = NULL
)

Arguments

dat

Data.frame/Tibble. A data.frame or a tibble.

lineup_id

Character. Column name of ids of lineup.

detected

Character. Column name of whether the lineup is detected by the subject.

n_sel

Character. Column name of the number of selections.

comb

Boolean. Whether to compute all the combinations. See also calc_p_value_comb().

n_eval

Integer. A vector of desired number of evaluations. Only used when comb = TRUE.

replace_0

Boolean. Whether to give treatment to evaluations with number of selections equal to 0.

replace_full

Boolean. Whether to give treatment to evaluations with number of selections equal to number of plots.

n_plot

Integer. Number of plots.

n_sim

Integer. Number of simulation draws.

cache_env

Environment. A provided environment for caching.

Value

If comb = TRUE, the function returns a tiible with columns lineup_id and p_value, where p_value is a list of vectors. If comb = FALSE, the p_value column is a vector.

Details

It is encouraged to provide a cache environment to boost up the performance. The cache environment will remember the result corresponding to the combinations of n_eval and n_sim. replace_0 and replace_full can be turned on to set evaluations with 0 or full selections to be false detection with only one selection.

Examples

dat <- data.frame(lineup_id = c(1,1,2),
                  detected = c(TRUE, FALSE, TRUE),
                  n_sel = c(1,1,2))
calc_p_value_multi(dat, comb = TRUE, n_eval = 1:2)
#> # A tibble: 3 x 4
#>   lineup_id p_value   n_eval total_eval
#>       <dbl> <list>     <int>      <int>
#> 1         1 <dbl [2]>      1          2
#> 2         2 <dbl [1]>      1          1
#> 3         1 <dbl [1]>      2          2
calc_p_value_multi(dat)
#> # A tibble: 2 x 2
#>   lineup_id p_value
#>       <dbl>   <dbl>
#> 1         1  0.0948
#> 2         2  0.102