Get the class method

This function construct an instance by using the provided class constructor, then get the method from the instance and set its evaluation environment to the container of the provided object environment.

class_method(env, cls, method_name, ..., container_name = "..method_env..")

Arguments

env

Environment. The object environment.

cls

Function. The class constructor.

method_name

Character. The method name.

...

Arguments passed to the class constructor.

container_name

Character. Container name of the object environment.

Value

The class method.

Details

If the body of the method depends on the arguments passed to the class constructor, those values needs to be provided in ....

Examples

# Define a derived class constructor
MY_CLASS <- function(..., env = new.env(parent = parent.frame())) {

  # Declare inheritance
  env <- inherit(env, BASE, "MY_CLASS", ...)

  # Methods of this derived class
  string_ <- function() "(MY_CLASS object)"

  # Register the method
  register_method(env, string = string_)

  # Return the instance environment
  return(env)
}

# Register the class constructor
register_class_ctor(MY_CLASS, "MY_CLASS", parent = BASE)

# Init an instance
my_instance <- MY_CLASS()

my_instance$string()
#> [1] "(MY_CLASS object)"

# Get the `string()` method from `BASE` class
BASE_string <- class_method(my_instance, BASE, "string")

BASE_string
#> function () 
#> paste0("<", self$class[1], " object>")
#> <environment: 0x7fe8c143fce0>

BASE_string()
#> [1] "<MY_CLASS object>"