Download the detailed information of lineups used in the polynomials study

This function downloads and returns a list containing the information of 588 lineups including their metadata and observations. See also polynomials.

Usage

get_polynomials_lineup()

Source

https://raw.githubusercontent.com/TengMCing/visage/master/data-raw/polynomials_lineup.rds

Value

A list with 588 items, where each item contains a named list metadata and a data.frame data:

Metadata contains metadata of the lineup:

type

Type of the model

formula

The main formula of the model

shape

Shape of the Hermite polynomials, please check POLY_MODEL$hermite

x_dist

Distribution of variable x

include_z

Whether to include variable z in the model

e_dist

Distribution of error term e

e_sigma

The standard deviation of the error term e

name

Name of the model

k

Number of residual plots in a lineup

n

Number of observations in a residual plot

effect_size

Effect size of the actual residual plot

answer

The answer of the lineup

Data contains data for drawing 20 residual plots:

y

Values of y

raw_x

Values of non-scaled x

x

Values of scaled x (between -1 to 1)

raw_z

Values of non-scaled z

z

Values of scaled z (between -1 to 1)

e

Values of the error term e

.resid

Value of residuals

.fitted

Fitted values

test_name

Name of the conventional test

statistic

Test statistics

p_value

P-value of the test

k

Position of the plot in a lineup

null

Whehter or not it is a null residual plot