Approximate the distribution of number of detections of a lineup with simulation

This function approximate the distribution of number of detections of a lineup for given number of evaluations, selections in each evaluation and plots in a lineup.

Usage

sim_dist(n_eval, n_sel, n_plot = 20, n_sim = 50000)

Arguments

n_eval

Integer. Number of evaluations.

n_sel

Integer. A vector of the number of selections.

n_plot

Integer. Number of plots in the lineup.

n_sim

Integer. Number of simulations draws.

Value

A named vector representing the probability mass function of the distribution.

Details

For a given lineup, plots are assumed to have weights \(W_i, i = 1, ..., N,\) where \(N\) is the number of plots, and \(W_i\) follows a uniform distribution independently. For each draw, weights for a lineup will be simulated. Then, for each evaluation of a draw, the function will sample same number of plots as the number of selection in the evaluation using the simulated weights without replacement. Finally, the distribution of the occurrences of plot 1 in a draw is the approximated distribution of number of detections of a lineup.

Examples

sim_dist(3, c(2,2,3))
#>       0       1       2       3 
#> 0.69912 0.25552 0.04234 0.00302 
sim_dist(1, c(1))
#>      0      1 
#> 0.9497 0.0503