Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 10.299767 10.610990  9.928266 11.045027 10.933550

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  8.761161 10.427850  8.385965 10.836072  9.525358
#> 
#> $rhs
#> $rhs$b
#> [1] 0.20210520 0.75766801 0.07703975 0.89374216 0.45683745
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1]  9.880694 13.271304 12.995929 12.012514  5.668541

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 10.812728 13.281212  7.674193  6.077962  7.691493
#> 
#> $rhs
#> $rhs$b
#> [1] 0.447802945 0.200969797 0.009704266 0.527646474 0.345226519
#> 
#> $rhs$cf
#> [1] 9.498254 8.757755 8.183958 9.737785 9.190525
#> 
#> $rhs$d
#> [1]  0.4381579  1.5078190 -0.1699217 -1.2199409 -0.4996774
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  3.3329892  0.8043213  2.8155017 -1.8133434  2.6818707
#> 
#> $rhs
#> $rhs$d
#> [1]  1.6664946  0.4021607  1.4077509 -0.9066717  1.3409353
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  0.656033 -1.729438 -4.446087  2.515464  4.655403
#> 
#> $rhs
#> $rhs$d
#> [1]  0.1640082 -0.4323594 -1.1115218  0.6288660  1.1638509
#> 
#> $rhs$cf3
#> [1]  0.3280165 -0.8647188 -2.2230437  1.2577320  2.3277017
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  -3.033614  -3.228193 -12.606794   4.543850  -8.668770
#> 
#> $rhs
#> $rhs$d
#> [1] -0.5056023 -0.5380321 -2.1011323  0.7573084 -1.4447951
#> 
#> $rhs$cf3
#> [1] -1.011205 -1.076064 -4.202265  1.514617 -2.889590
#> 
#> $rhs$cf4
#> [1] -2.022409 -2.152129 -8.404529  3.029234 -5.779180
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>