Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 10.299694  9.968732 10.488689  9.620319  9.957138

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  9.108933  9.234539  8.224437 10.289986  8.445484
#> 
#> $rhs
#> $rhs$b
#> [1] 0.31802929 0.35989777 0.02319721 0.71171334 0.09687942
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1] 10.818836 10.079797  6.538976  8.657741 12.109669

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 15.682353 13.603826  3.376833  7.921081  7.664068
#> 
#> $rhs
#> $rhs$b
#> [1] 0.58819202 0.49196656 0.54795792 0.74694000 0.08164131
#> 
#> $rhs$cf
#> [1]  9.919422  9.630745  9.798719 10.395665  8.399769
#> 
#> $rhs$d
#> [1]  1.9209770  1.3243603 -2.1406289 -0.8248616 -0.2452337
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  2.8835079 -1.5692354 -0.9462924 -2.3716680 -4.1755757
#> 
#> $rhs
#> $rhs$d
#> [1]  1.4417540 -0.7846177 -0.4731462 -1.1858340 -2.0877879
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -2.2067395  1.9661626  0.7215782  2.2635256  6.1143941
#> 
#> $rhs
#> $rhs$d
#> [1] -0.5516849  0.4915407  0.1803946  0.5658814  1.5285985
#> 
#> $rhs$cf3
#> [1] -1.1033698  0.9830813  0.3607891  1.1317628  3.0571970
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -1.896849  2.249998 -4.397445  5.216955 -4.287388
#> 
#> $rhs
#> $rhs$d
#> [1] -0.3161415  0.3749996 -0.7329075  0.8694925 -0.7145646
#> 
#> $rhs$cf3
#> [1] -0.6322830  0.7499992 -1.4658150  1.7389850 -1.4291292
#> 
#> $rhs$cf4
#> [1] -1.264566  1.499998 -2.931630  3.477970 -2.858258
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>