Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1]  8.580570 10.491488 10.947923  9.612318  8.377033

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  9.820120  8.307063  9.254005 10.348295  9.897292
#> 
#> $rhs
#> $rhs$b
#> [1] 0.55509163 0.05073902 0.36638635 0.73114986 0.58081543
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1]  7.405905  8.605803  2.173338 10.640622 12.179957

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 14.152338 16.217115 12.915264  7.496663  4.188830
#> 
#> $rhs
#> $rhs$b
#> [1] 0.95239005 0.61347039 0.03617756 0.94385663 0.78995900
#> 
#> $rhs$cf
#> [1] 11.012016  9.995257  8.263378 10.986415 10.524722
#> 
#> $rhs$d
#> [1]  1.046774  2.073953  1.550629 -1.163251 -2.111964
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  0.5094226  0.5124856  3.6288652 -3.4130771  4.4338568
#> 
#> $rhs
#> $rhs$d
#> [1]  0.2547113  0.2562428  1.8144326 -1.7065386  2.2169284
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -2.7672840  2.3607615 -4.4594548  0.3668064  2.2892951
#> 
#> $rhs
#> $rhs$d
#> [1] -0.69182101  0.59019038 -1.11486371  0.09170159  0.57232377
#> 
#> $rhs$cf3
#> [1] -1.3836420  1.1803808 -2.2297274  0.1834032  1.1446475
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  0.5035153 -4.3864000  2.2381792  2.7875446 10.3207744
#> 
#> $rhs
#> $rhs$d
#> [1]  0.08391922 -0.73106667  0.37302987  0.46459077  1.72012907
#> 
#> $rhs$cf3
#> [1]  0.1678384 -1.4621333  0.7460597  0.9291815  3.4402581
#> 
#> $rhs$cf4
#> [1]  0.3356769 -2.9242667  1.4921195  1.8583631  6.8805163
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>