Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1]  8.730178  9.872784  8.375920 10.448900  9.119745

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 10.414409  9.993539 10.466816  9.277689 10.824185
#> 
#> $rhs
#> $rhs$b
#> [1] 0.7531877 0.6128978 0.7706569 0.3742810 0.8897798
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1] 11.472919  9.596529 10.873776  7.357694 19.749831

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 11.421927 14.889540  7.266391  8.211639 10.310812
#> 
#> $rhs
#> $rhs$b
#> [1] 0.03710176 0.25721287 0.26614796 0.58599679 0.91915814
#> 
#> $rhs$cf
#> [1]  8.266151  8.926484  8.953289  9.912836 10.912320
#> 
#> $rhs$d
#> [1]  1.0519255  1.9876854 -0.5622995 -0.5670656 -0.2005027
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  1.6531912 -3.8492517  0.7027253  0.7425940  0.6000384
#> 
#> $rhs
#> $rhs$d
#> [1]  0.8265956 -1.9246259  0.3513626  0.3712970  0.3000192
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  1.176391 -4.720669  2.471896  4.200654  9.460267
#> 
#> $rhs
#> $rhs$d
#> [1]  0.2940977 -1.1801673  0.6179741  1.0501635  2.3650667
#> 
#> $rhs$cf3
#> [1]  0.5881954 -2.3603347  1.2359482  2.1003271  4.7301333
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  1.59091521 -0.02525087 -4.14100469  2.56756450 -3.56579732
#> 
#> $rhs
#> $rhs$d
#> [1]  0.265152535 -0.004208479 -0.690167448  0.427927417 -0.594299553
#> 
#> $rhs$cf3
#> [1]  0.530305070 -0.008416957 -1.380334895  0.855854834 -1.188599105
#> 
#> $rhs$cf4
#> [1]  1.06061014 -0.01683391 -2.76066979  1.71170967 -2.37719821
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>