Generating random values from the expression

This function generates random values from the expression. Random values will be generated independtly as long as they don't share the same symbol in the expression.

Arguments

n

Integer. Number of observations.

rhs

Boolean. Whether or not to keep the right hand side values of the expression.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- RAND_UNIFORM$instantiation()

# Define a closed form expression
cf <- CLOSED_FORM$instantiation(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 11.010426 10.170840  8.309093  9.643956  8.982610

# Generate 5 values, and keep RHS
cf$gen(5, rhs = TRUE)
#> $lhs
#> [1] 10.466548  8.811002 10.463364  8.985021  8.666516
#> 
#> $rhs
#> $rhs$b
#> [1] 0.7705676 0.2187189 0.7695063 0.2767253 0.1705570
#> 
#> 

d <- RAND_NORMAL$instantiation()

# Define a closed form expression with another closed form expression
cf2 <- CLOSED_FORM$instantiation(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1] 12.288573  9.726809  8.632150  8.160115  6.294046

# Generate 5 values, and keep RHS
cf2$gen(5, rhs = TRUE)
#> $lhs
#> [1] 10.183569  3.295011  8.192305  8.118646  9.922653
#> 
#> $rhs
#> $rhs$cf
#> $rhs$cf$b
#> [1] 0.4157632 0.1571073 0.1563594 0.3454534 0.9743058
#> 
#> 
#> $rhs$d
#> [1]  0.2604779 -1.7770522 -0.1438730 -0.3575198 -0.3850367
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- CLOSED_FORM$instantiation(~d * d)

# Both `d` will share the same values
cf3$gen(5, rhs = TRUE)
#> $lhs
#> [1] 4.2377404 2.2274887 1.2066247 0.3523011 0.7629823
#> 
#> $rhs
#> $rhs$d
#> [1] -2.0585773  1.4924774  1.0984647  0.5935496 -0.8734886
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- CLOSED_FORM$instantiation(~cf3 * cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs = TRUE)
#> $lhs
#> [1] 1.705847e-04 3.959580e-01 2.094124e+01 4.146599e-04 1.204101e-07
#> 
#> $rhs
#> $rhs$cf3
#> $rhs$cf3$d
#> [1] -0.11428390 -0.79325401 -2.13919615  0.14269968 -0.01862798
#> 
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- CLOSED_FORM$instantiation(~cf3 * cf4)

# Both `d` in `cf3` will share the same value, but is different with `d` in `cf4`
cf5$gen(5, rhs = TRUE)
#> $lhs
#> [1] 4.613427e-02 8.488113e-04 8.265951e-01 9.417128e-05 2.509453e+00
#> 
#> $rhs
#> $rhs$cf3
#> $rhs$cf3$d
#> [1]  0.4048541 -0.2077578  0.6046373  0.0297952 -1.0970330
#> 
#> 
#> $rhs$cf4
#> $rhs$cf4$cf3
#> $rhs$cf4$cf3$d
#> [1] -0.7283777 -0.3744761  1.2262407  0.5706982 -1.2016692
#> 
#> 
#> 
#>