Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1]  9.813065 10.599752 11.153466  9.938388  8.833995

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 10.169033 11.124924  9.624485  9.253514  9.948806
#> 
#> $rhs
#> $rhs$b
#> [1] 0.6713957 0.9900262 0.4898797 0.3662229 0.5979870
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1]  6.742092 12.325184  9.779852  7.587955  7.773059

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  9.941831  6.379883 13.151652 15.298761  5.648769
#> 
#> $rhs
#> $rhs$b
#> [1] 0.01849643 0.20541938 0.15075747 0.99759915 0.40326312
#> 
#> $rhs$cf
#> [1]  8.210335  8.771104  8.607118 11.147643  9.364635
#> 
#> $rhs$d
#> [1]  0.5771654 -0.7970734  1.5148448  1.3837060 -1.2386219
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -3.4027046 -1.4466144  0.2699191 -0.1524354  1.2047194
#> 
#> $rhs
#> $rhs$d
#> [1] -1.7013523 -0.7233072  0.1349596 -0.0762177  0.6023597
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  0.8365863  3.4022300 -3.9262931  1.2611847 -1.2970645
#> 
#> $rhs
#> $rhs$d
#> [1]  0.2091466  0.8505575 -0.9815733  0.3152962 -0.3242661
#> 
#> $rhs$cf3
#> [1]  0.4182931  1.7011150 -1.9631465  0.6305923 -0.6485322
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  5.75928218 -4.42633937  0.35253057 -2.34282121  0.08264526
#> 
#> $rhs
#> $rhs$d
#> [1]  0.95988036 -0.73772323  0.05875510 -0.39047020  0.01377421
#> 
#> $rhs$cf3
#> [1]  1.91976073 -1.47544646  0.11751019 -0.78094040  0.02754842
#> 
#> $rhs$cf4
#> [1]  3.83952145 -2.95089291  0.23502038 -1.56188080  0.05509684
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>