Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- RAND_UNIFORM$instantiation()

# Define a closed form expression
cf <- CLOSED_FORM$instantiation(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1]  9.333537  9.294093 10.837605 10.873525 10.818128

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  9.468549 10.010150 10.513162  9.619561 10.152280
#> 
#> $rhs
#> $rhs$b
#> [1] 0.4379012 0.6184350 0.7861055 0.4882384 0.6658116
#> 
#> 

d <- RAND_NORMAL$instantiation()

# Define a closed form expression with another closed form expression
cf2 <- CLOSED_FORM$instantiation(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1] 12.557650 10.554469  9.957628  9.472244  6.909465

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  7.865291  7.408152 10.767904  6.851972 11.809869
#> 
#> $rhs
#> $rhs$b
#> [1] 0.3769962 0.1972091 0.9792847 0.3335547 0.7779115
#> 
#> $rhs$cf
#> [1]  9.285834  8.746473 11.092700  9.155510 10.488580
#> 
#> $rhs$d
#> [1] -0.4735144 -0.4461068 -0.1082653 -0.7678459  0.4404297
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- CLOSED_FORM$instantiation(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -0.1088298 -1.9048907  3.1817097 -1.7026736  3.3643211
#> 
#> $rhs
#> $rhs$d
#> [1] -0.0544149 -0.9524454  1.5908548 -0.8513368  1.6821606
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- CLOSED_FORM$instantiation(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -0.4478959 -3.8364513 -2.4862029 -1.0980509  1.7125893
#> 
#> $rhs
#> $rhs$d
#> [1] -0.1119740 -0.9591128 -0.6215507 -0.2745127  0.4281473
#> 
#> $rhs$cf3
#> [1] -0.2239480 -1.9182256 -1.2431015 -0.5490255  0.8562946
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- CLOSED_FORM$instantiation(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  3.507017  2.376549 -1.601918 -2.284388 -1.278924
#> 
#> $rhs
#> $rhs$d
#> [1]  0.5845028  0.3960915 -0.2669864 -0.3807314 -0.2131541
#> 
#> $rhs$cf3
#> [1]  1.1690057  0.7921831 -0.5339727 -0.7614627 -0.4263081
#> 
#> $rhs$cf4
#> [1]  2.3380113  1.5843661 -1.0679455 -1.5229255 -0.8526162
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>