Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 8.300398 9.511534 9.131559 9.291747 8.991292

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  8.943354  9.633967 10.163608  9.072687 10.020766
#> 
#> $rhs
#> $rhs$b
#> [1] 0.2628362 0.4930404 0.6695875 0.3059470 0.6219736
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1] 11.803749  7.944224  8.003798  9.505910 12.088800

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 11.189886  9.837034  8.549043 16.409076  8.538539
#> 
#> $rhs
#> $rhs$b
#> [1] 0.1163603 0.6669287 0.7172548 0.8577353 0.9450329
#> 
#> $rhs$cf
#> [1]  8.503926 10.155632 10.306610 10.728051 10.989944
#> 
#> $rhs$d
#> [1]  0.8953198 -0.1061992 -0.5858555  1.8936750 -0.8171352
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -1.3260807 -0.1066304 -1.5408651 -2.6151111  1.1971806
#> 
#> $rhs
#> $rhs$d
#> [1] -0.66304037 -0.05331519 -0.77043253 -1.30755553  0.59859030
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -7.482020  2.630067 -2.478327 -1.464459  1.172896
#> 
#> $rhs
#> $rhs$d
#> [1] -1.8705050  0.6575167 -0.6195817 -0.3661147  0.2932240
#> 
#> $rhs$cf3
#> [1] -3.7410099  1.3150334 -1.2391635 -0.7322295  0.5864481
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]   3.4252294  -6.0624727   0.6096945  -8.5029511 -10.8609177
#> 
#> $rhs
#> $rhs$d
#> [1]  0.5708716 -1.0104121  0.1016158 -1.4171585 -1.8101530
#> 
#> $rhs$cf3
#> [1]  1.1417431 -2.0208242  0.2032315 -2.8343170 -3.6203059
#> 
#> $rhs$cf4
#> [1]  2.283486 -4.041648  0.406463 -5.668634 -7.240612
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>