Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- RAND_UNIFORM$instantiation()

# Define a closed form expression
cf <- CLOSED_FORM$instantiation(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1]  9.735472  8.412745  9.488554  9.508936 10.636754

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 10.763247  9.746489 10.193351  9.623151 10.365756
#> 
#> $rhs
#> $rhs$b
#> [1] 0.8694672 0.5305478 0.6795020 0.4894353 0.7369702
#> 
#> 

d <- RAND_NORMAL$instantiation()

# Define a closed form expression with another closed form expression
cf2 <- CLOSED_FORM$instantiation(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1] 12.664997  9.471498 16.671335 12.708179 11.739754

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  9.148821 11.422447 12.062549  9.223924 10.394759
#> 
#> $rhs
#> $rhs$b
#> [1] 0.19461987 0.33652143 0.65355753 0.02908544 0.90644696
#> 
#> $rhs$cf
#> [1]  8.738705  9.164410 10.115518  8.242102 10.874186
#> 
#> $rhs$d
#> [1]  0.1367054  0.7526792  0.6490102  0.3272740 -0.1598092
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- CLOSED_FORM$instantiation(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -1.6288773  0.7012105 -0.3353924 -1.2488007  1.7363038
#> 
#> $rhs
#> $rhs$d
#> [1] -0.8144387  0.3506053 -0.1676962 -0.6244004  0.8681519
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- CLOSED_FORM$instantiation(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -7.41533786  2.63647551  3.84824243  1.08798038 -0.01222782
#> 
#> $rhs
#> $rhs$d
#> [1] -1.853834464  0.659118878  0.962060608  0.271995094 -0.003056954
#> 
#> $rhs$cf3
#> [1] -3.707668928  1.318237756  1.924121215  0.543990188 -0.006113908
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- CLOSED_FORM$instantiation(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  4.6749090 -4.2940953  4.8192009  0.6377479  2.0132015
#> 
#> $rhs
#> $rhs$d
#> [1]  0.7791515 -0.7156825  0.8032002  0.1062913  0.3355336
#> 
#> $rhs$cf3
#> [1]  1.5583030 -1.4313651  1.6064003  0.2125826  0.6710672
#> 
#> $rhs$cf4
#> [1]  3.1166060 -2.8627302  3.2128006  0.4251653  1.3421344
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>