Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- RAND_UNIFORM$instantiation()

# Define a closed form expression
cf <- CLOSED_FORM$instantiation(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 10.194222  8.498901 10.871701  9.739792  9.152746

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 8.907524 9.521829 9.909696 8.487854 8.687116
#> 
#> $rhs
#> $rhs$b
#> [1] 0.2508928 0.4556610 0.5849502 0.1110028 0.1774236
#> 
#> 

d <- RAND_NORMAL$instantiation()

# Define a closed form expression with another closed form expression
cf2 <- CLOSED_FORM$instantiation(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1]  7.896556  8.428170 10.110837  7.004077 13.574517

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  5.661989 12.030573 11.161838 13.757583  5.367540
#> 
#> $rhs
#> $rhs$b
#> [1] 0.9936087 0.5240427 0.5103948 0.6421206 0.1603499
#> 
#> $rhs$cf
#> [1] 11.135672  9.726974  9.686030 10.081207  8.635895
#> 
#> $rhs$d
#> [1] -1.8245608  0.7678663  0.4919359  1.2254586 -1.0894516
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- CLOSED_FORM$instantiation(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  3.7072606  1.0223582 -1.0066186  0.4239765 -2.7156098
#> 
#> $rhs
#> $rhs$d
#> [1]  1.8536303  0.5111791 -0.5033093  0.2119883 -1.3578049
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- CLOSED_FORM$instantiation(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  2.7853037 -3.3805577 -8.3979241  7.1798752 -0.9724405
#> 
#> $rhs
#> $rhs$d
#> [1]  0.6963259 -0.8451394 -2.0994810  1.7949688 -0.2431101
#> 
#> $rhs$cf3
#> [1]  1.3926519 -1.6902789 -4.1989621  3.5899376 -0.4862203
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- CLOSED_FORM$instantiation(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -9.4228393  7.6207740 -0.8899273 -4.2608013 -9.8118773
#> 
#> $rhs
#> $rhs$d
#> [1] -1.5704732  1.2701290 -0.1483212 -0.7101335 -1.6353129
#> 
#> $rhs$cf3
#> [1] -3.1409464  2.5402580 -0.2966424 -1.4202671 -3.2706258
#> 
#> $rhs$cf4
#> [1] -6.2818928  5.0805160 -0.5932849 -2.8405342 -6.5412515
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>