Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 9.996550 9.377198 9.607804 8.366602 8.895924

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  8.759666 10.190394 11.144549  8.199652  8.762267
#> 
#> $rhs
#> $rhs$b
#> [1] 0.20160688 0.67851608 0.99656787 0.01493564 0.20247394
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1]  6.164394  7.368075  7.429594 12.608501  8.740480

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  3.042768 11.744538 12.567983 10.473553 10.426244
#> 
#> $rhs
#> $rhs$b
#> [1] 0.9867254 0.8573452 0.9323195 0.5679379 0.2688000
#> 
#> $rhs$cf
#> [1] 11.115022 10.726881 10.951804  9.858659  8.961245
#> 
#> $rhs$d
#> [1] -2.6907510  0.3392190  0.5387263  0.2049644  0.4883327
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -3.3494961  0.9826810 -5.3506900  0.8303152  0.4848374
#> 
#> $rhs
#> $rhs$d
#> [1] -1.6747480  0.4913405 -2.6753450  0.4151576  0.2424187
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -4.3865955 -0.8205749 -5.8272091 -5.5200441  0.6813035
#> 
#> $rhs
#> $rhs$d
#> [1] -1.0966489 -0.2051437 -1.4568023 -1.3800110  0.1703259
#> 
#> $rhs$cf3
#> [1] -2.1932977 -0.4102875 -2.9136046 -2.7600220  0.3406517
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  7.423801 -1.425465  2.135322 -9.109183 -6.031182
#> 
#> $rhs
#> $rhs$d
#> [1]  1.2373002 -0.2375775  0.3558870 -1.5181972 -1.0051971
#> 
#> $rhs$cf3
#> [1]  2.474600 -0.475155  0.711774 -3.036394 -2.010394
#> 
#> $rhs$cf4
#> [1]  4.9492009 -0.9503099  1.4235480 -6.0727889 -4.0207883
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>