Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1]  9.866227  9.533819  9.442240  8.640650 10.367487

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 8.185887 8.585152 8.445945 9.457648 8.549059
#> 
#> $rhs
#> $rhs$b
#> [1] 0.01034709 0.14343546 0.09703331 0.43426747 0.13140459
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1]  9.852648  9.974151 11.587360 13.097565 11.013894

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  7.176436  8.254419  9.099360 10.570456 11.560948
#> 
#> $rhs
#> $rhs$b
#> [1] 0.1114914 0.9970342 0.1151873 0.2817961 0.7971286
#> 
#> $rhs$cf
#> [1]  8.489320 11.145948  8.500407  9.000234 10.546231
#> 
#> $rhs$d
#> [1] -0.4376280 -0.9638429  0.1996510  0.5234075  0.3382390
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  1.4239150 -2.0235854  0.8160186 -0.8442326  3.6908621
#> 
#> $rhs
#> $rhs$d
#> [1]  0.7119575 -1.0117927  0.4080093 -0.4221163  1.8454310
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 4.2847485 0.5383889 1.1625429 3.6841114 1.4182360
#> 
#> $rhs
#> $rhs$d
#> [1] 1.0711871 0.1345972 0.2906357 0.9210278 0.3545590
#> 
#> $rhs$cf3
#> [1] 2.1423743 0.2691945 0.5812714 1.8420557 0.7091180
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 14.4711448 -2.7345010 -5.9153080  3.4273322  0.3241851
#> 
#> $rhs
#> $rhs$d
#> [1]  2.41185747 -0.45575017 -0.98588467  0.57122204  0.05403084
#> 
#> $rhs$cf3
#> [1]  4.8237149 -0.9115003 -1.9717693  1.1424441  0.1080617
#> 
#> $rhs$cf4
#> [1]  9.6474299 -1.8230007 -3.9435387  2.2848882  0.2161234
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>