Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1]  8.504926  9.035456  8.890373 10.934578  9.244574

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 11.047953 10.620983 10.825270  8.379911  8.546422
#> 
#> $rhs
#> $rhs$b
#> [1] 0.96436909 0.82204577 0.89014136 0.07502171 0.13052558
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1]  8.037778  5.976878 11.243668 12.316863 16.950176

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  9.413054  9.870053  7.431534 12.170943 11.194256
#> 
#> $rhs
#> $rhs$b
#> [1] 0.32925642 0.25755928 0.48150915 0.32947821 0.02731144
#> 
#> $rhs$cf
#> [1] 9.142615 8.927523 9.599373 9.143280 8.236780
#> 
#> $rhs$d
#> [1]  0.09014639  0.31417656 -0.72261312  1.00922110  0.98582530
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  2.5615607  1.6454271  1.1053771 -2.0013377  0.7364904
#> 
#> $rhs
#> $rhs$d
#> [1]  1.2807804  0.8227135  0.5526886 -1.0006688  0.3682452
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -1.475096 -2.787231  2.418054  1.914496  1.049173
#> 
#> $rhs
#> $rhs$d
#> [1] -0.3687739 -0.6968079  0.6045134  0.4786239  0.2622933
#> 
#> $rhs$cf3
#> [1] -0.7375478 -1.3936157  1.2090268  0.9572478  0.5245866
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -0.3516259  2.6644969 10.5016020  0.9307183  2.6740104
#> 
#> $rhs
#> $rhs$d
#> [1] -0.05860432  0.44408281  1.75026699  0.15511972  0.44566840
#> 
#> $rhs$cf3
#> [1] -0.1172086  0.8881656  3.5005340  0.3102394  0.8913368
#> 
#> $rhs$cf4
#> [1] -0.2344173  1.7763313  7.0010680  0.6204789  1.7826736
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>