Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 10.083071  8.937196  8.781802 10.474826 10.290357

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 10.378261  8.526266  8.198720  9.894443 10.101389
#> 
#> $rhs
#> $rhs$b
#> [1] 0.74113856 0.12380671 0.01462499 0.57986600 0.64884797
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1]  8.600739  9.455579  8.382511  7.181031 12.334919

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  6.485381  8.930756 12.106361  7.062871 11.128601
#> 
#> $rhs
#> $rhs$b
#> [1] 0.9875281 0.6805939 0.9788710 0.4168383 0.0668799
#> 
#> $rhs$cf
#> [1] 11.117430 10.196627 11.091458  9.405361  8.355485
#> 
#> $rhs$d
#> [1] -1.5440163 -0.4219570  0.3383008 -0.7808298  0.9243721
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -0.47141952 -1.65557201  1.45405312 -0.56348930 -0.04068322
#> 
#> $rhs
#> $rhs$d
#> [1] -0.23570976 -0.82778600  0.72702656 -0.28174465 -0.02034161
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -4.5309517 -2.8518840 -0.3722823 -2.2861048 -0.4953342
#> 
#> $rhs
#> $rhs$d
#> [1] -1.13273793 -0.71297099 -0.09307056 -0.57152620 -0.12383354
#> 
#> $rhs$cf3
#> [1] -2.2654759 -1.4259420 -0.1861411 -1.1430524 -0.2476671
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  1.999735 -9.055145 -7.227821 -2.795524  3.640855
#> 
#> $rhs
#> $rhs$d
#> [1]  0.3332891 -1.5091909 -1.2046368 -0.4659206  0.6068092
#> 
#> $rhs$cf3
#> [1]  0.6665782 -3.0183818 -2.4092735 -0.9318412  1.2136183
#> 
#> $rhs$cf4
#> [1]  1.333156 -6.036764 -4.818547 -1.863682  2.427237
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>