Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 10.612224  9.264712 11.022120  9.573208  9.884556

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 10.816373  8.208389  9.428574 10.673982 10.824328
#> 
#> $rhs
#> $rhs$b
#> [1] 0.88717572 0.01784792 0.42457627 0.83971209 0.88982747
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1] 7.413565 9.409794 7.593353 9.626572 6.096800

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 11.922953  7.744464 13.023004 16.452927  8.750067
#> 
#> $rhs
#> $rhs$b
#> [1] 0.9508887 0.8769301 0.9859285 0.6456592 0.8672203
#> 
#> $rhs$cf
#> [1] 11.00751 10.78564 11.11263 10.09182 10.75651
#> 
#> $rhs$d
#> [1]  0.3051472 -1.0137241  0.6367911  2.1203681 -0.6688132
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  0.7348426  0.1792436  0.6626880 -2.3172413 -1.1320679
#> 
#> $rhs
#> $rhs$d
#> [1]  0.3674213  0.0896218  0.3313440 -1.1586207 -0.5660340
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -3.576751  1.599913  1.767516 12.250369 -5.133075
#> 
#> $rhs
#> $rhs$d
#> [1] -0.8941877  0.3999782  0.4418790  3.0625924 -1.2832687
#> 
#> $rhs$cf3
#> [1] -1.7883755  0.7999564  0.8837580  6.1251847 -2.5665374
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  3.189835  2.716953 -4.957772  3.092746 -4.466523
#> 
#> $rhs
#> $rhs$d
#> [1]  0.5316391  0.4528255 -0.8262953  0.5154576 -0.7444206
#> 
#> $rhs$cf3
#> [1]  1.0632782  0.9056509 -1.6525906  1.0309152 -1.4888411
#> 
#> $rhs$cf4
#> [1]  2.126556  1.811302 -3.305181  2.061830 -2.977682
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>