Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- RAND_UNIFORM$instantiation()

# Define a closed form expression
cf <- CLOSED_FORM$instantiation(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 10.894910  8.483414 10.742744  8.527665 11.035049

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 10.239596 10.015338  9.930370  8.586752  8.187900
#> 
#> $rhs
#> $rhs$b
#> [1] 0.69491694 0.62016423 0.59184136 0.14396878 0.01101803
#> 
#> 

d <- RAND_NORMAL$instantiation()

# Define a closed form expression with another closed form expression
cf2 <- CLOSED_FORM$instantiation(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1]  5.303980 11.039621  3.109313 12.618713  9.536795

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 8.957806 7.195726 8.360213 7.167968 9.521663
#> 
#> $rhs
#> $rhs$b
#> [1] 0.4421652 0.8850594 0.1261303 0.6527459 0.9816537
#> 
#> $rhs$cf
#> [1]  9.481341 10.810024  8.533236 10.113083 11.099806
#> 
#> $rhs$d
#> [1] -0.17451180 -1.20476593 -0.05767451 -0.98170506 -0.52604779
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- CLOSED_FORM$instantiation(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  2.2166963 -0.7690226  0.6897635  1.2909091  0.1090786
#> 
#> $rhs
#> $rhs$d
#> [1]  1.10834816 -0.38451129  0.34488175  0.64545453  0.05453928
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- CLOSED_FORM$instantiation(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -5.1934864 -3.0943924 10.1486139 -0.1012362  4.1880750
#> 
#> $rhs
#> $rhs$d
#> [1] -1.29837159 -0.77359810  2.53715347 -0.02530905  1.04701875
#> 
#> $rhs$cf3
#> [1] -2.5967432 -1.5471962  5.0743069 -0.0506181  2.0940375
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- CLOSED_FORM$instantiation(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  3.9472542  9.0015336  0.1527773 -5.7874584  2.3285594
#> 
#> $rhs
#> $rhs$d
#> [1]  0.65787570  1.50025560  0.02546289 -0.96457640  0.38809323
#> 
#> $rhs$cf3
#> [1]  1.31575140  3.00051120  0.05092578 -1.92915279  0.77618646
#> 
#> $rhs$cf4
#> [1]  2.6315028  6.0010224  0.1018516 -3.8583056  1.5523729
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>