Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- RAND_UNIFORM$instantiation()

# Define a closed form expression
cf <- CLOSED_FORM$instantiation(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 10.544384  9.487726  8.383722  9.627029 10.311533

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  8.443388 10.611467  8.599006  8.830992  8.370293
#> 
#> $rhs
#> $rhs$b
#> [1] 0.09618082 0.81887371 0.14805336 0.22538223 0.07181591
#> 
#> 

d <- RAND_NORMAL$instantiation()

# Define a closed form expression with another closed form expression
cf2 <- CLOSED_FORM$instantiation(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1] 11.586543  9.878185 10.939350 10.482784 10.332948

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 13.259174  8.359825  9.123532  3.131745 13.722658
#> 
#> $rhs
#> $rhs$b
#> [1] 0.9046421 0.1229928 0.4670949 0.7521512 0.7931555
#> 
#> $rhs$cf
#> [1] 10.868772  8.523824  9.556130 10.411299 10.534312
#> 
#> $rhs$d
#> [1]  0.79680091 -0.05466649 -0.14419944 -2.42651795  1.06278207
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- CLOSED_FORM$instantiation(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  3.755486 -5.370205 -3.553076  1.885272  2.954869
#> 
#> $rhs
#> $rhs$d
#> [1]  1.8777432 -2.6851026 -1.7765380  0.9426362  1.4774343
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- CLOSED_FORM$instantiation(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -0.4585345  0.9126150 -1.8822751 -2.3424191 -1.4559741
#> 
#> $rhs
#> $rhs$d
#> [1] -0.1146336  0.2281538 -0.4705688 -0.5856048 -0.3639935
#> 
#> $rhs$cf3
#> [1] -0.2292673  0.4563075 -0.9411375 -1.1712096 -0.7279871
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- CLOSED_FORM$instantiation(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -9.7641412 -0.9501572 -1.2669891 -2.1753389  4.7629109
#> 
#> $rhs
#> $rhs$d
#> [1] -1.6273569 -0.1583595 -0.2111648 -0.3625565  0.7938185
#> 
#> $rhs$cf3
#> [1] -3.2547137 -0.3167191 -0.4223297 -0.7251130  1.5876370
#> 
#> $rhs$cf4
#> [1] -6.5094275 -0.6334382 -0.8446594 -1.4502259  3.1752740
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>