Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 11.015511  9.315640  9.927865 10.305397  9.692561

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  9.955955 10.051448 10.709410  9.415613 10.967166
#> 
#> $rhs
#> $rhs$b
#> [1] 0.6003699 0.6322010 0.8515214 0.4202557 0.9374401
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1] 11.488793  7.576213  8.582135 12.462396  9.336530

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 12.347136 11.427132  7.966452  4.672544  7.276163
#> 
#> $rhs
#> $rhs$b
#> [1] 0.9450225 0.7217795 0.1502168 0.1430796 0.4720244
#> 
#> $rhs$cf
#> [1] 10.989913 10.320184  8.605496  8.584084  9.570919
#> 
#> $rhs$d
#> [1]  0.4524075  0.3689827 -0.2130146 -1.3038469 -0.7649184
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -0.1800188  1.9932090  3.0392421 -1.6363935 -0.5757821
#> 
#> $rhs
#> $rhs$d
#> [1] -0.09000941  0.99660452  1.51962107 -0.81819675 -0.28789105
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  1.1630509 -0.3584084 -1.5600023  0.4617028  9.1107850
#> 
#> $rhs
#> $rhs$d
#> [1]  0.29076274 -0.08960209 -0.39000057  0.11542570  2.27769624
#> 
#> $rhs$cf3
#> [1]  0.5815255 -0.1792042 -0.7800011  0.2308514  4.5553925
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -1.738151  2.041698 -3.929971 -1.397050 -3.528333
#> 
#> $rhs
#> $rhs$d
#> [1] -0.2896919  0.3402830 -0.6549952 -0.2328417 -0.5880556
#> 
#> $rhs$cf3
#> [1] -0.5793837  0.6805660 -1.3099903 -0.4656834 -1.1761112
#> 
#> $rhs$cf4
#> [1] -1.1587674  1.3611319 -2.6199807 -0.9313669 -2.3522223
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>