Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- RAND_UNIFORM$instantiation()

# Define a closed form expression
cf <- CLOSED_FORM$instantiation(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 10.500088 10.847357 10.558964  8.381528 10.878803

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  8.901698 10.108610 10.496831 10.421495  8.459257
#> 
#> $rhs
#> $rhs$b
#> [1] 0.2489510 0.6512549 0.7806618 0.7555499 0.1014706
#> 
#> 

d <- RAND_NORMAL$instantiation()

# Define a closed form expression with another closed form expression
cf2 <- CLOSED_FORM$instantiation(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1] 15.371451  9.838817 15.022821  8.758966 14.500544

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  9.646788  5.684493  8.786705 11.563098 12.384000
#> 
#> $rhs
#> $rhs$b
#> [1] 0.39309331 0.02846141 0.86822226 0.23685191 0.86091923
#> 
#> $rhs$cf
#> [1]  9.334125  8.240230 10.759512  8.865401 10.737603
#> 
#> $rhs$d
#> [1]  0.1042207 -0.8519123 -0.6576023  0.8992321  0.5487990
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- CLOSED_FORM$instantiation(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 1.1149959 2.0768187 0.9205029 0.5358575 1.3654525
#> 
#> $rhs
#> $rhs$d
#> [1] 0.5574979 1.0384094 0.4602514 0.2679288 0.6827263
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- CLOSED_FORM$instantiation(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -4.801349  3.494435 -4.707083  2.484271 -6.352707
#> 
#> $rhs
#> $rhs$d
#> [1] -1.2003372  0.8736088 -1.1767707  0.6210676 -1.5881767
#> 
#> $rhs$cf3
#> [1] -2.400674  1.747218 -2.353541  1.242135 -3.176353
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- CLOSED_FORM$instantiation(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  1.26193131  5.21617815 -6.64329603 -0.07304661 -3.68494742
#> 
#> $rhs
#> $rhs$d
#> [1]  0.21032189  0.86936303 -1.10721601 -0.01217444 -0.61415790
#> 
#> $rhs$cf3
#> [1]  0.42064377  1.73872605 -2.21443201 -0.02434887 -1.22831581
#> 
#> $rhs$cf4
#> [1]  0.84128754  3.47745210 -4.42886402 -0.04869774 -2.45663161
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>