Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 9.148441 8.545462 9.620228 8.726942 8.820331

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  8.864475 11.042303  9.216371 10.657815  8.998506
#> 
#> $rhs
#> $rhs$b
#> [1] 0.2365432 0.9624857 0.3538417 0.8343232 0.2812203
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1]  9.941802  7.095574  8.845574 10.912820 12.590738

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 10.875528  7.684993 11.566200 13.107657 15.682378
#> 
#> $rhs
#> $rhs$b
#> [1] 0.89177353 0.09611645 0.87564087 0.53516558 0.51158392
#> 
#> $rhs$cf
#> [1] 10.830166  8.443195 10.781768  9.760342  9.689597
#> 
#> $rhs$d
#> [1]  0.01512049 -0.25273381  0.26147740  1.11577171  1.99759362
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  1.4787839  2.2190177  4.8377362  0.2112288 -0.8302090
#> 
#> $rhs
#> $rhs$d
#> [1]  0.7393920  1.1095088  2.4188681  0.1056144 -0.4151045
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  2.5540400 -0.4918491  4.1996777  2.8388643  4.2054849
#> 
#> $rhs
#> $rhs$d
#> [1]  0.6385100 -0.1229623  1.0499194  0.7097161  1.0513712
#> 
#> $rhs$cf3
#> [1]  1.2770200 -0.2459246  2.0998389  1.4194322  2.1027424
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  -4.028362  -3.207106 -14.777821  -8.900331  -2.617620
#> 
#> $rhs
#> $rhs$d
#> [1] -0.6713937 -0.5345176 -2.4629702 -1.4833886 -0.4362700
#> 
#> $rhs$cf3
#> [1] -1.342787 -1.069035 -4.925940 -2.966777 -0.872540
#> 
#> $rhs$cf4
#> [1] -2.685575 -2.138070 -9.851881 -5.933554 -1.745080
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>