Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- RAND_UNIFORM$instantiation()

# Define a closed form expression
cf <- CLOSED_FORM$instantiation(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 11.066185 10.317497  9.865020 11.025915  9.043694

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 10.199011  8.861031  9.679686  9.982426  8.177584
#> 
#> $rhs
#> $rhs$b
#> [1] 0.681388547 0.235395172 0.508280084 0.609193417 0.007579436
#> 
#> 

d <- RAND_NORMAL$instantiation()

# Define a closed form expression with another closed form expression
cf2 <- CLOSED_FORM$instantiation(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1] 17.337494  9.196123  5.533913  6.903085  5.301887

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  6.273432 11.529577  9.699952 11.967743 12.133953
#> 
#> $rhs
#> $rhs$b
#> [1] 0.1278948 0.9952910 0.4450225 0.0415371 0.2834195
#> 
#> $rhs$cf
#> [1]  8.538530 11.140718  9.489913  8.279457  9.005104
#> 
#> $rhs$d
#> [1] -0.75503266  0.12961958  0.07001303  1.22942883  1.04294973
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- CLOSED_FORM$instantiation(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  0.4813110 -2.0723833  0.2833041 -1.6834483  0.1228145
#> 
#> $rhs
#> $rhs$d
#> [1]  0.24065550 -1.03619165  0.14165204 -0.84172415  0.06140727
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- CLOSED_FORM$instantiation(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -0.3110275 -1.9205649  1.5336446  4.6719902 -3.2735349
#> 
#> $rhs
#> $rhs$d
#> [1] -0.07775688 -0.48014123  0.38341116  1.16799755 -0.81838374
#> 
#> $rhs$cf3
#> [1] -0.1555138 -0.9602825  0.7668223  2.3359951 -1.6367675
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- CLOSED_FORM$instantiation(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  4.026527 -3.025705  7.386388  4.200838 -3.121095
#> 
#> $rhs
#> $rhs$d
#> [1]  0.6710878 -0.5042841  1.2310647  0.7001397 -0.5201825
#> 
#> $rhs$cf3
#> [1]  1.342176 -1.008568  2.462129  1.400279 -1.040365
#> 
#> $rhs$cf4
#> [1]  2.684351 -2.017136  4.924259  2.800559 -2.080730
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>