Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 8.476098 9.765767 9.202718 8.366608 9.913803

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  8.966284 10.816413 10.129010  8.761970 10.990440
#> 
#> $rhs
#> $rhs$b
#> [1] 0.2704795 0.8871891 0.6580547 0.2023748 0.9451980
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1] 10.950585 14.155675 15.386574 12.313629  5.564361

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 10.575230 10.124563  6.194078 11.654103  3.908446
#> 
#> $rhs
#> $rhs$b
#> [1] 0.8985382 0.6030498 0.3518350 0.9005554 0.4663090
#> 
#> $rhs$cf
#> [1] 10.850460  9.963995  9.210350 10.856512  9.553772
#> 
#> $rhs$d
#> [1] -0.09174325  0.05352256 -1.00542416  0.26586383 -1.88177537
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  1.8442859 -0.1837443  0.7259187 -3.8842746  4.7184983
#> 
#> $rhs
#> $rhs$d
#> [1]  0.92214297 -0.09187213  0.36295936 -1.94213731  2.35924915
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -0.1025333 -3.4753799 -1.3900132 -2.4983596  0.1005271
#> 
#> $rhs
#> $rhs$d
#> [1] -0.02563331 -0.86884497 -0.34750331 -0.62458989  0.02513177
#> 
#> $rhs$cf3
#> [1] -0.05126663 -1.73768995 -0.69500662 -1.24917978  0.05026354
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  0.3792028 -2.9672675 11.0039575  2.9103422  5.8841665
#> 
#> $rhs
#> $rhs$d
#> [1]  0.06320046 -0.49454459  1.83399291  0.48505704  0.98069442
#> 
#> $rhs$cf3
#> [1]  0.1264009 -0.9890892  3.6679858  0.9701141  1.9613888
#> 
#> $rhs$cf4
#> [1]  0.2528018 -1.9781784  7.3359716  1.9402282  3.9227777
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>