Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1]  8.893087  8.921794 10.089801  9.372005  9.765577

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  9.639728  8.253363  9.703956 10.314210  8.861621
#> 
#> $rhs
#> $rhs$b
#> [1] 0.49496071 0.03283909 0.51637024 0.71978819 0.23559176
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1]  9.099651  8.379473 13.856099  6.491752 12.301107

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  7.512544 14.699771 11.414047  9.179862 11.881626
#> 
#> $rhs
#> $rhs$b
#> [1] 0.3491094 0.7445247 0.6848081 0.7552066 0.4255606
#> 
#> $rhs$cf
#> [1]  9.202174 10.388420 10.209270 10.420465  9.431527
#> 
#> $rhs$d
#> [1] -0.5632100  1.4371172  0.4015923 -0.4135344  0.8166995
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -2.8633244 -1.9961666 -2.7418902 -2.1940657  0.8745933
#> 
#> $rhs
#> $rhs$d
#> [1] -1.4316622 -0.9980833 -1.3709451 -1.0970328  0.4372967
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  -6.766689 -10.861015  -1.316905  -9.983790   1.597699
#> 
#> $rhs
#> $rhs$d
#> [1] -1.6916723 -2.7152536 -0.3292261 -2.4959475  0.3994248
#> 
#> $rhs$cf3
#> [1] -3.3833447 -5.4305073 -0.6584523 -4.9918950  0.7988495
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  1.910205  4.543729 -7.339972  6.649464 -3.549843
#> 
#> $rhs
#> $rhs$d
#> [1]  0.3183675  0.7572882 -1.2233287  1.1082440 -0.5916404
#> 
#> $rhs$cf3
#> [1]  0.636735  1.514576 -2.446657  2.216488 -1.183281
#> 
#> $rhs$cf4
#> [1]  1.273470  3.029153 -4.893315  4.432976 -2.366562
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>