Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 10.005670 10.306957  9.862574 10.871341  9.510012

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 11.040395 10.182256 11.088867  8.328141  8.248288
#> 
#> $rhs
#> $rhs$b
#> [1] 0.96184984 0.67580345 0.97800732 0.05776527 0.03114764
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1]  3.877272  8.665946  7.266858 11.004314 11.718836

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 13.86429 11.94724 12.69511 11.69867  8.00699
#> 
#> $rhs
#> $rhs$b
#> [1] 0.2408473 0.3853498 0.8262405 0.9147129 0.2676838
#> 
#> $rhs$cf
#> [1]  8.877388  9.310895 10.633567 10.898984  8.957897
#> 
#> $rhs$d
#> [1]  1.6623010  0.8787803  0.6871806  0.2665609 -0.3169690
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  3.1797684  2.5105415 -2.4806710 -1.2799019  0.1243006
#> 
#> $rhs
#> $rhs$d
#> [1]  1.58988418  1.25527073 -1.24033552 -0.63995095  0.06215029
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -0.6220038  7.3712183  3.7254116  7.9910506  0.1180475
#> 
#> $rhs
#> $rhs$d
#> [1] -0.15550094  1.84280457  0.93135289  1.99776264  0.02951187
#> 
#> $rhs$cf3
#> [1] -0.31100188  3.68560914  1.86270578  3.99552528  0.05902373
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -4.912389  1.793362 -2.775896 -2.175108  6.841910
#> 
#> $rhs
#> $rhs$d
#> [1] -0.8187315  0.2988936 -0.4626493 -0.3625180  1.1403183
#> 
#> $rhs$cf3
#> [1] -1.6374629  0.5977872 -0.9252987 -0.7250361  2.2806366
#> 
#> $rhs$cf4
#> [1] -3.274926  1.195574 -1.850597 -1.450072  4.561273
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>