Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 11.126905  9.946706 10.494012  8.526456  8.949670

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  8.842995  9.837795 10.702188 10.337452 11.111384
#> 
#> $rhs
#> $rhs$b
#> [1] 0.2293832 0.5609830 0.8491142 0.7275356 0.9855129
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1]  6.648458 14.677403  8.498807  9.032059 13.639496

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 9.956190 6.199017 6.118377 6.468297 3.394621
#> 
#> $rhs
#> $rhs$b
#> [1] 0.9809687 0.6770043 0.9876518 0.4194260 0.3810869
#> 
#> $rhs$cf
#> [1] 11.097751 10.185858 11.117801  9.413124  9.298106
#> 
#> $rhs$d
#> [1] -0.3805204 -1.3289471 -1.6664745 -0.9816087 -1.9678283
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -2.8975405 -1.3992008 -0.9515012 -0.3298938  0.2379660
#> 
#> $rhs
#> $rhs$d
#> [1] -1.4487702 -0.6996004 -0.4757506 -0.1649469  0.1189830
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -0.1941047  1.9506679 -2.2074668 -1.3946202 -3.8856649
#> 
#> $rhs
#> $rhs$d
#> [1] -0.04852618  0.48766697 -0.55186669 -0.34865505 -0.97141623
#> 
#> $rhs$cf3
#> [1] -0.09705235  0.97533395 -1.10373338 -0.69731010 -1.94283246
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]   5.255047  -8.029874   7.280512   3.461686 -12.960850
#> 
#> $rhs
#> $rhs$d
#> [1]  0.8758411 -1.3383123  1.2134186  0.5769477 -2.1601416
#> 
#> $rhs$cf3
#> [1]  1.751682 -2.676625  2.426837  1.153895 -4.320283
#> 
#> $rhs$cf4
#> [1]  3.503364 -5.353249  4.853675  2.307791 -8.640567
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>