Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 9.993465 9.269049 8.352682 9.138083 9.512398

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 10.130868  8.793714  8.514227 10.212017  9.952566
#> 
#> $rhs
#> $rhs$b
#> [1] 0.6586741 0.2129562 0.1197939 0.6857238 0.5992400
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1]  9.501503  4.418145 14.522578 11.796656 14.047091

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  7.865845 11.486582  3.791806  9.616758  9.231919
#> 
#> $rhs
#> $rhs$b
#> [1] 0.63654221 0.85907046 0.02122888 0.02237378 0.63578501
#> 
#> $rhs$cf
#> [1] 10.064472 10.732057  8.218532  8.221967 10.062201
#> 
#> $rhs$d
#> [1] -0.7328756  0.2515082 -1.4755752  0.4649304 -0.2767606
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 1.1073585 0.6641771 3.9752251 2.2450378 5.8573329
#> 
#> $rhs
#> $rhs$d
#> [1] 0.5536793 0.3320886 1.9876126 1.1225189 2.9286665
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -0.4357434  5.9435165 -2.1861058 -4.2000350 -4.3448176
#> 
#> $rhs
#> $rhs$d
#> [1] -0.1089358  1.4858791 -0.5465265 -1.0500088 -1.0862044
#> 
#> $rhs$cf3
#> [1] -0.2178717  2.9717582 -1.0930529 -2.1000175 -2.1724088
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  10.3696819 -18.8755121   7.4269980  -0.6745463  21.0680088
#> 
#> $rhs
#> $rhs$d
#> [1]  1.7282803 -3.1459187  1.2378330 -0.1124244  3.5113348
#> 
#> $rhs$cf3
#> [1]  3.4565606 -6.2918374  2.4756660 -0.2248488  7.0226696
#> 
#> $rhs$cf4
#> [1]   6.9131212 -12.5836748   4.9513320  -0.4496975  14.0453392
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>