Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- RAND_UNIFORM$instantiation()

# Define a closed form expression
cf <- CLOSED_FORM$instantiation(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 10.424698  9.716130  8.692861  9.371112  9.087651

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 10.517418  8.221414  9.308136 10.824770  9.857672
#> 
#> $rhs
#> $rhs$b
#> [1] 0.78752405 0.02218938 0.38443010 0.88997496 0.56760898
#> 
#> 

d <- RAND_NORMAL$instantiation()

# Define a closed form expression with another closed form expression
cf2 <- CLOSED_FORM$instantiation(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1] 10.91800  8.81154 15.79236 10.77229 15.01834

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  7.149142  9.754648  4.518631  3.923113 15.860837
#> 
#> $rhs
#> $rhs$b
#> [1] 0.8857730 0.2197367 0.1688148 0.1163695 0.9827962
#> 
#> $rhs$cf
#> [1] 10.812165  8.814056  8.661290  8.503954 11.103234
#> 
#> $rhs$d
#> [1] -1.2210077  0.3135309 -1.3808862 -1.5269469  1.5858677
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- CLOSED_FORM$instantiation(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -1.315511 -3.632588 -4.113638 -1.293151 -1.308708
#> 
#> $rhs
#> $rhs$d
#> [1] -0.6577555 -1.8162940 -2.0568191 -0.6465756 -0.6543538
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- CLOSED_FORM$instantiation(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  1.30223455 -2.38375454 -0.01076338 -2.17580065 -1.08965404
#> 
#> $rhs
#> $rhs$d
#> [1]  0.325558638 -0.595938635 -0.002690846 -0.543950164 -0.272413510
#> 
#> $rhs$cf3
#> [1]  0.651117277 -1.191877271 -0.005381691 -1.087900327 -0.544827020
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- CLOSED_FORM$instantiation(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  1.622086 -3.116108 -1.130030 -3.027557 -2.229847
#> 
#> $rhs
#> $rhs$d
#> [1]  0.2703477 -0.5193514 -0.1883384 -0.5045928 -0.3716411
#> 
#> $rhs$cf3
#> [1]  0.5406955 -1.0387027 -0.3766767 -1.0091855 -0.7432822
#> 
#> $rhs$cf4
#> [1]  1.0813910 -2.0774055 -0.7533534 -2.0183711 -1.4865644
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>