Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- RAND_UNIFORM$instantiation()

# Define a closed form expression
cf <- CLOSED_FORM$instantiation(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1]  8.369900 10.020093 10.196046 10.875523  9.237318

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 10.210178  8.507301  9.403618  8.571940 10.733729
#> 
#> $rhs
#> $rhs$b
#> [1] 0.6851110 0.1174853 0.4162574 0.1390315 0.8596279
#> 
#> 

d <- RAND_NORMAL$instantiation()

# Define a closed form expression with another closed form expression
cf2 <- CLOSED_FORM$instantiation(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1] 12.729279  9.148515 10.643323 12.284194  9.651607

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  8.301884  3.199980  7.966970  7.964853 11.367844
#> 
#> $rhs
#> $rhs$b
#> [1] 0.2600161 0.6176817 0.5046757 0.5284629 0.9430984
#> 
#> $rhs$cf
#> [1]  8.934894 10.007890  9.668873  9.740234 10.984141
#> 
#> $rhs$d
#> [1] -0.2110032 -2.2693035 -0.5673010 -0.5917936  0.1279010
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- CLOSED_FORM$instantiation(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  2.6012448 -1.1695942  4.2097359 -0.3359545 -1.6533894
#> 
#> $rhs
#> $rhs$d
#> [1]  1.3006224 -0.5847971  2.1048679 -0.1679772 -0.8266947
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- CLOSED_FORM$instantiation(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -4.719062  2.195157  3.517917  3.990859  2.942695
#> 
#> $rhs
#> $rhs$d
#> [1] -1.1797656  0.5487891  0.8794792  0.9977147  0.7356737
#> 
#> $rhs$cf3
#> [1] -2.359531  1.097578  1.758958  1.995429  1.471347
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- CLOSED_FORM$instantiation(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  5.718579 -4.301271 -1.394787 -3.202265  4.519968
#> 
#> $rhs
#> $rhs$d
#> [1]  0.9530965 -0.7168785 -0.2324645 -0.5337108  0.7533280
#> 
#> $rhs$cf3
#> [1]  1.906193 -1.433757 -0.464929 -1.067422  1.506656
#> 
#> $rhs$cf4
#> [1]  3.812386 -2.867514 -0.929858 -2.134843  3.013312
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>