Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 10.961839 10.762075  9.868249  9.287659  9.620204

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  8.372335  8.622980 10.471551  9.751025 10.293601
#> 
#> $rhs
#> $rhs$b
#> [1] 0.07249657 0.15604498 0.77223507 0.53205999 0.71291853
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1] 10.256028  8.039880  7.509018 16.305852 12.436472

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  6.847141 11.385818  9.687687 11.135717  8.786849
#> 
#> $rhs
#> $rhs$b
#> [1] 0.008468302 0.654114943 0.512118059 0.393999079 0.842538694
#> 
#> $rhs$cf
#> [1]  8.180250 10.117190  9.691200  9.336843 10.682462
#> 
#> $rhs$d
#> [1] -0.444369828  0.422875765 -0.001170945  0.599624725 -0.631870991
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -2.133771  2.042196 -2.753805  1.088972 -1.674010
#> 
#> $rhs
#> $rhs$d
#> [1] -1.0668856  1.0210982 -1.3769026  0.5444858 -0.8370049
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 4.144882 1.548357 1.865940 3.306647 5.104822
#> 
#> $rhs
#> $rhs$d
#> [1] 1.0362205 0.3870892 0.4664850 0.8266618 1.2762055
#> 
#> $rhs$cf3
#> [1] 2.0724410 0.7741785 0.9329701 1.6533237 2.5524110
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -4.187861  8.052551 -1.638399  1.109012 -8.383423
#> 
#> $rhs
#> $rhs$d
#> [1] -0.6979768  1.3420918 -0.2730665  0.1848353 -1.3972372
#> 
#> $rhs$cf3
#> [1] -1.3959535  2.6841836 -0.5461329  0.3696705 -2.7944743
#> 
#> $rhs$cf4
#> [1] -2.7919070  5.3683671 -1.0922659  0.7393411 -5.5889486
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>