Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1]  9.936314 11.086585  8.276026 10.962742  9.646397

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 10.712767  9.317654  8.350770  8.378785 10.964881
#> 
#> $rhs
#> $rhs$b
#> [1] 0.85264037 0.38760281 0.06530831 0.07464662 0.93667851
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1] 10.425571 11.407745 11.019584  2.767985 11.696716

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 13.507684 16.982234  8.670922  8.707899 12.728932
#> 
#> $rhs
#> $rhs$b
#> [1] 0.3425570 0.7893745 0.5262588 0.5727484 0.1977655
#> 
#> $rhs$cf
#> [1]  9.182516 10.522969  9.733622  9.873091  8.748142
#> 
#> $rhs$d
#> [1]  1.4417225  2.1530885 -0.3542332 -0.3883970  1.3269302
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -0.8721892  1.9614600 -0.8075060  0.8911309 -1.2017813
#> 
#> $rhs
#> $rhs$d
#> [1] -0.4360946  0.9807300 -0.4037530  0.4455654 -0.6008906
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -6.0997630  3.3729350  4.5002701 -3.8099938 -0.1953852
#> 
#> $rhs
#> $rhs$d
#> [1] -1.52494074  0.84323375  1.12506751 -0.95249844 -0.04884631
#> 
#> $rhs$cf3
#> [1] -3.04988148  1.68646749  2.25013503 -1.90499688 -0.09769262
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  9.433428  5.603208 -1.253776 -1.291352  3.235425
#> 
#> $rhs
#> $rhs$d
#> [1]  1.5722380  0.9338680 -0.2089627 -0.2152253  0.5392375
#> 
#> $rhs$cf3
#> [1]  3.1444761  1.8677360 -0.4179254 -0.4304505  1.0784750
#> 
#> $rhs$cf4
#> [1]  6.2889521  3.7354720 -0.8358508 -0.8609011  2.1569500
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>