Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- rand_uniform()

# Define a closed form expression
cf <- closed_form(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1]  8.959852 10.446422  9.861103  8.264035  9.633606

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 10.206462  9.827970  8.701981  9.533368 10.405271
#> 
#> $rhs
#> $rhs$b
#> [1] 0.6838723 0.5577083 0.1823785 0.4595073 0.7501418
#> 
#> 

d <- rand_normal()

# Define a closed form expression with another closed form expression
cf2 <- closed_form(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1] 11.816465  6.850042  2.661122 11.862288  7.399018

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 11.133017  9.049520  8.712975  9.089639  8.543909
#> 
#> $rhs
#> $rhs$b
#> [1] 0.8200285 0.1633284 0.7143805 0.1096769 0.1000993
#> 
#> $rhs$cf
#> [1] 10.614931  8.644831 10.297987  8.483876  8.455144
#> 
#> $rhs$d
#> [1]  0.17269536  0.13489642 -0.52833732  0.20192105  0.02958848
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- closed_form(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  1.630258 -2.634585  3.058232 -2.784523 -1.792654
#> 
#> $rhs
#> $rhs$d
#> [1]  0.8151291 -1.3172926  1.5291161 -1.3922617 -0.8963269
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- closed_form(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -2.165233 -7.491106 -6.102828  6.146870 -0.943537
#> 
#> $rhs
#> $rhs$d
#> [1] -0.5413082 -1.8727764 -1.5257070  1.5367174 -0.2358843
#> 
#> $rhs$cf3
#> [1] -1.0826164 -3.7455528 -3.0514140  3.0734348 -0.4717685
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- closed_form(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  5.3136212 -0.9000544  5.3162708 10.2709045 -2.0343197
#> 
#> $rhs
#> $rhs$d
#> [1]  0.8856035 -0.1500091  0.8860451  1.7118174 -0.3390533
#> 
#> $rhs$cf3
#> [1]  1.7712071 -0.3000181  1.7720903  3.4236348 -0.6781066
#> 
#> $rhs$cf4
#> [1]  3.5424141 -0.6000362  3.5441805  6.8472697 -1.3562131
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>