Generating random values from the expression

This function generates random values from the expression. Random values will be generated independently as long as they don't share the same symbol in the expression.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- RAND_UNIFORM$instantiation()

# Define a closed form expression
cf <- CLOSED_FORM$instantiation(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1]  9.132997  8.360757  8.249943  9.119460 11.105677

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 10.923268  9.687021  9.385026  9.415514  9.080337
#> 
#> $rhs
#> $rhs$b
#> [1] 0.9228074 0.5107250 0.4100603 0.4202229 0.3084970
#> 
#> 

d <- RAND_NORMAL$instantiation()

# Define a closed form expression with another closed form expression
cf2 <- CLOSED_FORM$instantiation(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1]  9.278894  7.361697  2.020514 13.857710 10.289820

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 11.946007 11.220423  7.974983 16.389219  7.841638
#> 
#> $rhs
#> $rhs$b
#> [1] 0.9169064 0.1377755 0.4454593 0.7562793 0.4114016
#> 
#> $rhs$cf
#> [1] 10.905565  8.568172  9.491223 10.423683  9.389050
#> 
#> $rhs$d
#> [1]  0.3468140  0.8840836 -0.5054136  1.9885120 -0.5158043
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- CLOSED_FORM$instantiation(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -0.6500048  0.5684983  0.7243094  1.2078874 -0.9101950
#> 
#> $rhs
#> $rhs$d
#> [1] -0.3250024  0.2842492  0.3621547  0.6039437 -0.4550975
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- CLOSED_FORM$instantiation(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  0.1693139 -0.4139042  0.3774017  6.3133598 -6.6664645
#> 
#> $rhs
#> $rhs$d
#> [1]  0.04232847 -0.10347605  0.09435042  1.57833994 -1.66661613
#> 
#> $rhs$cf3
#> [1]  0.08465693 -0.20695209  0.18870083  3.15667988 -3.33323226
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- CLOSED_FORM$instantiation(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -13.190082  -5.752520  -1.752577   5.182933   4.269874
#> 
#> $rhs
#> $rhs$d
#> [1] -2.1983469 -0.9587533 -0.2920961  0.8638222  0.7116456
#> 
#> $rhs$cf3
#> [1] -4.3966938 -1.9175066 -0.5841923  1.7276444  1.4232913
#> 
#> $rhs$cf4
#> [1] -8.793388 -3.835013 -1.168385  3.455289  2.846583
#> 
#>