Generating random values from the expression

This function generates random values from the expression. Random values or closed form expression will share the same value as long as they have the same name.

Arguments

n

Integer. Number of observations.

rhs_val

Boolean. Whether or not to keep the right hand side values of the expression. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

Numeric values.

Examples

# Constant variable
a <- 1

# Random uniform variable
b <- RAND_UNIFORM$instantiation()

# Define a closed form expression
cf <- CLOSED_FORM$instantiation(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)
#> [1] 10.839278  9.385675  8.728307 10.009449 10.621827

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  8.905833 10.667030 10.929172 11.088372 10.362810
#> 
#> $rhs
#> $rhs$b
#> [1] 0.2503293 0.8373948 0.9247754 0.9778421 0.7359883
#> 
#> 

d <- RAND_NORMAL$instantiation()

# Define a closed form expression with another closed form expression
cf2 <- CLOSED_FORM$instantiation(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)
#> [1]  7.791975  5.795214 12.406215  6.919078  9.647232

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] 12.399422  8.107502  8.136781  8.427744 10.034537
#> 
#> $rhs
#> $rhs$b
#> [1] 0.22628877 0.35157231 0.10586095 0.04569623 0.93568445
#> 
#> $rhs$cf
#> [1]  8.833712  9.209562  8.472428  8.291934 10.961899
#> 
#> $rhs$d
#> [1]  1.1885701 -0.3673535 -0.1118826  0.0452699 -0.3091206
#> 
#> 

# Define a closed form expression with two random variables of the same name
cf3 <- CLOSED_FORM$instantiation(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -1.14804884 -2.06072538 -0.09793635 -1.07594428  3.10508068
#> 
#> $rhs
#> $rhs$d
#> [1] -0.57402442 -1.03036269 -0.04896818 -0.53797214  1.55254034
#> 
#> 

# Define a closed form expression with two closed form expressions of the same name
cf4 <- CLOSED_FORM$instantiation(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)
#> $lhs
#> [1] -1.964809  1.910377  3.295173 -2.674064 -3.739805
#> 
#> $rhs
#> $rhs$d
#> [1] -0.4912022  0.4775942  0.8237931 -0.6685161 -0.9349513
#> 
#> $rhs$cf3
#> [1] -0.9824045  0.9551885  1.6475863 -1.3370322 -1.8699026
#> 
#> 

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- CLOSED_FORM$instantiation(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
#> $lhs
#> [1]  -2.230071   2.069147 -14.763515  -5.777574  -7.052856
#> 
#> $rhs
#> $rhs$d
#> [1] -0.3716786  0.3448579 -2.4605859 -0.9629290 -1.1754760
#> 
#> $rhs$cf3
#> [1] -0.7433571  0.6897157 -4.9211717 -1.9258580 -2.3509520
#> 
#> $rhs$cf4
#> [1] -1.486714  1.379431 -9.842343 -3.851716 -4.701904
#> 
#> 

# Control the value of `d`
cf5$gen(5, rhs_val = TRUE, computed = list(d = 1))
#> $lhs
#> [1] 6
#> 
#> $rhs
#> $rhs$d
#> [1] 1
#> 
#> $rhs$cf3
#> [1] 2
#> 
#> $rhs$cf4
#> [1] 4
#> 
#>