Generating random values from the expression of y

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

fit_model

Boolean. Whether or not to fit a null model to obtain the fitted values and the residuals. Default is FALSE.

test

Boolean. Whether or not to test the null model against the alternative model to obtain the test statistic and the p-value. Default is FALSE.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen(10)
#>           y         x          e
#> 1  1.534686 0.7947332 -0.8916479
#> 2  3.800603 0.2474789  2.4918787
#> 3  2.735456 0.4854358  1.0143720
#> 4  3.289825 0.7713075  0.9236024
#> 5  2.808395 0.2323987  1.5219869
#> 6  2.574567 0.2804379  1.2154835
#> 7  1.593113 0.9986827 -1.4029369
#> 8  2.301464 0.9906562 -0.6705916
#> 9  1.366268 0.4863475 -0.3566138
#> 10 1.910295 0.3855377  0.3761177

test$gen(10, fit_model = TRUE)
#>            y          x           e       .resid   .fitted
#> 1  1.0701605 0.11886454 -0.06283279  0.300257512 0.7699030
#> 2  1.0226563 0.31015774 -0.38369927 -0.436660984 1.4593173
#> 3  0.1372612 0.43342130 -1.48401413 -1.766293839 1.9035550
#> 4  1.2257683 0.08539148  0.13308513  0.576501100 0.6492672
#> 5  1.3225348 0.27264982 -0.02445289 -0.001605142 1.3241400
#> 6  2.1607838 0.26091042  0.83179916  0.878952237 1.2818316
#> 7  4.3757674 0.73605310  2.09794016  1.381537531 2.9942299
#> 8  1.1657304 0.55355876 -0.69425565 -1.170795991 2.3365264
#> 9  0.8097899 0.19837063 -0.42793158 -0.246650325 1.0564403
#> 10 2.8137036 0.55145532  0.95814530  0.484757902 2.3289457

test$gen(10, test = TRUE)
#>             y            x          e test_name statistic   p_value
#> 1   3.1994479 0.6367907636  1.1571547    F-test  0.208702 0.6616253
#> 2   2.2380921 0.3725143459  0.7268109    F-test  0.208702 0.6616253
#> 3   3.5126925 0.8812076394  0.8549580    F-test  0.208702 0.6616253
#> 4   2.0953629 0.5157145453  0.3136869    F-test  0.208702 0.6616253
#> 5   0.3077229 0.0004927143 -0.6927701    F-test  0.208702 0.6616253
#> 6   0.7860170 0.2920875810 -0.5913858    F-test  0.208702 0.6616253
#> 7   2.9840358 0.2211667569  1.7139543    F-test  0.208702 0.6616253
#> 8   3.2472865 0.8991053773  0.5397907    F-test  0.208702 0.6616253
#> 9  -0.2294420 0.5197316555 -2.0192946    F-test  0.208702 0.6616253
#> 10  1.6002643 0.5500034564 -0.2522430    F-test  0.208702 0.6616253