Generating random values from the expression of y

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

fit_model

Boolean. Whether or not to fit a null model to obtain the fitted values and the residuals. Default is TRUE.

test

Boolean. Whether or not to test the null model against the alternative model to obtain the test statistic and the p-value. Default is FALSE.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

dat <- test$gen(10)
dat
#>             y          x          e     .resid   .fitted
#> 1   0.4768398 0.12406246 -0.6626141 -0.4087669 0.8856068
#> 2   1.5356113 0.13922385  0.3770042  0.5961018 0.9395095
#> 3  -0.1139042 0.19634268 -1.3487973 -1.2564861 1.1425819
#> 4   4.5612620 0.94633002  1.7193915  0.7522786 3.8089834
#> 5   1.2438642 0.06621097  0.1732693  0.5639346 0.6799296
#> 6   3.0778036 0.49898295  1.3298367  0.8592563 2.2185473
#> 7   1.9652791 0.87855765 -0.6851420 -1.6027557 3.5680348
#> 8   2.9500129 0.65249599  0.8717659  0.1856865 2.7643264
#> 9   4.1760545 0.92009762  1.4093773  0.4603342 3.7157203
#> 10  2.6758294 0.66967793  0.5576829 -0.1495833 2.8254127

test$gen(10, computed = list(e = dat$e))
#>            y          e          x      .resid  .fitted
#> 1  0.5871418 -0.6626141 0.20693418 -0.99912513 1.586267
#> 2  2.0322183  0.3770042 0.45142741 -0.07894061 2.111159
#> 3  1.3411320 -1.3487973 0.89281346 -1.71761947 3.058751
#> 4  4.0895267  1.7193915 0.77284532  1.28832961 2.801197
#> 5  2.1935771  0.1732693 0.62707932 -0.29468123 2.488258
#> 6  4.2240388  1.3298367 0.96430943  1.01179572 3.212243
#> 7  1.8623718 -0.6851420 0.84071393 -1.08452946 2.946901
#> 8  1.9369159  0.8717659 0.06138224  0.66312810 1.273788
#> 9  3.8915038  1.4093773 0.81610279  0.99743917 2.894065
#> 10 1.8211526  0.5576829 0.21656795  0.21420330 1.606949

test$gen(10, test = TRUE)
#>            y          x          e      .resid   .fitted test_name statistic
#> 1  2.4312254 0.54534458  0.5884801  0.17436980 2.2568556    F-test 0.3046495
#> 2  0.6163656 0.11347760 -0.5099892  0.04432955 0.5720360    F-test 0.3046495
#> 3  2.0378019 0.34107327  0.5803976  0.57785901 1.4599429    F-test 0.3046495
#> 4  3.3357692 0.77511473  0.9598516  0.18252356 3.1532456    F-test 0.3046495
#> 5  3.0244741 0.91350327  0.2764826 -0.66865933 3.6931334    F-test 0.3046495
#> 6  4.7561969 0.92332872  1.9803322  1.02473190 3.7314650    F-test 0.3046495
#> 7  2.4498654 0.68246225  0.3016484 -0.34192004 2.7917855    F-test 0.3046495
#> 8  0.4098238 0.09222395 -0.6909054 -0.07929656 0.4891203    F-test 0.3046495
#> 9  3.3708020 0.82754784  0.8584187  0.01300184 3.3578001    F-test 0.3046495
#> 10 1.4925013 0.58701982 -0.4391108 -0.92693974 2.4194410    F-test 0.3046495
#>      p_value
#> 1  0.5981536
#> 2  0.5981536
#> 3  0.5981536
#> 4  0.5981536
#> 5  0.5981536
#> 6  0.5981536
#> 7  0.5981536
#> 8  0.5981536
#> 9  0.5981536
#> 10 0.5981536