Generating random values from the expression of y

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

fit_model

Boolean. Whether or not to fit a null model to obtain the fitted values and the residuals. Default is FALSE.

test

Boolean. Whether or not to test the null model against the alternative model to obtain the test statistic and the p-value. Default is FALSE.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen(10)
#>            y          x          e
#> 1  0.5760006 0.15083052 -0.5975798
#> 2  3.0721887 0.66157046  0.9729427
#> 3  2.4721408 0.01981987  1.4519281
#> 4  3.5679927 0.68900590  1.4042577
#> 5  0.7145766 0.46284779 -0.9624992
#> 6  1.5009003 0.62543878 -0.5157121
#> 7  0.7497842 0.31139902 -0.6585842
#> 8  0.9826025 0.61393397 -1.0082464
#> 9  1.2440050 0.78379129 -1.1541151
#> 10 2.8683923 0.99069463 -0.1037782

test$gen(10, fit_model = TRUE)
#>             y            x           e     .resid    .fitted
#> 1   2.0898247 0.6741937674 -0.03890634 -0.5135536  2.6033783
#> 2   0.7324111 0.4155377971 -0.85579835 -0.7085664  1.4409775
#> 3  -0.5897016 0.3485377487 -2.05971790 -1.7295806  1.1398790
#> 4   4.9002376 0.9757152284  1.97250220  0.9418208  3.9584168
#> 5   0.6421175 0.4775864971 -1.06355786 -1.0777071  1.7198245
#> 6   4.4708284 0.7957791274  2.04178481  1.3210450  3.1497833
#> 7   1.4071917 0.3465585969 -0.05946977  0.2762070  1.1309847
#> 8   2.6852082 0.8511480587  0.10960708 -0.7134033  3.3986115
#> 9   0.1206978 0.0005634895 -0.87986596  0.5446164 -0.4239185
#> 10  2.0430148 0.1803166787  0.83018400  1.6591218  0.3838930

test$gen(10, test = TRUE)
#>            y         x          e test_name statistic     p_value
#> 1  5.0451399 0.8870396  2.3712611    F-test  14.43889 0.006716057
#> 2  2.0538394 0.5592006  0.1819334    F-test  14.43889 0.006716057
#> 3  0.3611294 0.5248744 -1.4392381    F-test  14.43889 0.006716057
#> 4  2.4066728 0.7150018  0.1804436    F-test  14.43889 0.006716057
#> 5  0.8571916 0.1690724 -0.3404662    F-test  14.43889 0.006716057
#> 6  1.1850214 0.2849168 -0.1810730    F-test  14.43889 0.006716057
#> 7  1.7749636 0.4299195  0.1602133    F-test  14.43889 0.006716057
#> 8  0.8185567 0.1979310 -0.4185510    F-test  14.43889 0.006716057
#> 9  0.9273794 0.3284751 -0.5089916    F-test  14.43889 0.006716057
#> 10 1.7340999 0.1156005  0.6051359    F-test  14.43889 0.006716057