Generating random values from the expression of y

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

fit_model

Boolean. Whether or not to fit a null model to obtain the fitted values and the residuals. Default is FALSE.

test

Boolean. Whether or not to test the null model against the alternative model to obtain the test statistic and the p-value. Default is FALSE.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen(10)
#>            y         x           e
#> 1  0.7608824 0.7984483 -1.67508564
#> 2  1.3813913 0.5650453 -0.50293030
#> 3  1.7025304 0.3372136  0.25160372
#> 4  1.8561298 0.1519195  0.68113081
#> 5  1.3870227 0.2442790  0.08307153
#> 6  1.0738511 0.3662220 -0.42648951
#> 7  2.7390225 0.5378247  0.91194240
#> 8  2.0133795 0.2390067  0.71724857
#> 9  3.0460817 0.1319257  1.89675153
#> 10 3.1323793 0.5252842  1.33117166

test$gen(10, fit_model = TRUE)
#>             y          x           e     .resid   .fitted
#> 1   1.1568877 0.91218536 -1.58737975 -0.6863917 1.8432795
#> 2   3.2063380 0.28315956  1.84299906  2.1453621 1.0609758
#> 3   1.7057227 0.44144014  0.06941321  0.4478973 1.2578254
#> 4  -0.5729377 0.02958632 -1.60339941 -1.3185509 0.7456132
#> 5  -0.4654853 0.29789772 -1.85212612 -1.5447907 1.0793053
#> 6   2.2432879 0.46081288  0.57012655  0.9613691 1.2819188
#> 7   0.5016621 0.93710893 -2.31362000 -1.3726142 1.8742763
#> 8   1.8734643 0.60975443 -0.10809059  0.4063106 1.4671537
#> 9   0.4192381 0.10719239 -0.69944453 -0.4228918 0.8421299
#> 10  2.9564724 0.69419653  0.78036709  1.3843002 1.5721722

test$gen(10, test = TRUE)
#>             y          x          e test_name statistic  p_value
#> 1   0.2744480 0.04679983 -0.7745421    F-test 0.9315106 0.366622
#> 2   4.3948382 0.93505557  1.5854537    F-test 0.9315106 0.366622
#> 3   1.6081790 0.21811919  0.3424838    F-test 0.9315106 0.366622
#> 4   2.9575035 0.34194120  1.4986385    F-test 0.9315106 0.366622
#> 5   3.3480201 0.51696317  1.5638060    F-test 0.9315106 0.366622
#> 6   3.3179239 0.88978781  0.6364138    F-test 0.9315106 0.366622
#> 7   2.0059721 0.44774392  0.3577536    F-test 0.9315106 0.366622
#> 8  -0.4412908 0.02333251 -1.4651678    F-test 0.9315106 0.366622
#> 9   1.5276883 0.03994341  0.4861494    F-test 0.9315106 0.366622
#> 10  1.6756075 0.40028912  0.1150870    F-test 0.9315106 0.366622