Generating random values from the expression of y

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

fit_model

Boolean. Whether or not to fit a null model to obtain the fitted values and the residuals. Default is FALSE.

test

Boolean. Whether or not to test the null model against the alternative model to obtain the test statistic and the p-value. Default is FALSE.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen(10)
#>            y         x           e
#> 1  0.8214009 0.1999825 -0.41857461
#> 2  1.8751371 0.8559556 -0.71347843
#> 3  4.0344056 0.8266783  1.52433030
#> 4  2.7038400 0.2364812  1.41143548
#> 5  2.1045151 0.4970128  0.36048056
#> 6  1.3032544 0.2142220  0.04314125
#> 7  2.9461740 0.2679251  1.60646500
#> 8  2.8818143 0.8177970  0.39522534
#> 9  1.0124163 0.4658077 -0.67036831
#> 10 1.9047081 0.3722071  0.39396287

test$gen(10, fit_model = TRUE)
#>             y         x           e      .resid     .fitted
#> 1   0.1894324 0.2641066 -1.14442651 -0.13419803  0.32363043
#> 2   2.0644472 0.9088913 -0.67052759 -0.43348162  2.49792882
#> 3   1.8188927 0.2253960  0.54269330  1.62579946  0.19309321
#> 4   2.4105196 0.8223669 -0.08813469  0.20436239  2.20615723
#> 5   1.0967848 0.3655901 -0.40246153  0.43093861  0.66584615
#> 6   2.5626448 0.9008443 -0.14971982  0.09185178  2.47079305
#> 7  -0.7303894 0.2249870 -2.00599559 -0.92210350  0.19171412
#> 8   2.4875827 0.9154662 -0.26596189 -0.03251747  2.52010019
#> 9  -1.7124354 0.1372569 -2.86853182 -1.60831216 -0.10412325
#> 10  0.6990011 0.1448082 -0.46677650  0.77766053 -0.07865942

test$gen(10, test = TRUE)
#>             y           x          e test_name statistic    p_value
#> 1   4.2158134 0.499363681  2.4670856    F-test  5.259776 0.05552692
#> 2   1.1002773 0.997780820 -1.8930700    F-test  5.259776 0.05552692
#> 3   1.5208071 0.517265869 -0.2640227    F-test  5.259776 0.05552692
#> 4   0.8256761 0.173896635 -0.3784606    F-test  5.259776 0.05552692
#> 5   0.8725891 0.293721874 -0.5074053    F-test  5.259776 0.05552692
#> 6   0.6405051 0.364598760 -0.8570260    F-test  5.259776 0.05552692
#> 7  -0.4403356 0.145016182 -1.6063814    F-test  5.259776 0.05552692
#> 8  -0.5569490 0.114166124 -1.6841491    F-test  5.259776 0.05552692
#> 9   1.2795054 0.687445417 -0.8805212    F-test  5.259776 0.05552692
#> 10 -0.8892681 0.003327072 -1.8926063    F-test  5.259776 0.05552692