Generating random values from the expression of y

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

fit_model

Boolean. Whether or not to fit a null model to obtain the fitted values and the residuals. Default is FALSE.

test

Boolean. Whether or not to test the null model against the alternative model to obtain the test statistic and the p-value. Default is FALSE.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen(10)
#>            y          x          e
#> 1  1.6179494 0.86861530 -1.0051584
#> 2  2.3694383 0.57230917  0.4695913
#> 3  0.5371625 0.91818423 -2.2240840
#> 4  1.7013318 0.95331936 -1.1608054
#> 5  1.6358590 0.19026544  0.4093927
#> 6  0.6775916 0.50317296 -1.0787643
#> 7  2.3625974 0.08398873  1.2715546
#> 8  0.6652971 0.21794116 -0.6001424
#> 9  2.5254316 0.88244972 -0.1357356
#> 10 0.3775540 0.26869121 -0.9633322

test$gen(10, fit_model = TRUE)
#>             y         x          e      .resid   .fitted
#> 1   3.1180021 0.7150050  0.8917649  1.06218379 2.0558183
#> 2   2.7648570 0.8487561  0.1957141  0.40782537 2.3570316
#> 3  -0.1982669 0.5399839 -2.0298334 -1.85992984 1.6616629
#> 4   2.2050846 0.8847705 -0.4625047 -0.23305313 2.4381377
#> 5   1.9962851 0.7367966 -0.2833807 -0.10860881 2.1048939
#> 6   1.7292591 0.7102562 -0.4854609 -0.31586456 2.0451237
#> 7   1.1890872 0.3037289 -0.2068929  0.05948113 1.1296061
#> 8   0.9599263 0.1380700 -0.1972070  0.20339145 0.7565349
#> 9   1.5287868 0.6257202 -0.4884591 -0.32595805 1.8547449
#> 10  2.7061418 0.5106533  0.9347217  1.11053265 1.5956091

test$gen(10, test = TRUE)
#>             y          x          e test_name  statistic   p_value
#> 1   2.3400376 0.40977972  0.7623385    F-test 0.05865849 0.8155716
#> 2   0.8224201 0.64103513 -1.2295411    F-test 0.05865849 0.8155716
#> 3   1.9569083 0.07473496  0.8765881    F-test 0.05865849 0.8155716
#> 4  -0.8812547 0.01420764 -1.8956642    F-test 0.05865849 0.8155716
#> 5   3.7953871 0.98879215  0.8288850    F-test 0.05865849 0.8155716
#> 6  -0.8460539 0.15981345 -2.0314077    F-test 0.05865849 0.8155716
#> 7   1.3214763 0.42058872 -0.2760073    F-test 0.05865849 0.8155716
#> 8   0.2738900 0.98010104 -2.6668091    F-test 0.05865849 0.8155716
#> 9   2.7011544 0.95581038 -0.1682294    F-test 0.05865849 0.8155716
#> 10  1.4490404 0.12233418  0.3117406    F-test 0.05865849 0.8155716