Generating random values from the expression of y

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

fit_model

Boolean. Whether or not to fit a null model to obtain the fitted values and the residuals. Default is FALSE.

test

Boolean. Whether or not to test the null model against the alternative model to obtain the test statistic and the p-value.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. Default is FALSE.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen(10)
#>           y           x           e
#> 1  2.925163 0.225036649  1.64948510
#> 2  1.789461 0.009567951  0.77980127
#> 3  2.022316 0.700547618 -0.16899853
#> 4  2.044416 0.669712004 -0.07381033
#> 5  2.952211 0.697407611  0.76842600
#> 6  2.635562 0.887480228 -0.03953920
#> 7  2.664808 0.496538395  0.92171966
#> 8  1.175563 0.466095887 -0.50777858
#> 9  1.638140 0.527123817 -0.16684286
#> 10 1.775575 0.324784382  0.34530568

test$gen(10, fit_model = TRUE)
#>            y         x           e     .resid   .fitted
#> 1  2.7605235 0.5642415  0.87791343  0.7065591 2.0539644
#> 2  1.6192511 0.5868212 -0.31192934 -0.4914704 2.1107215
#> 3  1.3504170 0.2409112  0.05146762  0.1091874 1.2412297
#> 4  3.2426496 0.8914337  0.55656201  0.3662435 2.8764061
#> 5  0.3190654 0.1350470 -0.83421927 -0.6560602 0.9751256
#> 6  1.8141562 0.9409255 -1.01210995 -1.1866542 3.0008104
#> 7  3.4865857 0.9659215  0.58765990  0.4229443 3.0636414
#> 8  2.8509470 0.4016746  1.28792989  1.2056167 1.6453304
#> 9  1.0198733 0.2552144 -0.30047554 -0.2573095 1.2771828
#> 10 1.3922733 0.3881483 -0.14653405 -0.2190567 1.6113300

test$gen(10, test = TRUE)
#>             y         x          e test_name statistic   p_value
#> 1   0.2970643 0.5936622 -1.6490328    F-test   4.64908 0.0679911
#> 2  -0.1371437 0.5783512 -2.0499849    F-test   4.64908 0.0679911
#> 3   3.2281593 0.7618905  0.8857917    F-test   4.64908 0.0679911
#> 4  -0.1830568 0.5710146 -2.0801290    F-test   4.64908 0.0679911
#> 5   0.4521839 0.1632113 -0.7376653    F-test   4.64908 0.0679911
#> 6  -0.2212682 0.1033981 -1.3353575    F-test   4.64908 0.0679911
#> 7   1.3694338 0.1819084  0.1544347    F-test   4.64908 0.0679911
#> 8   0.7530157 0.1434140 -0.4109659    F-test   4.64908 0.0679911
#> 9   1.9823226 0.4728008  0.2859812    F-test   4.64908 0.0679911
#> 10  4.1267477 0.7828939  1.7309309    F-test   4.64908 0.0679911