Generating random values from the expression of y

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

fit_model

Boolean. Whether or not to fit a null model to obtain the fitted values and the residuals. Default is FALSE.

test

Boolean. Whether or not to test the null model against the alternative model to obtain the test statistic and the p-value. Default is FALSE.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen(10)
#>             y          x          e
#> 1   1.7244586 0.05796974  0.6631284
#> 2   3.0755969 0.74854960  0.7667208
#> 3   1.9228962 0.23839459  0.6276696
#> 4  -0.1459422 0.25437190 -1.4650191
#> 5   2.3251415 0.62458267  0.3104553
#> 6   4.7105242 0.96365175  1.8182477
#> 7   2.4054449 0.13979098  1.2461124
#> 8   2.5147888 0.59432719  0.5672368
#> 9   3.1589847 0.89485107  0.4633751
#> 10  1.6152330 0.17510942  0.4094603

test$gen(10, fit_model = TRUE)
#>            y          x           e      .resid  .fitted
#> 1  1.5450700 0.09463057  0.44148450  0.49817100 1.046899
#> 2  2.1691345 0.59309637  0.22427482  0.12436346 2.044771
#> 3  2.0067705 0.19401153  0.77511846  0.76092203 1.245848
#> 4  1.3701562 0.63252896 -0.66246560 -0.75355437 2.123711
#> 5  5.5184665 0.94845472  2.67044543  2.76230834 2.756158
#> 6  1.2481205 0.15396088  0.07045562  0.08244889 1.165672
#> 7  0.7771240 0.74881941 -1.53242596 -1.57938695 2.356511
#> 8  1.3603941 0.33511611 -0.08702478 -0.16792967 1.528324
#> 9  0.8062401 0.70920658 -1.40594049 -1.47097046 2.277211
#> 10 1.7124669 0.55516624 -0.15090886 -0.25637228 1.968839

test$gen(10, test = TRUE)
#>            y          x          e test_name  statistic  p_value
#> 1  0.6054319 0.09532227 -0.4989767    F-test 0.06896846 0.800406
#> 2  1.0762391 0.76330407 -1.2696981    F-test 0.06896846 0.800406
#> 3  1.7829027 0.39909417  0.2245324    F-test 0.06896846 0.800406
#> 4  1.1078005 0.19481781 -0.1249713    F-test 0.06896846 0.800406
#> 5  3.1307265 0.63407467  1.0946011    F-test 0.06896846 0.800406
#> 6  2.9011337 0.92195947  0.1291650    F-test 0.06896846 0.800406
#> 7  3.4499579 0.55927475  1.5778949    F-test 0.06896846 0.800406
#> 8  2.0340851 0.32484112  0.6037222    F-test 0.06896846 0.800406
#> 9  3.8434619 0.99196583  0.8674998    F-test 0.06896846 0.800406
#> 10 1.4420514 0.71879500 -0.7934099    F-test 0.06896846 0.800406