Generating random values from the expression of y

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

fit_model

Boolean. Whether or not to fit a null model to obtain the fitted values and the residuals. Default is FALSE.

test

Boolean. Whether or not to test the null model against the alternative model to obtain the test statistic and the p-value. Default is FALSE.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen(10)
#>             y          x          e
#> 1   2.0237205 0.04345858  0.9783733
#> 2   0.3772988 0.06616396 -0.6932428
#> 3   0.3918250 0.49320864 -1.3446384
#> 4   1.6321569 0.83069790 -0.8886000
#> 5  -0.2192540 0.03034755 -1.2505225
#> 6   1.9861499 0.51635834  0.2031656
#> 7   1.2229713 0.02155714  0.2009494
#> 8   2.9827279 0.35826689  1.4961058
#> 9   2.1178621 0.89032851 -0.5651513
#> 10 -0.4316971 0.06132390 -1.4967816

test$gen(10, fit_model = TRUE)
#>              y         x          e      .resid   .fitted
#> 1   2.44727418 0.4174118  0.8556298  1.36822967 1.0790445
#> 2   1.44882345 0.8025428 -0.9977942 -1.22098094 2.6698044
#> 3   1.04926810 0.5845146 -0.8769038 -0.71998430 1.7692524
#> 4   3.66896472 0.8891407  0.9892528  0.64147282 3.0274919
#> 5   2.35154173 0.4427662  0.7127335  1.16777220 1.1837695
#> 6   3.07425815 0.9243977  0.2953492 -0.09886066 3.1731188
#> 7   3.43722512 0.8308377  0.9160960  0.65055014 2.7866750
#> 8  -0.20830288 0.2893312 -1.5813466 -0.75831824 0.5500154
#> 9   2.66133487 0.6487369  0.5917385  0.62681635 2.0345185
#> 10 -0.07248983 0.5397142 -1.9034955 -1.65669704 1.5842072

test$gen(10, test = TRUE)
#>             y          x          e test_name   statistic   p_value
#> 1  2.40649195 0.57982989  0.4904594    F-test 0.006126554 0.9398019
#> 2  0.07402445 0.01304330 -0.9391890    F-test 0.006126554 0.9398019
#> 3  1.29135533 0.40106551 -0.2705637    F-test 0.006126554 0.9398019
#> 4  0.75287488 0.63046181 -1.2750690    F-test 0.006126554 0.9398019
#> 5  1.20961525 0.78422670 -1.1896230    F-test 0.006126554 0.9398019
#> 6  3.21725919 0.95390552  0.3534179    F-test 0.006126554 0.9398019
#> 7  2.75722441 0.95857097 -0.1202049    F-test 0.006126554 0.9398019
#> 8  1.72318470 0.05483998  0.6653373    F-test 0.006126554 0.9398019
#> 9  2.74157755 0.37978898  1.2175489    F-test 0.006126554 0.9398019
#> 10 2.47291758 0.48839867  0.7459857    F-test 0.006126554 0.9398019