Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>             y          x           e      .resid  .fitted test_name statistic
#> 1   2.1524129 0.10426730  1.03727395  0.72270735 1.429706    F-test 0.5281372
#> 2   0.3825304 0.23910918 -0.91375201 -1.16849680 1.551027    F-test 0.5281372
#> 3   3.4006813 0.50343045  1.64380860  1.61183571 1.788846    F-test 0.5281372
#> 4   1.5221951 0.02519942  0.49636065  0.16362944 1.358566    F-test 0.5281372
#> 5   3.0642905 0.52673036  1.26011526  1.25448125 1.809809    F-test 0.5281372
#> 6   0.1363263 0.87740027 -2.51090517 -1.98899195 2.125318    F-test 0.5281372
#> 7   3.0566197 0.99154866  0.08190234  0.82859846 2.228021    F-test 0.5281372
#> 8  -0.4628448 0.22389686 -1.73687143 -2.00018492 1.537340    F-test 0.5281372
#> 9   1.5586305 0.99476823 -1.42570157 -0.67228754 2.230918    F-test 0.5281372
#> 10  3.1378619 0.61491625  1.14482368  1.24870900 1.889153    F-test 0.5281372
#> 11  0.9312359 0.10426730  1.03727395 -0.49846967 1.429706    F-test 1.7379710
#> 12  3.4149286 0.23910918 -0.91375201  1.86390142 1.551027    F-test 1.7379710
#> 13  2.9325006 0.50343045  1.64380860  1.14365509 1.788846    F-test 1.7379710
#> 14  2.3809959 0.02519942  0.49636065  1.02243024 1.358566    F-test 1.7379710
#> 15 -0.6434562 0.52673036  1.26011526 -2.45326544 1.809809    F-test 1.7379710
#> 16  1.5383124 0.87740027 -2.51090517 -0.58700589 2.125318    F-test 1.7379710
#> 17  3.3713662 0.99154866  0.08190234  1.14334487 2.228021    F-test 1.7379710
#> 18  0.4105371 0.22389686 -1.73687143 -1.12680306 1.537340    F-test 1.7379710
#> 19  2.8703346 0.99476823 -1.42570157  0.63941654 2.230918    F-test 1.7379710
#> 20  0.7419488 0.61491625  1.14482368 -1.14720409 1.889153    F-test 1.7379710
#> 21  3.5513352 0.10426730  1.03727395  2.12162964 1.429706    F-test 0.1695475
#> 22  1.0966655 0.23910918 -0.91375201 -0.45436166 1.551027    F-test 0.1695475
#> 23  2.6584731 0.50343045  1.64380860  0.86962758 1.788846    F-test 0.1695475
#> 24  0.6624132 0.02519942  0.49636065 -0.69615241 1.358566    F-test 0.1695475
#> 25  1.7825485 0.52673036  1.26011526 -0.02726072 1.809809    F-test 0.1695475
#> 26  3.7252398 0.87740027 -2.51090517  1.59992146 2.125318    F-test 0.1695475
#> 27  2.5620681 0.99154866  0.08190234  0.33404679 2.228021    F-test 0.1695475
#> 28 -0.5055616 0.22389686 -1.73687143 -2.04290172 1.537340    F-test 0.1695475
#> 29  0.2619645 0.99476823 -1.42570157 -1.96895359 2.230918    F-test 0.1695475
#> 30  2.1535576 0.61491625  1.14482368  0.26440463 1.889153    F-test 0.1695475
#>      p_value k  null
#> 1  0.4909732 1 FALSE
#> 2  0.4909732 1 FALSE
#> 3  0.4909732 1 FALSE
#> 4  0.4909732 1 FALSE
#> 5  0.4909732 1 FALSE
#> 6  0.4909732 1 FALSE
#> 7  0.4909732 1 FALSE
#> 8  0.4909732 1 FALSE
#> 9  0.4909732 1 FALSE
#> 10 0.4909732 1 FALSE
#> 11 0.2288932 2  TRUE
#> 12 0.2288932 2  TRUE
#> 13 0.2288932 2  TRUE
#> 14 0.2288932 2  TRUE
#> 15 0.2288932 2  TRUE
#> 16 0.2288932 2  TRUE
#> 17 0.2288932 2  TRUE
#> 18 0.2288932 2  TRUE
#> 19 0.2288932 2  TRUE
#> 20 0.2288932 2  TRUE
#> 21 0.6928302 3  TRUE
#> 22 0.6928302 3  TRUE
#> 23 0.6928302 3  TRUE
#> 24 0.6928302 3  TRUE
#> 25 0.6928302 3  TRUE
#> 26 0.6928302 3  TRUE
#> 27 0.6928302 3  TRUE
#> 28 0.6928302 3  TRUE
#> 29 0.6928302 3  TRUE
#> 30 0.6928302 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e         x      .resid  .fitted test_name statistic   p_value k
#> 1   2.950930  1 0.5958694 -4.62844677 7.579377    F-test  1.999333 0.2002665 1
#> 2   3.397989  2 0.3049779 -3.35131672 6.749306    F-test  1.999333 0.2002665 1
#> 3   4.258895  3 0.2133690 -2.22900116 6.487896    F-test  1.999333 0.2002665 1
#> 4   6.144826  4 0.6810276 -1.67755261 7.822379    F-test  1.999333 0.2002665 1
#> 5   7.228760  5 0.7160427 -0.69353589 7.922296    F-test  1.999333 0.2002665 1
#> 6   7.355962  6 0.2784358  0.68239506 6.673567    F-test  1.999333 0.2002665 1
#> 7   9.606921  7 0.8626888  1.26616460 8.340756    F-test  1.999333 0.2002665 1
#> 8  10.960311  8 0.9867116  2.26565128 8.694660    F-test  1.999333 0.2002665 1
#> 9  10.292339  9 0.2364370  3.73861744 6.553722    F-test  1.999333 0.2002665 1
#> 10 11.426745 10 0.3226451  4.62702477 6.799720    F-test  1.999333 0.2002665 1
#> 11  7.520468  1 0.5958694 -0.05890898 7.579377    F-test  2.789333 0.1388208 2
#> 12  9.413713  2 0.3049779  2.66440638 6.749306    F-test  2.789333 0.1388208 2
#> 13  6.814522  3 0.2133690  0.32662547 6.487896    F-test  2.789333 0.1388208 2
#> 14  7.611933  4 0.6810276 -0.21044549 7.822379    F-test  2.789333 0.1388208 2
#> 15  4.206838  5 0.7160427 -3.71545771 7.922296    F-test  2.789333 0.1388208 2
#> 16  9.338717  6 0.2784358  2.66514949 6.673567    F-test  2.789333 0.1388208 2
#> 17  4.706172  7 0.8626888 -3.63458408 8.340756    F-test  2.789333 0.1388208 2
#> 18 14.257111  8 0.9867116  5.56245106 8.694660    F-test  2.789333 0.1388208 2
#> 19  6.289785  9 0.2364370 -0.26393748 6.553722    F-test  2.789333 0.1388208 2
#> 20  3.464421 10 0.3226451 -3.33529867 6.799720    F-test  2.789333 0.1388208 2
#> 21  7.277793  1 0.5958694 -0.30158317 7.579377    F-test  2.651874 0.1474520 3
#> 22  7.956858  2 0.3049779  1.20755232 6.749306    F-test  2.651874 0.1474520 3
#> 23  9.439614  3 0.2133690  2.95171800 6.487896    F-test  2.651874 0.1474520 3
#> 24 11.332830  4 0.6810276  3.51045130 7.822379    F-test  2.651874 0.1474520 3
#> 25 10.538009  5 0.7160427  2.61571303 7.922296    F-test  2.651874 0.1474520 3
#> 26  1.925303  6 0.2784358 -4.74826409 6.673567    F-test  2.651874 0.1474520 3
#> 27 10.063055  7 0.8626888  1.72229911 8.340756    F-test  2.651874 0.1474520 3
#> 28  3.876281  8 0.9867116 -4.81837916 8.694660    F-test  2.651874 0.1474520 3
#> 29  4.246785  9 0.2364370 -2.30693749 6.553722    F-test  2.651874 0.1474520 3
#> 30  6.967150 10 0.3226451  0.16743015 6.799720    F-test  2.651874 0.1474520 3
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE