Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

A data frame.

Examples

# Instantiate
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>              y         x            e       .resid   .fitted test_name
#> 1   2.75612172 0.8220634  0.258269992  0.562087389 2.1940343    F-test
#> 2   0.57119321 0.1392948 -0.587504655 -0.157623064 0.7288163    F-test
#> 3   1.56820604 0.4064649 -0.003472651  0.266044010 1.3021620    F-test
#> 4   1.63222172 0.2647789  0.297334987  0.634117278 0.9981044    F-test
#> 5   2.08740462 0.7218682 -0.155557192  0.108388841 1.9790158    F-test
#> 6   3.04627773 0.1955368  1.812506270  2.196766413 0.8495113    F-test
#> 7   1.40229919 0.5557542 -0.462317651 -0.220236754 1.6225359    F-test
#> 8   0.73852554 0.1465037 -0.429441459 -0.005760903 0.7442864    F-test
#> 9   1.28379616 0.8443289 -1.273424126 -0.958019823 2.2418160    F-test
#> 10 -1.73770962 0.1203001 -2.872481871 -2.425763387 0.6880538    F-test
#> 11  2.68904883 0.8220634  0.258269992  0.495014494 2.1940343    F-test
#> 12  0.57838104 0.1392948 -0.587504655 -0.150435234 0.7288163    F-test
#> 13  0.16796773 0.4064649 -0.003472651 -1.134194295 1.3021620    F-test
#> 14  0.12109886 0.2647789  0.297334987 -0.877005577 0.9981044    F-test
#> 15  0.08555908 0.7218682 -0.155557192 -1.893456692 1.9790158    F-test
#> 16 -0.15986367 0.1955368  1.812506270 -1.009374983 0.8495113    F-test
#> 17  2.75979164 0.5557542 -0.462317651  1.137255694 1.6225359    F-test
#> 18  1.48264141 0.1465037 -0.429441459  0.738354964 0.7442864    F-test
#> 19  3.35663983 0.8443289 -1.273424126  1.114823848 2.2418160    F-test
#> 20  2.26707155 0.1203001 -2.872481871  1.579017782 0.6880538    F-test
#> 21  2.17378228 0.8220634  0.258269992 -0.020252055 2.1940343    F-test
#> 22 -0.92897639 0.1392948 -0.587504655 -1.657792665 0.7288163    F-test
#> 23  3.23260170 0.4064649 -0.003472651  1.930439677 1.3021620    F-test
#> 24  0.09836826 0.2647789  0.297334987 -0.899736179 0.9981044    F-test
#> 25  1.83815780 0.7218682 -0.155557192 -0.140857972 1.9790158    F-test
#> 26  1.97038268 0.1955368  1.812506270  1.120871368 0.8495113    F-test
#> 27  3.03417744 0.5557542 -0.462317651  1.411641500 1.6225359    F-test
#> 28  0.42608364 0.1465037 -0.429441459 -0.318202808 0.7442864    F-test
#> 29  0.88456106 0.8443289 -1.273424126 -1.357254927 2.2418160    F-test
#> 30  0.61919783 0.1203001 -2.872481871 -0.068855938 0.6880538    F-test
#>    statistic    p_value k  null
#> 1  0.9669185 0.35820798 2 FALSE
#> 2  0.9669185 0.35820798 2 FALSE
#> 3  0.9669185 0.35820798 2 FALSE
#> 4  0.9669185 0.35820798 2 FALSE
#> 5  0.9669185 0.35820798 2 FALSE
#> 6  0.9669185 0.35820798 2 FALSE
#> 7  0.9669185 0.35820798 2 FALSE
#> 8  0.9669185 0.35820798 2 FALSE
#> 9  0.9669185 0.35820798 2 FALSE
#> 10 0.9669185 0.35820798 2 FALSE
#> 11 1.9558591 0.20466585 1  TRUE
#> 12 1.9558591 0.20466585 1  TRUE
#> 13 1.9558591 0.20466585 1  TRUE
#> 14 1.9558591 0.20466585 1  TRUE
#> 15 1.9558591 0.20466585 1  TRUE
#> 16 1.9558591 0.20466585 1  TRUE
#> 17 1.9558591 0.20466585 1  TRUE
#> 18 1.9558591 0.20466585 1  TRUE
#> 19 1.9558591 0.20466585 1  TRUE
#> 20 1.9558591 0.20466585 1  TRUE
#> 21 6.5241400 0.03786856 3  TRUE
#> 22 6.5241400 0.03786856 3  TRUE
#> 23 6.5241400 0.03786856 3  TRUE
#> 24 6.5241400 0.03786856 3  TRUE
#> 25 6.5241400 0.03786856 3  TRUE
#> 26 6.5241400 0.03786856 3  TRUE
#> 27 6.5241400 0.03786856 3  TRUE
#> 28 6.5241400 0.03786856 3  TRUE
#> 29 6.5241400 0.03786856 3  TRUE
#> 30 6.5241400 0.03786856 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e         x     .resid  .fitted test_name   statistic   p_value k
#> 1   3.076824  1 0.6518783 -3.2829367 6.359760    F-test 1.952901249 0.2049701 2
#> 2   3.418644  2 0.3177063 -4.4534066 7.872050    F-test 1.952901249 0.2049701 2
#> 3   5.919622  3 0.9729638  1.0129295 4.906693    F-test 1.952901249 0.2049701 2
#> 4   6.536585  4 0.8366318  1.0129230 5.523661    F-test 1.952901249 0.2049701 2
#> 5   7.233376  5 0.7179394  1.1725747 6.060802    F-test 1.952901249 0.2049701 2
#> 6   7.467472  6 0.3470372 -0.2718416 7.739314    F-test 1.952901249 0.2049701 2
#> 7   8.219779  7 0.1854045 -0.2510005 8.470780    F-test 1.952901249 0.2049701 2
#> 8   9.110513  8 0.1004275  0.2551709 8.855342    F-test 1.952901249 0.2049701 2
#> 9  10.301130  9 0.2423813  2.0881973 8.212933    F-test 1.952901249 0.2049701 2
#> 10 11.212453 10 0.1800387  2.7173899 8.495063    F-test 1.952901249 0.2049701 2
#> 11  8.718398  1 0.6518783  2.3586382 6.359760    F-test 0.006544068 0.9377894 1
#> 12  7.725840  2 0.3177063 -0.1462104 7.872050    F-test 0.006544068 0.9377894 1
#> 13  2.774579  3 0.9729638 -2.1321142 4.906693    F-test 0.006544068 0.9377894 1
#> 14  8.037602  4 0.8366318  2.5139405 5.523661    F-test 0.006544068 0.9377894 1
#> 15  5.604121  5 0.7179394 -0.4566811 6.060802    F-test 0.006544068 0.9377894 1
#> 16  4.572628  6 0.3470372 -3.1666856 7.739314    F-test 0.006544068 0.9377894 1
#> 17  8.858606  7 0.1854045  0.3878266 8.470780    F-test 0.006544068 0.9377894 1
#> 18  9.877196  8 0.1004275  1.0218543 8.855342    F-test 0.006544068 0.9377894 1
#> 19  5.020113  9 0.2423813 -3.1928193 8.212933    F-test 0.006544068 0.9377894 1
#> 20 11.307314 10 0.1800387  2.8122511 8.495063    F-test 0.006544068 0.9377894 1
#> 21  8.484116  1 0.6518783  2.1243561 6.359760    F-test 0.001576400 0.9694378 3
#> 22  9.498122  2 0.3177063  1.6260719 7.872050    F-test 0.001576400 0.9694378 3
#> 23  6.634346  3 0.9729638  1.7276527 4.906693    F-test 0.001576400 0.9694378 3
#> 24  2.456534  4 0.8366318 -3.0671271 5.523661    F-test 0.001576400 0.9694378 3
#> 25  5.259321  5 0.7179394 -0.8014812 6.060802    F-test 0.001576400 0.9694378 3
#> 26  6.403759  6 0.3470372 -1.3355544 7.739314    F-test 0.001576400 0.9694378 3
#> 27  6.969364  7 0.1854045 -1.5014155 8.470780    F-test 0.001576400 0.9694378 3
#> 28 10.249545  8 0.1004275  1.3942027 8.855342    F-test 0.001576400 0.9694378 3
#> 29 11.275783  9 0.2423813  3.0628505 8.212933    F-test 0.001576400 0.9694378 3
#> 30  5.265507 10 0.1800387 -3.2295557 8.495063    F-test 0.001576400 0.9694378 3
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE