Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

A data frame.

Examples

# Instantiate
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>             y          x           e        .resid   .fitted test_name
#> 1   2.4548708 0.56601224  0.56848871  0.3830774953 2.0717933    F-test
#> 2   0.6890208 0.05504359 -0.36905260 -0.0716070701 0.7606279    F-test
#> 3   1.3730824 0.32821141 -0.06285169 -0.0885046812 1.4615871    F-test
#> 4   2.8498624 0.75561036  0.52330507  0.2915529774 2.5583095    F-test
#> 5   3.6948824 0.53170203  1.88047334  1.7111304382 1.9837520    F-test
#> 6   3.9566357 0.91376818  1.20789525  0.9924871232 2.9641486    F-test
#> 7   1.4475926 0.72036313 -0.79169359 -1.0202711215 2.4678637    F-test
#> 8   0.8016632 0.12683997 -0.34126515 -0.1431969774 0.9448602    F-test
#> 9   1.6993502 0.58772859 -0.23380329 -0.4281681121 2.1275183    F-test
#> 10  0.9114202 0.74766457 -1.39524669 -1.6265000717 2.5379203    F-test
#> 11  2.3546278 0.56601224  0.56848871  0.2828345023 2.0717933    F-test
#> 12 -0.2936781 0.05504359 -0.36905260 -1.0543059419 0.7606279    F-test
#> 13  3.2000929 0.32821141 -0.06285169  1.7385057921 1.4615871    F-test
#> 14  1.8138637 0.75561036  0.52330507 -0.7444457538 2.5583095    F-test
#> 15  2.9577211 0.53170203  1.88047334  0.9739691206 1.9837520    F-test
#> 16  3.8107743 0.91376818  1.20789525  0.8466256828 2.9641486    F-test
#> 17  1.4209700 0.72036313 -0.79169359 -1.0468937098 2.4678637    F-test
#> 18  0.9438733 0.12683997 -0.34126515 -0.0009868701 0.9448602    F-test
#> 19  1.5222826 0.58772859 -0.23380329 -0.6052357221 2.1275183    F-test
#> 20  2.1478532 0.74766457 -1.39524669 -0.3900671000 2.5379203    F-test
#> 21  2.5321727 0.56601224  0.56848871  0.4603794148 2.0717933    F-test
#> 22  0.9888401 0.05504359 -0.36905260  0.2282122104 0.7606279    F-test
#> 23  0.7018914 0.32821141 -0.06285169 -0.7596956973 1.4615871    F-test
#> 24  2.3245072 0.75561036  0.52330507 -0.2338022528 2.5583095    F-test
#> 25  2.8803217 0.53170203  1.88047334  0.8965697189 1.9837520    F-test
#> 26  4.1091557 0.91376818  1.20789525  1.1450071512 2.9641486    F-test
#> 27  3.0290453 0.72036313 -0.79169359  0.5611816111 2.4678637    F-test
#> 28  1.1219764 0.12683997 -0.34126515  0.1771161799 0.9448602    F-test
#> 29  1.7923258 0.58772859 -0.23380329 -0.3351925305 2.1275183    F-test
#> 30  0.3981445 0.74766457 -1.39524669 -2.1397758057 2.5379203    F-test
#>       statistic   p_value k  null
#> 1  0.0001877969 0.9894486 3 FALSE
#> 2  0.0001877969 0.9894486 3 FALSE
#> 3  0.0001877969 0.9894486 3 FALSE
#> 4  0.0001877969 0.9894486 3 FALSE
#> 5  0.0001877969 0.9894486 3 FALSE
#> 6  0.0001877969 0.9894486 3 FALSE
#> 7  0.0001877969 0.9894486 3 FALSE
#> 8  0.0001877969 0.9894486 3 FALSE
#> 9  0.0001877969 0.9894486 3 FALSE
#> 10 0.0001877969 0.9894486 3 FALSE
#> 11 0.4179728330 0.5385619 1  TRUE
#> 12 0.4179728330 0.5385619 1  TRUE
#> 13 0.4179728330 0.5385619 1  TRUE
#> 14 0.4179728330 0.5385619 1  TRUE
#> 15 0.4179728330 0.5385619 1  TRUE
#> 16 0.4179728330 0.5385619 1  TRUE
#> 17 0.4179728330 0.5385619 1  TRUE
#> 18 0.4179728330 0.5385619 1  TRUE
#> 19 0.4179728330 0.5385619 1  TRUE
#> 20 0.4179728330 0.5385619 1  TRUE
#> 21 0.3341069184 0.5813603 2  TRUE
#> 22 0.3341069184 0.5813603 2  TRUE
#> 23 0.3341069184 0.5813603 2  TRUE
#> 24 0.3341069184 0.5813603 2  TRUE
#> 25 0.3341069184 0.5813603 2  TRUE
#> 26 0.3341069184 0.5813603 2  TRUE
#> 27 0.3341069184 0.5813603 2  TRUE
#> 28 0.3341069184 0.5813603 2  TRUE
#> 29 0.3341069184 0.5813603 2  TRUE
#> 30 0.3341069184 0.5813603 2  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e         x      .resid  .fitted test_name statistic    p_value k
#> 1   2.277754  1 0.2264670 -5.29678651 7.574541    F-test 6.7692404 0.03533368 2
#> 2   4.222004  2 0.7132618 -2.94874962 7.170754    F-test 6.7692404 0.03533368 2
#> 3   5.583323  3 0.8540026 -1.47068893 7.054012    F-test 6.7692404 0.03533368 2
#> 4   5.496576  4 0.3640464 -1.96384525 7.460421    F-test 6.7692404 0.03533368 2
#> 5   7.252248  5 0.7256623  0.09178026 7.160468    F-test 6.7692404 0.03533368 2
#> 6   7.794603  6 0.5220582  0.46524945 7.329354    F-test 6.7692404 0.03533368 2
#> 7   8.640886  7 0.4438678  1.24667545 7.394211    F-test 6.7692404 0.03533368 2
#> 8   9.835527  8 0.5418861  2.52261994 7.312907    F-test 6.7692404 0.03533368 2
#> 9  10.651929  9 0.4496992  3.26255454 7.389374    F-test 6.7692404 0.03533368 2
#> 10 11.533910 10 0.3853871  4.09119067 7.442720    F-test 6.7692404 0.03533368 2
#> 11  5.467986  1 0.2264670 -2.10655516 7.574541    F-test 1.6029618 0.24598880 1
#> 12  7.437880  2 0.7132618  0.26712655 7.170754    F-test 1.6029618 0.24598880 1
#> 13  4.661744  3 0.8540026 -2.39226845 7.054012    F-test 1.6029618 0.24598880 1
#> 14  7.017613  4 0.3640464 -0.44280793 7.460421    F-test 1.6029618 0.24598880 1
#> 15  8.357301  5 0.7256623  1.19683282 7.160468    F-test 1.6029618 0.24598880 1
#> 16  8.597558  6 0.5220582  1.26820451 7.329354    F-test 1.6029618 0.24598880 1
#> 17  8.019912  7 0.4438678  0.62570050 7.394211    F-test 1.6029618 0.24598880 1
#> 18  4.884322  8 0.5418861 -2.42858428 7.312907    F-test 1.6029618 0.24598880 1
#> 19 14.419879  9 0.4496992  7.03050483 7.389374    F-test 1.6029618 0.24598880 1
#> 20  4.424566 10 0.3853871 -3.01815339 7.442720    F-test 1.6029618 0.24598880 1
#> 21  9.344821  1 0.2264670  1.77027998 7.574541    F-test 0.8438687 0.38884903 3
#> 22  4.398381  2 0.7132618 -2.77237253 7.170754    F-test 0.8438687 0.38884903 3
#> 23  6.705706  3 0.8540026 -0.34830638 7.054012    F-test 0.8438687 0.38884903 3
#> 24 10.491909  4 0.3640464  3.03148741 7.460421    F-test 0.8438687 0.38884903 3
#> 25 13.659718  5 0.7256623  6.49924994 7.160468    F-test 0.8438687 0.38884903 3
#> 26  5.878157  6 0.5220582 -1.45119632 7.329354    F-test 0.8438687 0.38884903 3
#> 27  6.447806  7 0.4438678 -0.94640519 7.394211    F-test 0.8438687 0.38884903 3
#> 28  4.337466  8 0.5418861 -2.97544061 7.312907    F-test 0.8438687 0.38884903 3
#> 29  6.239399  9 0.4496992 -1.14997480 7.389374    F-test 0.8438687 0.38884903 3
#> 30  5.785398 10 0.3853871 -1.65732150 7.442720    F-test 0.8438687 0.38884903 3
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE