Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>           y         x          e      .resid  .fitted test_name   statistic
#> 1  1.942101 0.6872572 -0.2174791 -0.90287672 2.844977    F-test 0.146485445
#> 2  3.447465 0.4090573  1.8710794  1.19045425 2.257010    F-test 0.146485445
#> 3  1.934516 0.1831579  0.7178112  0.15493718 1.779579    F-test 0.146485445
#> 4  2.797291 0.5161244  1.0147820  0.31399749 2.483293    F-test 0.146485445
#> 5  1.475220 0.2654740  0.1392695 -0.47833145 1.953551    F-test 0.146485445
#> 6  2.196719 0.4169467  0.6059282 -0.07696481 2.273684    F-test 0.146485445
#> 7  2.044749 0.5852563  0.1169677 -0.58465256 2.629402    F-test 0.146485445
#> 8  3.715081 0.9582978  0.8384486  0.29726772 3.417813    F-test 0.146485445
#> 9  3.820390 0.6950205  1.6423160  0.95900524 2.861385    F-test 0.146485445
#> 10 1.747401 0.5809201 -0.1709875 -0.87283633 2.620237    F-test 0.146485445
#> 11 2.721116 0.6872572 -0.2174791 -0.12386138 2.844977    F-test 0.004730139
#> 12 3.500018 0.4090573  1.8710794  1.24300730 2.257010    F-test 0.004730139
#> 13 1.077341 0.1831579  0.7178112 -0.70223736 1.779579    F-test 0.004730139
#> 14 1.205552 0.5161244  1.0147820 -1.27774141 2.483293    F-test 0.004730139
#> 15 2.678100 0.2654740  0.1392695  0.72454832 1.953551    F-test 0.004730139
#> 16 2.228012 0.4169467  0.6059282 -0.04567247 2.273684    F-test 0.004730139
#> 17 2.111825 0.5852563  0.1169677 -0.51757643 2.629402    F-test 0.004730139
#> 18 3.522557 0.9582978  0.8384486  0.10474367 3.417813    F-test 0.004730139
#> 19 3.286145 0.6950205  1.6423160  0.42475992 2.861385    F-test 0.004730139
#> 20 2.790267 0.5809201 -0.1709875  0.17002984 2.620237    F-test 0.004730139
#> 21 3.535887 0.6872572 -0.2174791  0.69090964 2.844977    F-test 0.135547369
#> 22 3.037252 0.4090573  1.8710794  0.78024167 2.257010    F-test 0.135547369
#> 23 1.511763 0.1831579  0.7178112 -0.26781584 1.779579    F-test 0.135547369
#> 24 2.041972 0.5161244  1.0147820 -0.44132169 2.483293    F-test 0.135547369
#> 25 2.416099 0.2654740  0.1392695  0.46254783 1.953551    F-test 0.135547369
#> 26 1.692252 0.4169467  0.6059282 -0.58143196 2.273684    F-test 0.135547369
#> 27 3.106429 0.5852563  0.1169677  0.47702718 2.629402    F-test 0.135547369
#> 28 3.818106 0.9582978  0.8384486  0.40029282 3.417813    F-test 0.135547369
#> 29 1.309610 0.6950205  1.6423160 -1.55177501 2.861385    F-test 0.135547369
#> 30 2.651562 0.5809201 -0.1709875  0.03132536 2.620237    F-test 0.135547369
#>      p_value k  null
#> 1  0.7132733 2 FALSE
#> 2  0.7132733 2 FALSE
#> 3  0.7132733 2 FALSE
#> 4  0.7132733 2 FALSE
#> 5  0.7132733 2 FALSE
#> 6  0.7132733 2 FALSE
#> 7  0.7132733 2 FALSE
#> 8  0.7132733 2 FALSE
#> 9  0.7132733 2 FALSE
#> 10 0.7132733 2 FALSE
#> 11 0.9470913 1  TRUE
#> 12 0.9470913 1  TRUE
#> 13 0.9470913 1  TRUE
#> 14 0.9470913 1  TRUE
#> 15 0.9470913 1  TRUE
#> 16 0.9470913 1  TRUE
#> 17 0.9470913 1  TRUE
#> 18 0.9470913 1  TRUE
#> 19 0.9470913 1  TRUE
#> 20 0.9470913 1  TRUE
#> 21 0.7236295 3  TRUE
#> 22 0.7236295 3  TRUE
#> 23 0.7236295 3  TRUE
#> 24 0.7236295 3  TRUE
#> 25 0.7236295 3  TRUE
#> 26 0.7236295 3  TRUE
#> 27 0.7236295 3  TRUE
#> 28 0.7236295 3  TRUE
#> 29 0.7236295 3  TRUE
#> 30 0.7236295 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e         x     .resid  .fitted test_name statistic    p_value k
#> 1   3.049370  1 0.6398993 -4.5536703 7.603041    F-test 0.7553843 0.41358194 3
#> 2   4.571374  2 0.8495831 -3.6133017 8.184676    F-test 0.7553843 0.41358194 3
#> 3   4.126367  3 0.1134874 -2.0164759 6.142843    F-test 0.7553843 0.41358194 3
#> 4   6.335998  4 0.7593639 -1.5984226 7.934420    F-test 0.7553843 0.41358194 3
#> 5   7.283366  5 0.7382917 -0.5926021 7.875968    F-test 0.7553843 0.41358194 3
#> 6   8.990155  6 0.9967149  0.3973547 8.592801    F-test 0.7553843 0.41358194 3
#> 7   9.814923  7 0.9369840  1.3878080 8.427115    F-test 0.7553843 0.41358194 3
#> 8  10.793062  8 0.9293573  2.3871028 8.405960    F-test 0.7553843 0.41358194 3
#> 9  11.534897  9 0.8360004  3.3878974 8.147000    F-test 0.7553843 0.41358194 3
#> 10 11.285749 10 0.2319490  4.8143097 6.471440    F-test 0.7553843 0.41358194 3
#> 11  8.018534  1 0.6398993  0.4154937 7.603041    F-test 0.4009941 0.54669973 1
#> 12  4.799702  2 0.8495831 -3.3849745 8.184676    F-test 0.4009941 0.54669973 1
#> 13  5.686465  3 0.1134874 -0.4563773 6.142843    F-test 0.4009941 0.54669973 1
#> 14 13.725104  4 0.7593639  5.7906835 7.934420    F-test 0.4009941 0.54669973 1
#> 15  8.966287  5 0.7382917  1.0903188 7.875968    F-test 0.4009941 0.54669973 1
#> 16 10.134652  6 0.9967149  1.5418518 8.592801    F-test 0.4009941 0.54669973 1
#> 17 10.341938  7 0.9369840  1.9148228 8.427115    F-test 0.4009941 0.54669973 1
#> 18  2.974975  8 0.9293573 -5.4309842 8.405960    F-test 0.4009941 0.54669973 1
#> 19  7.351347  9 0.8360004 -0.7956531 8.147000    F-test 0.4009941 0.54669973 1
#> 20  5.786258 10 0.2319490 -0.6851814 6.471440    F-test 0.4009941 0.54669973 1
#> 21  9.706517  1 0.6398993  2.1034767 7.603041    F-test 4.9201506 0.06204722 2
#> 22  4.891635  2 0.8495831 -3.2930415 8.184676    F-test 4.9201506 0.06204722 2
#> 23  3.606101  3 0.1134874 -2.5367419 6.142843    F-test 4.9201506 0.06204722 2
#> 24 11.819771  4 0.7593639  3.8853508 7.934420    F-test 4.9201506 0.06204722 2
#> 25 10.448906  5 0.7382917  2.5729373 7.875968    F-test 4.9201506 0.06204722 2
#> 26  3.024735  6 0.9967149 -5.5680656 8.592801    F-test 4.9201506 0.06204722 2
#> 27 11.150426  7 0.9369840  2.7233111 8.427115    F-test 4.9201506 0.06204722 2
#> 28  8.995259  8 0.9293573  0.5892996 8.405960    F-test 4.9201506 0.06204722 2
#> 29  7.237389  9 0.8360004 -0.9096105 8.147000    F-test 4.9201506 0.06204722 2
#> 30  6.904523 10 0.2319490  0.4330839 6.471440    F-test 4.9201506 0.06204722 2
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE