Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

A data frame.

Examples

# Instantiate
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>             y          x           e      .resid   .fitted test_name statistic
#> 1   2.3802795 0.61341573  0.39058487 -0.08255143 2.4628309    F-test 2.6955139
#> 2   3.2399521 0.59762387  1.28517397  0.82328567 2.4166665    F-test 2.6955139
#> 3   2.2600751 0.37281205  0.74827419  0.50060217 1.7594729    F-test 2.6955139
#> 4   1.1171876 0.13962608 -0.04193389  0.03938850 1.0777991    F-test 2.6955139
#> 5  -0.1132013 0.06617128 -1.18375119 -0.97626960 0.8630683    F-test 2.6955139
#> 6   1.7736432 0.14745957  0.60443933  0.67294442 1.1006988    F-test 2.6955139
#> 7   2.3874673 0.81355915 -0.08797037 -0.66044393 3.0479112    F-test 2.6955139
#> 8   1.3098475 0.19327227  0.07922108  0.07522420 1.2346233    F-test 2.6955139
#> 9   1.8437381 0.38086118  0.31782170  0.06073515 1.7830030    F-test 2.6955139
#> 10  1.3042733 0.37203060 -0.20616404 -0.45291515 1.7571885    F-test 2.6955139
#> 11  2.5006127 0.61341573  0.39058487  0.03778186 2.4628309    F-test 0.3191423
#> 12  1.4115777 0.59762387  1.28517397 -1.00508877 2.4166665    F-test 0.3191423
#> 13  2.2109808 0.37281205  0.74827419  0.45150796 1.7594729    F-test 0.3191423
#> 14  1.8388152 0.13962608 -0.04193389  0.76101610 1.0777991    F-test 0.3191423
#> 15  1.0861349 0.06617128 -1.18375119  0.22306661 0.8630683    F-test 0.3191423
#> 16  0.5324496 0.14745957  0.60443933 -0.56824923 1.1006988    F-test 0.3191423
#> 17  3.5301710 0.81355915 -0.08797037  0.48225981 3.0479112    F-test 0.3191423
#> 18  0.5719455 0.19327227  0.07922108 -0.66267784 1.2346233    F-test 0.3191423
#> 19  2.1630537 0.38086118  0.31782170  0.38005072 1.7830030    F-test 0.3191423
#> 20  1.6575213 0.37203060 -0.20616404 -0.09966721 1.7571885    F-test 0.3191423
#> 21  2.8360924 0.61341573  0.39058487  0.37326154 2.4628309    F-test 1.9673973
#> 22  2.2069398 0.59762387  1.28517397 -0.20972671 2.4166665    F-test 1.9673973
#> 23  1.3824124 0.37281205  0.74827419 -0.37706052 1.7594729    F-test 1.9673973
#> 24  2.3269406 0.13962608 -0.04193389  1.24914148 1.0777991    F-test 1.9673973
#> 25  0.8506354 0.06617128 -1.18375119 -0.01243295 0.8630683    F-test 1.9673973
#> 26  0.9219745 0.14745957  0.60443933 -0.17872434 1.1006988    F-test 1.9673973
#> 27  3.4000458 0.81355915 -0.08797037  0.35213463 3.0479112    F-test 1.9673973
#> 28  0.9644151 0.19327227  0.07922108 -0.27020827 1.2346233    F-test 1.9673973
#> 29  1.7938090 0.38086118  0.31782170  0.01080601 1.7830030    F-test 1.9673973
#> 30  0.8199976 0.37203060 -0.20616404 -0.93719086 1.7571885    F-test 1.9673973
#>      p_value k  null
#> 1  0.1446348 2 FALSE
#> 2  0.1446348 2 FALSE
#> 3  0.1446348 2 FALSE
#> 4  0.1446348 2 FALSE
#> 5  0.1446348 2 FALSE
#> 6  0.1446348 2 FALSE
#> 7  0.1446348 2 FALSE
#> 8  0.1446348 2 FALSE
#> 9  0.1446348 2 FALSE
#> 10 0.1446348 2 FALSE
#> 11 0.5897608 1  TRUE
#> 12 0.5897608 1  TRUE
#> 13 0.5897608 1  TRUE
#> 14 0.5897608 1  TRUE
#> 15 0.5897608 1  TRUE
#> 16 0.5897608 1  TRUE
#> 17 0.5897608 1  TRUE
#> 18 0.5897608 1  TRUE
#> 19 0.5897608 1  TRUE
#> 20 0.5897608 1  TRUE
#> 21 0.2034852 3  TRUE
#> 22 0.2034852 3  TRUE
#> 23 0.2034852 3  TRUE
#> 24 0.2034852 3  TRUE
#> 25 0.2034852 3  TRUE
#> 26 0.2034852 3  TRUE
#> 27 0.2034852 3  TRUE
#> 28 0.2034852 3  TRUE
#> 29 0.2034852 3  TRUE
#> 30 0.2034852 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e          x     .resid  .fitted test_name statistic   p_value k
#> 1   3.446628  1 0.80254684 -3.0469076 6.493536    F-test 1.0439697 0.3409071 1
#> 2   4.937229  2 0.97892847 -1.1401652 6.077395    F-test 1.0439697 0.3409071 1
#> 3   4.228099  3 0.19144672 -3.7072195 7.935318    F-test 1.0439697 0.3409071 1
#> 4   5.436983  4 0.32884448 -2.1741693 7.611152    F-test 1.0439697 0.3409071 1
#> 5   6.596140  5 0.41985892 -0.8002796 7.396420    F-test 1.0439697 0.3409071 1
#> 6   8.459883  6 0.80762485  1.9783276 6.481555    F-test 1.0439697 0.3409071 1
#> 7   8.012582  7 0.01242722 -0.3451012 8.357683    F-test 1.0439697 0.3409071 1
#> 8   9.425201  8 0.32170596  1.7972061 7.627995    F-test 1.0439697 0.3409071 1
#> 9  10.618159  9 0.43175077  3.2497961 7.368363    F-test 1.0439697 0.3409071 1
#> 10 11.591234 10 0.41718810  4.1885126 7.402721    F-test 1.0439697 0.3409071 1
#> 11  5.769767  1 0.80254684 -0.7237686 6.493536    F-test 1.4764535 0.2637197 2
#> 12  5.716189  2 0.97892847 -0.3612060 6.077395    F-test 1.4764535 0.2637197 2
#> 13 11.403199  3 0.19144672  3.4678814 7.935318    F-test 1.4764535 0.2637197 2
#> 14  4.996199  4 0.32884448 -2.6149537 7.611152    F-test 1.4764535 0.2637197 2
#> 15 11.400734  5 0.41985892  4.0043136 7.396420    F-test 1.4764535 0.2637197 2
#> 16  5.514140  6 0.80762485 -0.9674152 6.481555    F-test 1.4764535 0.2637197 2
#> 17  4.145743  7 0.01242722 -4.2119402 8.357683    F-test 1.4764535 0.2637197 2
#> 18 10.598781  8 0.32170596  2.9707867 7.627995    F-test 1.4764535 0.2637197 2
#> 19  5.960139  9 0.43175077 -1.4082244 7.368363    F-test 1.4764535 0.2637197 2
#> 20  7.247248 10 0.41718810 -0.1554736 7.402721    F-test 1.4764535 0.2637197 2
#> 21  7.376167  1 0.80254684  0.8826312 6.493536    F-test 0.7870613 0.4044391 3
#> 22  1.871759  2 0.97892847 -4.2056360 6.077395    F-test 0.7870613 0.4044391 3
#> 23  6.572333  3 0.19144672 -1.3629853 7.935318    F-test 0.7870613 0.4044391 3
#> 24  6.716436  4 0.32884448 -0.8947169 7.611152    F-test 0.7870613 0.4044391 3
#> 25  3.993063  5 0.41985892 -3.4033571 7.396420    F-test 0.7870613 0.4044391 3
#> 26 10.863905  6 0.80762485  4.3823495 6.481555    F-test 0.7870613 0.4044391 3
#> 27  7.802372  7 0.01242722 -0.5553104 8.357683    F-test 0.7870613 0.4044391 3
#> 28 10.718261  8 0.32170596  3.0902669 7.627995    F-test 0.7870613 0.4044391 3
#> 29  7.641530  9 0.43175077  0.2731665 7.368363    F-test 0.7870613 0.4044391 3
#> 30  9.196313 10 0.41718810  1.7935916 7.402721    F-test 0.7870613 0.4044391 3
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE