Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>             y          x          e      .resid  .fitted test_name statistic
#> 1   4.1987937 0.93442148  1.3912287  1.19625718 3.002536    F-test  9.258098
#> 2   2.3394108 0.12475009  1.1990982  0.75110516 1.588306    F-test  9.258098
#> 3  -0.6043088 0.46403206 -2.2836666 -2.78522894 2.180920    F-test  9.258098
#> 4   1.6300114 0.30129474  0.2379382 -0.26665996 1.896671    F-test  9.258098
#> 5   2.3461069 0.18776085  1.1230920  0.64774210 1.698365    F-test  9.258098
#> 6   1.9797225 0.03244744  0.9462223  0.55263936 1.427083    F-test  9.258098
#> 7   2.4988515 0.73614464  0.2207979 -0.15736027 2.656212    F-test  9.258098
#> 8   1.5896542 0.57704576 -0.3203734 -0.78866390 2.378318    F-test  9.258098
#> 9   3.6620873 0.97892891  0.7248565  0.58181089 3.080276    F-test  9.258098
#> 10  1.7891258 0.08608329  0.6956322  0.26835838 1.520767    F-test  9.258098
#> 11  2.8706202 0.93442148  1.3912287 -0.13191627 3.002536    F-test  3.634777
#> 12  3.0073881 0.12475009  1.1990982  1.41908239 1.588306    F-test  3.634777
#> 13  1.0548089 0.46403206 -2.2836666 -1.12611132 2.180920    F-test  3.634777
#> 14  1.3687943 0.30129474  0.2379382 -0.52787709 1.896671    F-test  3.634777
#> 15  1.6554204 0.18776085  1.1230920 -0.04294442 1.698365    F-test  3.634777
#> 16  0.3319100 0.03244744  0.9462223 -1.09517315 1.427083    F-test  3.634777
#> 17  2.2171093 0.73614464  0.2207979 -0.43910245 2.656212    F-test  3.634777
#> 18  1.1335934 0.57704576 -0.3203734 -1.24472470 2.378318    F-test  3.634777
#> 19  4.6913482 0.97892891  0.7248565  1.61107180 3.080276    F-test  3.634777
#> 20  3.0984626 0.08608329  0.6956322  1.57769521 1.520767    F-test  3.634777
#> 21  2.0948397 0.93442148  1.3912287 -0.90769678 3.002536    F-test  4.480267
#> 22  2.5927054 0.12475009  1.1990982  1.00439976 1.588306    F-test  4.480267
#> 23  1.7401033 0.46403206 -2.2836666 -0.44081684 2.180920    F-test  4.480267
#> 24  3.7326346 0.30129474  0.2379382  1.83596322 1.896671    F-test  4.480267
#> 25  1.9908324 0.18776085  1.1230920  0.29246757 1.698365    F-test  4.480267
#> 26 -0.7265975 0.03244744  0.9462223 -2.15368070 1.427083    F-test  4.480267
#> 27  3.4766876 0.73614464  0.2207979  0.82047587 2.656212    F-test  4.480267
#> 28  2.9319097 0.57704576 -0.3203734  0.55359157 2.378318    F-test  4.480267
#> 29  2.5788219 0.97892891  0.7248565 -0.50145447 3.080276    F-test  4.480267
#> 30  1.0175182 0.08608329  0.6956322 -0.50324920 1.520767    F-test  4.480267
#>       p_value k  null
#> 1  0.01877500 2 FALSE
#> 2  0.01877500 2 FALSE
#> 3  0.01877500 2 FALSE
#> 4  0.01877500 2 FALSE
#> 5  0.01877500 2 FALSE
#> 6  0.01877500 2 FALSE
#> 7  0.01877500 2 FALSE
#> 8  0.01877500 2 FALSE
#> 9  0.01877500 2 FALSE
#> 10 0.01877500 2 FALSE
#> 11 0.09826086 1  TRUE
#> 12 0.09826086 1  TRUE
#> 13 0.09826086 1  TRUE
#> 14 0.09826086 1  TRUE
#> 15 0.09826086 1  TRUE
#> 16 0.09826086 1  TRUE
#> 17 0.09826086 1  TRUE
#> 18 0.09826086 1  TRUE
#> 19 0.09826086 1  TRUE
#> 20 0.09826086 1  TRUE
#> 21 0.07207567 3  TRUE
#> 22 0.07207567 3  TRUE
#> 23 0.07207567 3  TRUE
#> 24 0.07207567 3  TRUE
#> 25 0.07207567 3  TRUE
#> 26 0.07207567 3  TRUE
#> 27 0.07207567 3  TRUE
#> 28 0.07207567 3  TRUE
#> 29 0.07207567 3  TRUE
#> 30 0.07207567 3  TRUE