Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>             y          x           e      .resid   .fitted test_name  statistic
#> 1   1.3472766 0.14595376  0.18002033  0.37439566 0.9728809    F-test 5.75407360
#> 2   1.0748908 0.53118266 -0.73844689 -0.60309120 1.6779820    F-test 5.75407360
#> 3   2.3165739 0.79211228 -0.10298023  0.16100119 2.1555727    F-test 5.75407360
#> 4   1.1786290 0.54151436 -0.65612320 -0.51826359 1.6968925    F-test 5.75407360
#> 5   0.4202316 0.15159643 -0.75434631 -0.56297734 0.9832089    F-test 5.75407360
#> 6   1.5634032 0.18424976  0.34520545  0.52042745 1.0429757    F-test 5.75407360
#> 7   1.3030930 0.72319924 -0.94312335 -0.72634518 2.0294382    F-test 5.75407360
#> 8   3.3520757 0.87314273  0.71655469  1.04818941 2.3038863    F-test 5.75407360
#> 9   0.8184076 0.03464638 -0.21743915  0.04925738 0.7691502    F-test 5.75407360
#> 10  1.0815120 0.06467113  0.01265856  0.25740623 0.8241058    F-test 5.75407360
#> 11  0.3453933 0.14595376  0.18002033 -0.62748763 0.9728809    F-test 1.44433818
#> 12  1.8757109 0.53118266 -0.73844689  0.19772888 1.6779820    F-test 1.44433818
#> 13  2.2769264 0.79211228 -0.10298023  0.12135369 2.1555727    F-test 1.44433818
#> 14  0.6059496 0.54151436 -0.65612320 -1.09094297 1.6968925    F-test 1.44433818
#> 15  0.9079848 0.15159643 -0.75434631 -0.07522416 0.9832089    F-test 1.44433818
#> 16  0.8529754 0.18424976  0.34520545 -0.19000037 1.0429757    F-test 1.44433818
#> 17  2.8560575 0.72319924 -0.94312335  0.82661929 2.0294382    F-test 1.44433818
#> 18  2.1807625 0.87314273  0.71655469 -0.12312378 2.3038863    F-test 1.44433818
#> 19  1.5970988 0.03464638 -0.21743915  0.82794857 0.7691502    F-test 1.44433818
#> 20  0.9572343 0.06467113  0.01265856  0.13312848 0.8241058    F-test 1.44433818
#> 21  1.0676538 0.14595376  0.18002033  0.09477286 0.9728809    F-test 0.06002396
#> 22  1.5621735 0.53118266 -0.73844689 -0.11580847 1.6779820    F-test 0.06002396
#> 23  1.7045706 0.79211228 -0.10298023 -0.45100212 2.1555727    F-test 0.06002396
#> 24  2.3490892 0.54151436 -0.65612320  0.65219662 1.6968925    F-test 0.06002396
#> 25  1.5154787 0.15159643 -0.75434631  0.53226979 0.9832089    F-test 0.06002396
#> 26  0.9939507 0.18424976  0.34520545 -0.04902501 1.0429757    F-test 0.06002396
#> 27  1.1603647 0.72319924 -0.94312335 -0.86907349 2.0294382    F-test 0.06002396
#> 28  3.0079400 0.87314273  0.71655469  0.70405370 2.3038863    F-test 0.06002396
#> 29 -0.1001689 0.03464638 -0.21743915 -0.86931913 0.7691502    F-test 0.06002396
#> 30  1.1950411 0.06467113  0.01265856  0.37093525 0.8241058    F-test 0.06002396
#>       p_value k  null
#> 1  0.04755236 2 FALSE
#> 2  0.04755236 2 FALSE
#> 3  0.04755236 2 FALSE
#> 4  0.04755236 2 FALSE
#> 5  0.04755236 2 FALSE
#> 6  0.04755236 2 FALSE
#> 7  0.04755236 2 FALSE
#> 8  0.04755236 2 FALSE
#> 9  0.04755236 2 FALSE
#> 10 0.04755236 2 FALSE
#> 11 0.26851454 1  TRUE
#> 12 0.26851454 1  TRUE
#> 13 0.26851454 1  TRUE
#> 14 0.26851454 1  TRUE
#> 15 0.26851454 1  TRUE
#> 16 0.26851454 1  TRUE
#> 17 0.26851454 1  TRUE
#> 18 0.26851454 1  TRUE
#> 19 0.26851454 1  TRUE
#> 20 0.26851454 1  TRUE
#> 21 0.81348521 3  TRUE
#> 22 0.81348521 3  TRUE
#> 23 0.81348521 3  TRUE
#> 24 0.81348521 3  TRUE
#> 25 0.81348521 3  TRUE
#> 26 0.81348521 3  TRUE
#> 27 0.81348521 3  TRUE
#> 28 0.81348521 3  TRUE
#> 29 0.81348521 3  TRUE
#> 30 0.81348521 3  TRUE