Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>            y          x          e       .resid  .fitted test_name   statistic
#> 1  2.6783072 0.94674763 -0.1647715  0.381519964 2.296787    F-test 0.057441974
#> 2  3.1344316 0.86202227  0.5293269  0.915305863 2.219126    F-test 0.057441974
#> 3  1.8804859 0.68782023 -0.2804310 -0.178961515 2.059447    F-test 0.057441974
#> 4  1.6221107 0.79503336 -0.8050007 -0.535611194 2.157722    F-test 0.057441974
#> 5  1.3269922 0.02773054  0.2984926 -0.127399140 1.454391    F-test 0.057441974
#> 6  2.2951319 0.61809524  0.2949949  0.299596327 1.995536    F-test 0.057441974
#> 7  1.0998136 0.62164096 -0.9082648 -0.898972055 1.998786    F-test 0.057441974
#> 8  2.9664760 0.38759966  1.4286428  1.182218873 1.784257    F-test 0.057441974
#> 9  1.1978562 0.05661380  0.1380373 -0.283010270 1.480866    F-test 0.057441974
#> 10 1.4835707 0.88289433 -1.1788260 -0.754686854 2.238258    F-test 0.057441974
#> 11 1.9910035 0.94674763 -0.1647715 -0.305783745 2.296787    F-test 0.306832662
#> 12 2.8774985 0.86202227  0.5293269  0.658372847 2.219126    F-test 0.306832662
#> 13 1.3178424 0.68782023 -0.2804310 -0.741604951 2.059447    F-test 0.306832662
#> 14 1.6648575 0.79503336 -0.8050007 -0.492864401 2.157722    F-test 0.306832662
#> 15 2.2264448 0.02773054  0.2984926  0.772053471 1.454391    F-test 0.306832662
#> 16 2.1543585 0.61809524  0.2949949  0.158822903 1.995536    F-test 0.306832662
#> 17 2.5771551 0.62164096 -0.9082648  0.578369377 1.998786    F-test 0.306832662
#> 18 2.4476654 0.38759966  1.4286428  0.663408338 1.784257    F-test 0.306832662
#> 19 0.2376237 0.05661380  0.1380373 -1.243242817 1.480866    F-test 0.306832662
#> 20 2.1907266 0.88289433 -1.1788260 -0.047531021 2.238258    F-test 0.306832662
#> 21 2.2633449 0.94674763 -0.1647715 -0.033442325 2.296787    F-test 0.003604605
#> 22 1.0885086 0.86202227  0.5293269 -1.130617139 2.219126    F-test 0.003604605
#> 23 1.3096942 0.68782023 -0.2804310 -0.749753158 2.059447    F-test 0.003604605
#> 24 2.1586294 0.79503336 -0.8050007  0.000907533 2.157722    F-test 0.003604605
#> 25 1.9368353 0.02773054  0.2984926  0.482444043 1.454391    F-test 0.003604605
#> 26 2.6003027 0.61809524  0.2949949  0.604767124 1.995536    F-test 0.003604605
#> 27 2.1043058 0.62164096 -0.9082648  0.105520102 1.998786    F-test 0.003604605
#> 28 1.8104464 0.38759966  1.4286428  0.026189350 1.784257    F-test 0.003604605
#> 29 0.9415717 0.05661380  0.1380373 -0.539294798 1.480866    F-test 0.003604605
#> 30 3.4715369 0.88289433 -1.1788260  1.233279268 2.238258    F-test 0.003604605
#>      p_value k  null
#> 1  0.8174523 2 FALSE
#> 2  0.8174523 2 FALSE
#> 3  0.8174523 2 FALSE
#> 4  0.8174523 2 FALSE
#> 5  0.8174523 2 FALSE
#> 6  0.8174523 2 FALSE
#> 7  0.8174523 2 FALSE
#> 8  0.8174523 2 FALSE
#> 9  0.8174523 2 FALSE
#> 10 0.8174523 2 FALSE
#> 11 0.5968725 1  TRUE
#> 12 0.5968725 1  TRUE
#> 13 0.5968725 1  TRUE
#> 14 0.5968725 1  TRUE
#> 15 0.5968725 1  TRUE
#> 16 0.5968725 1  TRUE
#> 17 0.5968725 1  TRUE
#> 18 0.5968725 1  TRUE
#> 19 0.5968725 1  TRUE
#> 20 0.5968725 1  TRUE
#> 21 0.9538032 3  TRUE
#> 22 0.9538032 3  TRUE
#> 23 0.9538032 3  TRUE
#> 24 0.9538032 3  TRUE
#> 25 0.9538032 3  TRUE
#> 26 0.9538032 3  TRUE
#> 27 0.9538032 3  TRUE
#> 28 0.9538032 3  TRUE
#> 29 0.9538032 3  TRUE
#> 30 0.9538032 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e          x      .resid  .fitted test_name    statistic
#> 1   2.902312  1 0.57345785 -4.27824040 7.180552    F-test 0.2231276915
#> 2   4.292085  2 0.74180722 -2.37058476 6.662670    F-test 0.2231276915
#> 3   4.435982  3 0.32824023 -3.49891732 7.934899    F-test 0.2231276915
#> 4   6.442887  4 0.80110981 -0.03735452 6.480241    F-test 0.2231276915
#> 5   7.421296  5 0.79278633  0.91545028 6.505846    F-test 0.2231276915
#> 6   7.421228  6 0.31928497 -0.54121978 7.962448    F-test 0.2231276915
#> 7   8.591956  7 0.41758182  0.93189293 7.660063    F-test 0.2231276915
#> 8  10.258793  8 0.72832929  3.55466162 6.704131    F-test 0.2231276915
#> 9  10.054793  9 0.05208093  1.27036329 8.784430    F-test 0.2231276915
#> 10 11.637946 10 0.44230885  4.05394868 7.583997    F-test 0.2231276915
#> 11  8.063759  1 0.57345785  0.88320642 7.180552    F-test 0.0001963036
#> 12  5.726142  2 0.74180722 -0.93652757 6.662670    F-test 0.0001963036
#> 13  7.185707  3 0.32824023 -0.74919177 7.934899    F-test 0.0001963036
#> 14  7.233909  4 0.80110981  0.75366798 6.480241    F-test 0.0001963036
#> 15  8.742502  5 0.79278633  2.23665601 6.505846    F-test 0.0001963036
#> 16 13.845241  6 0.31928497  5.88279287 7.962448    F-test 0.0001963036
#> 17  3.370248  7 0.41758182 -4.28981529 7.660063    F-test 0.0001963036
#> 18  4.059642  8 0.72832929 -2.64448923 6.704131    F-test 0.0001963036
#> 19  7.762179  9 0.05208093 -1.02225068 8.784430    F-test 0.0001963036
#> 20  7.469949 10 0.44230885 -0.11404874 7.583997    F-test 0.0001963036
#> 21 12.494305  1 0.57345785  5.31375310 7.180552    F-test 0.2595217742
#> 22  3.914371  2 0.74180722 -2.74829923 6.662670    F-test 0.2595217742
#> 23 10.546166  3 0.32824023  2.61126639 7.934899    F-test 0.2595217742
#> 24  8.568711  4 0.80110981  2.08846962 6.480241    F-test 0.2595217742
#> 25  5.073141  5 0.79278633 -1.43270568 6.505846    F-test 0.2595217742
#> 26  9.168844  6 0.31928497  1.20639648 7.962448    F-test 0.2595217742
#> 27  6.530792  7 0.41758182 -1.12927124 7.660063    F-test 0.2595217742
#> 28  5.954942  8 0.72832929 -0.74918937 6.704131    F-test 0.2595217742
#> 29  7.473291  9 0.05208093 -1.31113889 8.784430    F-test 0.2595217742
#> 30  3.734716 10 0.44230885 -3.84928119 7.583997    F-test 0.2595217742
#>      p_value k  null
#> 1  0.6510390 1 FALSE
#> 2  0.6510390 1 FALSE
#> 3  0.6510390 1 FALSE
#> 4  0.6510390 1 FALSE
#> 5  0.6510390 1 FALSE
#> 6  0.6510390 1 FALSE
#> 7  0.6510390 1 FALSE
#> 8  0.6510390 1 FALSE
#> 9  0.6510390 1 FALSE
#> 10 0.6510390 1 FALSE
#> 11 0.9892123 2  TRUE
#> 12 0.9892123 2  TRUE
#> 13 0.9892123 2  TRUE
#> 14 0.9892123 2  TRUE
#> 15 0.9892123 2  TRUE
#> 16 0.9892123 2  TRUE
#> 17 0.9892123 2  TRUE
#> 18 0.9892123 2  TRUE
#> 19 0.9892123 2  TRUE
#> 20 0.9892123 2  TRUE
#> 21 0.6261117 3  TRUE
#> 22 0.6261117 3  TRUE
#> 23 0.6261117 3  TRUE
#> 24 0.6261117 3  TRUE
#> 25 0.6261117 3  TRUE
#> 26 0.6261117 3  TRUE
#> 27 0.6261117 3  TRUE
#> 28 0.6261117 3  TRUE
#> 29 0.6261117 3  TRUE
#> 30 0.6261117 3  TRUE