Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>               y           x          e      .resid   .fitted test_name
#> 1  -0.743191887 0.002184442 -1.7453811 -1.22485692 0.4816650    F-test
#> 2   1.221270030 0.586788856 -0.7098400 -0.52363444 1.7449045    F-test
#> 3   1.520941576 0.022294253  0.4981503  0.99582236 0.5251192    F-test
#> 4   0.685140785 0.622925462 -1.3258208 -1.13784928 1.8229901    F-test
#> 5   0.745966741 0.374509919 -0.7688009 -0.54023588 1.2862026    F-test
#> 6   3.621100766 0.586677313  1.6902332  1.87643733 1.7446634    F-test
#> 7  -0.006985442 0.447054974 -1.6538986 -1.44994667 1.4429612    F-test
#> 8   1.848486840 0.136152740  0.6937965  1.07733710 0.7711497    F-test
#> 9   1.685005064 0.679513917 -0.4562480 -0.26026387 1.9452689    F-test
#> 10  3.077231304 0.653955435  0.9956182  1.18719028 1.8900410    F-test
#> 11 -0.118354840 0.002184442 -1.7453811 -0.60001988 0.4816650    F-test
#> 12  0.490669802 0.586788856 -0.7098400 -1.25423467 1.7449045    F-test
#> 13 -0.180416811 0.022294253  0.4981503 -0.70553603 0.5251192    F-test
#> 14  3.841333311 0.622925462 -1.3258208  2.01834325 1.8229901    F-test
#> 15  0.869537503 0.374509919 -0.7688009 -0.41666512 1.2862026    F-test
#> 16  0.254861030 0.586677313  1.6902332 -1.48980241 1.7446634    F-test
#> 17  2.141939910 0.447054974 -1.6538986  0.69897868 1.4429612    F-test
#> 18  2.562094607 0.136152740  0.6937965  1.79094486 0.7711497    F-test
#> 19  1.826325796 0.679513917 -0.4562480 -0.11894314 1.9452689    F-test
#> 20  1.966975470 0.653955435  0.9956182  0.07693445 1.8900410    F-test
#> 21 -0.106060910 0.002184442 -1.7453811 -0.58772595 0.4816650    F-test
#> 22  1.645234837 0.586788856 -0.7098400 -0.09966963 1.7449045    F-test
#> 23 -0.220936049 0.022294253  0.4981503 -0.74605527 0.5251192    F-test
#> 24  1.265731890 0.622925462 -1.3258208 -0.55725817 1.8229901    F-test
#> 25  0.732506277 0.374509919 -0.7688009 -0.55369634 1.2862026    F-test
#> 26  2.011391883 0.586677313  1.6902332  0.26672844 1.7446634    F-test
#> 27  1.873151256 0.447054974 -1.6538986  0.43019003 1.4429612    F-test
#> 28  2.652996166 0.136152740  0.6937965  1.88184642 0.7711497    F-test
#> 29  3.836838336 0.679513917 -0.4562480  1.89156940 1.9452689    F-test
#> 30 -0.035887908 0.653955435  0.9956182 -1.92592893 1.8900410    F-test
#>    statistic   p_value k  null
#> 1  0.1677993 0.6943201 3 FALSE
#> 2  0.1677993 0.6943201 3 FALSE
#> 3  0.1677993 0.6943201 3 FALSE
#> 4  0.1677993 0.6943201 3 FALSE
#> 5  0.1677993 0.6943201 3 FALSE
#> 6  0.1677993 0.6943201 3 FALSE
#> 7  0.1677993 0.6943201 3 FALSE
#> 8  0.1677993 0.6943201 3 FALSE
#> 9  0.1677993 0.6943201 3 FALSE
#> 10 0.1677993 0.6943201 3 FALSE
#> 11 0.1831187 0.6815640 1  TRUE
#> 12 0.1831187 0.6815640 1  TRUE
#> 13 0.1831187 0.6815640 1  TRUE
#> 14 0.1831187 0.6815640 1  TRUE
#> 15 0.1831187 0.6815640 1  TRUE
#> 16 0.1831187 0.6815640 1  TRUE
#> 17 0.1831187 0.6815640 1  TRUE
#> 18 0.1831187 0.6815640 1  TRUE
#> 19 0.1831187 0.6815640 1  TRUE
#> 20 0.1831187 0.6815640 1  TRUE
#> 21 0.1661496 0.6957346 2  TRUE
#> 22 0.1661496 0.6957346 2  TRUE
#> 23 0.1661496 0.6957346 2  TRUE
#> 24 0.1661496 0.6957346 2  TRUE
#> 25 0.1661496 0.6957346 2  TRUE
#> 26 0.1661496 0.6957346 2  TRUE
#> 27 0.1661496 0.6957346 2  TRUE
#> 28 0.1661496 0.6957346 2  TRUE
#> 29 0.1661496 0.6957346 2  TRUE
#> 30 0.1661496 0.6957346 2  TRUE