Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>             y         x           e      .resid   .fitted test_name statistic
#> 1   1.6707208 0.6148570 -0.32218530 -0.23230582 1.9030266    F-test  3.406464
#> 2   1.5754567 0.6311590 -0.45406391 -0.38575170 1.9612084    F-test  3.406464
#> 3   0.8143069 0.5522277 -1.04287631 -0.86519581 1.6795027    F-test  3.406464
#> 4   0.9768056 0.5044713 -0.78215700 -0.53225431 1.5090599    F-test  3.406464
#> 5   4.1957443 0.8323683  1.67053902  1.51641940 2.6793249    F-test  3.406464
#> 6   2.0190183 0.7956838 -0.40977820 -0.52937953 2.5483978    F-test  3.406464
#> 7   2.6597121 0.3500791  1.18707757  1.70167818 0.9580339    F-test  3.406464
#> 8   0.2870087 0.3244301 -1.14267629 -0.57948377 0.8664924    F-test  3.406464
#> 9   1.1935822 0.5619864 -0.68423299 -0.52074930 1.7143315    F-test  3.406464
#> 10  1.6110469 0.4133994  0.02674838  0.42702267 1.1840242    F-test  3.406464
#> 11  2.5305310 0.6148570 -0.32218530  0.62750432 1.9030266    F-test  2.183820
#> 12  2.2278362 0.6311590 -0.45406391  0.26662778 1.9612084    F-test  2.183820
#> 13  1.1836393 0.5522277 -1.04287631 -0.49586344 1.6795027    F-test  2.183820
#> 14  2.5387016 0.5044713 -0.78215700  1.02964171 1.5090599    F-test  2.183820
#> 15  1.3763744 0.8323683  1.67053902 -1.30295048 2.6793249    F-test  2.183820
#> 16  3.0716499 0.7956838 -0.40977820  0.52325207 2.5483978    F-test  2.183820
#> 17  1.9522399 0.3500791  1.18707757  0.99420597 0.9580339    F-test  2.183820
#> 18 -0.5483761 0.3244301 -1.14267629 -1.41486849 0.8664924    F-test  2.183820
#> 19  2.2091956 0.5619864 -0.68423299  0.49486412 1.7143315    F-test  2.183820
#> 20  0.4616107 0.4133994  0.02674838 -0.72241357 1.1840242    F-test  2.183820
#> 21  3.4096151 0.6148570 -0.32218530  1.50658843 1.9030266    F-test  0.229706
#> 22  2.0831013 0.6311590 -0.45406391  0.12189294 1.9612084    F-test  0.229706
#> 23  1.1175947 0.5522277 -1.04287631 -0.56190797 1.6795027    F-test  0.229706
#> 24 -0.3138810 0.5044713 -0.78215700 -1.82294092 1.5090599    F-test  0.229706
#> 25  2.2175828 0.8323683  1.67053902 -0.46174214 2.6793249    F-test  0.229706
#> 26  2.4114423 0.7956838 -0.40977820 -0.13695544 2.5483978    F-test  0.229706
#> 27  0.8774917 0.3500791  1.18707757 -0.08054221 0.9580339    F-test  0.229706
#> 28  0.7101426 0.3244301 -1.14267629 -0.15634986 0.8664924    F-test  0.229706
#> 29  2.6617334 0.5619864 -0.68423299  0.94740194 1.7143315    F-test  0.229706
#> 30  1.8285795 0.4133994  0.02674838  0.64455523 1.1840242    F-test  0.229706
#>      p_value k  null
#> 1  0.1074490 2 FALSE
#> 2  0.1074490 2 FALSE
#> 3  0.1074490 2 FALSE
#> 4  0.1074490 2 FALSE
#> 5  0.1074490 2 FALSE
#> 6  0.1074490 2 FALSE
#> 7  0.1074490 2 FALSE
#> 8  0.1074490 2 FALSE
#> 9  0.1074490 2 FALSE
#> 10 0.1074490 2 FALSE
#> 11 0.1829923 1  TRUE
#> 12 0.1829923 1  TRUE
#> 13 0.1829923 1  TRUE
#> 14 0.1829923 1  TRUE
#> 15 0.1829923 1  TRUE
#> 16 0.1829923 1  TRUE
#> 17 0.1829923 1  TRUE
#> 18 0.1829923 1  TRUE
#> 19 0.1829923 1  TRUE
#> 20 0.1829923 1  TRUE
#> 21 0.6463527 3  TRUE
#> 22 0.6463527 3  TRUE
#> 23 0.6463527 3  TRUE
#> 24 0.6463527 3  TRUE
#> 25 0.6463527 3  TRUE
#> 26 0.6463527 3  TRUE
#> 27 0.6463527 3  TRUE
#> 28 0.6463527 3  TRUE
#> 29 0.6463527 3  TRUE
#> 30 0.6463527 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e         x     .resid  .fitted test_name statistic   p_value k
#> 1   2.478717  1 0.3536491 -4.1434704 6.622187    F-test 0.4728036 0.5138310 1
#> 2   3.940892  2 0.5912802 -3.6675030 7.608395    F-test 0.4728036 0.5138310 1
#> 3   5.145503  3 0.6813142 -2.8365481 7.982051    F-test 0.4728036 0.5138310 1
#> 4   5.747425  4 0.4987119 -1.4767962 7.224222    F-test 0.4728036 0.5138310 1
#> 5   6.241357  5 0.2009688  0.2528185 5.988539    F-test 0.4728036 0.5138310 1
#> 6   8.739832  6 0.9106141 -0.1938517 8.933684    F-test 0.4728036 0.5138310 1
#> 7   8.385507  7 0.2971868  1.9976476 6.387859    F-test 0.4728036 0.5138310 1
#> 8   9.638390  8 0.4425444  2.6472727 6.991117    F-test 0.4728036 0.5138310 1
#> 9  11.747994  9 0.9135041  2.8023160 8.945678    F-test 0.4728036 0.5138310 1
#> 10 11.662964 10 0.4554913  4.6181146 7.044849    F-test 0.4728036 0.5138310 1
#> 11  9.334657  1 0.3536491  2.7124702 6.622187    F-test 0.6888049 0.4339592 2
#> 12  5.250342  2 0.5912802 -2.3580537 7.608395    F-test 0.6888049 0.4339592 2
#> 13 13.656195  3 0.6813142  5.6741437 7.982051    F-test 0.6888049 0.4339592 2
#> 14  4.825830  4 0.4987119 -2.3983914 7.224222    F-test 0.6888049 0.4339592 2
#> 15  4.351143  5 0.2009688 -1.6373954 5.988539    F-test 0.6888049 0.4339592 2
#> 16  5.203033  6 0.9106141 -3.7306509 8.933684    F-test 0.6888049 0.4339592 2
#> 17  4.807101  7 0.2971868 -1.5807578 6.387859    F-test 0.6888049 0.4339592 2
#> 18 10.051369  8 0.4425444  3.0602517 6.991117    F-test 0.6888049 0.4339592 2
#> 19 10.013944  9 0.9135041  1.0682664 8.945678    F-test 0.6888049 0.4339592 2
#> 20  6.234966 10 0.4554913 -0.8098828 7.044849    F-test 0.6888049 0.4339592 2
#> 21  6.947376  1 0.3536491  0.3251893 6.622187    F-test 0.5677524 0.4757281 3
#> 22  3.434006  2 0.5912802 -4.1743896 7.608395    F-test 0.5677524 0.4757281 3
#> 23 10.404327  3 0.6813142  2.4222756 7.982051    F-test 0.5677524 0.4757281 3
#> 24  7.878906  4 0.4987119  0.6546844 7.224222    F-test 0.5677524 0.4757281 3
#> 25  4.717945  5 0.2009688 -1.2705935 5.988539    F-test 0.5677524 0.4757281 3
#> 26  7.801576  6 0.9106141 -1.1321072 8.933684    F-test 0.5677524 0.4757281 3
#> 27  4.191688  7 0.2971868 -2.1961713 6.387859    F-test 0.5677524 0.4757281 3
#> 28  5.758371  8 0.4425444 -1.2327465 6.991117    F-test 0.5677524 0.4757281 3
#> 29  8.624469  9 0.9135041 -0.3212094 8.945678    F-test 0.5677524 0.4757281 3
#> 30 13.969917 10 0.4554913  6.9250681 7.044849    F-test 0.5677524 0.4757281 3
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE