Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>              y          x           e      .resid   .fitted test_name statistic
#> 1   1.65388550 0.63761391 -0.39027991 -0.95642274 2.6103082    F-test 3.6623908
#> 2   1.13355835 0.14886621 -0.03746901  0.18803438 0.9455240    F-test 3.6623908
#> 3   1.69772429 0.11512164  0.56934966  0.86714192 0.8305824    F-test 3.6623908
#> 4   2.86394235 0.51899892  1.07558355  0.65766337 2.2062790    F-test 3.6623908
#> 5   2.51026090 0.56314182  0.62999037  0.15362130 2.3566396    F-test 3.6623908
#> 6   2.73308791 0.70022500  0.54254787 -0.09048772 2.8235756    F-test 3.6623908
#> 7   2.33835247 0.49117626  0.60592209  0.22684372 2.1115088    F-test 3.6623908
#> 8   0.19758532 0.04496795 -0.84940475 -0.39403788 0.5916232    F-test 3.6623908
#> 9  -0.25032838 0.04411741 -1.29639213 -0.83905442 0.5887260    F-test 3.6623908
#> 10  1.73533222 0.32592742  0.30317611  0.18669806 1.5486342    F-test 3.6623908
#> 11  1.81488753 0.63761391 -0.39027991 -0.79542071 2.6103082    F-test 0.4527863
#> 12  1.55984601 0.14886621 -0.03746901  0.61432204 0.9455240    F-test 0.4527863
#> 13 -0.05037642 0.11512164  0.56934966 -0.88095879 0.8305824    F-test 0.4527863
#> 14  2.43187299 0.51899892  1.07558355  0.22559401 2.2062790    F-test 0.4527863
#> 15  2.90544991 0.56314182  0.62999037  0.54881031 2.3566396    F-test 0.4527863
#> 16  3.45652166 0.70022500  0.54254787  0.63294603 2.8235756    F-test 0.4527863
#> 17  1.46137491 0.49117626  0.60592209 -0.65013384 2.1115088    F-test 0.4527863
#> 18  0.74617208 0.04496795 -0.84940475  0.15454888 0.5916232    F-test 0.4527863
#> 19  0.90387642 0.04411741 -1.29639213  0.31515038 0.5887260    F-test 0.4527863
#> 20  1.38377585 0.32592742  0.30317611 -0.16485830 1.5486342    F-test 0.4527863
#> 21  2.54378273 0.63761391 -0.39027991 -0.06652550 2.6103082    F-test 1.7470686
#> 22  1.51442081 0.14886621 -0.03746901  0.56889684 0.9455240    F-test 1.7470686
#> 23  0.48779618 0.11512164  0.56934966 -0.34278619 0.8305824    F-test 1.7470686
#> 24  2.17220091 0.51899892  1.07558355 -0.03407807 2.2062790    F-test 1.7470686
#> 25  1.70131202 0.56314182  0.62999037 -0.65532758 2.3566396    F-test 1.7470686
#> 26  2.48756914 0.70022500  0.54254787 -0.33600649 2.8235756    F-test 1.7470686
#> 27  3.36965236 0.49117626  0.60592209  1.25814360 2.1115088    F-test 1.7470686
#> 28  0.82720025 0.04496795 -0.84940475  0.23557705 0.5916232    F-test 1.7470686
#> 29 -0.10423457 0.04411741 -1.29639213 -0.69296061 0.5887260    F-test 1.7470686
#> 30  1.61370110 0.32592742  0.30317611  0.06506694 1.5486342    F-test 1.7470686
#>       p_value k  null
#> 1  0.09722154 3 FALSE
#> 2  0.09722154 3 FALSE
#> 3  0.09722154 3 FALSE
#> 4  0.09722154 3 FALSE
#> 5  0.09722154 3 FALSE
#> 6  0.09722154 3 FALSE
#> 7  0.09722154 3 FALSE
#> 8  0.09722154 3 FALSE
#> 9  0.09722154 3 FALSE
#> 10 0.09722154 3 FALSE
#> 11 0.52260066 1  TRUE
#> 12 0.52260066 1  TRUE
#> 13 0.52260066 1  TRUE
#> 14 0.52260066 1  TRUE
#> 15 0.52260066 1  TRUE
#> 16 0.52260066 1  TRUE
#> 17 0.52260066 1  TRUE
#> 18 0.52260066 1  TRUE
#> 19 0.52260066 1  TRUE
#> 20 0.52260066 1  TRUE
#> 21 0.22780263 2  TRUE
#> 22 0.22780263 2  TRUE
#> 23 0.22780263 2  TRUE
#> 24 0.22780263 2  TRUE
#> 25 0.22780263 2  TRUE
#> 26 0.22780263 2  TRUE
#> 27 0.22780263 2  TRUE
#> 28 0.22780263 2  TRUE
#> 29 0.22780263 2  TRUE
#> 30 0.22780263 2  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e          x     .resid  .fitted test_name  statistic    p_value
#> 1   2.836264  1 0.54223975 -3.5863960 6.422660    F-test 11.0657245 0.01265135
#> 2   3.376979  2 0.29182005 -3.6629326 7.039912    F-test 11.0657245 0.01265135
#> 3   5.186064  3 0.69835904 -0.8517816 6.037846    F-test 11.0657245 0.01265135
#> 4   5.185492  4 0.15991809 -2.1795409 7.365033    F-test 11.0657245 0.01265135
#> 5   6.120408  5 0.10861125 -1.3710898 7.491497    F-test 11.0657245 0.01265135
#> 6   7.556033  6 0.39779315  0.7773308 6.778702    F-test 11.0657245 0.01265135
#> 7   8.107418  7 0.09784434  0.5893813 7.518037    F-test 11.0657245 0.01265135
#> 8  10.391208  8 0.78109622  4.5572979 5.833910    F-test 11.0657245 0.01265135
#> 9  10.064306  9 0.06062996  2.4545408 7.609765    F-test 11.0657245 0.01265135
#> 10 11.009413 10 0.00932591  3.2731900 7.736223    F-test 11.0657245 0.01265135
#> 11  8.208268  1 0.54223975  1.7856080 6.422660    F-test  0.3690106 0.56272342
#> 12  1.895223  2 0.29182005 -5.1446891 7.039912    F-test  0.3690106 0.56272342
#> 13 11.001046  3 0.69835904  4.9631997 6.037846    F-test  0.3690106 0.56272342
#> 14  6.882456  4 0.15991809 -0.4825772 7.365033    F-test  0.3690106 0.56272342
#> 15  7.098082  5 0.10861125 -0.3934150 7.491497    F-test  0.3690106 0.56272342
#> 16  6.125209  6 0.39779315 -0.6534926 6.778702    F-test  0.3690106 0.56272342
#> 17  8.307797  7 0.09784434  0.7897606 7.518037    F-test  0.3690106 0.56272342
#> 18  2.404795  8 0.78109622 -3.4291146 5.833910    F-test  0.3690106 0.56272342
#> 19  8.080175  9 0.06062996  0.4704101 7.609765    F-test  0.3690106 0.56272342
#> 20  9.830533 10 0.00932591  2.0943101 7.736223    F-test  0.3690106 0.56272342
#> 21  1.370907  1 0.54223975 -5.0517525 6.422660    F-test  0.1258093 0.73326274
#> 22  4.026515  2 0.29182005 -3.0133965 7.039912    F-test  0.1258093 0.73326274
#> 23  7.119065  3 0.69835904  1.0812188 6.037846    F-test  0.1258093 0.73326274
#> 24  8.665468  4 0.15991809  1.3004350 7.365033    F-test  0.1258093 0.73326274
#> 25  8.018715  5 0.10861125  0.5272178 7.491497    F-test  0.1258093 0.73326274
#> 26  9.965651  6 0.39779315  3.1869488 6.778702    F-test  0.1258093 0.73326274
#> 27  8.871959  7 0.09784434  1.3539224 7.518037    F-test  0.1258093 0.73326274
#> 28  7.192075  8 0.78109622  1.3581655 5.833910    F-test  0.1258093 0.73326274
#> 29 10.419985  9 0.06062996  2.8102204 7.609765    F-test  0.1258093 0.73326274
#> 30  4.183243 10 0.00932591 -3.5529797 7.736223    F-test  0.1258093 0.73326274
#>    k  null
#> 1  1 FALSE
#> 2  1 FALSE
#> 3  1 FALSE
#> 4  1 FALSE
#> 5  1 FALSE
#> 6  1 FALSE
#> 7  1 FALSE
#> 8  1 FALSE
#> 9  1 FALSE
#> 10 1 FALSE
#> 11 2  TRUE
#> 12 2  TRUE
#> 13 2  TRUE
#> 14 2  TRUE
#> 15 2  TRUE
#> 16 2  TRUE
#> 17 2  TRUE
#> 18 2  TRUE
#> 19 2  TRUE
#> 20 2  TRUE
#> 21 3  TRUE
#> 22 3  TRUE
#> 23 3  TRUE
#> 24 3  TRUE
#> 25 3  TRUE
#> 26 3  TRUE
#> 27 3  TRUE
#> 28 3  TRUE
#> 29 3  TRUE
#> 30 3  TRUE