Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>            y          x           e      .resid  .fitted test_name statistic
#> 1  2.0729132 0.86896422 -0.55114983  0.60400279 1.468910    F-test 1.4597088
#> 2  1.0278370 0.38595940 -0.50708701 -0.73021548 1.758053    F-test 1.4597088
#> 3  0.8998594 0.62264067 -1.11046270 -0.71650817 1.616368    F-test 1.4597088
#> 4  3.6006928 0.08808394  2.50485010  1.66432253 1.936370    F-test 1.4597088
#> 5  1.2469867 0.16363125  0.05658024 -0.64415857 1.891145    F-test 1.4597088
#> 6  2.0904514 0.10668028  0.97239048  0.16521351 1.925238    F-test 1.4597088
#> 7  2.0181923 0.56746317  0.12871464  0.36879368 1.649399    F-test 1.4597088
#> 8  1.8567822 0.32631892  0.42397922  0.06302696 1.793755    F-test 1.4597088
#> 9  2.4014484 0.27198676  1.05548480  0.57516818 1.826280    F-test 1.4597088
#> 10 0.5321168 0.17930535 -0.67933894 -1.34964543 1.881762    F-test 1.4597088
#> 11 1.9189278 0.86896422 -0.55114983  0.45001736 1.468910    F-test 0.1832122
#> 12 1.3712888 0.38595940 -0.50708701 -0.38676376 1.758053    F-test 0.1832122
#> 13 1.0113866 0.62264067 -1.11046270 -0.60498099 1.616368    F-test 0.1832122
#> 14 0.9116427 0.08808394  2.50485010 -1.02472763 1.936370    F-test 0.1832122
#> 15 0.8490506 0.16363125  0.05658024 -1.04209461 1.891145    F-test 0.1832122
#> 16 3.3905550 0.10668028  0.97239048  1.46531705 1.925238    F-test 0.1832122
#> 17 1.5255692 0.56746317  0.12871464 -0.12382935 1.649399    F-test 0.1832122
#> 18 2.6895344 0.32631892  0.42397922  0.89577922 1.793755    F-test 0.1832122
#> 19 1.2949835 0.27198676  1.05548480 -0.53129669 1.826280    F-test 0.1832122
#> 20 2.7843416 0.17930535 -0.67933894  0.90257940 1.881762    F-test 0.1832122
#> 21 0.5866876 0.86896422 -0.55114983 -0.88222278 1.468910    F-test 6.5967210
#> 22 3.2638996 0.38595940 -0.50708701  1.50584709 1.758053    F-test 6.5967210
#> 23 1.9117221 0.62264067 -1.11046270  0.29535452 1.616368    F-test 6.5967210
#> 24 1.8431370 0.08808394  2.50485010 -0.09323329 1.936370    F-test 6.5967210
#> 25 0.8356633 0.16363125  0.05658024 -1.05548200 1.891145    F-test 6.5967210
#> 26 1.3555579 0.10668028  0.97239048 -0.56968001 1.925238    F-test 6.5967210
#> 27 1.3168929 0.56746317  0.12871464 -0.33250569 1.649399    F-test 6.5967210
#> 28 2.8853944 0.32631892  0.42397922  1.09163923 1.793755    F-test 6.5967210
#> 29 2.5649344 0.27198676  1.05548480  0.73865425 1.826280    F-test 6.5967210
#> 30 1.1833909 0.17930535 -0.67933894 -0.69837133 1.881762    F-test 6.5967210
#>       p_value k  null
#> 1  0.26620402 3 FALSE
#> 2  0.26620402 3 FALSE
#> 3  0.26620402 3 FALSE
#> 4  0.26620402 3 FALSE
#> 5  0.26620402 3 FALSE
#> 6  0.26620402 3 FALSE
#> 7  0.26620402 3 FALSE
#> 8  0.26620402 3 FALSE
#> 9  0.26620402 3 FALSE
#> 10 0.26620402 3 FALSE
#> 11 0.68148817 1  TRUE
#> 12 0.68148817 1  TRUE
#> 13 0.68148817 1  TRUE
#> 14 0.68148817 1  TRUE
#> 15 0.68148817 1  TRUE
#> 16 0.68148817 1  TRUE
#> 17 0.68148817 1  TRUE
#> 18 0.68148817 1  TRUE
#> 19 0.68148817 1  TRUE
#> 20 0.68148817 1  TRUE
#> 21 0.03709388 2  TRUE
#> 22 0.03709388 2  TRUE
#> 23 0.03709388 2  TRUE
#> 24 0.03709388 2  TRUE
#> 25 0.03709388 2  TRUE
#> 26 0.03709388 2  TRUE
#> 27 0.03709388 2  TRUE
#> 28 0.03709388 2  TRUE
#> 29 0.03709388 2  TRUE
#> 30 0.03709388 2  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e         x      .resid  .fitted test_name  statistic   p_value k
#> 1   2.516029  1 0.3752311 -4.61095973 7.126989    F-test 0.62245745 0.4560211 3
#> 2   3.267109  2 0.2191027 -2.56313035 5.830239    F-test 0.62245745 0.4560211 3
#> 3   4.360561  3 0.2813841 -1.98696650 6.347528    F-test 0.62245745 0.4560211 3
#> 4   5.710833  4 0.4802210 -2.28816740 7.999001    F-test 0.62245745 0.4560211 3
#> 5   6.314944  5 0.2516274  0.21456553 6.100378    F-test 0.62245745 0.4560211 3
#> 6   7.824942  6 0.5367943 -0.64393693 8.468879    F-test 0.62245745 0.4560211 3
#> 7   8.597268  7 0.4204719  1.09452426 7.502744    F-test 0.62245745 0.4560211 3
#> 8   9.376797  8 0.2917048  2.94354877 6.433248    F-test 0.62245745 0.4560211 3
#> 9  10.410730  9 0.3128532  3.80183111 6.608899    F-test 0.62245745 0.4560211 3
#> 10 11.613264 10 0.4291200  4.03869123 7.574573    F-test 0.62245745 0.4560211 3
#> 11  9.593911  1 0.3752311  2.46692169 7.126989    F-test 0.34974074 0.5728540 1
#> 12  8.440283  2 0.2191027  2.61004420 5.830239    F-test 0.34974074 0.5728540 1
#> 13  5.204542  3 0.2813841 -1.14298534 6.347528    F-test 0.34974074 0.5728540 1
#> 14  5.028302  4 0.4802210 -2.97069893 7.999001    F-test 0.34974074 0.5728540 1
#> 15  2.132714  5 0.2516274 -3.96766379 6.100378    F-test 0.34974074 0.5728540 1
#> 16 11.933565  6 0.5367943  3.46468542 8.468879    F-test 0.34974074 0.5728540 1
#> 17  5.860441  7 0.4204719 -1.64230353 7.502744    F-test 0.34974074 0.5728540 1
#> 18  5.976514  8 0.2917048 -0.45673392 6.433248    F-test 0.34974074 0.5728540 1
#> 19 10.825180  9 0.3128532  4.21628052 6.608899    F-test 0.34974074 0.5728540 1
#> 20  4.997026 10 0.4291200 -2.57754631 7.574573    F-test 0.34974074 0.5728540 1
#> 21  6.152332  1 0.3752311 -0.97465754 7.126989    F-test 0.08690354 0.7767022 2
#> 22  6.108887  2 0.2191027  0.27864806 5.830239    F-test 0.08690354 0.7767022 2
#> 23  4.105546  3 0.2813841 -2.24198205 6.347528    F-test 0.08690354 0.7767022 2
#> 24  3.413089  4 0.4802210 -4.58591198 7.999001    F-test 0.08690354 0.7767022 2
#> 25  6.140042  5 0.2516274  0.03966423 6.100378    F-test 0.08690354 0.7767022 2
#> 26  8.704884  6 0.5367943  0.23600457 8.468879    F-test 0.08690354 0.7767022 2
#> 27  9.712624  7 0.4204719  2.20987968 7.502744    F-test 0.08690354 0.7767022 2
#> 28  9.843670  8 0.2917048  3.41042220 6.433248    F-test 0.08690354 0.7767022 2
#> 29  3.353415  9 0.3128532 -3.25548435 6.608899    F-test 0.08690354 0.7767022 2
#> 30 12.457990 10 0.4291200  4.88341718 7.574573    F-test 0.08690354 0.7767022 2
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE