Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>              y          x          e      .resid  .fitted test_name  statistic
#> 1   0.32503538 0.90013694 -2.3853481 -1.58962245 1.914658    F-test 0.31361176
#> 2   1.89209758 0.68988282 -0.2737235  0.06139064 1.830707    F-test 0.31361176
#> 3   1.97275104 0.85328751 -0.6086360  0.07679939 1.895952    F-test 0.31361176
#> 4   2.98086852 0.94629646  0.1390951  1.04777999 1.933089    F-test 0.31361176
#> 5   0.86853326 0.01916379 -0.1509978 -0.69436701 1.562900    F-test 0.31361176
#> 6   2.59548709 0.34385632  1.1333936  0.90294263 1.692544    F-test 0.31361176
#> 7   1.62095035 0.58680458 -0.3101939 -0.16859921 1.789550    F-test 0.31361176
#> 8   2.47095920 0.44771189  0.8228014  0.73694698 1.734012    F-test 0.31361176
#> 9   0.20093358 0.54266324 -1.6362130 -1.57099109 1.771925    F-test 0.31361176
#> 10  2.97159919 0.54755799  1.1242214  1.19772012 1.773879    F-test 0.31361176
#> 11  3.06076940 0.90013694 -2.3853481  1.14611157 1.914658    F-test 0.46366604
#> 12  1.84189534 0.68988282 -0.2737235  0.01118840 1.830707    F-test 0.46366604
#> 13 -0.18496439 0.85328751 -0.6086360 -2.08091604 1.895952    F-test 0.46366604
#> 14  1.78242885 0.94629646  0.1390951 -0.15065969 1.933089    F-test 0.46366604
#> 15  1.06063095 0.01916379 -0.1509978 -0.50226932 1.562900    F-test 0.46366604
#> 16  2.17262168 0.34385632  1.1333936  0.48007723 1.692544    F-test 0.46366604
#> 17  2.13029619 0.58680458 -0.3101939  0.34074663 1.789550    F-test 0.46366604
#> 18  0.58602792 0.44771189  0.8228014 -1.14798430 1.734012    F-test 0.46366604
#> 19  2.79638723 0.54266324 -1.6362130  1.02446256 1.771925    F-test 0.46366604
#> 20  2.65312201 0.54755799  1.1242214  0.87924295 1.773879    F-test 0.46366604
#> 21  2.76152778 0.90013694 -2.3853481  0.84686995 1.914658    F-test 0.05710483
#> 22  0.80940693 0.68988282 -0.2737235 -1.02130001 1.830707    F-test 0.05710483
#> 23  3.11821354 0.85328751 -0.6086360  1.22226190 1.895952    F-test 0.05710483
#> 24  1.09541267 0.94629646  0.1390951 -0.83767586 1.933089    F-test 0.05710483
#> 25  1.48020714 0.01916379 -0.1509978 -0.08269312 1.562900    F-test 0.05710483
#> 26  2.16556504 0.34385632  1.1333936  0.47302058 1.692544    F-test 0.05710483
#> 27  1.36312685 0.58680458 -0.3101939 -0.42642271 1.789550    F-test 0.05710483
#> 28  3.05863631 0.44771189  0.8228014  1.32462409 1.734012    F-test 0.05710483
#> 29 -0.01664243 0.54266324 -1.6362130 -1.78856710 1.771925    F-test 0.05710483
#> 30  2.06376135 0.54755799  1.1242214  0.28988228 1.773879    F-test 0.05710483
#>      p_value k  null
#> 1  0.5929328 2 FALSE
#> 2  0.5929328 2 FALSE
#> 3  0.5929328 2 FALSE
#> 4  0.5929328 2 FALSE
#> 5  0.5929328 2 FALSE
#> 6  0.5929328 2 FALSE
#> 7  0.5929328 2 FALSE
#> 8  0.5929328 2 FALSE
#> 9  0.5929328 2 FALSE
#> 10 0.5929328 2 FALSE
#> 11 0.5177990 1  TRUE
#> 12 0.5177990 1  TRUE
#> 13 0.5177990 1  TRUE
#> 14 0.5177990 1  TRUE
#> 15 0.5177990 1  TRUE
#> 16 0.5177990 1  TRUE
#> 17 0.5177990 1  TRUE
#> 18 0.5177990 1  TRUE
#> 19 0.5177990 1  TRUE
#> 20 0.5177990 1  TRUE
#> 21 0.8179773 3  TRUE
#> 22 0.8179773 3  TRUE
#> 23 0.8179773 3  TRUE
#> 24 0.8179773 3  TRUE
#> 25 0.8179773 3  TRUE
#> 26 0.8179773 3  TRUE
#> 27 0.8179773 3  TRUE
#> 28 0.8179773 3  TRUE
#> 29 0.8179773 3  TRUE
#> 30 0.8179773 3  TRUE