Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>            y          x           e       .resid  .fitted test_name
#> 1  1.7853626 0.39772110  0.22945940 -0.004886653 1.790249    F-test
#> 2  1.9949335 0.69263796 -0.17745179 -0.125930387 2.120864    F-test
#> 3  3.0301783 0.82445818  0.52598882  0.761538162 2.268640    F-test
#> 4  1.5587663 0.82482776 -0.94640229 -0.710288131 2.269054    F-test
#> 5  2.6439616 0.73756401  0.36239687  0.472733587 2.171228    F-test
#> 6  0.6650443 0.07704672 -0.41793862 -0.765714949 1.430759    F-test
#> 7  2.5161808 0.62792095  0.49397511  0.467867475 2.048313    F-test
#> 8  1.5001931 0.74885096 -0.80943563 -0.683688039 2.183881    F-test
#> 9  2.1264802 0.05630500  1.06700492  0.718973303 1.407507    F-test
#> 10 1.7698552 0.49603158  0.02777629 -0.130604367 1.900460    F-test
#> 11 1.4775943 0.39772110  0.22945940 -0.312654895 1.790249    F-test
#> 12 1.7464844 0.69263796 -0.17745179 -0.374379480 2.120864    F-test
#> 13 3.2661275 0.82445818  0.52598882  0.997487360 2.268640    F-test
#> 14 1.0321890 0.82482776 -0.94640229 -1.236865395 2.269054    F-test
#> 15 2.2796777 0.73756401  0.36239687  0.108449760 2.171228    F-test
#> 16 1.5028534 0.07704672 -0.41793862  0.072094179 1.430759    F-test
#> 17 2.0119884 0.62792095  0.49397511 -0.036324901 2.048313    F-test
#> 18 2.6135136 0.74885096 -0.80943563  0.429632457 2.183881    F-test
#> 19 1.3571868 0.05630500  1.06700492 -0.050320079 1.407507    F-test
#> 20 2.3033406 0.49603158  0.02777629  0.402880993 1.900460    F-test
#> 21 0.7953350 0.39772110  0.22945940 -0.994914232 1.790249    F-test
#> 22 2.6026180 0.69263796 -0.17745179  0.481754102 2.120864    F-test
#> 23 3.2102324 0.82445818  0.52598882  0.941592325 2.268640    F-test
#> 24 1.3833551 0.82482776 -0.94640229 -0.885699306 2.269054    F-test
#> 25 2.2619208 0.73756401  0.36239687  0.090692865 2.171228    F-test
#> 26 1.7901262 0.07704672 -0.41793862  0.359366984 1.430759    F-test
#> 27 2.1294834 0.62792095  0.49397511  0.081170084 2.048313    F-test
#> 28 2.1526588 0.74885096 -0.80943563 -0.031222361 2.183881    F-test
#> 29 1.6001348 0.05630500  1.06700492  0.192627971 1.407507    F-test
#> 30 1.6650911 0.49603158  0.02777629 -0.235368433 1.900460    F-test
#>       statistic   p_value k  null
#> 1  0.0006327947 0.9806331 1 FALSE
#> 2  0.0006327947 0.9806331 1 FALSE
#> 3  0.0006327947 0.9806331 1 FALSE
#> 4  0.0006327947 0.9806331 1 FALSE
#> 5  0.0006327947 0.9806331 1 FALSE
#> 6  0.0006327947 0.9806331 1 FALSE
#> 7  0.0006327947 0.9806331 1 FALSE
#> 8  0.0006327947 0.9806331 1 FALSE
#> 9  0.0006327947 0.9806331 1 FALSE
#> 10 0.0006327947 0.9806331 1 FALSE
#> 11 0.0041491104 0.9504418 2  TRUE
#> 12 0.0041491104 0.9504418 2  TRUE
#> 13 0.0041491104 0.9504418 2  TRUE
#> 14 0.0041491104 0.9504418 2  TRUE
#> 15 0.0041491104 0.9504418 2  TRUE
#> 16 0.0041491104 0.9504418 2  TRUE
#> 17 0.0041491104 0.9504418 2  TRUE
#> 18 0.0041491104 0.9504418 2  TRUE
#> 19 0.0041491104 0.9504418 2  TRUE
#> 20 0.0041491104 0.9504418 2  TRUE
#> 21 1.2355225304 0.3030492 3  TRUE
#> 22 1.2355225304 0.3030492 3  TRUE
#> 23 1.2355225304 0.3030492 3  TRUE
#> 24 1.2355225304 0.3030492 3  TRUE
#> 25 1.2355225304 0.3030492 3  TRUE
#> 26 1.2355225304 0.3030492 3  TRUE
#> 27 1.2355225304 0.3030492 3  TRUE
#> 28 1.2355225304 0.3030492 3  TRUE
#> 29 1.2355225304 0.3030492 3  TRUE
#> 30 1.2355225304 0.3030492 3  TRUE