Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>              y         x          e      .resid  .fitted test_name  statistic
#> 1   1.58126843 0.1065516  0.4633635  0.23046619 1.350802    F-test 0.48432716
#> 2   1.96426760 0.8360403 -0.5707362 -0.56471991 2.528988    F-test 0.48432716
#> 3   4.45395660 0.7111066  2.2371774  2.12674751 2.327209    F-test 0.48432716
#> 4   2.94247755 0.8948626  0.2468358  0.31848711 2.623990    F-test 0.48432716
#> 5   1.93959105 0.4272241  0.3298464  0.07087590 1.868715    F-test 0.48432716
#> 6   1.99070313 0.1090572  0.8697524  0.63585414 1.354849    F-test 0.48432716
#> 7   1.60430957 0.7370334 -0.6759420 -0.76477342 2.369083    F-test 0.48432716
#> 8   1.33860827 0.5854059 -0.5894977 -0.78558368 2.124192    F-test 0.48432716
#> 9   0.37388565 0.3281133 -1.0618860 -1.33475720 1.708643    F-test 0.48432716
#> 10  1.53609039 0.1795415  0.3243137  0.06740336 1.468687    F-test 0.48432716
#> 11 -0.16096847 0.1065516  0.4633635 -1.51177071 1.350802    F-test 0.30250624
#> 12  0.79804455 0.8360403 -0.5707362 -1.73094296 2.528988    F-test 0.30250624
#> 13  3.19918162 0.7111066  2.2371774  0.87197253 2.327209    F-test 0.30250624
#> 14  3.51304205 0.8948626  0.2468358  0.88905162 2.623990    F-test 0.30250624
#> 15  2.51233147 0.4272241  0.3298464  0.64361632 1.868715    F-test 0.30250624
#> 16  2.10088332 0.1090572  0.8697524  0.74603433 1.354849    F-test 0.30250624
#> 17  1.96859680 0.7370334 -0.6759420 -0.40048619 2.369083    F-test 0.30250624
#> 18  2.33012293 0.5854059 -0.5894977  0.20593099 2.124192    F-test 0.30250624
#> 19  1.43543710 0.3281133 -1.0618860 -0.27320576 1.708643    F-test 0.30250624
#> 20  2.02848685 0.1795415  0.3243137  0.55979982 1.468687    F-test 0.30250624
#> 21  1.90968268 0.1065516  0.4633635  0.55888045 1.350802    F-test 0.09072929
#> 22  1.77765736 0.8360403 -0.5707362 -0.75133015 2.528988    F-test 0.09072929
#> 23  2.10550812 0.7111066  2.2371774 -0.22170097 2.327209    F-test 0.09072929
#> 24  3.02953876 0.8948626  0.2468358  0.40554832 2.623990    F-test 0.09072929
#> 25  1.62531829 0.4272241  0.3298464 -0.24339686 1.868715    F-test 0.09072929
#> 26  0.02464691 0.1090572  0.8697524 -1.33020208 1.354849    F-test 0.09072929
#> 27  1.99195948 0.7370334 -0.6759420 -0.37712351 2.369083    F-test 0.09072929
#> 28  3.80674121 0.5854059 -0.5894977  1.68254927 2.124192    F-test 0.09072929
#> 29  0.73722320 0.3281133 -1.0618860 -0.97141965 1.708643    F-test 0.09072929
#> 30  2.71688221 0.1795415  0.3243137  1.24819518 1.468687    F-test 0.09072929
#>      p_value k  null
#> 1  0.5089083 1 FALSE
#> 2  0.5089083 1 FALSE
#> 3  0.5089083 1 FALSE
#> 4  0.5089083 1 FALSE
#> 5  0.5089083 1 FALSE
#> 6  0.5089083 1 FALSE
#> 7  0.5089083 1 FALSE
#> 8  0.5089083 1 FALSE
#> 9  0.5089083 1 FALSE
#> 10 0.5089083 1 FALSE
#> 11 0.5994173 2  TRUE
#> 12 0.5994173 2  TRUE
#> 13 0.5994173 2  TRUE
#> 14 0.5994173 2  TRUE
#> 15 0.5994173 2  TRUE
#> 16 0.5994173 2  TRUE
#> 17 0.5994173 2  TRUE
#> 18 0.5994173 2  TRUE
#> 19 0.5994173 2  TRUE
#> 20 0.5994173 2  TRUE
#> 21 0.7720027 3  TRUE
#> 22 0.7720027 3  TRUE
#> 23 0.7720027 3  TRUE
#> 24 0.7720027 3  TRUE
#> 25 0.7720027 3  TRUE
#> 26 0.7720027 3  TRUE
#> 27 0.7720027 3  TRUE
#> 28 0.7720027 3  TRUE
#> 29 0.7720027 3  TRUE
#> 30 0.7720027 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e          x      .resid  .fitted test_name statistic    p_value
#> 1   2.575777  1 0.40872293 -4.45860672 7.034384    F-test 0.7608675 0.41197506
#> 2   4.022377  2 0.62799681 -3.68846932 7.710846    F-test 0.7608675 0.41197506
#> 3   4.356292  3 0.27864770 -2.27680850 6.633101    F-test 0.7608675 0.41197506
#> 4   5.852123  4 0.54982032 -1.61754819 7.469671    F-test 0.7608675 0.41197506
#> 5   7.983983  5 0.99465152 -0.85799629 8.841979    F-test 0.7608675 0.41197506
#> 6   7.125790  6 0.11301674  1.00366189 6.122128    F-test 0.7608675 0.41197506
#> 7   9.148056  7 0.68239403  1.26939356 7.878662    F-test 0.7608675 0.41197506
#> 8   9.069730  8 0.06544703  3.09435574 5.975375    F-test 0.7608675 0.41197506
#> 9  10.984027  9 0.61086750  3.32602456 7.658002    F-test 0.7608675 0.41197506
#> 10 12.401371 10 0.78505677  4.20599328 8.195378    F-test 0.7608675 0.41197506
#> 11  4.677371  1 0.40872293 -2.35701315 7.034384    F-test 3.9424124 0.08746312
#> 12  7.182992  2 0.62799681 -0.52785376 7.710846    F-test 3.9424124 0.08746312
#> 13  3.835867  3 0.27864770 -2.79723403 6.633101    F-test 3.9424124 0.08746312
#> 14  4.815597  4 0.54982032 -2.65407437 7.469671    F-test 3.9424124 0.08746312
#> 15 11.058638  5 0.99465152  2.21665880 8.841979    F-test 3.9424124 0.08746312
#> 16  4.198574  6 0.11301674 -1.92355325 6.122128    F-test 3.9424124 0.08746312
#> 17  5.723981  7 0.68239403 -2.15468157 7.878662    F-test 3.9424124 0.08746312
#> 18 12.466740  8 0.06544703  6.49136560 5.975375    F-test 3.9424124 0.08746312
#> 19  9.483996  9 0.61086750  1.82599444 7.658002    F-test 3.9424124 0.08746312
#> 20 10.075769 10 0.78505677  1.88039130 8.195378    F-test 3.9424124 0.08746312
#> 21 10.738932  1 0.40872293  3.70454752 7.034384    F-test 0.9033491 0.37353873
#> 22 10.810178  2 0.62799681  3.09933199 7.710846    F-test 0.9033491 0.37353873
#> 23  2.624991  3 0.27864770 -4.00810960 6.633101    F-test 0.9033491 0.37353873
#> 24  7.555429  4 0.54982032  0.08575767 7.469671    F-test 0.9033491 0.37353873
#> 25  6.406956  5 0.99465152 -2.43502368 8.841979    F-test 0.9033491 0.37353873
#> 26 10.138223  6 0.11301674  4.01609533 6.122128    F-test 0.9033491 0.37353873
#> 27  6.057774  7 0.68239403 -1.82088802 7.878662    F-test 0.9033491 0.37353873
#> 28  2.009062  8 0.06544703 -3.96631310 5.975375    F-test 0.9033491 0.37353873
#> 29  7.430848  9 0.61086750 -0.22715452 7.658002    F-test 0.9033491 0.37353873
#> 30  9.747134 10 0.78505677  1.55175642 8.195378    F-test 0.9033491 0.37353873
#>    k  null
#> 1  2 FALSE
#> 2  2 FALSE
#> 3  2 FALSE
#> 4  2 FALSE
#> 5  2 FALSE
#> 6  2 FALSE
#> 7  2 FALSE
#> 8  2 FALSE
#> 9  2 FALSE
#> 10 2 FALSE
#> 11 1  TRUE
#> 12 1  TRUE
#> 13 1  TRUE
#> 14 1  TRUE
#> 15 1  TRUE
#> 16 1  TRUE
#> 17 1  TRUE
#> 18 1  TRUE
#> 19 1  TRUE
#> 20 1  TRUE
#> 21 3  TRUE
#> 22 3  TRUE
#> 23 3  TRUE
#> 24 3  TRUE
#> 25 3  TRUE
#> 26 3  TRUE
#> 27 3  TRUE
#> 28 3  TRUE
#> 29 3  TRUE
#> 30 3  TRUE