Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

A data frame.

Examples

# Instantiate
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>              y          x           e      .resid   .fitted test_name
#> 1   1.27991025 0.15970332  0.09470178  0.43384581 0.8460644    F-test
#> 2  -0.08275312 0.04165680 -1.12614521 -0.63923733 0.5564842    F-test
#> 3   2.67878588 0.83686868  0.14156802  0.17156541 2.5072205    F-test
#> 4   2.91385297 0.85574050  0.32582066  0.36033797 2.5535150    F-test
#> 5   1.58153297 0.34230680  0.12205222  0.28752348 1.2940095    F-test
#> 6   0.64419697 0.34300642 -0.81646285 -0.65152875 1.2957257    F-test
#> 7   1.42742620 0.06131952  0.36234660  0.82270731 0.6047189    F-test
#> 8   0.68089825 0.58673421 -1.25009298 -1.21271672 1.8936150    F-test
#> 9   2.89918283 0.91564820  0.14512301  0.19870811 2.7004747    F-test
#> 10  1.85834879 0.47909056  0.14973047  0.22879472 1.6295541    F-test
#> 11  0.47344190 0.15970332  0.09470178 -0.37262254 0.8460644    F-test
#> 12  1.17511591 0.04165680 -1.12614521  0.61863171 0.5564842    F-test
#> 13  1.62808864 0.83686868  0.14156802 -0.87913183 2.5072205    F-test
#> 14  3.55567568 0.85574050  0.32582066  1.00216068 2.5535150    F-test
#> 15  1.78466547 0.34230680  0.12205222  0.49065598 1.2940095    F-test
#> 16  0.31998279 0.34300642 -0.81646285 -0.97574293 1.2957257    F-test
#> 17  0.57598482 0.06131952  0.36234660 -0.02873407 0.6047189    F-test
#> 18  1.92785743 0.58673421 -1.25009298  0.03424245 1.8936150    F-test
#> 19  2.71721997 0.91564820  0.14512301  0.01674524 2.7004747    F-test
#> 20  1.72334938 0.47909056  0.14973047  0.09379531 1.6295541    F-test
#> 21  1.35039047 0.15970332  0.09470178  0.50432602 0.8460644    F-test
#> 22  0.38119592 0.04165680 -1.12614521 -0.17528829 0.5564842    F-test
#> 23  2.57496334 0.83686868  0.14156802  0.06774287 2.5072205    F-test
#> 24  2.92595540 0.85574050  0.32582066  0.37244040 2.5535150    F-test
#> 25  2.40375228 0.34230680  0.12205222  1.10974279 1.2940095    F-test
#> 26  1.52247272 0.34300642 -0.81646285  0.22674700 1.2957257    F-test
#> 27  0.08834026 0.06131952  0.36234660 -0.51637862 0.6047189    F-test
#> 28  0.75194887 0.58673421 -1.25009298 -1.14166611 1.8936150    F-test
#> 29  2.72268997 0.91564820  0.14512301  0.02221525 2.7004747    F-test
#> 30  1.15967276 0.47909056  0.14973047 -0.46988131 1.6295541    F-test
#>       statistic   p_value k  null
#> 1  0.9954153879 0.3516537 3 FALSE
#> 2  0.9954153879 0.3516537 3 FALSE
#> 3  0.9954153879 0.3516537 3 FALSE
#> 4  0.9954153879 0.3516537 3 FALSE
#> 5  0.9954153879 0.3516537 3 FALSE
#> 6  0.9954153879 0.3516537 3 FALSE
#> 7  0.9954153879 0.3516537 3 FALSE
#> 8  0.9954153879 0.3516537 3 FALSE
#> 9  0.9954153879 0.3516537 3 FALSE
#> 10 0.9954153879 0.3516537 3 FALSE
#> 11 0.3762303605 0.5590230 1  TRUE
#> 12 0.3762303605 0.5590230 1  TRUE
#> 13 0.3762303605 0.5590230 1  TRUE
#> 14 0.3762303605 0.5590230 1  TRUE
#> 15 0.3762303605 0.5590230 1  TRUE
#> 16 0.3762303605 0.5590230 1  TRUE
#> 17 0.3762303605 0.5590230 1  TRUE
#> 18 0.3762303605 0.5590230 1  TRUE
#> 19 0.3762303605 0.5590230 1  TRUE
#> 20 0.3762303605 0.5590230 1  TRUE
#> 21 0.0009921387 0.9757515 2  TRUE
#> 22 0.0009921387 0.9757515 2  TRUE
#> 23 0.0009921387 0.9757515 2  TRUE
#> 24 0.0009921387 0.9757515 2  TRUE
#> 25 0.0009921387 0.9757515 2  TRUE
#> 26 0.0009921387 0.9757515 2  TRUE
#> 27 0.0009921387 0.9757515 2  TRUE
#> 28 0.0009921387 0.9757515 2  TRUE
#> 29 0.0009921387 0.9757515 2  TRUE
#> 30 0.0009921387 0.9757515 2  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e           x     .resid  .fitted test_name  statistic   p_value
#> 1   2.358930  1 0.280339932 -4.1338029 6.492733    F-test 1.20326279 0.3089662
#> 2   3.035707  2 0.034515451 -2.1170908 5.152798    F-test 1.20326279 0.3089662
#> 3   5.076156  3 0.651588329 -3.4401720 8.516328    F-test 1.20326279 0.3089662
#> 4   5.774825  4 0.512336229 -1.9824701 7.757295    F-test 1.20326279 0.3089662
#> 5   6.008775  5 0.008699151  0.9966963 5.012079    F-test 1.20326279 0.3089662
#> 6   7.017613  6 0.017313468  1.9585800 5.059033    F-test 1.20326279 0.3089662
#> 7   9.023176  7 0.628351140  0.6335095 8.389667    F-test 1.20326279 0.3089662
#> 8  10.767779  8 0.920485370  0.7857517 9.982027    F-test 1.20326279 0.3089662
#> 9  10.237602  9 0.198285220  4.1921313 6.045471    F-test 1.20326279 0.3089662
#> 10 12.447170 10 0.802754758  3.1068670 9.340303    F-test 1.20326279 0.3089662
#> 11  8.549708  1 0.280339932  2.0569747 6.492733    F-test 1.15686512 0.3177849
#> 12  3.171532  2 0.034515451 -1.9812660 5.152798    F-test 1.15686512 0.3177849
#> 13 11.218468  3 0.651588329  2.7021403 8.516328    F-test 1.15686512 0.3177849
#> 14  4.685719  4 0.512336229 -3.0715756 7.757295    F-test 1.15686512 0.3177849
#> 15  6.722091  5 0.008699151  1.7100128 5.012079    F-test 1.15686512 0.3177849
#> 16  5.431324  6 0.017313468  0.3722910 5.059033    F-test 1.15686512 0.3177849
#> 17  2.998708  7 0.628351140 -5.3909587 8.389667    F-test 1.15686512 0.3177849
#> 18 13.392297  8 0.920485370  3.4102701 9.982027    F-test 1.15686512 0.3177849
#> 19  7.075299  9 0.198285220  1.0298283 6.045471    F-test 1.15686512 0.3177849
#> 20  8.502586 10 0.802754758 -0.8377169 9.340303    F-test 1.15686512 0.3177849
#> 21  3.814577  1 0.280339932 -2.6781562 6.492733    F-test 0.05043468 0.8287232
#> 22  1.862340  2 0.034515451 -3.2904579 5.152798    F-test 0.05043468 0.8287232
#> 23  9.618399  3 0.651588329  1.1020711 8.516328    F-test 0.05043468 0.8287232
#> 24 10.353487  4 0.512336229  2.5961919 7.757295    F-test 0.05043468 0.8287232
#> 25  3.760009  5 0.008699151 -1.2520691 5.012079    F-test 0.05043468 0.8287232
#> 26  9.158637  6 0.017313468  4.0996037 5.059033    F-test 0.05043468 0.8287232
#> 27  4.121871  7 0.628351140 -4.2677962 8.389667    F-test 0.05043468 0.8287232
#> 28 10.785557  8 0.920485370  0.8035301 9.982027    F-test 0.05043468 0.8287232
#> 29  8.724717  9 0.198285220  2.6792465 6.045471    F-test 0.05043468 0.8287232
#> 30  9.548139 10 0.802754758  0.2078361 9.340303    F-test 0.05043468 0.8287232
#>    k  null
#> 1  2 FALSE
#> 2  2 FALSE
#> 3  2 FALSE
#> 4  2 FALSE
#> 5  2 FALSE
#> 6  2 FALSE
#> 7  2 FALSE
#> 8  2 FALSE
#> 9  2 FALSE
#> 10 2 FALSE
#> 11 1  TRUE
#> 12 1  TRUE
#> 13 1  TRUE
#> 14 1  TRUE
#> 15 1  TRUE
#> 16 1  TRUE
#> 17 1  TRUE
#> 18 1  TRUE
#> 19 1  TRUE
#> 20 1  TRUE
#> 21 3  TRUE
#> 22 3  TRUE
#> 23 3  TRUE
#> 24 3  TRUE
#> 25 3  TRUE
#> 26 3  TRUE
#> 27 3  TRUE
#> 28 3  TRUE
#> 29 3  TRUE
#> 30 3  TRUE