Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>             y          x          e      .resid  .fitted test_name  statistic
#> 1  0.73599159 0.13823391 -0.4213509 -0.67641902 1.412411    F-test 2.47598884
#> 2  1.92433700 0.77705635 -0.4565359 -0.20279447 2.127131    F-test 2.47598884
#> 3  2.72873808 0.01128015  1.7173307  1.45836458 1.270373    F-test 2.47598884
#> 4  1.50757171 0.43310427 -0.1131119 -0.23474278 1.742314    F-test 2.47598884
#> 5  2.51132974 0.72620478  0.2577516  0.44109150 2.070238    F-test 2.47598884
#> 6  2.00002182 0.72687676 -0.2552048 -0.07096823 2.070990    F-test 2.47598884
#> 7  2.51401716 0.98899114 -0.4530775  0.14977094 2.364246    F-test 2.47598884
#> 8  0.97866220 0.17237429 -0.2234250 -0.47194501 1.450607    F-test 2.47598884
#> 9  0.61659013 0.49777815 -1.1289711 -1.19808213 1.814672    F-test 2.47598884
#> 10 2.89270046 0.74116496  0.6022100  0.80572462 2.086976    F-test 2.47598884
#> 11 0.05829224 0.13823391 -0.4213509 -1.35411838 1.412411    F-test 0.07086163
#> 12 3.11138800 0.77705635 -0.4565359  0.98425653 2.127131    F-test 0.07086163
#> 13 1.48268347 0.01128015  1.7173307  0.21230997 1.270373    F-test 0.07086163
#> 14 2.38870101 0.43310427 -0.1131119  0.64638652 1.742314    F-test 0.07086163
#> 15 2.25675429 0.72620478  0.2577516  0.18651605 2.070238    F-test 0.07086163
#> 16 1.41693498 0.72687676 -0.2552048 -0.65405508 2.070990    F-test 0.07086163
#> 17 2.29731130 0.98899114 -0.4530775 -0.06693492 2.364246    F-test 0.07086163
#> 18 2.23110133 0.17237429 -0.2234250  0.78049412 1.450607    F-test 0.07086163
#> 19 1.99425141 0.49777815 -1.1289711  0.17957915 1.814672    F-test 0.07086163
#> 20 1.17254188 0.74116496  0.6022100 -0.91443396 2.086976    F-test 0.07086163
#> 21 1.35945134 0.13823391 -0.4213509 -0.05295927 1.412411    F-test 1.95863066
#> 22 2.45435321 0.77705635 -0.4565359  0.32722174 2.127131    F-test 1.95863066
#> 23 0.62601757 0.01128015  1.7173307 -0.64435592 1.270373    F-test 1.95863066
#> 24 2.41149638 0.43310427 -0.1131119  0.66918189 1.742314    F-test 1.95863066
#> 25 1.00722426 0.72620478  0.2577516 -1.06301398 2.070238    F-test 1.95863066
#> 26 1.24318826 0.72687676 -0.2552048 -0.82780179 2.070990    F-test 1.95863066
#> 27 1.87741149 0.98899114 -0.4530775 -0.48683473 2.364246    F-test 1.95863066
#> 28 1.49146535 0.17237429 -0.2234250  0.04085814 1.450607    F-test 1.95863066
#> 29 2.60219543 0.49777815 -1.1289711  0.78752317 1.814672    F-test 1.95863066
#> 30 3.33715659 0.74116496  0.6022100  1.25018075 2.086976    F-test 1.95863066
#>      p_value k  null
#> 1  0.1595957 1 FALSE
#> 2  0.1595957 1 FALSE
#> 3  0.1595957 1 FALSE
#> 4  0.1595957 1 FALSE
#> 5  0.1595957 1 FALSE
#> 6  0.1595957 1 FALSE
#> 7  0.1595957 1 FALSE
#> 8  0.1595957 1 FALSE
#> 9  0.1595957 1 FALSE
#> 10 0.1595957 1 FALSE
#> 11 0.7977569 2  TRUE
#> 12 0.7977569 2  TRUE
#> 13 0.7977569 2  TRUE
#> 14 0.7977569 2  TRUE
#> 15 0.7977569 2  TRUE
#> 16 0.7977569 2  TRUE
#> 17 0.7977569 2  TRUE
#> 18 0.7977569 2  TRUE
#> 19 0.7977569 2  TRUE
#> 20 0.7977569 2  TRUE
#> 21 0.2043814 3  TRUE
#> 22 0.2043814 3  TRUE
#> 23 0.2043814 3  TRUE
#> 24 0.2043814 3  TRUE
#> 25 0.2043814 3  TRUE
#> 26 0.2043814 3  TRUE
#> 27 0.2043814 3  TRUE
#> 28 0.2043814 3  TRUE
#> 29 0.2043814 3  TRUE
#> 30 0.2043814 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e          x      .resid   .fitted test_name statistic    p_value
#> 1   2.088779  1 0.08204759 -2.21106316  4.299843    F-test  0.287958 0.60815621
#> 2   3.016361  2 0.01610171 -0.79126958  3.807631    F-test  0.287958 0.60815621
#> 3   4.391182  3 0.30073856 -1.54094295  5.932125    F-test  0.287958 0.60815621
#> 4   6.433592  4 0.79753316 -3.20654698  9.640139    F-test  0.287958 0.60815621
#> 5   6.723277  5 0.48654820 -0.59570827  7.318986    F-test  0.287958 0.60815621
#> 6   7.906206  6 0.57527014 -0.07498946  7.981195    F-test  0.287958 0.60815621
#> 7   8.698825  7 0.47407671  1.47292551  7.225900    F-test  0.287958 0.60815621
#> 8  10.740534  8 0.91086288  0.25451562 10.486018    F-test  0.287958 0.60815621
#> 9  11.825079  9 0.94051328  1.11775309 10.707325    F-test  0.287958 0.60815621
#> 10 11.359918 10 0.28097218  5.57532619  5.784591    F-test  0.287958 0.60815621
#> 11  5.556256  1 0.08204759  1.25641310  4.299843    F-test  2.824544 0.13672052
#> 12  5.760179  2 0.01610171  1.95254806  3.807631    F-test  2.824544 0.13672052
#> 13  7.610687  3 0.30073856  1.67856213  5.932125    F-test  2.824544 0.13672052
#> 14 12.547853  4 0.79753316  2.90771386  9.640139    F-test  2.824544 0.13672052
#> 15  4.378611  5 0.48654820 -2.94037488  7.318986    F-test  2.824544 0.13672052
#> 16  6.793498  6 0.57527014 -1.18769708  7.981195    F-test  2.824544 0.13672052
#> 17  5.643191  7 0.47407671 -1.58270916  7.225900    F-test  2.824544 0.13672052
#> 18  7.946638  8 0.91086288 -2.53938086 10.486018    F-test  2.824544 0.13672052
#> 19 13.877835  9 0.94051328  3.17050941 10.707325    F-test  2.824544 0.13672052
#> 20  3.069007 10 0.28097218 -2.71558458  5.784591    F-test  2.824544 0.13672052
#> 21  4.387577  1 0.08204759  0.08773492  4.299843    F-test 11.823554 0.01085868
#> 22  7.805145  2 0.01610171  3.99751438  3.807631    F-test 11.823554 0.01085868
#> 23  5.233792  3 0.30073856 -0.69833334  5.932125    F-test 11.823554 0.01085868
#> 24  7.026392  4 0.79753316 -2.61374702  9.640139    F-test 11.823554 0.01085868
#> 25  5.791689  5 0.48654820 -1.52729636  7.318986    F-test 11.823554 0.01085868
#> 26  5.275572  6 0.57527014 -2.70562341  7.981195    F-test 11.823554 0.01085868
#> 27  7.615603  7 0.47407671  0.38970315  7.225900    F-test 11.823554 0.01085868
#> 28 12.831531  8 0.91086288  2.34551283 10.486018    F-test 11.823554 0.01085868
#> 29 13.736408  9 0.94051328  3.02908220 10.707325    F-test 11.823554 0.01085868
#> 30  3.480044 10 0.28097218 -2.30454735  5.784591    F-test 11.823554 0.01085868
#>    k  null
#> 1  3 FALSE
#> 2  3 FALSE
#> 3  3 FALSE
#> 4  3 FALSE
#> 5  3 FALSE
#> 6  3 FALSE
#> 7  3 FALSE
#> 8  3 FALSE
#> 9  3 FALSE
#> 10 3 FALSE
#> 11 1  TRUE
#> 12 1  TRUE
#> 13 1  TRUE
#> 14 1  TRUE
#> 15 1  TRUE
#> 16 1  TRUE
#> 17 1  TRUE
#> 18 1  TRUE
#> 19 1  TRUE
#> 20 1  TRUE
#> 21 2  TRUE
#> 22 2  TRUE
#> 23 2  TRUE
#> 24 2  TRUE
#> 25 2  TRUE
#> 26 2  TRUE
#> 27 2  TRUE
#> 28 2  TRUE
#> 29 2  TRUE
#> 30 2  TRUE