Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>             y          x          e      .resid    .fitted test_name statistic
#> 1   0.8239816 0.00959796 -0.1857085  0.74299986 0.08098176    F-test 0.2206878
#> 2   3.1374483 0.62216904  1.1281849  1.67953978 1.45790847    F-test 0.2206878
#> 3  -0.7010949 0.71833140 -2.9354263 -2.37515548 1.67406058    F-test 0.2206878
#> 4   3.2466358 0.92538708  0.4649075  1.10715908 2.13947677    F-test 0.2206878
#> 5   2.1824783 0.95161530 -0.6747087 -0.01595384 2.19843211    F-test 0.2206878
#> 6  -0.2411181 0.01701463 -1.2584222 -0.33877093 0.09765283    F-test 0.2206878
#> 7   2.1826605 0.45852393  0.5138924  1.09259070 1.09006983    F-test 0.2206878
#> 8   0.8832201 0.32004185 -0.5392485  0.10442795 0.77879220    F-test 0.2206878
#> 9  -0.6924946 0.46022547 -2.3645275 -1.78638908 1.09389452    F-test 0.2206878
#> 10  0.5090783 0.29367547 -0.8708425 -0.21044802 0.71952631    F-test 0.2206878
#> 11 -0.5322861 0.00959796 -0.1857085 -0.61326785 0.08098176    F-test 0.7788517
#> 12  2.1906366 0.62216904  1.1281849  0.73272808 1.45790847    F-test 0.7788517
#> 13  1.7926887 0.71833140 -2.9354263  0.11862816 1.67406058    F-test 0.7788517
#> 14  3.3750284 0.92538708  0.4649075  1.23555164 2.13947677    F-test 0.7788517
#> 15  1.8278625 0.95161530 -0.6747087 -0.37056964 2.19843211    F-test 0.7788517
#> 16  1.3249080 0.01701463 -1.2584222  1.22725514 0.09765283    F-test 0.7788517
#> 17 -0.3892191 0.45852393  0.5138924 -1.47928888 1.09006983    F-test 0.7788517
#> 18  1.2883497 0.32004185 -0.5392485  0.50955748 0.77879220    F-test 0.7788517
#> 19 -1.5001088 0.46022547 -2.3645275 -2.59400336 1.09389452    F-test 0.7788517
#> 20  1.9529355 0.29367547 -0.8708425  1.23340923 0.71952631    F-test 0.7788517
#> 21 -0.4731188 0.00959796 -0.1857085 -0.55410057 0.08098176    F-test 0.9924528
#> 22 -0.5968056 0.62216904  1.1281849 -2.05471403 1.45790847    F-test 0.9924528
#> 23  3.5127795 0.71833140 -2.9354263  1.83871888 1.67406058    F-test 0.9924528
#> 24  2.2204501 0.92538708  0.4649075  0.08097338 2.13947677    F-test 0.9924528
#> 25  1.0536364 0.95161530 -0.6747087 -1.14479570 2.19843211    F-test 0.9924528
#> 26 -0.5436813 0.01701463 -1.2584222 -0.64133413 0.09765283    F-test 0.9924528
#> 27  0.5421796 0.45852393  0.5138924 -0.54789022 1.09006983    F-test 0.9924528
#> 28  1.8069584 0.32004185 -0.5392485  1.02816620 0.77879220    F-test 0.9924528
#> 29  3.0426343 0.46022547 -2.3645275  1.94873978 1.09389452    F-test 0.9924528
#> 30  0.7657627 0.29367547 -0.8708425  0.04623642 0.71952631    F-test 0.9924528
#>      p_value k  null
#> 1  0.6527990 1 FALSE
#> 2  0.6527990 1 FALSE
#> 3  0.6527990 1 FALSE
#> 4  0.6527990 1 FALSE
#> 5  0.6527990 1 FALSE
#> 6  0.6527990 1 FALSE
#> 7  0.6527990 1 FALSE
#> 8  0.6527990 1 FALSE
#> 9  0.6527990 1 FALSE
#> 10 0.6527990 1 FALSE
#> 11 0.4067764 2  TRUE
#> 12 0.4067764 2  TRUE
#> 13 0.4067764 2  TRUE
#> 14 0.4067764 2  TRUE
#> 15 0.4067764 2  TRUE
#> 16 0.4067764 2  TRUE
#> 17 0.4067764 2  TRUE
#> 18 0.4067764 2  TRUE
#> 19 0.4067764 2  TRUE
#> 20 0.4067764 2  TRUE
#> 21 0.3523263 3  TRUE
#> 22 0.3523263 3  TRUE
#> 23 0.3523263 3  TRUE
#> 24 0.3523263 3  TRUE
#> 25 0.3523263 3  TRUE
#> 26 0.3523263 3  TRUE
#> 27 0.3523263 3  TRUE
#> 28 0.3523263 3  TRUE
#> 29 0.3523263 3  TRUE
#> 30 0.3523263 3  TRUE