Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

A data frame.

Examples

# Instantiate
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>             y         x           e      .resid  .fitted test_name  statistic
#> 1   1.7836079 0.7979771 -0.65113659 -0.66855744 2.452165    F-test 1.14033234
#> 2   0.7662124 0.5240618 -1.03249020 -1.14439776 1.910610    F-test 1.14033234
#> 3   1.2225168 0.4073513 -0.35076960 -0.45734617 1.679863    F-test 1.14033234
#> 4   3.0044330 0.8748758  0.36414958  0.40023197 2.604201    F-test 1.14033234
#> 5   3.9651692 0.6821080  1.81778988  1.74208747 2.223082    F-test 1.14033234
#> 6   1.3122208 0.4238244 -0.29123081 -0.40021105 1.712432    F-test 1.14033234
#> 7   0.1078672 0.2937947 -1.27224286 -1.34748416 1.455351    F-test 1.14033234
#> 8   2.7872440 0.1806393  1.57397406  1.55561102 1.231633    F-test 1.14033234
#> 9   1.5883097 0.2076196  0.33758415  0.30333425 1.284975    F-test 1.14033234
#> 10  1.1989802 0.1556609  0.01908899  0.01673187 1.182248    F-test 1.14033234
#> 11  2.1732909 0.7979771 -0.65113659 -0.27887448 2.452165    F-test 0.25376302
#> 12  1.3206386 0.5240618 -1.03249020 -0.58997157 1.910610    F-test 0.25376302
#> 13  0.2341105 0.4073513 -0.35076960 -1.44575250 1.679863    F-test 0.25376302
#> 14  3.1214486 0.8748758  0.36414958  0.51724764 2.604201    F-test 0.25376302
#> 15  2.4791684 0.6821080  1.81778988  0.25608671 2.223082    F-test 0.25376302
#> 16  2.0420946 0.4238244 -0.29123081  0.32966283 1.712432    F-test 0.25376302
#> 17  1.8275223 0.2937947 -1.27224286  0.37217092 1.455351    F-test 0.25376302
#> 18  1.3360239 0.1806393  1.57397406  0.10439100 1.231633    F-test 0.25376302
#> 19  3.4233063 0.2076196  0.33758415  2.13833083 1.284975    F-test 0.25376302
#> 20 -0.2210430 0.1556609  0.01908899 -1.40329138 1.182248    F-test 0.25376302
#> 21  1.9260065 0.7979771 -0.65113659 -0.52615887 2.452165    F-test 0.00482028
#> 22  3.6775698 0.5240618 -1.03249020  1.76695961 1.910610    F-test 0.00482028
#> 23  1.0666621 0.4073513 -0.35076960 -0.61320092 1.679863    F-test 0.00482028
#> 24  2.8324750 0.8748758  0.36414958  0.22827405 2.604201    F-test 0.00482028
#> 25  1.9584860 0.6821080  1.81778988 -0.26459573 2.223082    F-test 0.00482028
#> 26  1.8786401 0.4238244 -0.29123081  0.16620832 1.712432    F-test 0.00482028
#> 27  0.0357527 0.2937947 -1.27224286 -1.41959868 1.455351    F-test 0.00482028
#> 28  0.3857672 0.1806393  1.57397406 -0.84586577 1.231633    F-test 0.00482028
#> 29  1.0755235 0.2076196  0.33758415 -0.20945192 1.284975    F-test 0.00482028
#> 30  2.8996783 0.1556609  0.01908899  1.71742993 1.182248    F-test 0.00482028
#>      p_value k  null
#> 1  0.3210192 2 FALSE
#> 2  0.3210192 2 FALSE
#> 3  0.3210192 2 FALSE
#> 4  0.3210192 2 FALSE
#> 5  0.3210192 2 FALSE
#> 6  0.3210192 2 FALSE
#> 7  0.3210192 2 FALSE
#> 8  0.3210192 2 FALSE
#> 9  0.3210192 2 FALSE
#> 10 0.3210192 2 FALSE
#> 11 0.6299011 1  TRUE
#> 12 0.6299011 1  TRUE
#> 13 0.6299011 1  TRUE
#> 14 0.6299011 1  TRUE
#> 15 0.6299011 1  TRUE
#> 16 0.6299011 1  TRUE
#> 17 0.6299011 1  TRUE
#> 18 0.6299011 1  TRUE
#> 19 0.6299011 1  TRUE
#> 20 0.6299011 1  TRUE
#> 21 0.9465905 3  TRUE
#> 22 0.9465905 3  TRUE
#> 23 0.9465905 3  TRUE
#> 24 0.9465905 3  TRUE
#> 25 0.9465905 3  TRUE
#> 26 0.9465905 3  TRUE
#> 27 0.9465905 3  TRUE
#> 28 0.9465905 3  TRUE
#> 29 0.9465905 3  TRUE
#> 30 0.9465905 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e         x      .resid  .fitted test_name    statistic   p_value
#> 1   2.666550  1 0.4573663 -4.70241767 7.368968    F-test 0.0004247577 0.9841322
#> 2   4.170048  2 0.6916575 -3.49620066 7.666248    F-test 0.0004247577 0.9841322
#> 3   5.587642  3 0.8555966 -2.28662020 7.874262    F-test 0.0004247577 0.9841322
#> 4   6.850553  4 0.9493284 -1.14264112 7.993194    F-test 0.0004247577 0.9841322
#> 5   6.143815  5 0.1275468 -0.80666153 6.950477    F-test 0.0004247577 0.9841322
#> 6   8.986391  6 0.9954567  0.93466703 8.051724    F-test 0.0004247577 0.9841322
#> 7   8.174277  7 0.1513658  1.19357816 6.980699    F-test 0.0004247577 0.9841322
#> 8   9.515046  8 0.3746692  2.25100848 7.264038    F-test 0.0004247577 0.9841322
#> 9  11.172553  9 0.6927082  3.50497140 7.667581    F-test 0.0004247577 0.9841322
#> 10 12.267677 10 0.7319405  4.55031609 7.717361    F-test 0.0004247577 0.9841322
#> 11  5.727229  1 0.4573663 -1.64173877 7.368968    F-test 0.6025773673 0.4630133
#> 12 10.679605  2 0.6916575  3.01335682 7.666248    F-test 0.6025773673 0.4630133
#> 13 11.907685  3 0.8555966  4.03342262 7.874262    F-test 0.6025773673 0.4630133
#> 14  2.010298  4 0.9493284 -5.98289580 7.993194    F-test 0.6025773673 0.4630133
#> 15  8.215823  5 0.1275468  1.26534674 6.950477    F-test 0.6025773673 0.4630133
#> 16  7.386322  6 0.9954567 -0.66540159 8.051724    F-test 0.6025773673 0.4630133
#> 17  5.787026  7 0.1513658 -1.19367323 6.980699    F-test 0.6025773673 0.4630133
#> 18  5.257482  8 0.3746692 -2.00655575 7.264038    F-test 0.6025773673 0.4630133
#> 19  7.599522  9 0.6927082 -0.06805955 7.667581    F-test 0.6025773673 0.4630133
#> 20 10.963560 10 0.7319405  3.24619851 7.717361    F-test 0.6025773673 0.4630133
#> 21 10.442714  1 0.4573663  3.07374595 7.368968    F-test 0.0547876219 0.8216310
#> 22  7.054327  2 0.6916575 -0.61192114 7.666248    F-test 0.0547876219 0.8216310
#> 23 11.426797  3 0.8555966  3.55253489 7.874262    F-test 0.0547876219 0.8216310
#> 24  3.797211  4 0.9493284 -4.19598304 7.993194    F-test 0.0547876219 0.8216310
#> 25  3.735635  5 0.1275468 -3.21484114 6.950477    F-test 0.0547876219 0.8216310
#> 26  9.919120  6 0.9954567  1.86739618 8.051724    F-test 0.0547876219 0.8216310
#> 27  9.992471  7 0.1513658  3.01177193 6.980699    F-test 0.0547876219 0.8216310
#> 28  5.364538  8 0.3746692 -1.89949995 7.264038    F-test 0.0547876219 0.8216310
#> 29  9.528614  9 0.6927082  1.86103244 7.667581    F-test 0.0547876219 0.8216310
#> 30  4.273125 10 0.7319405 -3.44423612 7.717361    F-test 0.0547876219 0.8216310
#>    k  null
#> 1  3 FALSE
#> 2  3 FALSE
#> 3  3 FALSE
#> 4  3 FALSE
#> 5  3 FALSE
#> 6  3 FALSE
#> 7  3 FALSE
#> 8  3 FALSE
#> 9  3 FALSE
#> 10 3 FALSE
#> 11 1  TRUE
#> 12 1  TRUE
#> 13 1  TRUE
#> 14 1  TRUE
#> 15 1  TRUE
#> 16 1  TRUE
#> 17 1  TRUE
#> 18 1  TRUE
#> 19 1  TRUE
#> 20 1  TRUE
#> 21 2  TRUE
#> 22 2  TRUE
#> 23 2  TRUE
#> 24 2  TRUE
#> 25 2  TRUE
#> 26 2  TRUE
#> 27 2  TRUE
#> 28 2  TRUE
#> 29 2  TRUE
#> 30 2  TRUE