Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>            y           x           e      .resid  .fitted test_name
#> 1  1.6562857 0.403267271  0.09039393 -0.35275151 2.009037    F-test
#> 2  1.4788673 0.891437371 -1.20723066 -0.90742313 2.386290    F-test
#> 3  3.8304908 0.704786218  1.62898101  1.58844265 2.242048    F-test
#> 4  2.4750736 0.002782991  1.47228288  0.77552688 1.699547    F-test
#> 5  2.2571421 0.973919069 -0.66529534 -0.19288942 2.450032    F-test
#> 6  0.7752492 0.185274709 -0.44435218 -1.06532537 1.840575    F-test
#> 7  3.3598969 0.526509125  1.55617587  1.25561952 2.104277    F-test
#> 8  0.3578862 0.524207954 -1.44111578 -1.74461286 2.102499    F-test
#> 9  2.7474287 0.755625758  0.42083262  0.46609216 2.281337    F-test
#> 10 2.3801891 0.654086649  0.29827314  0.17732109 2.202868    F-test
#> 11 3.1008870 0.403267271  0.09039393  1.09184977 2.009037    F-test
#> 12 1.9139412 0.891437371 -1.20723066 -0.47234924 2.386290    F-test
#> 13 3.3501011 0.704786218  1.62898101  1.10805287 2.242048    F-test
#> 14 1.6419922 0.002782991  1.47228288 -0.05755452 1.699547    F-test
#> 15 3.9932898 0.973919069 -0.66529534  1.54325828 2.450032    F-test
#> 16 2.0198099 0.185274709 -0.44435218  0.17923534 1.840575    F-test
#> 17 2.3595561 0.526509125  1.55617587  0.25527873 2.104277    F-test
#> 18 1.3628620 0.524207954 -1.44111578 -0.73963701 2.102499    F-test
#> 19 0.5012796 0.755625758  0.42083262 -1.78005686 2.281337    F-test
#> 20 1.0747907 0.654086649  0.29827314 -1.12807736 2.202868    F-test
#> 21 3.0407541 0.403267271  0.09039393  1.03171689 2.009037    F-test
#> 22 2.0434261 0.891437371 -1.20723066 -0.34286429 2.386290    F-test
#> 23 2.6119040 0.704786218  1.62898101  0.36985584 2.242048    F-test
#> 24 2.5525155 0.002782991  1.47228288  0.85296876 1.699547    F-test
#> 25 3.2621338 0.973919069 -0.66529534  0.81210232 2.450032    F-test
#> 26 0.4534963 0.185274709 -0.44435218 -1.38707829 1.840575    F-test
#> 27 1.6025853 0.526509125  1.55617587 -0.50169205 2.104277    F-test
#> 28 1.5282649 0.524207954 -1.44111578 -0.57423408 2.102499    F-test
#> 29 0.5907850 0.755625758  0.42083262 -1.69055153 2.281337    F-test
#> 30 3.6326445 0.654086649  0.29827314  1.42977645 2.202868    F-test
#>       statistic   p_value k  null
#> 1  2.578777e-05 0.9960899 2 FALSE
#> 2  2.578777e-05 0.9960899 2 FALSE
#> 3  2.578777e-05 0.9960899 2 FALSE
#> 4  2.578777e-05 0.9960899 2 FALSE
#> 5  2.578777e-05 0.9960899 2 FALSE
#> 6  2.578777e-05 0.9960899 2 FALSE
#> 7  2.578777e-05 0.9960899 2 FALSE
#> 8  2.578777e-05 0.9960899 2 FALSE
#> 9  2.578777e-05 0.9960899 2 FALSE
#> 10 2.578777e-05 0.9960899 2 FALSE
#> 11 3.474306e-01 0.5740943 1  TRUE
#> 12 3.474306e-01 0.5740943 1  TRUE
#> 13 3.474306e-01 0.5740943 1  TRUE
#> 14 3.474306e-01 0.5740943 1  TRUE
#> 15 3.474306e-01 0.5740943 1  TRUE
#> 16 3.474306e-01 0.5740943 1  TRUE
#> 17 3.474306e-01 0.5740943 1  TRUE
#> 18 3.474306e-01 0.5740943 1  TRUE
#> 19 3.474306e-01 0.5740943 1  TRUE
#> 20 3.474306e-01 0.5740943 1  TRUE
#> 21 2.138309e-01 0.6578109 3  TRUE
#> 22 2.138309e-01 0.6578109 3  TRUE
#> 23 2.138309e-01 0.6578109 3  TRUE
#> 24 2.138309e-01 0.6578109 3  TRUE
#> 25 2.138309e-01 0.6578109 3  TRUE
#> 26 2.138309e-01 0.6578109 3  TRUE
#> 27 2.138309e-01 0.6578109 3  TRUE
#> 28 2.138309e-01 0.6578109 3  TRUE
#> 29 2.138309e-01 0.6578109 3  TRUE
#> 30 2.138309e-01 0.6578109 3  TRUE