Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>             y          x          e      .resid  .fitted test_name statistic
#> 1   1.3991426 0.98920216 -1.5685804 -1.44615482 2.845297    F-test 5.1708450
#> 2   3.1761561 0.55202262  1.3194045  1.09108908 2.085067    F-test 5.1708450
#> 3   2.1717432 0.61910434  0.1693486 -0.02997516 2.201718    F-test 5.1708450
#> 4   3.3062190 0.82414812  0.8028508  0.74794118 2.558278    F-test 5.1708450
#> 5   2.7011319 0.76776265  0.3439098  0.24090523 2.460227    F-test 5.1708450
#> 6   0.2967885 0.04330748 -0.7483945 -0.90365171 1.200440    F-test 5.1708450
#> 7   0.3283376 0.02339089 -0.6956004 -0.83746882 1.165806    F-test 5.1708450
#> 8   2.1186712 0.82620818 -0.3901569 -0.44318897 2.561860    F-test 5.1708450
#> 9   2.5427031 0.11419569  1.4154668  1.21899231 1.323711    F-test 5.1708450
#> 10  3.0793273 0.91589220  0.3245765  0.36151167 2.717816    F-test 5.1708450
#> 11  4.4236347 0.98920216 -1.5685804  1.57833725 2.845297    F-test 0.4478477
#> 12  3.0247608 0.55202262  1.3194045  0.93969380 2.085067    F-test 0.4478477
#> 13  1.8613121 0.61910434  0.1693486 -0.34040623 2.201718    F-test 0.4478477
#> 14  2.1421763 0.82414812  0.8028508 -0.41610149 2.558278    F-test 0.4478477
#> 15  1.8149789 0.76776265  0.3439098 -0.64524779 2.460227    F-test 0.4478477
#> 16  0.2089881 0.04330748 -0.7483945 -0.99145208 1.200440    F-test 0.4478477
#> 17  1.8183863 0.02339089 -0.6956004  0.65257994 1.165806    F-test 0.4478477
#> 18  1.3413099 0.82620818 -0.3901569 -1.22055029 2.561860    F-test 0.4478477
#> 19  1.8235592 0.11419569  1.4154668  0.49984844 1.323711    F-test 0.4478477
#> 20  2.6611140 0.91589220  0.3245765 -0.05670155 2.717816    F-test 0.4478477
#> 21  3.7123781 0.98920216 -1.5685804  0.86708064 2.845297    F-test 0.5720960
#> 22  3.1106034 0.55202262  1.3194045  1.02553640 2.085067    F-test 0.5720960
#> 23  3.2914750 0.61910434  0.1693486  1.08975669 2.201718    F-test 0.5720960
#> 24  1.3110835 0.82414812  0.8028508 -1.24719433 2.558278    F-test 0.5720960
#> 25  1.9642562 0.76776265  0.3439098 -0.49597052 2.460227    F-test 0.5720960
#> 26  1.8403895 0.04330748 -0.7483945  0.63994928 1.200440    F-test 0.5720960
#> 27 -0.1173327 0.02339089 -0.6956004 -1.28313911 1.165806    F-test 0.5720960
#> 28  2.1695708 0.82620818 -0.3901569 -0.39228938 2.561860    F-test 0.5720960
#> 29  1.5442976 0.11419569  1.4154668  0.22058677 1.323711    F-test 0.5720960
#> 30  2.2934991 0.91589220  0.3245765 -0.42431645 2.717816    F-test 0.5720960
#>       p_value k  null
#> 1  0.05714377 3 FALSE
#> 2  0.05714377 3 FALSE
#> 3  0.05714377 3 FALSE
#> 4  0.05714377 3 FALSE
#> 5  0.05714377 3 FALSE
#> 6  0.05714377 3 FALSE
#> 7  0.05714377 3 FALSE
#> 8  0.05714377 3 FALSE
#> 9  0.05714377 3 FALSE
#> 10 0.05714377 3 FALSE
#> 11 0.52480861 1  TRUE
#> 12 0.52480861 1  TRUE
#> 13 0.52480861 1  TRUE
#> 14 0.52480861 1  TRUE
#> 15 0.52480861 1  TRUE
#> 16 0.52480861 1  TRUE
#> 17 0.52480861 1  TRUE
#> 18 0.52480861 1  TRUE
#> 19 0.52480861 1  TRUE
#> 20 0.52480861 1  TRUE
#> 21 0.47410847 2  TRUE
#> 22 0.47410847 2  TRUE
#> 23 0.47410847 2  TRUE
#> 24 0.47410847 2  TRUE
#> 25 0.47410847 2  TRUE
#> 26 0.47410847 2  TRUE
#> 27 0.47410847 2  TRUE
#> 28 0.47410847 2  TRUE
#> 29 0.47410847 2  TRUE
#> 30 0.47410847 2  TRUE