Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

A data frame.

Examples

# Instantiate
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>              y         x           e      .resid  .fitted test_name  statistic
#> 1   1.51838396 0.1974564  0.28193856  0.44224301 1.076141    F-test 1.53437454
#> 2   0.87270708 0.5163358 -0.91023133 -0.87011600 1.742823    F-test 1.53437454
#> 3   3.69765457 0.7683006  1.33906823  1.42804792 2.269607    F-test 1.53437454
#> 4   0.87901381 0.2908329 -0.49640280 -0.39234963 1.271363    F-test 1.53437454
#> 5   3.30890332 0.3825573  1.77999589  1.84577126 1.463132    F-test 1.53437454
#> 6   2.27993940 0.6966322  0.09801075  0.16017001 2.119769    F-test 1.53437454
#> 7   1.77632315 0.4089742  0.20008905  0.25796124 1.518362    F-test 1.53437454
#> 8   1.27951192 0.6525893 -0.79895023 -0.74817687 2.027689    F-test 1.53437454
#> 9   0.02572197 0.5397591 -1.80537696 -1.76607226 1.791794    F-test 1.53437454
#> 10  0.90706333 0.2875701 -0.46320336 -0.35747868 1.264542    F-test 1.53437454
#> 11 -0.04882586 0.1974564  0.28193856 -1.12496681 1.076141    F-test 0.04135294
#> 12  2.53693770 0.5163358 -0.91023133  0.79411463 1.742823    F-test 0.04135294
#> 13  2.82075544 0.7683006  1.33906823  0.55114880 2.269607    F-test 0.04135294
#> 14  3.33903066 0.2908329 -0.49640280  2.06766721 1.271363    F-test 0.04135294
#> 15  0.22088307 0.3825573  1.77999589 -1.24224899 1.463132    F-test 0.04135294
#> 16  1.97086393 0.6966322  0.09801075 -0.14890546 2.119769    F-test 0.04135294
#> 17  1.72225277 0.4089742  0.20008905  0.20389087 1.518362    F-test 0.04135294
#> 18  1.94550047 0.6525893 -0.79895023 -0.08218833 2.027689    F-test 0.04135294
#> 19  0.32232096 0.5397591 -1.80537696 -1.46947327 1.791794    F-test 0.04135294
#> 20  1.71550338 0.2875701 -0.46320336  0.45096137 1.264542    F-test 0.04135294
#> 21  0.81393490 0.1974564  0.28193856 -0.26220605 1.076141    F-test 2.81783391
#> 22  1.68872797 0.5163358 -0.91023133 -0.05409511 1.742823    F-test 2.81783391
#> 23  1.14094471 0.7683006  1.33906823 -1.12866193 2.269607    F-test 2.81783391
#> 24  0.86897761 0.2908329 -0.49640280 -0.40238584 1.271363    F-test 2.81783391
#> 25  0.81846516 0.3825573  1.77999589 -0.64466690 1.463132    F-test 2.81783391
#> 26  0.98887757 0.6966322  0.09801075 -1.13089181 2.119769    F-test 2.81783391
#> 27  3.26666181 0.4089742  0.20008905  1.74829990 1.518362    F-test 2.81783391
#> 28  3.63889535 0.6525893 -0.79895023  1.61120655 2.027689    F-test 2.81783391
#> 29  2.80734136 0.5397591 -1.80537696  1.01554713 1.791794    F-test 2.81783391
#> 30  0.51239607 0.2875701 -0.46320336 -0.75214594 1.264542    F-test 2.81783391
#>      p_value k  null
#> 1  0.2553805 3 FALSE
#> 2  0.2553805 3 FALSE
#> 3  0.2553805 3 FALSE
#> 4  0.2553805 3 FALSE
#> 5  0.2553805 3 FALSE
#> 6  0.2553805 3 FALSE
#> 7  0.2553805 3 FALSE
#> 8  0.2553805 3 FALSE
#> 9  0.2553805 3 FALSE
#> 10 0.2553805 3 FALSE
#> 11 0.8446432 1  TRUE
#> 12 0.8446432 1  TRUE
#> 13 0.8446432 1  TRUE
#> 14 0.8446432 1  TRUE
#> 15 0.8446432 1  TRUE
#> 16 0.8446432 1  TRUE
#> 17 0.8446432 1  TRUE
#> 18 0.8446432 1  TRUE
#> 19 0.8446432 1  TRUE
#> 20 0.8446432 1  TRUE
#> 21 0.1371174 2  TRUE
#> 22 0.1371174 2  TRUE
#> 23 0.1371174 2  TRUE
#> 24 0.1371174 2  TRUE
#> 25 0.1371174 2  TRUE
#> 26 0.1371174 2  TRUE
#> 27 0.1371174 2  TRUE
#> 28 0.1371174 2  TRUE
#> 29 0.1371174 2  TRUE
#> 30 0.1371174 2  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e          x     .resid  .fitted test_name statistic     p_value
#> 1   2.769907  1 0.50990446 -4.4451480 7.215055    F-test  1.363856 0.281103800
#> 2   3.512654  2 0.37330087 -3.7627244 7.275379    F-test  1.363856 0.281103800
#> 3   4.794422  3 0.52196965 -2.4153050 7.209727    F-test  1.363856 0.281103800
#> 4   5.391068  4 0.30066740 -1.9163853 7.307454    F-test  1.363856 0.281103800
#> 5   7.759133  5 0.91743896  0.7240446 7.035089    F-test  1.363856 0.281103800
#> 6   7.862889  6 0.55493540  0.6677193 7.195169    F-test  1.363856 0.281103800
#> 7   8.373844  7 0.28983766  1.0616075 7.312236    F-test  1.363856 0.281103800
#> 8  10.292988  8 0.74217074  3.1805015 7.112487    F-test  1.363856 0.281103800
#> 9  10.035924  9 0.03471863  2.6110280 7.424896    F-test  1.363856 0.281103800
#> 10 11.558595 10 0.39921922  4.2946619 7.263933    F-test  1.363856 0.281103800
#> 11  7.561661  1 0.50990446  0.3466059 7.215055    F-test  1.039131 0.341955090
#> 12  2.072460  2 0.37330087 -5.2029186 7.275379    F-test  1.039131 0.341955090
#> 13 10.396093  3 0.52196965  3.1863664 7.209727    F-test  1.039131 0.341955090
#> 14  3.999196  4 0.30066740 -3.3082573 7.307454    F-test  1.039131 0.341955090
#> 15  6.709480  5 0.91743896 -0.3256085 7.035089    F-test  1.039131 0.341955090
#> 16  6.310529  6 0.55493540 -0.8846402 7.195169    F-test  1.039131 0.341955090
#> 17  5.205853  7 0.28983766 -2.1063826 7.312236    F-test  1.039131 0.341955090
#> 18  9.427594  8 0.74217074  2.3151070 7.112487    F-test  1.039131 0.341955090
#> 19 11.837117  9 0.03471863  4.4122210 7.424896    F-test  1.039131 0.341955090
#> 20  8.831440 10 0.39921922  1.5675068 7.263933    F-test  1.039131 0.341955090
#> 21  7.733928  1 0.50990446  0.5188728 7.215055    F-test 12.612240 0.009326588
#> 22 10.027100  2 0.37330087  2.7517217 7.275379    F-test 12.612240 0.009326588
#> 23  7.630949  3 0.52196965  0.4212224 7.209727    F-test 12.612240 0.009326588
#> 24  6.937435  4 0.30066740 -0.3700183 7.307454    F-test 12.612240 0.009326588
#> 25  1.091910  5 0.91743896 -5.9431788 7.035089    F-test 12.612240 0.009326588
#> 26  9.649033  6 0.55493540  2.4538631 7.195169    F-test 12.612240 0.009326588
#> 27  8.405919  7 0.28983766  1.0936825 7.312236    F-test 12.612240 0.009326588
#> 28 10.654973  8 0.74217074  3.5424860 7.112487    F-test 12.612240 0.009326588
#> 29  3.232279  9 0.03471863 -4.1926173 7.424896    F-test 12.612240 0.009326588
#> 30  6.987899 10 0.39921922 -0.2760340 7.263933    F-test 12.612240 0.009326588
#>    k  null
#> 1  3 FALSE
#> 2  3 FALSE
#> 3  3 FALSE
#> 4  3 FALSE
#> 5  3 FALSE
#> 6  3 FALSE
#> 7  3 FALSE
#> 8  3 FALSE
#> 9  3 FALSE
#> 10 3 FALSE
#> 11 1  TRUE
#> 12 1  TRUE
#> 13 1  TRUE
#> 14 1  TRUE
#> 15 1  TRUE
#> 16 1  TRUE
#> 17 1  TRUE
#> 18 1  TRUE
#> 19 1  TRUE
#> 20 1  TRUE
#> 21 2  TRUE
#> 22 2  TRUE
#> 23 2  TRUE
#> 24 2  TRUE
#> 25 2  TRUE
#> 26 2  TRUE
#> 27 2  TRUE
#> 28 2  TRUE
#> 29 2  TRUE
#> 30 2  TRUE