Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiate
x <- bandicoot::rand_uniform()
e <- bandicoot::rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>              y          x           e        .resid   .fitted test_name
#> 1   1.97831879 0.61652512 -0.01830956  0.0715446121 1.9067742    F-test
#> 2   3.75634876 0.79223268  1.33648346  1.3872547396 2.3690940    F-test
#> 3   0.03009393 0.30895030 -1.37430666 -1.0673926609 1.0974866    F-test
#> 4   2.04806017 0.78072131 -0.34218691 -0.2907452600 2.3388054    F-test
#> 5   2.20966300 0.94189034 -0.61938475 -0.5532086539 2.7628717    F-test
#> 6   2.53442604 0.48681937  0.81061358  0.9689322659 1.5654938    F-test
#> 7  -0.13357949 0.08634694 -1.22738223 -0.6453544937 0.5117750    F-test
#> 8   1.23274091 0.72129803 -1.00882797 -0.9497106153 2.1824515    F-test
#> 9   1.39641781 0.28505445  0.03010731  0.3618057220 1.0346121    F-test
#> 10  1.37071500 0.14033988  0.21067983  0.7168743442 0.6538407    F-test
#> 11  3.80008230 0.61652512 -0.01830956  1.8933081201 1.9067742    F-test
#> 12  2.82996208 0.79223268  1.33648346  0.4608680643 2.3690940    F-test
#> 13  0.91916037 0.30895030 -1.37430666 -0.1783262278 1.0974866    F-test
#> 14  2.59885437 0.78072131 -0.34218691  0.2600489472 2.3388054    F-test
#> 15  1.87419812 0.94189034 -0.61938475 -0.8886735286 2.7628717    F-test
#> 16  0.66030270 0.48681937  0.81061358 -0.9051910732 1.5654938    F-test
#> 17  1.04374793 0.08634694 -1.22738223  0.5319729193 0.5117750    F-test
#> 18  1.85730141 0.72129803 -1.00882797 -0.3251501096 2.1824515    F-test
#> 19  0.37886182 0.28505445  0.03010731 -0.6557502668 1.0346121    F-test
#> 20  0.46073381 0.14033988  0.21067983 -0.1931068449 0.6538407    F-test
#> 21  3.04478979 0.61652512 -0.01830956  1.1380156114 1.9067742    F-test
#> 22  1.38322259 0.79223268  1.33648346 -0.9858714230 2.3690940    F-test
#> 23 -0.12270455 0.30895030 -1.37430666 -1.2201911487 1.0974866    F-test
#> 24  1.52121117 0.78072131 -0.34218691 -0.8175942551 2.3388054    F-test
#> 25  2.85828890 0.94189034 -0.61938475  0.0954172537 2.7628717    F-test
#> 26  1.56640341 0.48681937  0.81061358  0.0009096339 1.5654938    F-test
#> 27  0.85318721 0.08634694 -1.22738223  0.3414122051 0.5117750    F-test
#> 28  3.37286879 0.72129803 -1.00882797  1.1904172657 2.1824515    F-test
#> 29  1.58838969 0.28505445  0.03010731  0.5537776089 1.0346121    F-test
#> 30  0.35754791 0.14033988  0.21067983 -0.2962927517 0.6538407    F-test
#>     statistic   p_value k  null
#> 1  0.28179742 0.6119446 1 FALSE
#> 2  0.28179742 0.6119446 1 FALSE
#> 3  0.28179742 0.6119446 1 FALSE
#> 4  0.28179742 0.6119446 1 FALSE
#> 5  0.28179742 0.6119446 1 FALSE
#> 6  0.28179742 0.6119446 1 FALSE
#> 7  0.28179742 0.6119446 1 FALSE
#> 8  0.28179742 0.6119446 1 FALSE
#> 9  0.28179742 0.6119446 1 FALSE
#> 10 0.28179742 0.6119446 1 FALSE
#> 11 0.16227172 0.6990925 2  TRUE
#> 12 0.16227172 0.6990925 2  TRUE
#> 13 0.16227172 0.6990925 2  TRUE
#> 14 0.16227172 0.6990925 2  TRUE
#> 15 0.16227172 0.6990925 2  TRUE
#> 16 0.16227172 0.6990925 2  TRUE
#> 17 0.16227172 0.6990925 2  TRUE
#> 18 0.16227172 0.6990925 2  TRUE
#> 19 0.16227172 0.6990925 2  TRUE
#> 20 0.16227172 0.6990925 2  TRUE
#> 21 0.08710795 0.7764483 3  TRUE
#> 22 0.08710795 0.7764483 3  TRUE
#> 23 0.08710795 0.7764483 3  TRUE
#> 24 0.08710795 0.7764483 3  TRUE
#> 25 0.08710795 0.7764483 3  TRUE
#> 26 0.08710795 0.7764483 3  TRUE
#> 27 0.08710795 0.7764483 3  TRUE
#> 28 0.08710795 0.7764483 3  TRUE
#> 29 0.08710795 0.7764483 3  TRUE
#> 30 0.08710795 0.7764483 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e         x     .resid  .fitted test_name   statistic   p_value k
#> 1   2.875610  1 0.5609478 -4.1166625 6.992273    F-test 0.759854067 0.4122713 1
#> 2   4.450793  2 0.8041444 -2.0368791 6.487672    F-test 0.759854067 0.4122713 1
#> 3   4.292265  3 0.2363865 -3.3734298 7.665695    F-test 0.759854067 0.4122713 1
#> 4   6.893648  4 0.9641201  0.7379045 6.155743    F-test 0.759854067 0.4122713 1
#> 5   6.376433  5 0.2914755 -1.1749592 7.551393    F-test 0.759854067 0.4122713 1
#> 6   7.123133  6 0.1108464 -0.8030409 7.926174    F-test 0.759854067 0.4122713 1
#> 7   8.449844  7 0.3365667  0.9920097 7.457834    F-test 0.759854067 0.4122713 1
#> 8  10.394094  8 0.7822225  3.8609377 6.533157    F-test 0.759854067 0.4122713 1
#> 9  10.224434  9 0.1887914  2.4599850 7.764449    F-test 0.759854067 0.4122713 1
#> 10 11.222101 10 0.1870960  3.4541347 7.767966    F-test 0.759854067 0.4122713 1
#> 11  4.443471  1 0.5609478 -2.5488017 6.992273    F-test 0.008563755 0.9288614 2
#> 12  9.699152  2 0.8041444  3.2114800 6.487672    F-test 0.008563755 0.9288614 2
#> 13  8.466203  3 0.2363865  0.8005081 7.665695    F-test 0.008563755 0.9288614 2
#> 14  4.500816  4 0.9641201 -1.6549271 6.155743    F-test 0.008563755 0.9288614 2
#> 15  3.555449  5 0.2914755 -3.9959436 7.551393    F-test 0.008563755 0.9288614 2
#> 16  6.741293  6 0.1108464 -1.1848815 7.926174    F-test 0.008563755 0.9288614 2
#> 17 10.305749  7 0.3365667  2.8479146 7.457834    F-test 0.008563755 0.9288614 2
#> 18  6.720272  8 0.7822225  0.1871154 6.533157    F-test 0.008563755 0.9288614 2
#> 19 12.133252  9 0.1887914  4.3688032 7.764449    F-test 0.008563755 0.9288614 2
#> 20  5.736699 10 0.1870960 -2.0312674 7.767966    F-test 0.008563755 0.9288614 2
#> 21  6.747079  1 0.5609478 -0.2451933 6.992273    F-test 0.339809041 0.5782268 3
#> 22 10.479868  2 0.8041444  3.9921965 6.487672    F-test 0.339809041 0.5782268 3
#> 23  9.737411  3 0.2363865  2.0717159 7.665695    F-test 0.339809041 0.5782268 3
#> 24  4.443528  4 0.9641201 -1.7122151 6.155743    F-test 0.339809041 0.5782268 3
#> 25  1.759998  5 0.2914755 -5.7913951 7.551393    F-test 0.339809041 0.5782268 3
#> 26 11.039306  6 0.1108464  3.1131322 7.926174    F-test 0.339809041 0.5782268 3
#> 27  6.205631  7 0.3365667 -1.2522032 7.457834    F-test 0.339809041 0.5782268 3
#> 28  6.339118  8 0.7822225 -0.1940390 6.533157    F-test 0.339809041 0.5782268 3
#> 29  8.429232  9 0.1887914  0.6647835 7.764449    F-test 0.339809041 0.5782268 3
#> 30  7.121184 10 0.1870960 -0.6467824 7.767966    F-test 0.339809041 0.5782268 3
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE