Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>              y         x          e       .resid   .fitted test_name statistic
#> 1   0.24395700 0.1052743 -0.8724000 -0.244907380 0.4888644    F-test 0.3840768
#> 2   1.04089166 0.1613575 -0.1465021  0.432938499 0.6079532    F-test 0.3840768
#> 3   3.10497376 0.4416872  1.4681990  1.901760021 1.2032137    F-test 0.3840768
#> 4   2.21502557 0.3401168  0.7592293  1.227489459 0.9875361    F-test 0.3840768
#> 5  -0.29643886 0.3866511 -1.8325891 -1.382787296 1.0863484    F-test 0.3840768
#> 6   1.41108569 0.4633520 -0.2669614  0.161868285 1.2492174    F-test 0.3840768
#> 7  -0.42331550 0.2671277 -1.7618005 -1.255864397 0.8325489    F-test 0.3840768
#> 8   0.39697286 0.3070679 -1.0043857 -0.520386133 0.9173590    F-test 0.3840768
#> 9   1.69485022 0.9010435 -1.0180727 -0.483774593 2.1786248    F-test 0.3840768
#> 10  1.58731053 0.5454972 -0.2557538  0.163663535 1.4236470    F-test 0.3840768
#> 11  0.96329094 0.1052743 -0.8724000  0.474426556 0.4888644    F-test 4.7413361
#> 12  1.72878485 0.1613575 -0.1465021  1.120831688 0.6079532    F-test 4.7413361
#> 13  1.50055654 0.4416872  1.4681990  0.297342801 1.2032137    F-test 4.7413361
#> 14  2.17641718 0.3401168  0.7592293  1.188881070 0.9875361    F-test 4.7413361
#> 15  0.25357361 0.3866511 -1.8325891 -0.832774826 1.0863484    F-test 4.7413361
#> 16  0.44393541 0.4633520 -0.2669614 -0.805282001 1.2492174    F-test 4.7413361
#> 17  0.84174127 0.2671277 -1.7618005  0.009192375 0.8325489    F-test 4.7413361
#> 18 -0.27892364 0.3070679 -1.0043857 -1.196282634 0.9173590    F-test 4.7413361
#> 19  3.39744301 0.9010435 -1.0180727  1.218818203 2.1786248    F-test 4.7413361
#> 20 -0.05150623 0.5454972 -0.2557538 -1.475153231 1.4236470    F-test 4.7413361
#> 21  0.35783398 0.1052743 -0.8724000 -0.131030398 0.4888644    F-test 0.3607653
#> 22  1.27224777 0.1613575 -0.1465021  0.664294606 0.6079532    F-test 0.3607653
#> 23  0.27428827 0.4416872  1.4681990 -0.928925468 1.2032137    F-test 0.3607653
#> 24  1.51842219 0.3401168  0.7592293  0.530886082 0.9875361    F-test 0.3607653
#> 25  1.46990682 0.3866511 -1.8325891  0.383558385 1.0863484    F-test 0.3607653
#> 26  1.07363687 0.4633520 -0.2669614 -0.175580539 1.2492174    F-test 0.3607653
#> 27 -0.39191598 0.2671277 -1.7618005 -1.224464877 0.8325489    F-test 0.3607653
#> 28  0.22587282 0.3070679 -1.0043857 -0.691486171 0.9173590    F-test 0.3607653
#> 29  1.47795845 0.9010435 -1.0180727 -0.700666359 2.1786248    F-test 0.3607653
#> 30  3.69706173 0.5454972 -0.2557538  2.273414738 1.4236470    F-test 0.3607653
#>       p_value k  null
#> 1  0.55505763 1 FALSE
#> 2  0.55505763 1 FALSE
#> 3  0.55505763 1 FALSE
#> 4  0.55505763 1 FALSE
#> 5  0.55505763 1 FALSE
#> 6  0.55505763 1 FALSE
#> 7  0.55505763 1 FALSE
#> 8  0.55505763 1 FALSE
#> 9  0.55505763 1 FALSE
#> 10 0.55505763 1 FALSE
#> 11 0.06588724 2  TRUE
#> 12 0.06588724 2  TRUE
#> 13 0.06588724 2  TRUE
#> 14 0.06588724 2  TRUE
#> 15 0.06588724 2  TRUE
#> 16 0.06588724 2  TRUE
#> 17 0.06588724 2  TRUE
#> 18 0.06588724 2  TRUE
#> 19 0.06588724 2  TRUE
#> 20 0.06588724 2  TRUE
#> 21 0.56701200 3  TRUE
#> 22 0.56701200 3  TRUE
#> 23 0.56701200 3  TRUE
#> 24 0.56701200 3  TRUE
#> 25 0.56701200 3  TRUE
#> 26 0.56701200 3  TRUE
#> 27 0.56701200 3  TRUE
#> 28 0.56701200 3  TRUE
#> 29 0.56701200 3  TRUE
#> 30 0.56701200 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e         x     .resid  .fitted test_name statistic   p_value k
#> 1   2.947627  1 0.5943615 -4.3931339 7.340761    F-test 1.0102255 0.3483208 1
#> 2   4.212338  2 0.7092717 -3.0501267 7.262465    F-test 1.0102255 0.3483208 1
#> 3   5.768935  3 0.9208923 -1.3493385 7.118273    F-test 1.0102255 0.3483208 1
#> 4   5.872901  4 0.5596704 -1.4914970 7.364398    F-test 1.0102255 0.3483208 1
#> 5   6.448145  5 0.3355505 -1.0689619 7.517107    F-test 1.0102255 0.3483208 1
#> 6   7.161862  6 0.1417651 -0.4872832 7.649146    F-test 1.0102255 0.3483208 1
#> 7   8.965720  7 0.6025971  1.6305710 7.335149    F-test 1.0102255 0.3483208 1
#> 8   9.395788  8 0.3036095  1.8569182 7.538870    F-test 1.0102255 0.3483208 1
#> 9  10.359361  9 0.2806157  2.8048236 7.554537    F-test 1.0102255 0.3483208 1
#> 10 12.686318 10 0.8915163  5.5480284 7.138289    F-test 1.0102255 0.3483208 1
#> 11  6.199824  1 0.5943615 -1.1409373 7.340761    F-test 0.1556715 0.7049184 2
#> 12  4.114263  2 0.7092717 -3.1482023 7.262465    F-test 0.1556715 0.7049184 2
#> 13  6.804363  3 0.9208923 -0.3139105 7.118273    F-test 0.1556715 0.7049184 2
#> 14 12.252244  4 0.5596704  4.8878458 7.364398    F-test 0.1556715 0.7049184 2
#> 15  6.183188  5 0.3355505 -1.3339185 7.517107    F-test 0.1556715 0.7049184 2
#> 16  9.803663  6 0.1417651  2.1545172 7.649146    F-test 0.1556715 0.7049184 2
#> 17 12.063336  7 0.6025971  4.7281865 7.335149    F-test 0.1556715 0.7049184 2
#> 18  4.130271  8 0.3036095 -3.4085989 7.538870    F-test 0.1556715 0.7049184 2
#> 19  5.992386  9 0.2806157 -1.5621514 7.554537    F-test 0.1556715 0.7049184 2
#> 20  6.275459 10 0.8915163 -0.8628305 7.138289    F-test 0.1556715 0.7049184 2
#> 21  8.607194  1 0.5943615  1.2664327 7.340761    F-test 0.1229378 0.7361843 3
#> 22  3.047140  2 0.7092717 -4.2153249 7.262465    F-test 0.1229378 0.7361843 3
#> 23 11.662920  3 0.9208923  4.5446462 7.118273    F-test 0.1229378 0.7361843 3
#> 24  7.054707  4 0.5596704 -0.3096915 7.364398    F-test 0.1229378 0.7361843 3
#> 25  3.876177  5 0.3355505 -3.6409291 7.517107    F-test 0.1229378 0.7361843 3
#> 26  6.870394  6 0.1417651 -0.7787514 7.649146    F-test 0.1229378 0.7361843 3
#> 27  8.358310  7 0.6025971  1.0231605 7.335149    F-test 0.1229378 0.7361843 3
#> 28 11.315746  8 0.3036095  3.7768763 7.538870    F-test 0.1229378 0.7361843 3
#> 29  8.733464  9 0.2806157  1.1789264 7.554537    F-test 0.1229378 0.7361843 3
#> 30  4.292944 10 0.8915163 -2.8453453 7.138289    F-test 0.1229378 0.7361843 3
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE