Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

A data frame.

Examples

# Instantiate
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>             y         x           e       .resid   .fitted test_name statistic
#> 1   0.3637564 0.3724283 -1.14737475 -0.784841514 1.1485979    F-test 9.7419137
#> 2   0.8200465 0.3716376 -0.68970563 -0.327002831 1.1470493    F-test 9.7419137
#> 3   0.8923093 0.2630896 -0.43999648 -0.042147645 0.9344570    F-test 9.7419137
#> 4   1.7036068 0.1010937  0.59229315  1.086420632 0.6171861    F-test 9.7419137
#> 5   2.5471000 0.8934546 -0.14461586  0.378066382 2.1690336    F-test 9.7419137
#> 6   0.7237058 0.5561219 -1.14168754 -0.784657876 1.5083637    F-test 9.7419137
#> 7   0.6650800 0.6855203 -1.49037842 -1.096711978 1.7617920    F-test 9.7419137
#> 8   2.6609875 0.9506816 -0.19348960  0.379874299 2.2811132    F-test 9.7419137
#> 9   1.5497648 0.3801743  0.02505797  0.385996241 1.1637685    F-test 9.7419137
#> 10  2.8240739 0.8168842  0.33988983  0.805004292 2.0190696    F-test 9.7419137
#> 11  1.5280174 0.3724283 -1.14737475  0.379419548 1.1485979    F-test 1.4002157
#> 12  1.2990292 0.3716376 -0.68970563  0.151979839 1.1470493    F-test 1.4002157
#> 13  1.3847893 0.2630896 -0.43999648  0.450332319 0.9344570    F-test 1.4002157
#> 14  1.1131654 0.1010937  0.59229315  0.495979265 0.6171861    F-test 1.4002157
#> 15  2.5724697 0.8934546 -0.14461586  0.403436092 2.1690336    F-test 1.4002157
#> 16  1.5070438 0.5561219 -1.14168754 -0.001319943 1.5083637    F-test 1.4002157
#> 17  0.7937875 0.6855203 -1.49037842 -0.968004526 1.7617920    F-test 1.4002157
#> 18  2.2020071 0.9506816 -0.19348960 -0.079106142 2.2811132    F-test 1.4002157
#> 19 -0.4228694 0.3801743  0.02505797 -1.586637911 1.1637685    F-test 1.4002157
#> 20  2.7729911 0.8168842  0.33988983  0.753921460 2.0190696    F-test 1.4002157
#> 21  2.0923349 0.3724283 -1.14737475  0.943736996 1.1485979    F-test 0.1190538
#> 22 -0.1506273 0.3716376 -0.68970563 -1.297676590 1.1470493    F-test 0.1190538
#> 23  1.0909279 0.2630896 -0.43999648  0.156470950 0.9344570    F-test 0.1190538
#> 24  0.7310361 0.1010937  0.59229315  0.113849992 0.6171861    F-test 0.1190538
#> 25  2.3556211 0.8934546 -0.14461586  0.186587496 2.1690336    F-test 0.1190538
#> 26  2.4293168 0.5561219 -1.14168754  0.920953067 1.5083637    F-test 0.1190538
#> 27  2.3343755 0.6855203 -1.49037842  0.572583470 1.7617920    F-test 0.1190538
#> 28  1.8884456 0.9506816 -0.19348960 -0.392667614 2.2811132    F-test 0.1190538
#> 29  0.3315767 0.3801743  0.02505797 -0.832191817 1.1637685    F-test 0.1190538
#> 30  1.6474237 0.8168842  0.33988983 -0.371645948 2.0190696    F-test 0.1190538
#>       p_value k  null
#> 1  0.01681535 2 FALSE
#> 2  0.01681535 2 FALSE
#> 3  0.01681535 2 FALSE
#> 4  0.01681535 2 FALSE
#> 5  0.01681535 2 FALSE
#> 6  0.01681535 2 FALSE
#> 7  0.01681535 2 FALSE
#> 8  0.01681535 2 FALSE
#> 9  0.01681535 2 FALSE
#> 10 0.01681535 2 FALSE
#> 11 0.27531170 1  TRUE
#> 12 0.27531170 1  TRUE
#> 13 0.27531170 1  TRUE
#> 14 0.27531170 1  TRUE
#> 15 0.27531170 1  TRUE
#> 16 0.27531170 1  TRUE
#> 17 0.27531170 1  TRUE
#> 18 0.27531170 1  TRUE
#> 19 0.27531170 1  TRUE
#> 20 0.27531170 1  TRUE
#> 21 0.74019842 3  TRUE
#> 22 0.74019842 3  TRUE
#> 23 0.74019842 3  TRUE
#> 24 0.74019842 3  TRUE
#> 25 0.74019842 3  TRUE
#> 26 0.74019842 3  TRUE
#> 27 0.74019842 3  TRUE
#> 28 0.74019842 3  TRUE
#> 29 0.74019842 3  TRUE
#> 30 0.74019842 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>             y  e          x     .resid  .fitted test_name statistic   p_value k
#> 1   2.0932520  1 0.08587713 -3.1102554 5.203507    F-test 0.0221496 0.8858866 1
#> 2   3.8809612  2 0.56346661 -4.3892794 8.270241    F-test 0.0221496 0.8858866 1
#> 3   4.1674147  3 0.14607637 -1.4226486 5.590063    F-test 0.0221496 0.8858866 1
#> 4   5.2063411  4 0.17553026 -0.5728536 5.779195    F-test 0.0221496 0.8858866 1
#> 5   6.4402310  5 0.33080143 -0.3360026 6.776234    F-test 0.0221496 0.8858866 1
#> 6   7.2000070  6 0.17082565  1.4510219 5.748985    F-test 0.0221496 0.8858866 1
#> 7   9.2861038  7 0.73939657 -0.1138314 9.399935    F-test 0.0221496 0.8858866 1
#> 8  10.3430505  8 0.76216105  0.7969383 9.546112    F-test 0.0221496 0.8858866 1
#> 9  11.0743737  9 0.65081436  2.2432491 8.831125    F-test 0.0221496 0.8858866 1
#> 10 11.1993102 10 0.17030604  5.4536616 5.745649    F-test 0.0221496 0.8858866 1
#> 11  3.0850940  1 0.08587713 -2.1184134 5.203507    F-test 0.1184508 0.7408283 2
#> 12 11.7805361  2 0.56346661  3.5102955 8.270241    F-test 0.1184508 0.7408283 2
#> 13  9.9257363  3 0.14607637  4.3356730 5.590063    F-test 0.1184508 0.7408283 2
#> 14  3.1686968  4 0.17553026 -2.6104979 5.779195    F-test 0.1184508 0.7408283 2
#> 15  2.6642122  5 0.33080143 -4.1120214 6.776234    F-test 0.1184508 0.7408283 2
#> 16  7.0951047  6 0.17082565  1.3461195 5.748985    F-test 0.1184508 0.7408283 2
#> 17  9.0236770  7 0.73939657 -0.3762582 9.399935    F-test 0.1184508 0.7408283 2
#> 18  7.3482140  8 0.76216105 -2.1978983 9.546112    F-test 0.1184508 0.7408283 2
#> 19 10.3585648  9 0.65081436  1.5274403 8.831125    F-test 0.1184508 0.7408283 2
#> 20  6.4412095 10 0.17030604  0.6955609 5.745649    F-test 0.1184508 0.7408283 2
#> 21  7.1388695  1 0.08587713  1.9353621 5.203507    F-test 1.4647587 0.2654512 3
#> 22  6.3239759  2 0.56346661 -1.9462648 8.270241    F-test 1.4647587 0.2654512 3
#> 23  9.4600582  3 0.14607637  3.8699949 5.590063    F-test 1.4647587 0.2654512 3
#> 24  5.2876384  4 0.17553026 -0.4915564 5.779195    F-test 1.4647587 0.2654512 3
#> 25  5.0134373  5 0.33080143 -1.7627963 6.776234    F-test 1.4647587 0.2654512 3
#> 26  7.2133353  6 0.17082565  1.4643502 5.748985    F-test 1.4647587 0.2654512 3
#> 27 12.3671247  7 0.73939657  2.9671895 9.399935    F-test 1.4647587 0.2654512 3
#> 28  8.3855658  8 0.76216105 -1.1605465 9.546112    F-test 1.4647587 0.2654512 3
#> 29  9.4666336  9 0.65081436  0.6355090 8.831125    F-test 1.4647587 0.2654512 3
#> 30  0.2344067 10 0.17030604 -5.5112419 5.745649    F-test 1.4647587 0.2654512 3
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE