Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>              y          x          e      .resid   .fitted test_name statistic
#> 1   2.40962949 0.97714551 -0.5223294 -0.19589197 2.6055215    F-test 2.7262043
#> 2   3.85363471 0.96876638  0.9463600  1.26322822 2.5904065    F-test 2.7262043
#> 3   1.99690746 0.05852332  0.9349592  1.04847137 0.9484361    F-test 2.7262043
#> 4   0.35263947 0.57654511 -1.5563099 -1.53024642 1.8828859    F-test 2.7262043
#> 5   1.91520685 0.79315624 -0.5070462 -0.35841980 2.2736266    F-test 2.7262043
#> 6   0.04240357 0.18377724 -1.1751477 -1.13197572 1.1743793    F-test 2.7262043
#> 7   1.30026257 0.32695651 -0.1335945 -0.13239513 1.4326577    F-test 2.7262043
#> 8   2.11392498 0.53231764  0.2982453  0.31082019 1.8031048    F-test 2.7262043
#> 9   2.59229557 0.95985072 -0.2888685  0.01797187 2.5743237    F-test 2.7262043
#> 10  1.91143604 0.19964267  0.6719362  0.70843739 1.2029987    F-test 2.7262043
#> 11  2.31221776 0.97714551 -0.5223294 -0.29330370 2.6055215    F-test 4.9334049
#> 12  3.77402982 0.96876638  0.9463600  1.18362333 2.5904065    F-test 4.9334049
#> 13  1.19009898 0.05852332  0.9349592  0.24166289 0.9484361    F-test 4.9334049
#> 14 -0.13046372 0.57654511 -1.5563099 -2.01334961 1.8828859    F-test 4.9334049
#> 15  2.18923667 0.79315624 -0.5070462 -0.08438998 2.2736266    F-test 4.9334049
#> 16  1.36053944 0.18377724 -1.1751477  0.18616015 1.1743793    F-test 4.9334049
#> 17  1.63639730 0.32695651 -0.1335945  0.20373960 1.4326577    F-test 4.9334049
#> 18  1.07772847 0.53231764  0.2982453 -0.72537632 1.8031048    F-test 4.9334049
#> 19  3.07293346 0.95985072 -0.2888685  0.49860976 2.5743237    F-test 4.9334049
#> 20  2.00562252 0.19964267  0.6719362  0.80262386 1.2029987    F-test 4.9334049
#> 21  3.48004039 0.97714551 -0.5223294  0.87451894 2.6055215    F-test 0.4583359
#> 22  1.39789833 0.96876638  0.9463600 -1.19250816 2.5904065    F-test 0.4583359
#> 23 -0.59392610 0.05852332  0.9349592 -1.54236219 0.9484361    F-test 0.4583359
#> 24  1.59865218 0.57654511 -1.5563099 -0.28423371 1.8828859    F-test 0.4583359
#> 25  2.19984215 0.79315624 -0.5070462 -0.07378450 2.2736266    F-test 0.4583359
#> 26  2.24767723 0.18377724 -1.1751477  1.07329794 1.1743793    F-test 0.4583359
#> 27  2.46121848 0.32695651 -0.1335945  1.02856079 1.4326577    F-test 0.4583359
#> 28  1.44632766 0.53231764  0.2982453 -0.35677712 1.8031048    F-test 0.4583359
#> 29  2.80488925 0.95985072 -0.2888685  0.23056555 2.5743237    F-test 0.4583359
#> 30  1.44572112 0.19964267  0.6719362  0.24272247 1.2029987    F-test 0.4583359
#>       p_value k  null
#> 1  0.14269715 3 FALSE
#> 2  0.14269715 3 FALSE
#> 3  0.14269715 3 FALSE
#> 4  0.14269715 3 FALSE
#> 5  0.14269715 3 FALSE
#> 6  0.14269715 3 FALSE
#> 7  0.14269715 3 FALSE
#> 8  0.14269715 3 FALSE
#> 9  0.14269715 3 FALSE
#> 10 0.14269715 3 FALSE
#> 11 0.06177444 1  TRUE
#> 12 0.06177444 1  TRUE
#> 13 0.06177444 1  TRUE
#> 14 0.06177444 1  TRUE
#> 15 0.06177444 1  TRUE
#> 16 0.06177444 1  TRUE
#> 17 0.06177444 1  TRUE
#> 18 0.06177444 1  TRUE
#> 19 0.06177444 1  TRUE
#> 20 0.06177444 1  TRUE
#> 21 0.52014081 2  TRUE
#> 22 0.52014081 2  TRUE
#> 23 0.52014081 2  TRUE
#> 24 0.52014081 2  TRUE
#> 25 0.52014081 2  TRUE
#> 26 0.52014081 2  TRUE
#> 27 0.52014081 2  TRUE
#> 28 0.52014081 2  TRUE
#> 29 0.52014081 2  TRUE
#> 30 0.52014081 2  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>             y  e         x      .resid  .fitted test_name    statistic
#> 1   2.8074348  1 0.5283165 -4.58677868 7.394213    F-test 8.112259e-05
#> 2   3.2632593  2 0.2164211 -2.19820228 5.461462    F-test 8.112259e-05
#> 3   5.1786990  3 0.6952820 -3.25016564 8.428865    F-test 8.112259e-05
#> 4   5.3983899  4 0.3052266 -0.61338099 6.011771    F-test 8.112259e-05
#> 5   6.2365024  5 0.1974973  0.89230835 5.344194    F-test 8.112259e-05
#> 6   8.5085428  6 0.8261006 -0.73097832 9.239521    F-test 8.112259e-05
#> 7   8.4673385  7 0.3469584  2.19696417 6.270374    F-test 8.112259e-05
#> 8  10.7725093  8 0.9221495  0.93779254 9.834717    F-test 8.112259e-05
#> 9  11.1662662  9 0.6900698  2.76970004 8.396566    F-test 8.112259e-05
#> 10 11.7256866 10 0.4877685  4.58274081 7.142946    F-test 8.112259e-05
#> 11  8.9724792  1 0.5283165  1.57826575 7.394213    F-test 1.068078e-02
#> 12  3.9868406  2 0.2164211 -1.47462093 5.461462    F-test 1.068078e-02
#> 13 12.5727275  3 0.6952820  4.14386291 8.428865    F-test 1.068078e-02
#> 14  8.5100091  4 0.3052266  2.49823821 6.011771    F-test 1.068078e-02
#> 15  5.5764887  5 0.1974973  0.23229467 5.344194    F-test 1.068078e-02
#> 16  4.2473669  6 0.8261006 -4.99215420 9.239521    F-test 1.068078e-02
#> 17  6.2198404  7 0.3469584 -0.05053399 6.270374    F-test 1.068078e-02
#> 18 12.5084994  8 0.9221495  2.67378271 9.834717    F-test 1.068078e-02
#> 19  7.0118702  9 0.6900698 -1.38469602 8.396566    F-test 1.068078e-02
#> 20  3.9185066 10 0.4877685 -3.22443911 7.142946    F-test 1.068078e-02
#> 21  4.8193743  1 0.5283165 -2.57483917 7.394213    F-test 1.770136e+00
#> 22  7.3264481  2 0.2164211  1.86498652 5.461462    F-test 1.770136e+00
#> 23  6.6154369  3 0.6952820 -1.81342773 8.428865    F-test 1.770136e+00
#> 24  0.6523003  4 0.3052266 -5.35947058 6.011771    F-test 1.770136e+00
#> 25  7.5208729  5 0.1974973  2.17667883 5.344194    F-test 1.770136e+00
#> 26 10.6674949  6 0.8261006  1.42797375 9.239521    F-test 1.770136e+00
#> 27  8.3117711  7 0.3469584  2.04139678 6.270374    F-test 1.770136e+00
#> 28 12.1116009  8 0.9221495  2.27688420 9.834717    F-test 1.770136e+00
#> 29  5.7053253  9 0.6900698 -2.69124089 8.396566    F-test 1.770136e+00
#> 30  9.7940040 10 0.4877685  2.65105828 7.142946    F-test 1.770136e+00
#>      p_value k  null
#> 1  0.9930650 1 FALSE
#> 2  0.9930650 1 FALSE
#> 3  0.9930650 1 FALSE
#> 4  0.9930650 1 FALSE
#> 5  0.9930650 1 FALSE
#> 6  0.9930650 1 FALSE
#> 7  0.9930650 1 FALSE
#> 8  0.9930650 1 FALSE
#> 9  0.9930650 1 FALSE
#> 10 0.9930650 1 FALSE
#> 11 0.9205854 2  TRUE
#> 12 0.9205854 2  TRUE
#> 13 0.9205854 2  TRUE
#> 14 0.9205854 2  TRUE
#> 15 0.9205854 2  TRUE
#> 16 0.9205854 2  TRUE
#> 17 0.9205854 2  TRUE
#> 18 0.9205854 2  TRUE
#> 19 0.9205854 2  TRUE
#> 20 0.9205854 2  TRUE
#> 21 0.2250704 3  TRUE
#> 22 0.2250704 3  TRUE
#> 23 0.2250704 3  TRUE
#> 24 0.2250704 3  TRUE
#> 25 0.2250704 3  TRUE
#> 26 0.2250704 3  TRUE
#> 27 0.2250704 3  TRUE
#> 28 0.2250704 3  TRUE
#> 29 0.2250704 3  TRUE
#> 30 0.2250704 3  TRUE