Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

A data frame.

Examples

# Instantiate
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>              y          x          e      .resid  .fitted test_name   statistic
#> 1   1.47956533 0.64061066 -0.5714274 -1.20171194 2.681277    F-test 0.225289698
#> 2   3.58086939 0.82607532  1.0723936  0.41312997 3.167739    F-test 0.225289698
#> 3   0.77290299 0.17272975 -0.4296623 -0.68115201 1.454055    F-test 0.225289698
#> 4   2.91735052 0.59646075  0.9651243  0.35187571 2.565475    F-test 0.225289698
#> 5   1.24606046 0.07347736  0.1671842  0.05233826 1.193722    F-test 0.225289698
#> 6   1.90288076 0.10371481  0.7884092  0.62984763 1.273033    F-test 0.225289698
#> 7   3.91001557 0.46648604  2.2259203  1.68545626 2.224559    F-test 0.225289698
#> 8   2.37908051 0.74908367  0.0688705 -0.58671463 2.965795    F-test 0.225289698
#> 9   1.69419985 0.07092772  0.6182414  0.50716519 1.187035    F-test 0.225289698
#> 10  0.08976939 0.09874736 -1.0187290 -1.17023443 1.260004    F-test 0.225289698
#> 11  4.03482425 0.64061066 -0.5714274  1.35354699 2.681277    F-test 0.448903579
#> 12  2.56037395 0.82607532  1.0723936 -0.60736548 3.167739    F-test 0.448903579
#> 13 -0.31337335 0.17272975 -0.4296623 -1.76742836 1.454055    F-test 0.448903579
#> 14  2.36885451 0.59646075  0.9651243 -0.19662030 2.565475    F-test 0.448903579
#> 15  0.65977760 0.07347736  0.1671842 -0.53394460 1.193722    F-test 0.448903579
#> 16  2.21509809 0.10371481  0.7884092  0.94206496 1.273033    F-test 0.448903579
#> 17  2.86628093 0.46648604  2.2259203  0.64172162 2.224559    F-test 0.448903579
#> 18  2.48686215 0.74908367  0.0688705 -0.47893300 2.965795    F-test 0.448903579
#> 19  1.39390979 0.07092772  0.6182414  0.20687514 1.187035    F-test 0.448903579
#> 20  1.70008685 0.09874736 -1.0187290  0.44008303 1.260004    F-test 0.448903579
#> 21  1.53063302 0.64061066 -0.5714274 -1.15064425 2.681277    F-test 0.008898187
#> 22  3.62864252 0.82607532  1.0723936  0.46090310 3.167739    F-test 0.008898187
#> 23  2.60941713 0.17272975 -0.4296623  1.15536212 1.454055    F-test 0.008898187
#> 24  2.46485763 0.59646075  0.9651243 -0.10061718 2.565475    F-test 0.008898187
#> 25  0.67197660 0.07347736  0.1671842 -0.52174559 1.193722    F-test 0.008898187
#> 26 -0.06359016 0.10371481  0.7884092 -1.33662329 1.273033    F-test 0.008898187
#> 27  3.04024977 0.46648604  2.2259203  0.81569046 2.224559    F-test 0.008898187
#> 28  2.93618856 0.74908367  0.0688705 -0.02960659 2.965795    F-test 0.008898187
#> 29  2.42304465 0.07092772  0.6182414  1.23601000 1.187035    F-test 0.008898187
#> 30  0.73127505 0.09874736 -1.0187290 -0.52872877 1.260004    F-test 0.008898187
#>      p_value k  null
#> 1  0.6494894 3 FALSE
#> 2  0.6494894 3 FALSE
#> 3  0.6494894 3 FALSE
#> 4  0.6494894 3 FALSE
#> 5  0.6494894 3 FALSE
#> 6  0.6494894 3 FALSE
#> 7  0.6494894 3 FALSE
#> 8  0.6494894 3 FALSE
#> 9  0.6494894 3 FALSE
#> 10 0.6494894 3 FALSE
#> 11 0.5243351 1  TRUE
#> 12 0.5243351 1  TRUE
#> 13 0.5243351 1  TRUE
#> 14 0.5243351 1  TRUE
#> 15 0.5243351 1  TRUE
#> 16 0.5243351 1  TRUE
#> 17 0.5243351 1  TRUE
#> 18 0.5243351 1  TRUE
#> 19 0.5243351 1  TRUE
#> 20 0.5243351 1  TRUE
#> 21 0.9274902 2  TRUE
#> 22 0.9274902 2  TRUE
#> 23 0.9274902 2  TRUE
#> 24 0.9274902 2  TRUE
#> 25 0.9274902 2  TRUE
#> 26 0.9274902 2  TRUE
#> 27 0.9274902 2  TRUE
#> 28 0.9274902 2  TRUE
#> 29 0.9274902 2  TRUE
#> 30 0.9274902 2  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e         x      .resid  .fitted test_name  statistic   p_value k
#> 1   3.902075  1 0.9669951 -4.43521509 8.337290    F-test 0.04694692 0.8346430 1
#> 2   4.242834  2 0.7218160 -3.53033122 7.773166    F-test 0.04694692 0.8346430 1
#> 3   4.753321  3 0.5016592 -2.51329320 7.266614    F-test 0.04694692 0.8346430 1
#> 4   5.569992  4 0.4055340 -1.47545136 7.045443    F-test 0.04694692 0.8346430 1
#> 5   6.987687  5 0.6125139 -0.53398895 7.521676    F-test 0.04694692 0.8346430 1
#> 6   8.202623  6 0.7052481  0.46757793 7.735045    F-test 0.04694692 0.8346430 1
#> 7   8.887704  7 0.5666320  1.47159578 7.416108    F-test 0.04694692 0.8346430 1
#> 8  10.112327  8 0.6671877  2.46485394 7.647473    F-test 0.04694692 0.8346430 1
#> 9  10.685197  9 0.4670560  3.49820027 7.186997    F-test 0.04694692 0.8346430 1
#> 10 12.996913 10 0.9989707  4.58605190 8.410861    F-test 0.04694692 0.8346430 1
#> 11 12.883282  1 0.9669951  4.54599221 8.337290    F-test 0.34213248 0.5769603 2
#> 12  7.835867  2 0.7218160  0.06270113 7.773166    F-test 0.34213248 0.5769603 2
#> 13  3.796603  3 0.5016592 -3.47001171 7.266614    F-test 0.34213248 0.5769603 2
#> 14  6.036791  4 0.4055340 -1.00865191 7.045443    F-test 0.34213248 0.5769603 2
#> 15  5.842556  5 0.6125139 -1.67912045 7.521676    F-test 0.34213248 0.5769603 2
#> 16  8.339117  6 0.7052481  0.60407213 7.735045    F-test 0.34213248 0.5769603 2
#> 17 11.882526  7 0.5666320  4.46641844 7.416108    F-test 0.34213248 0.5769603 2
#> 18  7.900209  8 0.6671877  0.25273547 7.647473    F-test 0.34213248 0.5769603 2
#> 19  8.349078  9 0.4670560  1.16208120 7.186997    F-test 0.34213248 0.5769603 2
#> 20  3.474645 10 0.9989707 -4.93621652 8.410861    F-test 0.34213248 0.5769603 2
#> 21  6.673168  1 0.9669951 -1.66412217 8.337290    F-test 1.42359981 0.2716785 3
#> 22 12.741255  2 0.7218160  4.96808954 7.773166    F-test 1.42359981 0.2716785 3
#> 23  2.016298  3 0.5016592 -5.25031644 7.266614    F-test 1.42359981 0.2716785 3
#> 24  8.372591  4 0.4055340  1.32714816 7.045443    F-test 1.42359981 0.2716785 3
#> 25  7.494745  5 0.6125139 -0.02693133 7.521676    F-test 1.42359981 0.2716785 3
#> 26  7.093615  6 0.7052481 -0.64142990 7.735045    F-test 1.42359981 0.2716785 3
#> 27  6.395986  7 0.5666320 -1.02012179 7.416108    F-test 1.42359981 0.2716785 3
#> 28 12.271490  8 0.6671877  4.62401703 7.647473    F-test 1.42359981 0.2716785 3
#> 29  6.445196  9 0.4670560 -0.74180137 7.186997    F-test 1.42359981 0.2716785 3
#> 30  6.836330 10 0.9989707 -1.57453174 8.410861    F-test 1.42359981 0.2716785 3
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE