Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>            y          x           e      .resid  .fitted test_name   statistic
#> 1  4.3274812 0.94409101  1.49208233  1.50564742 2.821834    F-test 1.192910236
#> 2  1.6710373 0.63210664 -0.36062813 -0.90462363 2.575661    F-test 1.192910236
#> 3  0.9468402 0.52870269 -0.86138907 -1.54722939 2.494070    F-test 1.192910236
#> 4  2.2879747 0.94827096 -0.55951410 -0.53715730 2.825132    F-test 1.192910236
#> 5  3.2861336 0.90619661  0.55874471  0.49420062 2.791933    F-test 1.192910236
#> 6  1.0952089 0.02235661  0.07235246 -0.99932577 2.094535    F-test 1.192910236
#> 7  1.5791826 0.44408265 -0.06210943 -0.84811706 2.427300    F-test 1.192910236
#> 8  3.2852178 0.26261111  1.95364214  1.00110918 2.284109    F-test 1.192910236
#> 9  1.8751899 0.27211320  0.52903111 -0.41641644 2.291606    F-test 1.192910236
#> 10 4.4266671 0.12402264  3.28726285  2.25191236 2.174755    F-test 1.192910236
#> 11 2.8721828 0.94409101  1.49208233  0.05034904 2.821834    F-test 0.006701991
#> 12 2.9445443 0.63210664 -0.36062813  0.36888335 2.575661    F-test 0.006701991
#> 13 2.1711496 0.52870269 -0.86138907 -0.32291996 2.494070    F-test 0.006701991
#> 14 1.7793608 0.94827096 -0.55951410 -1.04577113 2.825132    F-test 0.006701991
#> 15 4.1189626 0.90619661  0.55874471  1.32702959 2.791933    F-test 0.006701991
#> 16 0.5733345 0.02235661  0.07235246 -1.52120012 2.094535    F-test 0.006701991
#> 17 1.1844130 0.44408265 -0.06210943 -1.24288667 2.427300    F-test 0.006701991
#> 18 1.9093613 0.26261111  1.95364214 -0.37474738 2.284109    F-test 0.006701991
#> 19 2.4337665 0.27211320  0.52903111  0.14216021 2.291606    F-test 0.006701991
#> 20 4.7938578 0.12402264  3.28726285  2.61910307 2.174755    F-test 0.006701991
#> 21 3.0070133 0.94409101  1.49208233  0.18517954 2.821834    F-test 0.016096270
#> 22 4.3083370 0.63210664 -0.36062813  1.73267601 2.575661    F-test 0.016096270
#> 23 1.6770650 0.52870269 -0.86138907 -0.81700452 2.494070    F-test 0.016096270
#> 24 4.4998572 0.94827096 -0.55951410  1.67472528 2.825132    F-test 0.016096270
#> 25 0.2054885 0.90619661  0.55874471 -2.58644445 2.791933    F-test 0.016096270
#> 26 1.7242341 0.02235661  0.07235246 -0.37030053 2.094535    F-test 0.016096270
#> 27 2.0312735 0.44408265 -0.06210943 -0.39602622 2.427300    F-test 0.016096270
#> 28 2.8159999 0.26261111  1.95364214  0.53189126 2.284109    F-test 0.016096270
#> 29 1.9976767 0.27211320  0.52903111 -0.29392967 2.291606    F-test 0.016096270
#> 30 2.5139880 0.12402264  3.28726285  0.33923330 2.174755    F-test 0.016096270
#>      p_value k  null
#> 1  0.3109017 2 FALSE
#> 2  0.3109017 2 FALSE
#> 3  0.3109017 2 FALSE
#> 4  0.3109017 2 FALSE
#> 5  0.3109017 2 FALSE
#> 6  0.3109017 2 FALSE
#> 7  0.3109017 2 FALSE
#> 8  0.3109017 2 FALSE
#> 9  0.3109017 2 FALSE
#> 10 0.3109017 2 FALSE
#> 11 0.9370452 1  TRUE
#> 12 0.9370452 1  TRUE
#> 13 0.9370452 1  TRUE
#> 14 0.9370452 1  TRUE
#> 15 0.9370452 1  TRUE
#> 16 0.9370452 1  TRUE
#> 17 0.9370452 1  TRUE
#> 18 0.9370452 1  TRUE
#> 19 0.9370452 1  TRUE
#> 20 0.9370452 1  TRUE
#> 21 0.9026099 3  TRUE
#> 22 0.9026099 3  TRUE
#> 23 0.9026099 3  TRUE
#> 24 0.9026099 3  TRUE
#> 25 0.9026099 3  TRUE
#> 26 0.9026099 3  TRUE
#> 27 0.9026099 3  TRUE
#> 28 0.9026099 3  TRUE
#> 29 0.9026099 3  TRUE
#> 30 0.9026099 3  TRUE