Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>              y         x          e      .resid  .fitted test_name  statistic
#> 1   2.59111315 0.5378264  0.7640296  0.79764161 1.793472    F-test 0.51275387
#> 2   1.62104489 0.2056050  0.3731665  0.04124856 1.579796    F-test 0.51275387
#> 3   3.65103215 0.5431796  1.8128085  1.85411757 1.796915    F-test 0.51275387
#> 4   2.58327736 0.9898247 -0.3863004  0.49909352 2.084184    F-test 0.51275387
#> 5   0.18091265 0.6143559 -1.8108764 -1.66178045 1.842693    F-test 0.51275387
#> 6   0.64714678 0.6842119 -1.5052111 -1.24047572 1.887623    F-test 0.51275387
#> 7   3.56090082 0.4675610  1.8747265  1.81262193 1.748279    F-test 0.51275387
#> 8   0.70397807 0.8999241 -2.0058095 -1.32238431 2.026362    F-test 0.51275387
#> 9   2.55713865 0.8714754 -0.0738060  0.54907372 2.008065    F-test 0.51275387
#> 10  0.21755292 0.1541615 -0.9603743 -1.32915643 1.546709    F-test 0.51275387
#> 11  0.85899732 0.5378264  0.7640296 -0.93447422 1.793472    F-test 0.01903595
#> 12 -0.04502568 0.2056050  0.3731665 -1.62482201 1.579796    F-test 0.01903595
#> 13  4.14530216 0.5431796  1.8128085  2.34838758 1.796915    F-test 0.01903595
#> 14  2.04876534 0.9898247 -0.3863004 -0.03541851 2.084184    F-test 0.01903595
#> 15  2.64626914 0.6143559 -1.8108764  0.80357603 1.842693    F-test 0.01903595
#> 16  0.47151538 0.6842119 -1.5052111 -1.41610712 1.887623    F-test 0.01903595
#> 17  0.88174086 0.4675610  1.8747265 -0.86653803 1.748279    F-test 0.01903595
#> 18  2.90129038 0.8999241 -2.0058095  0.87492800 2.026362    F-test 0.01903595
#> 19  1.39192617 0.8714754 -0.0738060 -0.61613877 2.008065    F-test 0.01903595
#> 20  3.01331638 0.1541615 -0.9603743  1.46660703 1.546709    F-test 0.01903595
#> 21  1.63631958 0.5378264  0.7640296 -0.15715196 1.793472    F-test 0.01251208
#> 22  4.19374090 0.2056050  0.3731665  2.61394457 1.579796    F-test 0.01251208
#> 23  0.80585327 0.5431796  1.8128085 -0.99106131 1.796915    F-test 0.01251208
#> 24  1.80909637 0.9898247 -0.3863004 -0.27508748 2.084184    F-test 0.01251208
#> 25  1.39134355 0.6143559 -1.8108764 -0.45134955 1.842693    F-test 0.01251208
#> 26  2.23972498 0.6842119 -1.5052111  0.35210248 1.887623    F-test 0.01251208
#> 27  2.59251071 0.4675610  1.8747265  0.84423182 1.748279    F-test 0.01251208
#> 28  3.03934101 0.8999241 -2.0058095  1.01297864 2.026362    F-test 0.01251208
#> 29  1.36998647 0.8714754 -0.0738060 -0.63807847 2.008065    F-test 0.01251208
#> 30 -0.76381940 0.1541615 -0.9603743 -2.31052874 1.546709    F-test 0.01251208
#>      p_value k  null
#> 1  0.4971356 3 FALSE
#> 2  0.4971356 3 FALSE
#> 3  0.4971356 3 FALSE
#> 4  0.4971356 3 FALSE
#> 5  0.4971356 3 FALSE
#> 6  0.4971356 3 FALSE
#> 7  0.4971356 3 FALSE
#> 8  0.4971356 3 FALSE
#> 9  0.4971356 3 FALSE
#> 10 0.4971356 3 FALSE
#> 11 0.8941484 1  TRUE
#> 12 0.8941484 1  TRUE
#> 13 0.8941484 1  TRUE
#> 14 0.8941484 1  TRUE
#> 15 0.8941484 1  TRUE
#> 16 0.8941484 1  TRUE
#> 17 0.8941484 1  TRUE
#> 18 0.8941484 1  TRUE
#> 19 0.8941484 1  TRUE
#> 20 0.8941484 1  TRUE
#> 21 0.9140764 2  TRUE
#> 22 0.9140764 2  TRUE
#> 23 0.9140764 2  TRUE
#> 24 0.9140764 2  TRUE
#> 25 0.9140764 2  TRUE
#> 26 0.9140764 2  TRUE
#> 27 0.9140764 2  TRUE
#> 28 0.9140764 2  TRUE
#> 29 0.9140764 2  TRUE
#> 30 0.9140764 2  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e          x       .resid  .fitted test_name  statistic   p_value
#> 1   3.752633  1 0.91514415 -2.200164498 5.952797    F-test 1.30711025 0.2905074
#> 2   4.051139  2 0.64067484 -2.755033644 6.806173    F-test 1.30711025 0.2905074
#> 3   5.291676  3 0.74164233 -1.200570694 6.492246    F-test 1.30711025 0.2905074
#> 4   5.595948  4 0.41975418 -1.897107783 7.493056    F-test 1.30711025 0.2905074
#> 5   6.271570  5 0.22219790 -1.835724350 8.107294    F-test 1.30711025 0.2905074
#> 6   7.105066  6 0.09587421 -1.394991546 8.500058    F-test 1.30711025 0.2905074
#> 7   9.250878  7 0.72510333  2.707209053 6.543669    F-test 1.30711025 0.2905074
#> 8   9.100276  8 0.09184088  0.587677646 8.512598    F-test 1.30711025 0.2905074
#> 9  10.719404  9 0.48458312  3.427913422 7.291490    F-test 1.30711025 0.2905074
#> 10 11.771344 10 0.51061545  4.560792394 7.210551    F-test 1.30711025 0.2905074
#> 11  6.512431  1 0.91514415  0.559633434 5.952797    F-test 0.01270391 0.9134234
#> 12  6.807683  2 0.64067484  0.001510413 6.806173    F-test 0.01270391 0.9134234
#> 13  9.371701  3 0.74164233  2.879454215 6.492246    F-test 0.01270391 0.9134234
#> 14 10.214974  4 0.41975418  2.721918536 7.493056    F-test 0.01270391 0.9134234
#> 15  6.750375  5 0.22219790 -1.356918726 8.107294    F-test 0.01270391 0.9134234
#> 16  6.586974  6 0.09587421 -1.913083199 8.500058    F-test 0.01270391 0.9134234
#> 17  1.495407  7 0.72510333 -5.048262456 6.543669    F-test 0.01270391 0.9134234
#> 18 10.442091  8 0.09184088  1.929492773 8.512598    F-test 0.01270391 0.9134234
#> 19  4.928222  9 0.48458312 -2.363268509 7.291490    F-test 0.01270391 0.9134234
#> 20  9.800075 10 0.51061545  2.589523518 7.210551    F-test 0.01270391 0.9134234
#> 21  7.696680  1 0.91514415  1.743882465 5.952797    F-test 0.30053317 0.6005860
#> 22  7.163016  2 0.64067484  0.356842970 6.806173    F-test 0.30053317 0.6005860
#> 23  2.805451  3 0.74164233 -3.686795480 6.492246    F-test 0.30053317 0.6005860
#> 24  6.406117  4 0.41975418 -1.086938427 7.493056    F-test 0.30053317 0.6005860
#> 25  7.612963  5 0.22219790 -0.494331028 8.107294    F-test 0.30053317 0.6005860
#> 26 13.637184  6 0.09587421  5.137126585 8.500058    F-test 0.30053317 0.6005860
#> 27  8.248045  7 0.72510333  1.704375675 6.543669    F-test 0.30053317 0.6005860
#> 28  4.668087  8 0.09184088 -3.844510743 8.512598    F-test 0.30053317 0.6005860
#> 29  6.769824  9 0.48458312 -0.521666495 7.291490    F-test 0.30053317 0.6005860
#> 30  7.902566 10 0.51061545  0.692014477 7.210551    F-test 0.30053317 0.6005860
#>    k  null
#> 1  2 FALSE
#> 2  2 FALSE
#> 3  2 FALSE
#> 4  2 FALSE
#> 5  2 FALSE
#> 6  2 FALSE
#> 7  2 FALSE
#> 8  2 FALSE
#> 9  2 FALSE
#> 10 2 FALSE
#> 11 1  TRUE
#> 12 1  TRUE
#> 13 1  TRUE
#> 14 1  TRUE
#> 15 1  TRUE
#> 16 1  TRUE
#> 17 1  TRUE
#> 18 1  TRUE
#> 19 1  TRUE
#> 20 1  TRUE
#> 21 3  TRUE
#> 22 3  TRUE
#> 23 3  TRUE
#> 24 3  TRUE
#> 25 3  TRUE
#> 26 3  TRUE
#> 27 3  TRUE
#> 28 3  TRUE
#> 29 3  TRUE
#> 30 3  TRUE