Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>            y          x          e      .resid  .fitted test_name statistic
#> 1  0.8905435 0.36812052 -0.6130897 -0.94387466 1.834418    F-test 4.7006617
#> 2  2.9452846 0.80027088  0.5045803  0.86012356 2.085161    F-test 4.7006617
#> 3  1.9405845 0.64786929 -0.1270194 -0.05614986 1.996734    F-test 4.7006617
#> 4  1.0548852 0.64990301 -1.0173918 -0.94302923 1.997914    F-test 4.7006617
#> 5  3.0307290 0.09613561  1.9253513  1.35412229 1.676607    F-test 4.7006617
#> 6  2.1924278 0.44363087  0.5519886  0.31419692 1.878231    F-test 4.7006617
#> 7  2.1111577 0.12757307  0.9673097  0.41631029 1.694847    F-test 4.7006617
#> 8  0.2593444 0.34882572 -1.2111607 -1.56387850 1.823223    F-test 4.7006617
#> 9  2.1573871 0.93059301 -0.6392093 -0.00338969 2.160777    F-test 4.7006617
#> 10 2.7346219 0.94485708 -0.1029900  0.56556888 2.169053    F-test 4.7006617
#> 11 2.2092906 0.36812052 -0.6130897  0.37487241 1.834418    F-test 0.1964657
#> 12 3.1114899 0.80027088  0.5045803  1.02632885 2.085161    F-test 0.1964657
#> 13 1.4143033 0.64786929 -0.1270194 -0.58243115 1.996734    F-test 0.1964657
#> 14 0.1484892 0.64990301 -1.0173918 -1.84942520 1.997914    F-test 0.1964657
#> 15 2.0976324 0.09613561  1.9253513  0.42102572 1.676607    F-test 0.1964657
#> 16 3.1700772 0.44363087  0.5519886  1.29184632 1.878231    F-test 0.1964657
#> 17 1.3172662 0.12757307  0.9673097 -0.37758119 1.694847    F-test 0.1964657
#> 18 1.2408549 0.34882572 -1.2111607 -0.58236803 1.823223    F-test 0.1964657
#> 19 2.4927083 0.93059301 -0.6392093  0.33193153 2.160777    F-test 0.1964657
#> 20 2.1148538 0.94485708 -0.1029900 -0.05419925 2.169053    F-test 0.1964657
#> 21 1.4476136 0.36812052 -0.6130897 -0.38680459 1.834418    F-test 1.4480471
#> 22 1.2720421 0.80027088  0.5045803 -0.81311894 2.085161    F-test 1.4480471
#> 23 3.6585682 0.64786929 -0.1270194  1.66183380 1.996734    F-test 1.4480471
#> 24 1.0285442 0.64990301 -1.0173918 -0.96937016 1.997914    F-test 1.4480471
#> 25 2.1861073 0.09613561  1.9253513  0.50950062 1.676607    F-test 1.4480471
#> 26 1.1260868 0.44363087  0.5519886 -0.75214411 1.878231    F-test 1.4480471
#> 27 2.3817631 0.12757307  0.9673097  0.68691570 1.694847    F-test 1.4480471
#> 28 0.9889635 0.34882572 -1.2111607 -0.83425944 1.823223    F-test 1.4480471
#> 29 3.1026717 0.93059301 -0.6392093  0.94189491 2.160777    F-test 1.4480471
#> 30 2.1246053 0.94485708 -0.1029900 -0.04444778 2.169053    F-test 1.4480471
#>       p_value k  null
#> 1  0.06680408 3 FALSE
#> 2  0.06680408 3 FALSE
#> 3  0.06680408 3 FALSE
#> 4  0.06680408 3 FALSE
#> 5  0.06680408 3 FALSE
#> 6  0.06680408 3 FALSE
#> 7  0.06680408 3 FALSE
#> 8  0.06680408 3 FALSE
#> 9  0.06680408 3 FALSE
#> 10 0.06680408 3 FALSE
#> 11 0.67096419 1  TRUE
#> 12 0.67096419 1  TRUE
#> 13 0.67096419 1  TRUE
#> 14 0.67096419 1  TRUE
#> 15 0.67096419 1  TRUE
#> 16 0.67096419 1  TRUE
#> 17 0.67096419 1  TRUE
#> 18 0.67096419 1  TRUE
#> 19 0.67096419 1  TRUE
#> 20 0.67096419 1  TRUE
#> 21 0.26795435 2  TRUE
#> 22 0.26795435 2  TRUE
#> 23 0.26795435 2  TRUE
#> 24 0.26795435 2  TRUE
#> 25 0.26795435 2  TRUE
#> 26 0.26795435 2  TRUE
#> 27 0.26795435 2  TRUE
#> 28 0.26795435 2  TRUE
#> 29 0.26795435 2  TRUE
#> 30 0.26795435 2  TRUE