Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>              y          x          e       .resid   .fitted test_name
#> 1   3.13287784 0.83214365  0.6082711  0.535673675 2.5972042    F-test
#> 2   1.71331973 0.56300546 -0.1666609 -0.248754071 1.9620738    F-test
#> 3   2.63734072 0.64183883  0.5835448  0.489230646 2.1481101    F-test
#> 4  -0.01198322 0.03986533 -1.0534378 -0.739515824 0.7275326    F-test
#> 5   2.43027984 0.44670575  0.7840281  0.742657909 1.6876219    F-test
#> 6   1.57801575 0.17041308  0.3785621  0.542407817 1.0356079    F-test
#> 7   0.67167617 0.61800344 -1.3282555 -1.420185553 2.0918617    F-test
#> 8   2.11579621 0.85073136 -0.4586790 -0.525272473 2.6410687    F-test
#> 9   2.10737456 0.46588634  0.4244381  0.374488984 1.7328856    F-test
#> 10  2.16703268 0.54422898  0.3266185  0.249268889 1.9177638    F-test
#> 11  2.87383837 0.83214365  0.6082711  0.276634202 2.5972042    F-test
#> 12  2.45482862 0.56300546 -0.1666609  0.492754816 1.9620738    F-test
#> 13  1.49062540 0.64183883  0.5835448 -0.657484668 2.1481101    F-test
#> 14  0.93601201 0.03986533 -1.0534378  0.208479409 0.7275326    F-test
#> 15  1.01373638 0.44670575  0.7840281 -0.673885551 1.6876219    F-test
#> 16  1.92589621 0.17041308  0.3785621  0.890288273 1.0356079    F-test
#> 17  1.58148075 0.61800344 -1.3282555 -0.510380981 2.0918617    F-test
#> 18  3.61622155 0.85073136 -0.4586790  0.975152869 2.6410687    F-test
#> 19  0.63159520 0.46588634  0.4244381 -1.101290375 1.7328856    F-test
#> 20  2.01749579 0.54422898  0.3266185  0.099732006 1.9177638    F-test
#> 21  2.66184461 0.83214365  0.6082711  0.064640438 2.5972042    F-test
#> 22  1.55571042 0.56300546 -0.1666609 -0.406363389 1.9620738    F-test
#> 23  3.55648869 0.64183883  0.5835448  1.408378621 2.1481101    F-test
#> 24  0.10497730 0.03986533 -1.0534378 -0.622555302 0.7275326    F-test
#> 25  1.20874850 0.44670575  0.7840281 -0.478873428 1.6876219    F-test
#> 26  1.98623833 0.17041308  0.3785621  0.950630395 1.0356079    F-test
#> 27  1.27892962 0.61800344 -1.3282555 -0.812932105 2.0918617    F-test
#> 28  2.33249136 0.85073136 -0.4586790 -0.308577325 2.6410687    F-test
#> 29  1.72310743 0.46588634  0.4244381 -0.009778146 1.7328856    F-test
#> 30  2.13319403 0.54422898  0.3266185  0.215430240 1.9177638    F-test
#>     statistic   p_value k  null
#> 1  0.47980005 0.5108316 1 FALSE
#> 2  0.47980005 0.5108316 1 FALSE
#> 3  0.47980005 0.5108316 1 FALSE
#> 4  0.47980005 0.5108316 1 FALSE
#> 5  0.47980005 0.5108316 1 FALSE
#> 6  0.47980005 0.5108316 1 FALSE
#> 7  0.47980005 0.5108316 1 FALSE
#> 8  0.47980005 0.5108316 1 FALSE
#> 9  0.47980005 0.5108316 1 FALSE
#> 10 0.47980005 0.5108316 1 FALSE
#> 11 4.27153198 0.0775860 2  TRUE
#> 12 4.27153198 0.0775860 2  TRUE
#> 13 4.27153198 0.0775860 2  TRUE
#> 14 4.27153198 0.0775860 2  TRUE
#> 15 4.27153198 0.0775860 2  TRUE
#> 16 4.27153198 0.0775860 2  TRUE
#> 17 4.27153198 0.0775860 2  TRUE
#> 18 4.27153198 0.0775860 2  TRUE
#> 19 4.27153198 0.0775860 2  TRUE
#> 20 4.27153198 0.0775860 2  TRUE
#> 21 0.07424169 0.7931206 3  TRUE
#> 22 0.07424169 0.7931206 3  TRUE
#> 23 0.07424169 0.7931206 3  TRUE
#> 24 0.07424169 0.7931206 3  TRUE
#> 25 0.07424169 0.7931206 3  TRUE
#> 26 0.07424169 0.7931206 3  TRUE
#> 27 0.07424169 0.7931206 3  TRUE
#> 28 0.07424169 0.7931206 3  TRUE
#> 29 0.07424169 0.7931206 3  TRUE
#> 30 0.07424169 0.7931206 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e          x     .resid  .fitted test_name statistic    p_value k
#> 1   3.701032  1 0.89679358 -3.2736005 6.974633    F-test 0.4112403 0.54175971 2
#> 2   3.072875  2 0.06822071 -4.5518150 7.624690    F-test 0.4112403 0.54175971 2
#> 3   5.267607  3 0.73191182 -1.8363841 7.103991    F-test 0.4112403 0.54175971 2
#> 4   5.912657  4 0.57826570 -1.3118770 7.224534    F-test 0.4112403 0.54175971 2
#> 5   6.306728  5 0.24614211 -1.1783735 7.485102    F-test 0.4112403 0.54175971 2
#> 6   7.627184  6 0.43658120  0.2914918 7.335693    F-test 0.4112403 0.54175971 2
#> 7   9.256291  7 0.72731056  2.1486905 7.107601    F-test 0.4112403 0.54175971 2
#> 8  10.459003  8 0.80728840  3.4141489 7.044854    F-test 0.4112403 0.54175971 2
#> 9  10.042515  9 0.04084627  2.3963482 7.646166    F-test 0.4112403 0.54175971 2
#> 10 11.359404 10 0.28064347  3.9013707 7.458033    F-test 0.4112403 0.54175971 2
#> 11  3.009410  1 0.89679358 -3.9652228 6.974633    F-test 5.1509770 0.05751340 1
#> 12  8.226765  2 0.06822071  0.6020755 7.624690    F-test 5.1509770 0.05751340 1
#> 13 11.509874  3 0.73191182  4.4058836 7.103991    F-test 5.1509770 0.05751340 1
#> 14  9.274249  4 0.57826570  2.0497150 7.224534    F-test 5.1509770 0.05751340 1
#> 15 10.076372  5 0.24614211  2.5912702 7.485102    F-test 5.1509770 0.05751340 1
#> 16  7.293283  6 0.43658120 -0.0424100 7.335693    F-test 5.1509770 0.05751340 1
#> 17  8.804051  7 0.72731056  1.6964502 7.107601    F-test 5.1509770 0.05751340 1
#> 18  3.910549  8 0.80728840 -3.1343053 7.044854    F-test 5.1509770 0.05751340 1
#> 19  3.635596  9 0.04084627 -4.0105704 7.646166    F-test 5.1509770 0.05751340 1
#> 20  7.265147 10 0.28064347 -0.1928860 7.458033    F-test 5.1509770 0.05751340 1
#> 21  5.897286  1 0.89679358 -1.0773469 6.974633    F-test 6.0359080 0.04366915 3
#> 22  2.983569  2 0.06822071 -4.6411211 7.624690    F-test 6.0359080 0.04366915 3
#> 23  6.507829  3 0.73191182 -0.5961623 7.103991    F-test 6.0359080 0.04366915 3
#> 24  9.426858  4 0.57826570  2.2023245 7.224534    F-test 6.0359080 0.04366915 3
#> 25 11.762828  5 0.24614211  4.2777262 7.485102    F-test 6.0359080 0.04366915 3
#> 26 11.545476  6 0.43658120  4.2097838 7.335693    F-test 6.0359080 0.04366915 3
#> 27  7.613226  7 0.72731056  0.5056248 7.107601    F-test 6.0359080 0.04366915 3
#> 28  4.020493  8 0.80728840 -3.0243611 7.044854    F-test 6.0359080 0.04366915 3
#> 29  7.022565  9 0.04084627 -0.6236011 7.646166    F-test 6.0359080 0.04366915 3
#> 30  6.225167 10 0.28064347 -1.2328668 7.458033    F-test 6.0359080 0.04366915 3
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE