Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>            y          x           e      .resid  .fitted test_name statistic
#> 1  3.0568316 0.63202117  1.02535971  0.85546467 2.201367    F-test 0.8398592
#> 2  0.2182737 0.24957461 -1.09358842 -1.30735070 1.525624    F-test 0.8398592
#> 3  2.2462034 0.07737075  1.16284640  1.02484499 1.221358    F-test 0.8398592
#> 4  3.3261890 0.62580464  1.30875292  1.13580599 2.190383    F-test 0.8398592
#> 5  2.6478373 0.71559794  0.42015897  0.29879904 2.349038    F-test 0.8398592
#> 6  2.0573260 0.65787849 -0.03335662 -0.18972814 2.247054    F-test 0.8398592
#> 7  1.8385916 0.91772941 -0.92136512 -0.86759163 2.706183    F-test 0.8398592
#> 8  2.5678696 0.35592553  1.08526110  0.85433442 1.713535    F-test 0.8398592
#> 9  0.2364076 0.06348717 -0.83111017 -0.96041995 1.196828    F-test 0.8398592
#> 10 1.3715300 0.64012678 -0.67835906 -0.84415870 2.215689    F-test 0.8398592
#> 11 3.6978544 0.63202117  1.02535971  1.49648744 2.201367    F-test 0.1171632
#> 12 0.7292127 0.24957461 -1.09358842 -0.79641163 1.525624    F-test 0.1171632
#> 13 0.3744286 0.07737075  1.16284640 -0.84692977 1.221358    F-test 0.1171632
#> 14 1.9508149 0.62580464  1.30875292 -0.23956813 2.190383    F-test 0.1171632
#> 15 1.4085605 0.71559794  0.42015897 -0.94047782 2.349038    F-test 0.1171632
#> 16 2.6126996 0.65787849 -0.03335662  0.36564552 2.247054    F-test 0.1171632
#> 17 3.1452972 0.91772941 -0.92136512  0.43911401 2.706183    F-test 0.1171632
#> 18 2.2754843 0.35592553  1.08526110  0.56194912 1.713535    F-test 0.1171632
#> 19 2.3702828 0.06348717 -0.83111017  1.17345521 1.196828    F-test 0.1171632
#> 20 1.0024248 0.64012678 -0.67835906 -1.21326396 2.215689    F-test 0.1171632
#> 21 2.6332271 0.63202117  1.02535971  0.43186009 2.201367    F-test 1.3686824
#> 22 2.1379517 0.24957461 -1.09358842  0.61232737 1.525624    F-test 1.3686824
#> 23 0.6054621 0.07737075  1.16284640 -0.61589633 1.221358    F-test 1.3686824
#> 24 2.2195766 0.62580464  1.30875292  0.02919359 2.190383    F-test 1.3686824
#> 25 1.1108822 0.71559794  0.42015897 -1.23815611 2.349038    F-test 1.3686824
#> 26 4.3769249 0.65787849 -0.03335662  2.12987080 2.247054    F-test 1.3686824
#> 27 2.1444661 0.91772941 -0.92136512 -0.56171706 2.706183    F-test 1.3686824
#> 28 1.9701959 0.35592553  1.08526110  0.25666070 1.713535    F-test 1.3686824
#> 29 0.8829770 0.06348717 -0.83111017 -0.31385057 1.196828    F-test 1.3686824
#> 30 1.4853962 0.64012678 -0.67835906 -0.73029248 2.215689    F-test 1.3686824
#>      p_value k  null
#> 1  0.3899172 3 FALSE
#> 2  0.3899172 3 FALSE
#> 3  0.3899172 3 FALSE
#> 4  0.3899172 3 FALSE
#> 5  0.3899172 3 FALSE
#> 6  0.3899172 3 FALSE
#> 7  0.3899172 3 FALSE
#> 8  0.3899172 3 FALSE
#> 9  0.3899172 3 FALSE
#> 10 0.3899172 3 FALSE
#> 11 0.7421793 1  TRUE
#> 12 0.7421793 1  TRUE
#> 13 0.7421793 1  TRUE
#> 14 0.7421793 1  TRUE
#> 15 0.7421793 1  TRUE
#> 16 0.7421793 1  TRUE
#> 17 0.7421793 1  TRUE
#> 18 0.7421793 1  TRUE
#> 19 0.7421793 1  TRUE
#> 20 0.7421793 1  TRUE
#> 21 0.2803247 2  TRUE
#> 22 0.2803247 2  TRUE
#> 23 0.2803247 2  TRUE
#> 24 0.2803247 2  TRUE
#> 25 0.2803247 2  TRUE
#> 26 0.2803247 2  TRUE
#> 27 0.2803247 2  TRUE
#> 28 0.2803247 2  TRUE
#> 29 0.2803247 2  TRUE
#> 30 0.2803247 2  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e          x      .resid  .fitted test_name statistic   p_value k
#> 1   3.677375  1 0.88829936 -1.88032116 5.557696    F-test 0.1621877 0.6991657 3
#> 2   3.543563  2 0.39082137 -4.54269202 8.086255    F-test 0.1621877 0.6991657 3
#> 3   4.922834  3 0.58297470 -2.18675240 7.109587    F-test 0.1621877 0.6991657 3
#> 4   6.237997  4 0.71983467 -0.17596430 6.413961    F-test 0.1621877 0.6991657 3
#> 5   6.999295  5 0.61771873  0.06630393 6.932991    F-test 0.1621877 0.6991657 3
#> 6   8.402880  6 0.78564388  2.32341130 6.079469    F-test 0.1621877 0.6991657 3
#> 7   8.835266  7 0.54176123  1.51620195 7.319065    F-test 0.1621877 0.6991657 3
#> 8  10.047686  8 0.63916007  3.22367556 6.824010    F-test 0.1621877 0.6991657 3
#> 9  10.081151  9 0.07545744  0.39197908 9.689172    F-test 0.1621877 0.6991657 3
#> 10 11.057897 10 0.05488460  1.26415805 9.793739    F-test 0.1621877 0.6991657 3
#> 11  3.522821  1 0.88829936 -2.03487565 5.557696    F-test 0.0051417 0.9448419 1
#> 12  4.859509  2 0.39082137 -3.22674606 8.086255    F-test 0.0051417 0.9448419 1
#> 13  6.753576  3 0.58297470 -0.35601039 7.109587    F-test 0.0051417 0.9448419 1
#> 14 10.185242  4 0.71983467  3.77128078 6.413961    F-test 0.0051417 0.9448419 1
#> 15  9.780891  5 0.61771873  2.84789931 6.932991    F-test 0.0051417 0.9448419 1
#> 16  4.625976  6 0.78564388 -1.45349242 6.079469    F-test 0.0051417 0.9448419 1
#> 17  5.674836  7 0.54176123 -1.64422858 7.319065    F-test 0.0051417 0.9448419 1
#> 18  8.097665  8 0.63916007  1.27365441 6.824010    F-test 0.0051417 0.9448419 1
#> 19  8.512854  9 0.07545744 -1.17631847 9.689172    F-test 0.0051417 0.9448419 1
#> 20 11.792576 10 0.05488460  1.99883707 9.793739    F-test 0.0051417 0.9448419 1
#> 21  4.967811  1 0.88829936 -0.58988568 5.557696    F-test 1.6219838 0.2434720 2
#> 22  6.932613  2 0.39082137 -1.15364157 8.086255    F-test 1.6219838 0.2434720 2
#> 23  6.580210  3 0.58297470 -0.52937666 7.109587    F-test 1.6219838 0.2434720 2
#> 24  5.813691  4 0.71983467 -0.60026956 6.413961    F-test 1.6219838 0.2434720 2
#> 25  8.440978  5 0.61771873  1.50798674 6.932991    F-test 1.6219838 0.2434720 2
#> 26  3.463423  6 0.78564388 -2.61604626 6.079469    F-test 1.6219838 0.2434720 2
#> 27 13.146095  7 0.54176123  5.82703069 7.319065    F-test 1.6219838 0.2434720 2
#> 28  6.442635  8 0.63916007 -0.38137498 6.824010    F-test 1.6219838 0.2434720 2
#> 29 10.010692  9 0.07545744  0.32152015 9.689172    F-test 1.6219838 0.2434720 2
#> 30  8.007796 10 0.05488460 -1.78594287 9.793739    F-test 1.6219838 0.2434720 2
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE