Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>              y         x          e      .resid   .fitted test_name  statistic
#> 1   3.02823342 0.5318002  1.2136217  0.61033642 2.4178970    F-test 0.05069810
#> 2   5.10061212 0.8890679  2.4211026  1.41000322 3.6906089    F-test 0.05069810
#> 3   2.41306907 0.6047531  0.4425896 -0.26471157 2.6777806    F-test 0.05069810
#> 4   1.94459990 0.2160942  0.6818090  0.65135788 1.2932420    F-test 0.05069810
#> 5   2.34613839 0.9390450 -0.4747121 -1.52250646 3.8686448    F-test 0.05069810
#> 6   0.81509566 0.2629830 -0.5170475 -0.64518078 1.4602764    F-test 0.05069810
#> 7   3.24017383 0.5459126  1.3962407  0.77200378 2.4681700    F-test 0.05069810
#> 8   1.48958041 0.1033926  0.3754978  0.59782092 0.8917595    F-test 0.05069810
#> 9  -0.07342265 0.2696605 -1.4157999 -1.55748655 1.4840639    F-test 0.05069810
#> 10  2.03960876 0.4401047  0.4058119 -0.05163686 2.0912456    F-test 0.05069810
#> 11  0.96402209 0.5318002  1.2136217 -1.45387491 2.4178970    F-test 3.29151131
#> 12  3.59017813 0.8890679  2.4211026 -0.10043078 3.6906089    F-test 3.29151131
#> 13  2.92601096 0.6047531  0.4425896  0.24823032 2.6777806    F-test 3.29151131
#> 14  2.08801410 0.2160942  0.6818090  0.79477208 1.2932420    F-test 3.29151131
#> 15  5.20401346 0.9390450 -0.4747121  1.33536861 3.8686448    F-test 3.29151131
#> 16  1.63713775 0.2629830 -0.5170475  0.17686131 1.4602764    F-test 3.29151131
#> 17  0.48150958 0.5459126  1.3962407 -1.98666046 2.4681700    F-test 3.29151131
#> 18  1.00807495 0.1033926  0.3754978  0.11631546 0.8917595    F-test 3.29151131
#> 19  1.89477902 0.2696605 -1.4157999  0.41071512 1.4840639    F-test 3.29151131
#> 20  2.54994888 0.4401047  0.4058119  0.45870325 2.0912456    F-test 3.29151131
#> 21  0.88225467 0.5318002  1.2136217 -1.53564233 2.4178970    F-test 0.01318888
#> 22  4.79866954 0.8890679  2.4211026  1.10806064 3.6906089    F-test 0.01318888
#> 23  3.73339586 0.6047531  0.4425896  1.05561522 2.6777806    F-test 0.01318888
#> 24  0.64066332 0.2160942  0.6818090 -0.65257869 1.2932420    F-test 0.01318888
#> 25  2.92984942 0.9390450 -0.4747121 -0.93879543 3.8686448    F-test 0.01318888
#> 26  0.81214456 0.2629830 -0.5170475 -0.64813187 1.4602764    F-test 0.01318888
#> 27  2.66610077 0.5459126  1.3962407  0.19793072 2.4681700    F-test 0.01318888
#> 28  1.14369659 0.1033926  0.3754978  0.25193709 0.8917595    F-test 0.01318888
#> 29  3.00458745 0.2696605 -1.4157999  1.52052355 1.4840639    F-test 0.01318888
#> 30  1.73232673 0.4401047  0.4058119 -0.35891889 2.0912456    F-test 0.01318888
#>      p_value k  null
#> 1  0.8282850 3 FALSE
#> 2  0.8282850 3 FALSE
#> 3  0.8282850 3 FALSE
#> 4  0.8282850 3 FALSE
#> 5  0.8282850 3 FALSE
#> 6  0.8282850 3 FALSE
#> 7  0.8282850 3 FALSE
#> 8  0.8282850 3 FALSE
#> 9  0.8282850 3 FALSE
#> 10 0.8282850 3 FALSE
#> 11 0.1125119 1  TRUE
#> 12 0.1125119 1  TRUE
#> 13 0.1125119 1  TRUE
#> 14 0.1125119 1  TRUE
#> 15 0.1125119 1  TRUE
#> 16 0.1125119 1  TRUE
#> 17 0.1125119 1  TRUE
#> 18 0.1125119 1  TRUE
#> 19 0.1125119 1  TRUE
#> 20 0.1125119 1  TRUE
#> 21 0.9117945 2  TRUE
#> 22 0.9117945 2  TRUE
#> 23 0.9117945 2  TRUE
#> 24 0.9117945 2  TRUE
#> 25 0.9117945 2  TRUE
#> 26 0.9117945 2  TRUE
#> 27 0.9117945 2  TRUE
#> 28 0.9117945 2  TRUE
#> 29 0.9117945 2  TRUE
#> 30 0.9117945 2  TRUE