Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>              y         x           e       .resid   .fitted test_name
#> 1   0.54654753 0.6158083 -1.44848068 -1.623831739 2.1703793    F-test
#> 2   1.19279611 0.1414790  0.03130074  1.029024904 0.1637712    F-test
#> 3   0.93058532 0.3839746 -0.60082572 -0.259041755 1.1896271    F-test
#> 4   4.29519273 0.8392819  1.75151678  1.179428335 3.1157644    F-test
#> 5   1.58472390 0.6111289 -0.39988359 -0.565859608 2.1505835    F-test
#> 6  -1.15557906 0.1951155 -2.38876456 -1.546254358 0.3906753    F-test
#> 7   3.17775001 0.5040678  1.41959794  1.480079297 1.6976707    F-test
#> 8   2.61560076 0.4701228  0.92446253  1.061531244 1.5540695    F-test
#> 9   1.16474368 0.1799245 -0.04755368  0.838332277 0.3264114    F-test
#> 10  0.05454394 0.4923152 -1.68014551 -1.593408596 1.6479525    F-test
#> 11  3.80437276 0.6158083 -1.44848068  1.633993485 2.1703793    F-test
#> 12  0.30945092 0.1414790  0.03130074  0.145679707 0.1637712    F-test
#> 13  0.80332582 0.3839746 -0.60082572 -0.386301252 1.1896271    F-test
#> 14  1.60925461 0.8392819  1.75151678 -1.506509785 3.1157644    F-test
#> 15  4.09448097 0.6111289 -0.39988359  1.943897460 2.1505835    F-test
#> 16  1.71447338 0.1951155 -2.38876456  1.323798084 0.3906753    F-test
#> 17  1.03532948 0.5040678  1.41959794 -0.662341231 1.6976707    F-test
#> 18  1.54633595 0.4701228  0.92446253 -0.007733568 1.5540695    F-test
#> 19 -1.27498932 0.1799245 -0.04755368 -1.601400726 0.3264114    F-test
#> 20  0.76487036 0.4923152 -1.68014551 -0.883082174 1.6479525    F-test
#> 21  2.07155069 0.6158083 -1.44848068 -0.098828581 2.1703793    F-test
#> 22  0.13542474 0.1414790  0.03130074 -0.028346465 0.1637712    F-test
#> 23 -0.21322408 0.3839746 -0.60082572 -1.402851155 1.1896271    F-test
#> 24  3.25372277 0.8392819  1.75151678  0.137958374 3.1157644    F-test
#> 25  0.80866687 0.6111289 -0.39988359 -1.341916633 2.1505835    F-test
#> 26  1.12286403 0.1951155 -2.38876456  0.732188732 0.3906753    F-test
#> 27  3.72660634 0.5040678  1.41959794  2.028935635 1.6976707    F-test
#> 28  0.03948722 0.4701228  0.92446253 -1.514582293 1.5540695    F-test
#> 29 -0.06410584 0.1799245 -0.04755368 -0.390517245 0.3264114    F-test
#> 30  3.52591217 0.4923152 -1.68014551  1.877959632 1.6479525    F-test
#>       statistic   p_value k  null
#> 1  0.7488443256 0.4155121 2 FALSE
#> 2  0.7488443256 0.4155121 2 FALSE
#> 3  0.7488443256 0.4155121 2 FALSE
#> 4  0.7488443256 0.4155121 2 FALSE
#> 5  0.7488443256 0.4155121 2 FALSE
#> 6  0.7488443256 0.4155121 2 FALSE
#> 7  0.7488443256 0.4155121 2 FALSE
#> 8  0.7488443256 0.4155121 2 FALSE
#> 9  0.7488443256 0.4155121 2 FALSE
#> 10 0.7488443256 0.4155121 2 FALSE
#> 11 0.5484723555 0.4830388 1  TRUE
#> 12 0.5484723555 0.4830388 1  TRUE
#> 13 0.5484723555 0.4830388 1  TRUE
#> 14 0.5484723555 0.4830388 1  TRUE
#> 15 0.5484723555 0.4830388 1  TRUE
#> 16 0.5484723555 0.4830388 1  TRUE
#> 17 0.5484723555 0.4830388 1  TRUE
#> 18 0.5484723555 0.4830388 1  TRUE
#> 19 0.5484723555 0.4830388 1  TRUE
#> 20 0.5484723555 0.4830388 1  TRUE
#> 21 0.0001118631 0.9918564 3  TRUE
#> 22 0.0001118631 0.9918564 3  TRUE
#> 23 0.0001118631 0.9918564 3  TRUE
#> 24 0.0001118631 0.9918564 3  TRUE
#> 25 0.0001118631 0.9918564 3  TRUE
#> 26 0.0001118631 0.9918564 3  TRUE
#> 27 0.0001118631 0.9918564 3  TRUE
#> 28 0.0001118631 0.9918564 3  TRUE
#> 29 0.0001118631 0.9918564 3  TRUE
#> 30 0.0001118631 0.9918564 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>             y  e           x     .resid  .fitted test_name   statistic
#> 1   2.6204077  1 0.432956442 -4.6502297 7.270637    F-test 0.084449989
#> 2   4.1623337  2 0.688416475 -3.1630182 7.325352    F-test 0.084449989
#> 3   5.6436555  3 0.876101549 -1.7218948 7.365550    F-test 0.084449989
#> 4   5.0020424  4 0.002038264 -2.1763009 7.178343    F-test 0.084449989
#> 5   6.8097702  5 0.529451402 -0.4815346 7.291305    F-test 0.084449989
#> 6   8.4237981  6 0.793753485  1.0758851 7.347913    F-test 0.084449989
#> 7   8.2894709  7 0.234486854  1.0613417 7.228129    F-test 0.084449989
#> 8   9.0291651  8 0.028360730  1.8451840 7.183981    F-test 0.084449989
#> 9  10.7297589  9 0.489827707  3.4469408 7.282818    F-test 0.084449989
#> 10 12.0815961 10 0.653948057  4.7636267 7.317969    F-test 0.084449989
#> 11 10.0551570  1 0.432956442  2.7845196 7.270637    F-test 0.002806649
#> 12  8.3647579  2 0.688416475  1.0394060 7.325352    F-test 0.002806649
#> 13  5.7109218  3 0.876101549 -1.6546285 7.365550    F-test 0.002806649
#> 14  7.4792104  4 0.002038264  0.3008671 7.178343    F-test 0.002806649
#> 15  0.1095045  5 0.529451402 -7.1818003 7.291305    F-test 0.002806649
#> 16  8.9342836  6 0.793753485  1.5863706 7.347913    F-test 0.002806649
#> 17  5.8410535  7 0.234486854 -1.3870757 7.228129    F-test 0.002806649
#> 18  7.2137411  8 0.028360730  0.0297600 7.183981    F-test 0.002806649
#> 19 10.5478111  9 0.489827707  3.2649930 7.282818    F-test 0.002806649
#> 20  8.5355576 10 0.653948057  1.2175881 7.317969    F-test 0.002806649
#> 21  5.5801238  1 0.432956442 -1.6905137 7.270637    F-test 7.049780994
#> 22  2.5466454  2 0.688416475 -4.7787065 7.325352    F-test 7.049780994
#> 23 12.7592231  3 0.876101549  5.3936728 7.365550    F-test 7.049780994
#> 24  9.5151318  4 0.002038264  2.3367884 7.178343    F-test 7.049780994
#> 25  8.8376938  5 0.529451402  1.5463891 7.291305    F-test 7.049780994
#> 26  8.6926514  6 0.793753485  1.3447384 7.347913    F-test 7.049780994
#> 27  6.6490438  7 0.234486854 -0.5790854 7.228129    F-test 7.049780994
#> 28  8.3038854  8 0.028360730  1.1199043 7.183981    F-test 7.049780994
#> 29  3.7580282  9 0.489827707 -3.5247900 7.282818    F-test 7.049780994
#> 30  6.1495718 10 0.653948057 -1.1683976 7.317969    F-test 7.049780994
#>       p_value k  null
#> 1  0.77977611 2 FALSE
#> 2  0.77977611 2 FALSE
#> 3  0.77977611 2 FALSE
#> 4  0.77977611 2 FALSE
#> 5  0.77977611 2 FALSE
#> 6  0.77977611 2 FALSE
#> 7  0.77977611 2 FALSE
#> 8  0.77977611 2 FALSE
#> 9  0.77977611 2 FALSE
#> 10 0.77977611 2 FALSE
#> 11 0.95922978 1  TRUE
#> 12 0.95922978 1  TRUE
#> 13 0.95922978 1  TRUE
#> 14 0.95922978 1  TRUE
#> 15 0.95922978 1  TRUE
#> 16 0.95922978 1  TRUE
#> 17 0.95922978 1  TRUE
#> 18 0.95922978 1  TRUE
#> 19 0.95922978 1  TRUE
#> 20 0.95922978 1  TRUE
#> 21 0.03269676 3  TRUE
#> 22 0.03269676 3  TRUE
#> 23 0.03269676 3  TRUE
#> 24 0.03269676 3  TRUE
#> 25 0.03269676 3  TRUE
#> 26 0.03269676 3  TRUE
#> 27 0.03269676 3  TRUE
#> 28 0.03269676 3  TRUE
#> 29 0.03269676 3  TRUE
#> 30 0.03269676 3  TRUE