Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

A data frame.

Examples

# Instantiate
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>              y         x           e      .resid   .fitted test_name
#> 1   2.59017759 0.2054507  1.34251687  0.86714309 1.7230345    F-test
#> 2   2.08327091 0.1667789  0.88867683  0.29860666 1.7846643    F-test
#> 3   1.26118346 0.1786049  0.05067889 -0.50463419 1.7658176    F-test
#> 4   1.62767993 0.3399675  0.17213448  0.11901947 1.5086605    F-test
#> 5   2.05937505 0.2927121  0.68098256  0.47540553 1.5839695    F-test
#> 6  -0.90439068 0.2711454 -2.24905590 -2.52273023 1.6183395    F-test
#> 7   1.29190392 0.9515701 -1.56515178  0.75792985 0.5339741    F-test
#> 8  -0.03609839 0.8172393 -2.52121780 -0.78415000 0.7480516    F-test
#> 9   2.52099843 0.1795405  1.30922311  0.75667190 1.7643265    F-test
#> 10  2.14080119 0.2801036  0.78223963  0.53673793 1.6040633    F-test
#> 11  3.48538437 0.2054507  1.34251687  1.76234987 1.7230345    F-test
#> 12  1.94757084 0.1667789  0.88867683  0.16290659 1.7846643    F-test
#> 13  0.94854199 0.1786049  0.05067889 -0.81727566 1.7658176    F-test
#> 14  1.46467028 0.3399675  0.17213448 -0.04399019 1.5086605    F-test
#> 15  1.86443112 0.2927121  0.68098256  0.28046160 1.5839695    F-test
#> 16  2.74805531 0.2711454 -2.24905590  1.12971577 1.6183395    F-test
#> 17 -0.24054936 0.9515701 -1.56515178 -0.77452343 0.5339741    F-test
#> 18  1.58119738 0.8172393 -2.52121780  0.83314577 0.7480516    F-test
#> 19  0.28421084 0.1795405  1.30922311 -1.48011569 1.7643265    F-test
#> 20  0.55138864 0.2801036  0.78223963 -1.05267463 1.6040633    F-test
#> 21 -0.40837839 0.2054507  1.34251687 -2.13141289 1.7230345    F-test
#> 22  2.70957009 0.1667789  0.88867683  0.92490584 1.7846643    F-test
#> 23  1.95257859 0.1786049  0.05067889  0.18676094 1.7658176    F-test
#> 24  0.90531877 0.3399675  0.17213448 -0.60334170 1.5086605    F-test
#> 25  1.86357896 0.2927121  0.68098256  0.27960944 1.5839695    F-test
#> 26  3.54057822 0.2711454 -2.24905590  1.92223868 1.6183395    F-test
#> 27  0.57670628 0.9515701 -1.56515178  0.04273221 0.5339741    F-test
#> 28  0.64915940 0.8172393 -2.52121780 -0.09889221 0.7480516    F-test
#> 29  1.37757799 0.1795405  1.30922311 -0.38674854 1.7643265    F-test
#> 30  1.46821148 0.2801036  0.78223963 -0.13585178 1.6040633    F-test
#>      statistic   p_value k  null
#> 1  0.847969973 0.3877613 2 FALSE
#> 2  0.847969973 0.3877613 2 FALSE
#> 3  0.847969973 0.3877613 2 FALSE
#> 4  0.847969973 0.3877613 2 FALSE
#> 5  0.847969973 0.3877613 2 FALSE
#> 6  0.847969973 0.3877613 2 FALSE
#> 7  0.847969973 0.3877613 2 FALSE
#> 8  0.847969973 0.3877613 2 FALSE
#> 9  0.847969973 0.3877613 2 FALSE
#> 10 0.847969973 0.3877613 2 FALSE
#> 11 0.753121526 0.4142481 1  TRUE
#> 12 0.753121526 0.4142481 1  TRUE
#> 13 0.753121526 0.4142481 1  TRUE
#> 14 0.753121526 0.4142481 1  TRUE
#> 15 0.753121526 0.4142481 1  TRUE
#> 16 0.753121526 0.4142481 1  TRUE
#> 17 0.753121526 0.4142481 1  TRUE
#> 18 0.753121526 0.4142481 1  TRUE
#> 19 0.753121526 0.4142481 1  TRUE
#> 20 0.753121526 0.4142481 1  TRUE
#> 21 0.008212538 0.9303308 3  TRUE
#> 22 0.008212538 0.9303308 3  TRUE
#> 23 0.008212538 0.9303308 3  TRUE
#> 24 0.008212538 0.9303308 3  TRUE
#> 25 0.008212538 0.9303308 3  TRUE
#> 26 0.008212538 0.9303308 3  TRUE
#> 27 0.008212538 0.9303308 3  TRUE
#> 28 0.008212538 0.9303308 3  TRUE
#> 29 0.008212538 0.9303308 3  TRUE
#> 30 0.008212538 0.9303308 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e          x      .resid  .fitted test_name statistic    p_value
#> 1   3.387651  1 0.77970724 -3.44177722 6.829428    F-test 4.4325328 0.07328905
#> 2   4.217306  2 0.71132403 -2.74418050 6.961486    F-test 4.4325328 0.07328905
#> 3   5.363155  3 0.77010026 -1.48482573 6.847980    F-test 4.4325328 0.07328905
#> 4   5.151860  4 0.13392459 -2.92467378 8.076534    F-test 4.4325328 0.07328905
#> 5   7.970987  5 0.99029750  1.54824121 6.422745    F-test 4.4325328 0.07328905
#> 6   7.140906  6 0.12522479 -0.95242879 8.093335    F-test 4.4325328 0.07328905
#> 7   8.541780  7 0.38982023  0.95941997 7.582360    F-test 4.4325328 0.07328905
#> 8  10.846494  8 0.94792756  4.34192585 6.504568    F-test 4.4325328 0.07328905
#> 9  10.094236  9 0.08671636  1.92653551 8.167701    F-test 4.4325328 0.07328905
#> 10 11.037335 10 0.03603636  2.77176347 8.265572    F-test 4.4325328 0.07328905
#> 11  8.431448  1 0.77970724  1.60202049 6.829428    F-test 0.1193225 0.73991833
#> 12  8.972900  2 0.71132403  2.01141313 6.961486    F-test 0.1193225 0.73991833
#> 13  5.121573  3 0.77010026 -1.72640750 6.847980    F-test 0.1193225 0.73991833
#> 14 12.946189  4 0.13392459  4.86965462 8.076534    F-test 0.1193225 0.73991833
#> 15  6.302440  5 0.99029750 -0.12030538 6.422745    F-test 0.1193225 0.73991833
#> 16  4.093824  6 0.12522479 -3.99951055 8.093335    F-test 0.1193225 0.73991833
#> 17  7.505302  7 0.38982023 -0.07705801 7.582360    F-test 0.1193225 0.73991833
#> 18  5.301458  8 0.94792756 -1.20311000 6.504568    F-test 0.1193225 0.73991833
#> 19  4.948150  9 0.08671636 -3.21955059 8.167701    F-test 0.1193225 0.73991833
#> 20 10.128425 10 0.03603636  1.86285377 8.265572    F-test 0.1193225 0.73991833
#> 21  1.242231  1 0.77970724 -5.58719709 6.829428    F-test 3.3049771 0.11190257
#> 22  8.538566  2 0.71132403  1.57707943 6.961486    F-test 3.3049771 0.11190257
#> 23  7.162973  3 0.77010026  0.31499304 6.847980    F-test 3.3049771 0.11190257
#> 24  7.175296  4 0.13392459 -0.90123849 8.076534    F-test 3.3049771 0.11190257
#> 25  7.483073  5 0.99029750  1.06032750 6.422745    F-test 3.3049771 0.11190257
#> 26  7.638950  6 0.12522479 -0.45438503 8.093335    F-test 3.3049771 0.11190257
#> 27  4.475097  7 0.38982023 -3.10726347 7.582360    F-test 3.3049771 0.11190257
#> 28  9.727655  8 0.94792756  3.22308648 6.504568    F-test 3.3049771 0.11190257
#> 29 10.850044  9 0.08671636  2.68234334 8.167701    F-test 3.3049771 0.11190257
#> 30  9.457826 10 0.03603636  1.19225429 8.265572    F-test 3.3049771 0.11190257
#>    k  null
#> 1  2 FALSE
#> 2  2 FALSE
#> 3  2 FALSE
#> 4  2 FALSE
#> 5  2 FALSE
#> 6  2 FALSE
#> 7  2 FALSE
#> 8  2 FALSE
#> 9  2 FALSE
#> 10 2 FALSE
#> 11 1  TRUE
#> 12 1  TRUE
#> 13 1  TRUE
#> 14 1  TRUE
#> 15 1  TRUE
#> 16 1  TRUE
#> 17 1  TRUE
#> 18 1  TRUE
#> 19 1  TRUE
#> 20 1  TRUE
#> 21 3  TRUE
#> 22 3  TRUE
#> 23 3  TRUE
#> 24 3  TRUE
#> 25 3  TRUE
#> 26 3  TRUE
#> 27 3  TRUE
#> 28 3  TRUE
#> 29 3  TRUE
#> 30 3  TRUE