Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>             y          x           e      .resid  .fitted test_name  statistic
#> 1   2.6228482 0.18154175  1.40834903  1.17030745 1.452541    F-test 0.41644555
#> 2   1.5036627 0.48742188 -0.22133927 -0.36345823 1.867121    F-test 0.41644555
#> 3   2.3809621 0.90085040 -0.33141972 -0.04650664 2.427469    F-test 0.41644555
#> 4   1.6497784 0.15213006  0.47450482  0.23710136 1.412677    F-test 0.41644555
#> 5   2.6383991 0.98502142 -0.31688952  0.09684762 2.541551    F-test 0.41644555
#> 6  -0.5772119 0.33779949 -2.02911989 -2.24153940 1.664328    F-test 0.41644555
#> 7   1.4495438 0.01599718  0.43329066  0.22137685 1.228167    F-test 0.41644555
#> 8   3.1276247 0.66710573  1.01548891  1.01696601 2.110659    F-test 0.41644555
#> 9   2.0032876 0.60255328  0.03766390 -0.01987872 2.023166    F-test 0.41644555
#> 10  1.8521981 0.52895559  0.04344849 -0.07121630 1.923414    F-test 0.41644555
#> 11  0.7375509 0.18154175  1.40834903 -0.71498987 1.452541    F-test 0.50587746
#> 12  2.5753841 0.48742188 -0.22133927  0.70826320 1.867121    F-test 0.50587746
#> 13  1.8920228 0.90085040 -0.33141972 -0.53544601 2.427469    F-test 0.50587746
#> 14  0.1316250 0.15213006  0.47450482 -1.28105208 1.412677    F-test 0.50587746
#> 15  3.1483239 0.98502142 -0.31688952  0.60677244 2.541551    F-test 0.50587746
#> 16  1.0161078 0.33779949 -2.02911989 -0.64821976 1.664328    F-test 0.50587746
#> 17  2.9085036 0.01599718  0.43329066  1.68033666 1.228167    F-test 0.50587746
#> 18  2.1795302 0.66710573  1.01548891  0.06887149 2.110659    F-test 0.50587746
#> 19  1.3329920 0.60255328  0.03766390 -0.69017440 2.023166    F-test 0.50587746
#> 20  2.7290527 0.52895559  0.04344849  0.80563833 1.923414    F-test 0.50587746
#> 21  1.8909664 0.18154175  1.40834903  0.43842570 1.452541    F-test 0.01849372
#> 22  0.8944278 0.48742188 -0.22133927 -0.97269310 1.867121    F-test 0.01849372
#> 23  2.1968378 0.90085040 -0.33141972 -0.23063093 2.427469    F-test 0.01849372
#> 24  0.6199712 0.15213006  0.47450482 -0.79270591 1.412677    F-test 0.01849372
#> 25  2.7521714 0.98502142 -0.31688952  0.21061992 2.541551    F-test 0.01849372
#> 26  3.1770921 0.33779949 -2.02911989  1.51276461 1.664328    F-test 0.01849372
#> 27  1.3039296 0.01599718  0.43329066  0.07576271 1.228167    F-test 0.01849372
#> 28  2.5979554 0.66710573  1.01548891  0.48729674 2.110659    F-test 0.01849372
#> 29  2.8877975 0.60255328  0.03766390  0.86463112 2.023166    F-test 0.01849372
#> 30  0.3299435 0.52895559  0.04344849 -1.59347085 1.923414    F-test 0.01849372
#>      p_value k  null
#> 1  0.5392840 2 FALSE
#> 2  0.5392840 2 FALSE
#> 3  0.5392840 2 FALSE
#> 4  0.5392840 2 FALSE
#> 5  0.5392840 2 FALSE
#> 6  0.5392840 2 FALSE
#> 7  0.5392840 2 FALSE
#> 8  0.5392840 2 FALSE
#> 9  0.5392840 2 FALSE
#> 10 0.5392840 2 FALSE
#> 11 0.4999366 1  TRUE
#> 12 0.4999366 1  TRUE
#> 13 0.4999366 1  TRUE
#> 14 0.4999366 1  TRUE
#> 15 0.4999366 1  TRUE
#> 16 0.4999366 1  TRUE
#> 17 0.4999366 1  TRUE
#> 18 0.4999366 1  TRUE
#> 19 0.4999366 1  TRUE
#> 20 0.4999366 1  TRUE
#> 21 0.8956562 3  TRUE
#> 22 0.8956562 3  TRUE
#> 23 0.8956562 3  TRUE
#> 24 0.8956562 3  TRUE
#> 25 0.8956562 3  TRUE
#> 26 0.8956562 3  TRUE
#> 27 0.8956562 3  TRUE
#> 28 0.8956562 3  TRUE
#> 29 0.8956562 3  TRUE
#> 30 0.8956562 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e          x     .resid  .fitted test_name statistic    p_value k
#> 1   2.042008  1 0.04037731 -4.0309251 6.072933    F-test 0.8330431 0.39174395 3
#> 2   3.892674  2 0.56895946 -3.7565581 7.649232    F-test 0.8330431 0.39174395 3
#> 3   4.702987  3 0.47621061 -2.6696564 7.372644    F-test 0.8330431 0.39174395 3
#> 4   5.175538  4 0.15233236 -1.2312596 6.406797    F-test 0.8330431 0.39174395 3
#> 5   6.848256  5 0.54797721 -0.7384045 7.586661    F-test 0.8330431 0.39174395 3
#> 6   8.998446  6 0.99948201  0.0653404 8.933106    F-test 0.8330431 0.39174395 3
#> 7   8.391903  7 0.30118859  1.5411977 6.850705    F-test 0.8330431 0.39174395 3
#> 8   9.053739  8 0.05112522  2.9487546 6.104984    F-test 0.8330431 0.39174395 3
#> 9  11.787432  9 0.92738654  3.0693244 8.718108    F-test 0.8330431 0.39174395 3
#> 10 11.150215 10 0.13262538  4.8021866 6.348028    F-test 0.8330431 0.39174395 3
#> 11  8.523276  1 0.04037731  2.4503429 6.072933    F-test 4.5798697 0.06962783 1
#> 12 11.863800  2 0.56895946  4.2145680 7.649232    F-test 4.5798697 0.06962783 1
#> 13 10.715586  3 0.47621061  3.3429429 7.372644    F-test 4.5798697 0.06962783 1
#> 14  4.674134  4 0.15233236 -1.7326634 6.406797    F-test 4.5798697 0.06962783 1
#> 15 10.411360  5 0.54797721  2.8246992 7.586661    F-test 4.5798697 0.06962783 1
#> 16  7.696342  6 0.99948201 -1.2367639 8.933106    F-test 4.5798697 0.06962783 1
#> 17  7.056598  7 0.30118859  0.2058924 6.850705    F-test 4.5798697 0.06962783 1
#> 18  3.845556  8 0.05112522 -2.2594286 6.104984    F-test 4.5798697 0.06962783 1
#> 19  4.883802  9 0.92738654 -3.8343056 8.718108    F-test 4.5798697 0.06962783 1
#> 20  2.372744 10 0.13262538 -3.9752839 6.348028    F-test 4.5798697 0.06962783 1
#> 21 10.842512  1 0.04037731  4.7695791 6.072933    F-test 0.9147656 0.37071003 2
#> 22  6.205043  2 0.56895946 -1.4441896 7.649232    F-test 0.9147656 0.37071003 2
#> 23  9.566576  3 0.47621061  2.1939326 7.372644    F-test 0.9147656 0.37071003 2
#> 24  5.285005  4 0.15233236 -1.1217921 6.406797    F-test 0.9147656 0.37071003 2
#> 25  2.871395  5 0.54797721 -4.7152656 7.586661    F-test 0.9147656 0.37071003 2
#> 26 12.807625  6 0.99948201  3.8745189 8.933106    F-test 0.9147656 0.37071003 2
#> 27  8.675789  7 0.30118859  1.8250832 6.850705    F-test 0.9147656 0.37071003 2
#> 28  3.436284  8 0.05112522 -2.6687002 6.104984    F-test 0.9147656 0.37071003 2
#> 29  6.721456  9 0.92738654 -1.9966522 8.718108    F-test 0.9147656 0.37071003 2
#> 30  5.631514 10 0.13262538 -0.7165140 6.348028    F-test 0.9147656 0.37071003 2
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE