Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>             y          x           e      .resid   .fitted test_name statistic
#> 1   1.2742650 0.26534197 -0.06148333  0.12681417 1.1474508    F-test 0.4202478
#> 2   2.3307452 0.64287667  0.27457809  0.02206683 2.3086783    F-test 0.4202478
#> 3   3.8527713 0.98992483  0.88289530  0.47663644 3.3761349    F-test 0.4202478
#> 4   2.9835075 0.49223622  1.24897474  1.13817151 1.8453359    F-test 0.4202478
#> 5  -0.1650463 0.39740423 -1.72038061 -1.71869638 1.5536501    F-test 0.4202478
#> 6   0.9481379 0.03901952 -0.09240416  0.49681341 0.4513245    F-test 0.4202478
#> 7   1.4571058 0.41471352 -0.12959501 -0.14978452 1.6068903    F-test 0.4202478
#> 8   2.3207078 0.68611419  0.16384094 -0.12096122 2.4416690    F-test 0.4202478
#> 9   1.8223799 0.79758183 -0.61133873 -0.96214318 2.7845230    F-test 0.4202478
#> 10  3.2308899 0.71802049  0.99731597  0.69108293 2.5398069    F-test 0.4202478
#> 11  1.9107348 0.26534197 -0.06148333  0.76328392 1.1474508    F-test 3.1246029
#> 12  0.7712395 0.64287667  0.27457809 -1.53743889 2.3086783    F-test 3.1246029
#> 13  4.7624339 0.98992483  0.88289530  1.38629909 3.3761349    F-test 3.1246029
#> 14  1.5498682 0.49223622  1.24897474 -0.29546776 1.8453359    F-test 3.1246029
#> 15  1.9382013 0.39740423 -1.72038061  0.38455122 1.5536501    F-test 3.1246029
#> 16  0.6139146 0.03901952 -0.09240416  0.16259013 0.4513245    F-test 3.1246029
#> 17  1.6640957 0.41471352 -0.12959501  0.05720533 1.6068903    F-test 3.1246029
#> 18  1.4208509 0.68611419  0.16384094 -1.02081811 2.4416690    F-test 3.1246029
#> 19  2.8450170 0.79758183 -0.61133873  0.06049399 2.7845230    F-test 3.1246029
#> 20  2.5791080 0.71802049  0.99731597  0.03930106 2.5398069    F-test 3.1246029
#> 21  2.1460191 0.26534197 -0.06148333  0.99856828 1.1474508    F-test 0.8902746
#> 22  2.2032660 0.64287667  0.27457809 -0.10541236 2.3086783    F-test 0.8902746
#> 23  3.5837829 0.98992483  0.88289530  0.20764807 3.3761349    F-test 0.8902746
#> 24  1.5701190 0.49223622  1.24897474 -0.27521697 1.8453359    F-test 0.8902746
#> 25  1.2758949 0.39740423 -1.72038061 -0.27775520 1.5536501    F-test 0.8902746
#> 26  0.7265410 0.03901952 -0.09240416  0.27521656 0.4513245    F-test 0.8902746
#> 27  0.5383361 0.41471352 -0.12959501 -1.06855422 1.6068903    F-test 0.8902746
#> 28  2.7014094 0.68611419  0.16384094  0.25974034 2.4416690    F-test 0.8902746
#> 29  4.1358846 0.79758183 -0.61133873  1.35136156 2.7845230    F-test 0.8902746
#> 30  1.1742109 0.71802049  0.99731597 -1.36559606 2.5398069    F-test 0.8902746
#>      p_value k  null
#> 1  0.5374898 3 FALSE
#> 2  0.5374898 3 FALSE
#> 3  0.5374898 3 FALSE
#> 4  0.5374898 3 FALSE
#> 5  0.5374898 3 FALSE
#> 6  0.5374898 3 FALSE
#> 7  0.5374898 3 FALSE
#> 8  0.5374898 3 FALSE
#> 9  0.5374898 3 FALSE
#> 10 0.5374898 3 FALSE
#> 11 0.1204483 1  TRUE
#> 12 0.1204483 1  TRUE
#> 13 0.1204483 1  TRUE
#> 14 0.1204483 1  TRUE
#> 15 0.1204483 1  TRUE
#> 16 0.1204483 1  TRUE
#> 17 0.1204483 1  TRUE
#> 18 0.1204483 1  TRUE
#> 19 0.1204483 1  TRUE
#> 20 0.1204483 1  TRUE
#> 21 0.3768205 2  TRUE
#> 22 0.3768205 2  TRUE
#> 23 0.3768205 2  TRUE
#> 24 0.3768205 2  TRUE
#> 25 0.3768205 2  TRUE
#> 26 0.3768205 2  TRUE
#> 27 0.3768205 2  TRUE
#> 28 0.3768205 2  TRUE
#> 29 0.3768205 2  TRUE
#> 30 0.3768205 2  TRUE