Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

A data frame.

Examples

# Instantiate
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>            y         x           e     .resid  .fitted test_name    statistic
#> 1  3.1803984 0.3388418  1.72674285  1.0731514 2.107247    F-test 0.1857837766
#> 2  1.0488034 0.5486612 -0.80088688 -1.4230210 2.471824    F-test 0.1857837766
#> 3  2.2086413 0.6236156  0.19612914 -0.3934223 2.602064    F-test 0.1857837766
#> 4  3.5817827 0.5365393  1.75736889  1.1310209 2.450762    F-test 0.1857837766
#> 5  2.1262810 0.7041334 -0.07365639 -0.6156885 2.741969    F-test 0.1857837766
#> 6  1.0862245 0.2605673 -0.24223808 -0.8850144 1.971239    F-test 0.1857837766
#> 7  1.9640045 0.4093415  0.38710261 -0.2657411 2.229746    F-test 0.1857837766
#> 8  4.0864425 0.6358843  2.04620925  1.4630611 2.623381    F-test 0.1857837766
#> 9  2.9356979 0.7509530  0.62081451  0.1123758 2.823322    F-test 0.1857837766
#> 10 2.9662265 0.9464127  0.12411687 -0.1967219 3.162948    F-test 0.1857837766
#> 11 1.4703794 0.3388418  1.72674285 -0.6368676 2.107247    F-test 0.0004468419
#> 12 3.4936272 0.5486612 -0.80088688  1.0218028 2.471824    F-test 0.0004468419
#> 13 3.5862640 0.6236156  0.19612914  0.9842004 2.602064    F-test 0.0004468419
#> 14 2.1901925 0.5365393  1.75736889 -0.2605693 2.450762    F-test 0.0004468419
#> 15 2.3719404 0.7041334 -0.07365639 -0.3700291 2.741969    F-test 0.0004468419
#> 16 1.7506431 0.2605673 -0.24223808 -0.2205959 1.971239    F-test 0.0004468419
#> 17 3.3559530 0.4093415  0.38710261  1.1262074 2.229746    F-test 0.0004468419
#> 18 0.6165727 0.6358843  2.04620925 -2.0068087 2.623381    F-test 0.0004468419
#> 19 2.9710994 0.7509530  0.62081451  0.1477773 2.823322    F-test 0.0004468419
#> 20 3.3778310 0.9464127  0.12411687  0.2148826 3.162948    F-test 0.0004468419
#> 21 2.6412880 0.3388418  1.72674285  0.5340410 2.107247    F-test 0.0178009092
#> 22 1.9560824 0.5486612 -0.80088688 -0.5157420 2.471824    F-test 0.0178009092
#> 23 1.8921934 0.6236156  0.19612914 -0.7098702 2.602064    F-test 0.0178009092
#> 24 1.5605860 0.5365393  1.75736889 -0.8901757 2.450762    F-test 0.0178009092
#> 25 1.8946494 0.7041334 -0.07365639 -0.8473201 2.741969    F-test 0.0178009092
#> 26 1.0172647 0.2605673 -0.24223808 -0.9539743 1.971239    F-test 0.0178009092
#> 27 3.8562992 0.4093415  0.38710261  1.6265536 2.229746    F-test 0.0178009092
#> 28 3.3868125 0.6358843  2.04620925  0.7634311 2.623381    F-test 0.0178009092
#> 29 3.9672793 0.7509530  0.62081451  1.1439573 2.823322    F-test 0.0178009092
#> 30 3.0120478 0.9464127  0.12411687 -0.1509006 3.162948    F-test 0.0178009092
#>      p_value k  null
#> 1  0.6794110 3 FALSE
#> 2  0.6794110 3 FALSE
#> 3  0.6794110 3 FALSE
#> 4  0.6794110 3 FALSE
#> 5  0.6794110 3 FALSE
#> 6  0.6794110 3 FALSE
#> 7  0.6794110 3 FALSE
#> 8  0.6794110 3 FALSE
#> 9  0.6794110 3 FALSE
#> 10 0.6794110 3 FALSE
#> 11 0.9837250 1  TRUE
#> 12 0.9837250 1  TRUE
#> 13 0.9837250 1  TRUE
#> 14 0.9837250 1  TRUE
#> 15 0.9837250 1  TRUE
#> 16 0.9837250 1  TRUE
#> 17 0.9837250 1  TRUE
#> 18 0.9837250 1  TRUE
#> 19 0.9837250 1  TRUE
#> 20 0.9837250 1  TRUE
#> 21 0.8976158 2  TRUE
#> 22 0.8976158 2  TRUE
#> 23 0.8976158 2  TRUE
#> 24 0.8976158 2  TRUE
#> 25 0.8976158 2  TRUE
#> 26 0.8976158 2  TRUE
#> 27 0.8976158 2  TRUE
#> 28 0.8976158 2  TRUE
#> 29 0.8976158 2  TRUE
#> 30 0.8976158 2  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e         x      .resid  .fitted test_name  statistic   p_value k
#> 1   2.930534  1 0.5865240 -4.72116446 7.651699    F-test 5.49617868 0.0515108 2
#> 2   3.827091  2 0.5378297 -3.60571810 7.432809    F-test 5.49617868 0.0515108 2
#> 3   4.390649  3 0.3004052 -1.97489089 6.365539    F-test 5.49617868 0.0515108 2
#> 4   5.290584  4 0.2352444 -0.78204447 6.072629    F-test 5.49617868 0.0515108 2
#> 5   6.412857  5 0.3141601 -0.01451365 6.427370    F-test 5.49617868 0.0515108 2
#> 6   8.845808  6 0.9476907 -0.42940492 9.275213    F-test 5.49617868 0.0515108 2
#> 7   9.514558  7 0.8283665  0.77572962 8.738828    F-test 5.49617868 0.0515108 2
#> 8   9.514876  8 0.3745720  2.81594256 6.698934    F-test 5.49617868 0.0515108 2
#> 9  10.231485  9 0.1938913  4.34474639 5.886739    F-test 5.49617868 0.0515108 2
#> 10 12.808768 10 0.9348406  3.59131791 9.217450    F-test 5.49617868 0.0515108 2
#> 11  7.525815  1 0.5865240 -0.12588405 7.651699    F-test 1.86781121 0.2139998 1
#> 12 10.351565  2 0.5378297  2.91875667 7.432809    F-test 1.86781121 0.2139998 1
#> 13  2.531799  3 0.3004052 -3.83374023 6.365539    F-test 1.86781121 0.2139998 1
#> 14  6.817356  4 0.2352444  0.74472726 6.072629    F-test 1.86781121 0.2139998 1
#> 15  5.701409  5 0.3141601 -0.72596177 6.427370    F-test 1.86781121 0.2139998 1
#> 16  3.640230  6 0.9476907 -5.63498305 9.275213    F-test 1.86781121 0.2139998 1
#> 17 12.918988  7 0.8283665  4.18016039 8.738828    F-test 1.86781121 0.2139998 1
#> 18  9.003504  8 0.3745720  2.30456991 6.698934    F-test 1.86781121 0.2139998 1
#> 19  5.128583  9 0.1938913 -0.75815577 5.886739    F-test 1.86781121 0.2139998 1
#> 20 10.147960 10 0.9348406  0.93051063 9.217450    F-test 1.86781121 0.2139998 1
#> 21  5.847570  1 0.5865240 -1.80412921 7.651699    F-test 0.02110964 0.8885758 3
#> 22  4.007284  2 0.5378297 -3.42552474 7.432809    F-test 0.02110964 0.8885758 3
#> 23  6.617568  3 0.3004052  0.25202817 6.365539    F-test 0.02110964 0.8885758 3
#> 24  8.697201  4 0.2352444  2.62457245 6.072629    F-test 0.02110964 0.8885758 3
#> 25  5.186656  5 0.3141601 -1.24071400 6.427370    F-test 0.02110964 0.8885758 3
#> 26  9.396679  6 0.9476907  0.12146594 9.275213    F-test 0.02110964 0.8885758 3
#> 27 14.940545  7 0.8283665  6.20171704 8.738828    F-test 0.02110964 0.8885758 3
#> 28  8.847659  8 0.3745720  2.14872505 6.698934    F-test 0.02110964 0.8885758 3
#> 29  4.401893  9 0.1938913 -1.48484540 5.886739    F-test 0.02110964 0.8885758 3
#> 30  5.824154 10 0.9348406 -3.39329528 9.217450    F-test 0.02110964 0.8885758 3
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE