Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>             y         x          e      .resid   .fitted test_name   statistic
#> 1   1.7428859 0.5502392 -0.1101165 -0.18327566 1.9261616    F-test 0.850737365
#> 2   3.4020834 0.8939304  0.7090415  0.51178848 2.8902949    F-test 0.850737365
#> 3   1.0029103 0.1880406 -0.2204895  0.09279976 0.9101105    F-test 0.850737365
#> 4   1.3953646 0.9122610 -1.3491165 -1.54635193 2.9417165    F-test 0.850737365
#> 5   1.4811346 0.2543077  0.1621544  0.38512938 1.0960052    F-test 0.850737365
#> 6   1.6767776 0.4520512  0.0203762  0.02605624 1.6507214    F-test 0.850737365
#> 7   2.4147230 0.7468775  0.1100195 -0.06305452 2.4777776    F-test 0.850737365
#> 8  -0.1593419 0.2979389 -1.5460483 -1.37774258 1.2184007    F-test 0.850737365
#> 9   3.2521125 0.8376284  0.7128627  0.51975754 2.7323549    F-test 0.850737365
#> 10  3.6459542 0.5805038  1.7284656  1.63489330 2.0110609    F-test 0.850737365
#> 11  2.5784299 0.5502392 -0.1101165  0.65226826 1.9261616    F-test 0.032973140
#> 12  3.8319480 0.8939304  0.7090415  0.94165309 2.8902949    F-test 0.032973140
#> 13  1.6417356 0.1880406 -0.2204895  0.73162504 0.9101105    F-test 0.032973140
#> 14  3.2881300 0.9122610 -1.3491165  0.34641348 2.9417165    F-test 0.032973140
#> 15 -0.4374930 0.2543077  0.1621544 -1.53349820 1.0960052    F-test 0.032973140
#> 16  2.7045433 0.4520512  0.0203762  1.05382194 1.6507214    F-test 0.032973140
#> 17  1.8760428 0.7468775  0.1100195 -0.60173472 2.4777776    F-test 0.032973140
#> 18  1.4978226 0.2979389 -1.5460483  0.27942195 1.2184007    F-test 0.032973140
#> 19  1.6084840 0.8376284  0.7128627 -1.12387091 2.7323549    F-test 0.032973140
#> 20  1.2649610 0.5805038  1.7284656 -0.74609992 2.0110609    F-test 0.032973140
#> 21  1.2606819 0.5502392 -0.1101165 -0.66547975 1.9261616    F-test 0.001672555
#> 22  4.3027890 0.8939304  0.7090415  1.41249414 2.8902949    F-test 0.001672555
#> 23  0.7282446 0.1880406 -0.2204895 -0.18186592 0.9101105    F-test 0.001672555
#> 24  1.5696393 0.9122610 -1.3491165 -1.37207722 2.9417165    F-test 0.001672555
#> 25  0.9186867 0.2543077  0.1621544 -0.17731850 1.0960052    F-test 0.001672555
#> 26  2.4022509 0.4520512  0.0203762  0.75152953 1.6507214    F-test 0.001672555
#> 27  3.5587124 0.7468775  0.1100195  1.08093480 2.4777776    F-test 0.001672555
#> 28  1.7475521 0.2979389 -1.5460483  0.52915137 1.2184007    F-test 0.001672555
#> 29  2.4575582 0.8376284  0.7128627 -0.27479669 2.7323549    F-test 0.001672555
#> 30  0.9084891 0.5805038  1.7284656 -1.10257177 2.0110609    F-test 0.001672555
#>      p_value k  null
#> 1  0.3870301 2 FALSE
#> 2  0.3870301 2 FALSE
#> 3  0.3870301 2 FALSE
#> 4  0.3870301 2 FALSE
#> 5  0.3870301 2 FALSE
#> 6  0.3870301 2 FALSE
#> 7  0.3870301 2 FALSE
#> 8  0.3870301 2 FALSE
#> 9  0.3870301 2 FALSE
#> 10 0.3870301 2 FALSE
#> 11 0.8610546 1  TRUE
#> 12 0.8610546 1  TRUE
#> 13 0.8610546 1  TRUE
#> 14 0.8610546 1  TRUE
#> 15 0.8610546 1  TRUE
#> 16 0.8610546 1  TRUE
#> 17 0.8610546 1  TRUE
#> 18 0.8610546 1  TRUE
#> 19 0.8610546 1  TRUE
#> 20 0.8610546 1  TRUE
#> 21 0.9685201 3  TRUE
#> 22 0.9685201 3  TRUE
#> 23 0.9685201 3  TRUE
#> 24 0.9685201 3  TRUE
#> 25 0.9685201 3  TRUE
#> 26 0.9685201 3  TRUE
#> 27 0.9685201 3  TRUE
#> 28 0.9685201 3  TRUE
#> 29 0.9685201 3  TRUE
#> 30 0.9685201 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>             y  e          x      .resid   .fitted test_name statistic
#> 1   2.0124652  1 0.01231358 -2.59255618  4.605021    F-test  1.323229
#> 2   3.3075587  2 0.24669857 -2.80634781  6.113907    F-test  1.323229
#> 3   4.2360478  3 0.19717132 -1.55902032  5.795068    F-test  1.323229
#> 4   5.5812468  4 0.41172737 -1.59505508  7.176302    F-test  1.323229
#> 5   7.9269819  5 0.97545989 -2.87842463 10.805407    F-test  1.323229
#> 6   7.9410324  6 0.59134432 -0.39157798  8.332610    F-test  1.323229
#> 7   8.4721513  7 0.34979485  1.69454838  6.777603    F-test  1.323229
#> 8   9.1291415  8 0.11574463  3.85826861  5.270873    F-test  1.323229
#> 9  10.6915350  9 0.47032729  3.13798823  7.553547    F-test  1.323229
#> 10 12.2053573 10 0.70638192  3.13217678  9.073181    F-test  1.323229
#> 11  8.3965069  1 0.01231358  3.79148550  4.605021    F-test 11.850210
#> 12  2.6249337  2 0.24669857 -3.48897286  6.113907    F-test 11.850210
#> 13  4.5475690  3 0.19717132 -1.24749917  5.795068    F-test 11.850210
#> 14  4.4214319  4 0.41172737 -2.75486992  7.176302    F-test 11.850210
#> 15 14.4626647  5 0.97545989  3.65725820 10.805407    F-test 11.850210
#> 16  8.2672561  6 0.59134432 -0.06535434  8.332610    F-test 11.850210
#> 17  8.1836229  7 0.34979485  1.40602004  6.777603    F-test 11.850210
#> 18  7.8017720  8 0.11574463  2.53089916  5.270873    F-test 11.850210
#> 19  5.0737874  9 0.47032729 -2.47975938  7.553547    F-test 11.850210
#> 20  7.7239733 10 0.70638192 -1.34920722  9.073181    F-test 11.850210
#> 21 -0.2065365  1 0.01231358 -4.81155789  4.605021    F-test  4.218876
#> 22 10.0343232  2 0.24669857  3.92041668  6.113907    F-test  4.218876
#> 23  7.3417912  3 0.19717132  1.54672307  5.795068    F-test  4.218876
#> 24  7.0022630  4 0.41172737 -0.17403889  7.176302    F-test  4.218876
#> 25  8.4279345  5 0.97545989 -2.37747201 10.805407    F-test  4.218876
#> 26  8.7514141  6 0.59134432  0.41880370  8.332610    F-test  4.218876
#> 27  4.5401142  7 0.34979485 -2.23748876  6.777603    F-test  4.218876
#> 28  5.3371945  8 0.11574463  0.06632170  5.270873    F-test  4.218876
#> 29 11.2470022  9 0.47032729  3.69345540  7.553547    F-test  4.218876
#> 30  9.0280176 10 0.70638192 -0.04516299  9.073181    F-test  4.218876
#>       p_value k  null
#> 1  0.28778966 2 FALSE
#> 2  0.28778966 2 FALSE
#> 3  0.28778966 2 FALSE
#> 4  0.28778966 2 FALSE
#> 5  0.28778966 2 FALSE
#> 6  0.28778966 2 FALSE
#> 7  0.28778966 2 FALSE
#> 8  0.28778966 2 FALSE
#> 9  0.28778966 2 FALSE
#> 10 0.28778966 2 FALSE
#> 11 0.01080179 1  TRUE
#> 12 0.01080179 1  TRUE
#> 13 0.01080179 1  TRUE
#> 14 0.01080179 1  TRUE
#> 15 0.01080179 1  TRUE
#> 16 0.01080179 1  TRUE
#> 17 0.01080179 1  TRUE
#> 18 0.01080179 1  TRUE
#> 19 0.01080179 1  TRUE
#> 20 0.01080179 1  TRUE
#> 21 0.07906335 3  TRUE
#> 22 0.07906335 3  TRUE
#> 23 0.07906335 3  TRUE
#> 24 0.07906335 3  TRUE
#> 25 0.07906335 3  TRUE
#> 26 0.07906335 3  TRUE
#> 27 0.07906335 3  TRUE
#> 28 0.07906335 3  TRUE
#> 29 0.07906335 3  TRUE
#> 30 0.07906335 3  TRUE