Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>             y          x           e      .resid   .fitted test_name  statistic
#> 1   4.5995932 0.94476497  1.76224735  1.55245888 3.0471343    F-test 2.87465001
#> 2   1.9239535 0.86068385 -0.67750707 -0.88555040 2.8095039    F-test 2.87465001
#> 3   2.5335331 0.70877890  0.32238664  0.15334362 2.3801894    F-test 2.87465001
#> 4   0.9172495 0.30486176 -0.48055291 -0.32138758 1.2386371    F-test 2.87465001
#> 5   1.1332305 0.04555588  0.08559925  0.62744473 0.5057857    F-test 2.87465001
#> 6   1.3335367 0.41889793 -0.26083673 -0.22738995 1.5609266    F-test 2.87465001
#> 7   1.4566440 0.45700410 -0.20921282 -0.21197843 1.6686225    F-test 2.87465001
#> 8   1.9441000 0.71448016 -0.28086208 -0.45220240 2.3963024    F-test 2.87465001
#> 9   0.5146839 0.03400378 -0.52047611  0.04154677 0.4731372    F-test 2.87465001
#> 10  2.4536575 0.83253261 -0.07198568 -0.27628524 2.7299427    F-test 2.87465001
#> 11  3.4434286 0.94476497  1.76224735  0.39629429 3.0471343    F-test 1.66446705
#> 12  2.8763735 0.86068385 -0.67750707  0.06686959 2.8095039    F-test 1.66446705
#> 13  1.2506287 0.70877890  0.32238664 -1.12956073 2.3801894    F-test 1.66446705
#> 14  1.5429538 0.30486176 -0.48055291  0.30431662 1.2386371    F-test 1.66446705
#> 15  0.6841210 0.04555588  0.08559925  0.17833524 0.5057857    F-test 1.66446705
#> 16  1.8240627 0.41889793 -0.26083673  0.26313603 1.5609266    F-test 1.66446705
#> 17  0.9505170 0.45700410 -0.20921282 -0.71810549 1.6686225    F-test 1.66446705
#> 18  1.6834666 0.71448016 -0.28086208 -0.71283581 2.3963024    F-test 1.66446705
#> 19  0.6361883 0.03400378 -0.52047611  0.16305116 0.4731372    F-test 1.66446705
#> 20  3.9184418 0.83253261 -0.07198568  1.18849909 2.7299427    F-test 1.66446705
#> 21  2.7612912 0.94476497  1.76224735 -0.28584308 3.0471343    F-test 0.03598124
#> 22  3.2346965 0.86068385 -0.67750707  0.42519263 2.8095039    F-test 0.03598124
#> 23  3.5089894 0.70877890  0.32238664  1.12879999 2.3801894    F-test 0.03598124
#> 24  1.2645851 0.30486176 -0.48055291  0.02594793 1.2386371    F-test 0.03598124
#> 25 -0.1308722 0.04555588  0.08559925 -0.63665797 0.5057857    F-test 0.03598124
#> 26  1.4931556 0.41889793 -0.26083673 -0.06777106 1.5609266    F-test 0.03598124
#> 27  1.0492833 0.45700410 -0.20921282 -0.61933916 1.6686225    F-test 0.03598124
#> 28  2.8155283 0.71448016 -0.28086208  0.41922587 2.3963024    F-test 0.03598124
#> 29  1.1477027 0.03400378 -0.52047611  0.67456555 0.4731372    F-test 0.03598124
#> 30  1.6658220 0.83253261 -0.07198568 -1.06412070 2.7299427    F-test 0.03598124
#>      p_value k  null
#> 1  0.1338052 1 FALSE
#> 2  0.1338052 1 FALSE
#> 3  0.1338052 1 FALSE
#> 4  0.1338052 1 FALSE
#> 5  0.1338052 1 FALSE
#> 6  0.1338052 1 FALSE
#> 7  0.1338052 1 FALSE
#> 8  0.1338052 1 FALSE
#> 9  0.1338052 1 FALSE
#> 10 0.1338052 1 FALSE
#> 11 0.2379825 2  TRUE
#> 12 0.2379825 2  TRUE
#> 13 0.2379825 2  TRUE
#> 14 0.2379825 2  TRUE
#> 15 0.2379825 2  TRUE
#> 16 0.2379825 2  TRUE
#> 17 0.2379825 2  TRUE
#> 18 0.2379825 2  TRUE
#> 19 0.2379825 2  TRUE
#> 20 0.2379825 2  TRUE
#> 21 0.8549374 3  TRUE
#> 22 0.8549374 3  TRUE
#> 23 0.8549374 3  TRUE
#> 24 0.8549374 3  TRUE
#> 25 0.8549374 3  TRUE
#> 26 0.8549374 3  TRUE
#> 27 0.8549374 3  TRUE
#> 28 0.8549374 3  TRUE
#> 29 0.8549374 3  TRUE
#> 30 0.8549374 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e          x      .resid  .fitted test_name statistic   p_value k
#> 1   2.729195  1 0.48954305 -4.28696235 7.016158    F-test 0.3949140 0.5496741 1
#> 2   4.055640  2 0.64264603 -3.67133524 7.726975    F-test 0.3949140 0.5496741 1
#> 3   4.106897  3 0.09740887 -1.08868323 5.195581    F-test 0.3949140 0.5496741 1
#> 4   6.097557  4 0.66084330 -1.71390321 7.811460    F-test 0.3949140 0.5496741 1
#> 5   7.900785  5 0.96655535 -1.33001748 9.230802    F-test 0.3949140 0.5496741 1
#> 6   8.755754  6 0.91624657 -0.24147712 8.997231    F-test 0.3949140 0.5496741 1
#> 7   8.083203  7 0.07723777  2.98127205 5.101931    F-test 0.3949140 0.5496741 1
#> 8  10.129117  8 0.67435809  2.25491087 7.874206    F-test 0.3949140 0.5496741 1
#> 9  11.358393  9 0.76822432  3.04839034 8.310003    F-test 0.3949140 0.5496741 1
#> 10 12.359098 10 0.76850208  4.04780538 8.311292    F-test 0.3949140 0.5496741 1
#> 11  5.499759  1 0.48954305 -1.51639920 7.016158    F-test 0.4935816 0.5050189 2
#> 12 11.651513  2 0.64264603  3.92453797 7.726975    F-test 0.4935816 0.5050189 2
#> 13  3.227530  3 0.09740887 -1.96805064 5.195581    F-test 0.4935816 0.5050189 2
#> 14  6.260149  4 0.66084330 -1.55131120 7.811460    F-test 0.4935816 0.5050189 2
#> 15 14.348462  5 0.96655535  5.11765958 9.230802    F-test 0.4935816 0.5050189 2
#> 16  5.472913  6 0.91624657 -3.52431841 8.997231    F-test 0.4935816 0.5050189 2
#> 17  7.998051  7 0.07723777  2.89611912 5.101931    F-test 0.4935816 0.5050189 2
#> 18  6.894269  8 0.67435809 -0.97993678 7.874206    F-test 0.4935816 0.5050189 2
#> 19  8.259874  9 0.76822432 -0.05012871 8.310003    F-test 0.4935816 0.5050189 2
#> 20  5.963120 10 0.76850208 -2.34817173 8.311292    F-test 0.4935816 0.5050189 2
#> 21  5.889146  1 0.48954305 -1.12701186 7.016158    F-test 0.8237446 0.3942585 3
#> 22  5.692998  2 0.64264603 -2.03397698 7.726975    F-test 0.8237446 0.3942585 3
#> 23  6.362653  3 0.09740887  1.16707277 5.195581    F-test 0.8237446 0.3942585 3
#> 24  4.147523  4 0.66084330 -3.66393700 7.811460    F-test 0.8237446 0.3942585 3
#> 25  7.907003  5 0.96655535 -1.32379914 9.230802    F-test 0.8237446 0.3942585 3
#> 26 14.011684  6 0.91624657  5.01445249 8.997231    F-test 0.8237446 0.3942585 3
#> 27  5.582278  7 0.07723777  0.48034626 5.101931    F-test 0.8237446 0.3942585 3
#> 28 11.372140  8 0.67435809  3.49793417 7.874206    F-test 0.8237446 0.3942585 3
#> 29  4.486898  9 0.76822432 -3.82310418 8.310003    F-test 0.8237446 0.3942585 3
#> 30 10.123316 10 0.76850208  1.81202347 8.311292    F-test 0.8237446 0.3942585 3
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE