Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

A data frame.

Examples

# Instantiate
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>              y          x          e      .resid   .fitted test_name  statistic
#> 1   0.35578786 0.68470726 -1.7977434 -2.05234744 2.4081353    F-test 0.27765754
#> 2   3.41657275 0.83735821  0.8780458  0.55752337 2.8590494    F-test 0.27765754
#> 3   3.67672407 0.87779764  1.0283977  0.69822104 2.9785030    F-test 0.27765754
#> 4   2.40447163 0.84275888 -0.1485298 -0.47053077 2.8750024    F-test 0.27765754
#> 5   1.99823224 0.45361672  0.3388474  0.27271296 1.7255193    F-test 0.27765754
#> 6   0.42004017 0.15067345 -0.7533358 -0.41061804 0.8306582    F-test 0.27765754
#> 7   0.09960734 0.27297568 -1.2478841 -1.09231820 1.1919255    F-test 0.27765754
#> 8   0.53309883 0.03210788 -0.5000400  0.05267026 0.4804286    F-test 0.27765754
#> 9   2.25336414 0.17419368  1.0488270  1.35322974 0.9001344    F-test 0.27765754
#> 10  3.42828547 0.66056726  1.3313691  1.09145707 2.3368284    F-test 0.27765754
#> 11  1.03096377 0.68470726 -1.7977434 -1.37717153 2.4081353    F-test 0.07456393
#> 12  3.84879702 0.83735821  0.8780458  0.98974764 2.8590494    F-test 0.07456393
#> 13  4.55967495 0.87779764  1.0283977  1.58117191 2.9785030    F-test 0.07456393
#> 14  1.43640887 0.84275888 -0.1485298 -1.43859352 2.8750024    F-test 0.07456393
#> 15  2.49815800 0.45361672  0.3388474  0.77263873 1.7255193    F-test 0.07456393
#> 16  0.51308805 0.15067345 -0.7533358 -0.31757015 0.8306582    F-test 0.07456393
#> 17  1.75824790 0.27297568 -1.2478841  0.56632236 1.1919255    F-test 0.07456393
#> 18  0.09145548 0.03210788 -0.5000400 -0.38897309 0.4804286    F-test 0.07456393
#> 19  1.41433473 0.17419368  1.0488270  0.51420033 0.9001344    F-test 0.07456393
#> 20  1.43505572 0.66056726  1.3313691 -0.90177267 2.3368284    F-test 0.07456393
#> 21  2.67029381 0.68470726 -1.7977434  0.26215851 2.4081353    F-test 1.37551411
#> 22  4.71497217 0.83735821  0.8780458  1.85592278 2.8590494    F-test 1.37551411
#> 23  2.83387022 0.87779764  1.0283977 -0.14463281 2.9785030    F-test 1.37551411
#> 24  1.82620838 0.84275888 -0.1485298 -1.04879402 2.8750024    F-test 1.37551411
#> 25  0.76708449 0.45361672  0.3388474 -0.95843478 1.7255193    F-test 1.37551411
#> 26 -0.18001213 0.15067345 -0.7533358 -1.01067034 0.8306582    F-test 1.37551411
#> 27  0.28415225 0.27297568 -1.2478841 -0.90777329 1.1919255    F-test 1.37551411
#> 28  1.49069435 0.03210788 -0.5000400  1.01026577 0.4804286    F-test 1.37551411
#> 29  2.01564763 0.17419368  1.0488270  1.11551323 0.9001344    F-test 1.37551411
#> 30  2.16327334 0.66056726  1.3313691 -0.17355505 2.3368284    F-test 1.37551411
#>      p_value k  null
#> 1  0.6145210 2 FALSE
#> 2  0.6145210 2 FALSE
#> 3  0.6145210 2 FALSE
#> 4  0.6145210 2 FALSE
#> 5  0.6145210 2 FALSE
#> 6  0.6145210 2 FALSE
#> 7  0.6145210 2 FALSE
#> 8  0.6145210 2 FALSE
#> 9  0.6145210 2 FALSE
#> 10 0.6145210 2 FALSE
#> 11 0.7926846 1  TRUE
#> 12 0.7926846 1  TRUE
#> 13 0.7926846 1  TRUE
#> 14 0.7926846 1  TRUE
#> 15 0.7926846 1  TRUE
#> 16 0.7926846 1  TRUE
#> 17 0.7926846 1  TRUE
#> 18 0.7926846 1  TRUE
#> 19 0.7926846 1  TRUE
#> 20 0.7926846 1  TRUE
#> 21 0.2792273 3  TRUE
#> 22 0.2792273 3  TRUE
#> 23 0.2792273 3  TRUE
#> 24 0.2792273 3  TRUE
#> 25 0.2792273 3  TRUE
#> 26 0.2792273 3  TRUE
#> 27 0.2792273 3  TRUE
#> 28 0.2792273 3  TRUE
#> 29 0.2792273 3  TRUE
#> 30 0.2792273 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e         x      .resid  .fitted test_name statistic   p_value k
#> 1   2.399301  1 0.3057921 -4.84566195 7.244963    F-test 0.3041551 0.5984446 1
#> 2   4.963599  2 0.9878169 -2.91723236 7.880831    F-test 0.3041551 0.5984446 1
#> 3   5.595449  3 0.8584732 -2.16479171 7.760241    F-test 0.3041551 0.5984446 1
#> 4   5.613602  4 0.4293021 -1.74651188 7.360114    F-test 0.3041551 0.5984446 1
#> 5   6.145132  5 0.1285955 -0.93462579 7.079758    F-test 0.3041551 0.5984446 1
#> 6   8.583918  6 0.8542221  0.82763985 7.756278    F-test 0.3041551 0.5984446 1
#> 7   9.758036  7 0.9170517  1.94318029 7.814855    F-test 0.3041551 0.5984446 1
#> 8   9.297286  8 0.2397877  2.11386053 7.183425    F-test 0.3041551 0.5984446 1
#> 9  10.324893  9 0.2582168  3.12428548 7.200607    F-test 0.3041551 0.5984446 1
#> 10 12.226404 10 0.7150735  4.59985755 7.626546    F-test 0.3041551 0.5984446 1
#> 11  3.930945  1 0.3057921 -3.31401758 7.244963    F-test 0.6494528 0.4468165 2
#> 12  2.257143  2 0.9878169 -5.62368807 7.880831    F-test 0.6494528 0.4468165 2
#> 13  6.196315  3 0.8584732 -1.56392636 7.760241    F-test 0.6494528 0.4468165 2
#> 14  9.006238  4 0.4293021  1.64612385 7.360114    F-test 0.6494528 0.4468165 2
#> 15  8.675790  5 0.1285955  1.59603237 7.079758    F-test 0.6494528 0.4468165 2
#> 16 11.678159  6 0.8542221  3.92188103 7.756278    F-test 0.6494528 0.4468165 2
#> 17 10.640615  7 0.9170517  2.82575945 7.814855    F-test 0.6494528 0.4468165 2
#> 18  7.289727  8 0.2397877  0.10630120 7.183425    F-test 0.6494528 0.4468165 2
#> 19  5.575952  9 0.2582168 -1.62465526 7.200607    F-test 0.6494528 0.4468165 2
#> 20  9.656735 10 0.7150735  2.03018938 7.626546    F-test 0.6494528 0.4468165 2
#> 21  7.158330  1 0.3057921 -0.08663336 7.244963    F-test 0.2178442 0.6548656 3
#> 22 10.061229  2 0.9878169  2.18039809 7.880831    F-test 0.2178442 0.6548656 3
#> 23  7.118016  3 0.8584732 -0.64222457 7.760241    F-test 0.2178442 0.6548656 3
#> 24  7.627141  4 0.4293021  0.26702641 7.360114    F-test 0.2178442 0.6548656 3
#> 25  4.778372  5 0.1285955 -2.30138610 7.079758    F-test 0.2178442 0.6548656 3
#> 26  5.943460  6 0.8542221 -1.81281812 7.756278    F-test 0.2178442 0.6548656 3
#> 27  5.114301  7 0.9170517 -2.70055400 7.814855    F-test 0.2178442 0.6548656 3
#> 28 13.056405  8 0.2397877  5.87297969 7.183425    F-test 0.2178442 0.6548656 3
#> 29  3.309688  9 0.2582168 -3.89091906 7.200607    F-test 0.2178442 0.6548656 3
#> 30 10.740677 10 0.7150735  3.11413102 7.626546    F-test 0.2178442 0.6548656 3
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE