Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>            y         x          e      .resid  .fitted test_name statistic
#> 1  3.0494851 0.4817672  1.3356183  0.58786756 2.461618    F-test 0.4197312
#> 2  1.0877339 0.6902096 -1.0788651 -1.70159504 2.789329    F-test 0.4197312
#> 3  3.7183826 0.1339645  2.5664717  1.80357749 1.914805    F-test 0.4197312
#> 4  3.6961307 0.7962763  1.2657985  0.74004472 2.956086    F-test 0.4197312
#> 5  3.5607036 0.8521672  0.9823473  0.51674629 3.043957    F-test 0.4197312
#> 6  3.9075527 0.9452775  1.0687257  0.71720833 3.190344    F-test 0.4197312
#> 7  2.1269489 0.5155621  0.3455825 -0.38780073 2.514750    F-test 0.4197312
#> 8  0.5231804 0.1133804 -0.6030552 -1.35926269 1.882443    F-test 0.4197312
#> 9  2.0739366 0.7321994 -0.1943789 -0.78140837 2.855345    F-test 0.4197312
#> 10 3.0991355 0.9733712  0.1783128 -0.13537755 3.234513    F-test 0.4197312
#> 11 0.7112802 0.4817672  1.3356183 -1.75033742 2.461618    F-test 1.0301378
#> 12 4.3032315 0.6902096 -1.0788651  1.51390258 2.789329    F-test 1.0301378
#> 13 2.1567166 0.1339645  2.5664717  0.24191152 1.914805    F-test 1.0301378
#> 14 2.8656374 0.7962763  1.2657985 -0.09044859 2.956086    F-test 1.0301378
#> 15 3.3594277 0.8521672  0.9823473  0.31547048 3.043957    F-test 1.0301378
#> 16 1.8760387 0.9452775  1.0687257 -1.31430570 3.190344    F-test 1.0301378
#> 17 3.1829782 0.5155621  0.3455825  0.66822854 2.514750    F-test 1.0301378
#> 18 1.3952047 0.1133804 -0.6030552 -0.48723840 1.882443    F-test 1.0301378
#> 19 4.3336959 0.7321994 -0.1943789  1.47835098 2.855345    F-test 1.0301378
#> 20 2.6589791 0.9733712  0.1783128 -0.57553398 3.234513    F-test 1.0301378
#> 21 1.5817266 0.4817672  1.3356183 -0.87989100 2.461618    F-test 2.4466503
#> 22 2.7333629 0.6902096 -1.0788651 -0.05596601 2.789329    F-test 2.4466503
#> 23 2.1750024 0.1339645  2.5664717  0.26019729 1.914805    F-test 2.4466503
#> 24 1.9174784 0.7962763  1.2657985 -1.03860764 2.956086    F-test 2.4466503
#> 25 2.3919477 0.8521672  0.9823473 -0.65200953 3.043957    F-test 2.4466503
#> 26 5.7763958 0.9452775  1.0687257  2.58605143 3.190344    F-test 2.4466503
#> 27 2.8189781 0.5155621  0.3455825  0.30422847 2.514750    F-test 2.4466503
#> 28 2.4809671 0.1133804 -0.6030552  0.59852399 1.882443    F-test 2.4466503
#> 29 1.8863129 0.7321994 -0.1943789 -0.96903203 2.855345    F-test 2.4466503
#> 30 3.0810181 0.9733712  0.1783128 -0.15349497 3.234513    F-test 2.4466503
#>      p_value k  null
#> 1  0.5377328 1 FALSE
#> 2  0.5377328 1 FALSE
#> 3  0.5377328 1 FALSE
#> 4  0.5377328 1 FALSE
#> 5  0.5377328 1 FALSE
#> 6  0.5377328 1 FALSE
#> 7  0.5377328 1 FALSE
#> 8  0.5377328 1 FALSE
#> 9  0.5377328 1 FALSE
#> 10 0.5377328 1 FALSE
#> 11 0.3439162 2  TRUE
#> 12 0.3439162 2  TRUE
#> 13 0.3439162 2  TRUE
#> 14 0.3439162 2  TRUE
#> 15 0.3439162 2  TRUE
#> 16 0.3439162 2  TRUE
#> 17 0.3439162 2  TRUE
#> 18 0.3439162 2  TRUE
#> 19 0.3439162 2  TRUE
#> 20 0.3439162 2  TRUE
#> 21 0.1617529 3  TRUE
#> 22 0.1617529 3  TRUE
#> 23 0.1617529 3  TRUE
#> 24 0.1617529 3  TRUE
#> 25 0.1617529 3  TRUE
#> 26 0.1617529 3  TRUE
#> 27 0.1617529 3  TRUE
#> 28 0.1617529 3  TRUE
#> 29 0.1617529 3  TRUE
#> 30 0.1617529 3  TRUE