Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>            y          x            e      .resid  .fitted test_name  statistic
#> 1  1.8132150 0.04869052  0.762153761 -0.25415972 2.067375    F-test 0.02301926
#> 2  1.6622531 0.31038178  0.255534463 -0.13408434 1.796337    F-test 0.02301926
#> 3  0.8837370 0.03822332 -0.155947332 -1.19447876 2.078216    F-test 0.02301926
#> 4  1.7126644 0.52783742 -0.093785407  0.14154880 1.571116    F-test 0.02301926
#> 5  1.4624771 0.64748146 -0.604236609  0.01527854 1.447199    F-test 0.02301926
#> 6  3.0584253 0.08761985  1.963128248  1.03137023 2.027055    F-test 0.02301926
#> 7  3.5166864 0.12037518  2.381821010  1.52355642 1.993130    F-test 0.02301926
#> 8  1.3236000 0.12968946  0.177091231 -0.65988297 1.983483    F-test 0.02301926
#> 9  1.4965674 0.36168993  0.004057862 -0.24662946 1.743197    F-test 0.02301926
#> 10 1.5045796 0.37723334 -0.014958722 -0.22251873 1.727098    F-test 0.02301926
#> 11 2.9052336 0.04869052  0.762153761  0.83785887 2.067375    F-test 6.34651421
#> 12 1.1828597 0.31038178  0.255534463 -0.61347770 1.796337    F-test 6.34651421
#> 13 2.1817788 0.03822332 -0.155947332  0.10356300 2.078216    F-test 6.34651421
#> 14 2.2586828 0.52783742 -0.093785407  0.68756719 1.571116    F-test 6.34651421
#> 15 2.0434005 0.64748146 -0.604236609  0.59620199 1.447199    F-test 6.34651421
#> 16 3.1620259 0.08761985  1.963128248  1.13497077 2.027055    F-test 6.34651421
#> 17 1.7313329 0.12037518  2.381821010 -0.26179709 1.993130    F-test 6.34651421
#> 18 0.9665270 0.12968946  0.177091231 -1.01695601 1.983483    F-test 6.34651421
#> 19 0.8158923 0.36168993  0.004057862 -0.92730453 1.743197    F-test 6.34651421
#> 20 1.1864719 0.37723334 -0.014958722 -0.54062648 1.727098    F-test 6.34651421
#> 21 0.1624849 0.04869052  0.762153761 -1.90488990 2.067375    F-test 0.05534993
#> 22 1.9751406 0.31038178  0.255534463  0.17880320 1.796337    F-test 0.05534993
#> 23 2.7322725 0.03822332 -0.155947332  0.65405668 2.078216    F-test 0.05534993
#> 24 1.8441771 0.52783742 -0.093785407  0.27306158 1.571116    F-test 0.05534993
#> 25 1.3419723 0.64748146 -0.604236609 -0.10522626 1.447199    F-test 0.05534993
#> 26 2.4019300 0.08761985  1.963128248  0.37487489 2.027055    F-test 0.05534993
#> 27 2.8117971 0.12037518  2.381821010  0.81866710 1.993130    F-test 0.05534993
#> 28 2.3662231 0.12968946  0.177091231  0.38274006 1.983483    F-test 0.05534993
#> 29 1.1799192 0.36168993  0.004057862 -0.56327769 1.743197    F-test 0.05534993
#> 30 1.6182887 0.37723334 -0.014958722 -0.10880966 1.727098    F-test 0.05534993
#>       p_value k  null
#> 1  0.88368713 3 FALSE
#> 2  0.88368713 3 FALSE
#> 3  0.88368713 3 FALSE
#> 4  0.88368713 3 FALSE
#> 5  0.88368713 3 FALSE
#> 6  0.88368713 3 FALSE
#> 7  0.88368713 3 FALSE
#> 8  0.88368713 3 FALSE
#> 9  0.88368713 3 FALSE
#> 10 0.88368713 3 FALSE
#> 11 0.03985544 1  TRUE
#> 12 0.03985544 1  TRUE
#> 13 0.03985544 1  TRUE
#> 14 0.03985544 1  TRUE
#> 15 0.03985544 1  TRUE
#> 16 0.03985544 1  TRUE
#> 17 0.03985544 1  TRUE
#> 18 0.03985544 1  TRUE
#> 19 0.03985544 1  TRUE
#> 20 0.03985544 1  TRUE
#> 21 0.82073690 2  TRUE
#> 22 0.82073690 2  TRUE
#> 23 0.82073690 2  TRUE
#> 24 0.82073690 2  TRUE
#> 25 0.82073690 2  TRUE
#> 26 0.82073690 2  TRUE
#> 27 0.82073690 2  TRUE
#> 28 0.82073690 2  TRUE
#> 29 0.82073690 2  TRUE
#> 30 0.82073690 2  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e         x     .resid  .fitted test_name statistic   p_value k
#> 1   3.284817  1 0.7388774 -3.3478740 6.632691    F-test 2.3671570 0.1678024 2
#> 2   4.131582  2 0.6754073 -2.8058217 6.937404    F-test 2.3671570 0.1678024 2
#> 3   4.723543  3 0.4866829 -3.1199054 7.843448    F-test 2.3671570 0.1678024 2
#> 4   6.651429  4 0.8789231  0.6910816 5.960347    F-test 2.3671570 0.1678024 2
#> 5   6.588890  5 0.4159092 -1.5943350 8.183225    F-test 2.3671570 0.1678024 2
#> 6   8.879248  6 0.9591944  3.3042745 5.574974    F-test 2.3671570 0.1678024 2
#> 7   8.685821  7 0.4673783  0.7496933 7.936127    F-test 2.3671570 0.1678024 2
#> 8   9.833906  8 0.5411080  2.2517461 7.582160    F-test 2.3671570 0.1678024 2
#> 9  10.211798  9 0.1795572  0.8938734 9.317925    F-test 2.3671570 0.1678024 2
#> 10 11.473655 10 0.3506790  2.9772671 8.496388    F-test 2.3671570 0.1678024 2
#> 11  8.826663  1 0.7388774  2.1939714 6.632691    F-test 1.8430599 0.2167312 1
#> 12  6.722624  2 0.6754073 -0.2147799 6.937404    F-test 1.8430599 0.2167312 1
#> 13  7.123991  3 0.4866829 -0.7194571 7.843448    F-test 1.8430599 0.2167312 1
#> 14  5.489384  4 0.8789231 -0.4709636 5.960347    F-test 1.8430599 0.2167312 1
#> 15  6.635329  5 0.4159092 -1.5478961 8.183225    F-test 1.8430599 0.2167312 1
#> 16  7.009829  6 0.9591944  1.4348555 5.574974    F-test 1.8430599 0.2167312 1
#> 17 10.821011  7 0.4673783  2.8848838 7.936127    F-test 1.8430599 0.2167312 1
#> 18  2.012676  8 0.5411080 -5.5694836 7.582160    F-test 1.8430599 0.2167312 1
#> 19 12.143708  9 0.1795572  2.8257835 9.317925    F-test 1.8430599 0.2167312 1
#> 20  7.679474 10 0.3506790 -0.8169139 8.496388    F-test 1.8430599 0.2167312 1
#> 21  4.707091  1 0.7388774 -1.9256002 6.632691    F-test 0.5034401 0.5009365 3
#> 22  5.442228  2 0.6754073 -1.4951761 6.937404    F-test 0.5034401 0.5009365 3
#> 23  9.286304  3 0.4866829  1.4428557 7.843448    F-test 0.5034401 0.5009365 3
#> 24  8.880984  4 0.8789231  2.9206368 5.960347    F-test 0.5034401 0.5009365 3
#> 25  7.875210  5 0.4159092 -0.3080152 8.183225    F-test 0.5034401 0.5009365 3
#> 26  5.090990  6 0.9591944 -0.4839837 5.574974    F-test 0.5034401 0.5009365 3
#> 27  3.296696  7 0.4673783 -4.6394310 7.936127    F-test 0.5034401 0.5009365 3
#> 28 11.282673  8 0.5411080  3.7005137 7.582160    F-test 0.5034401 0.5009365 3
#> 29 11.390515  9 0.1795572  2.0725901 9.317925    F-test 0.5034401 0.5009365 3
#> 30  7.211998 10 0.3506790 -1.2843901 8.496388    F-test 0.5034401 0.5009365 3
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE