Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>              y          x          e       .resid    .fitted test_name
#> 1  -0.32454256 0.30085960 -1.7159187 -0.434260151  0.1097176    F-test
#> 2  -0.52749678 0.02907391 -1.5574160  0.308986461 -0.8364832    F-test
#> 3   2.07334473 0.93499657 -0.7358704 -0.244071624  2.3174164    F-test
#> 4  -2.14478899 0.14800529 -3.3146998 -1.722356142 -0.4224328    F-test
#> 5   4.43669341 0.60143044  2.4735444  3.280561705  1.1561317    F-test
#> 6  -0.03381031 0.34569943 -1.4990178 -0.299634317  0.2658240    F-test
#> 7   0.60198561 0.62633143 -1.4166369 -0.640836957  1.2428226    F-test
#> 8   2.28069256 0.91547932 -0.4728891  0.031224028  2.2494685    F-test
#> 9   2.86881262 0.57854717  0.9555486  1.792347252  1.0764654    F-test
#> 10  0.23124910 0.93091577 -2.5662708 -2.071960255  2.3032094    F-test
#> 11  1.29737785 0.30085960 -1.7159187  1.187660257  0.1097176    F-test
#> 12 -2.19788689 0.02907391 -1.5574160 -1.361403658 -0.8364832    F-test
#> 13  0.03682263 0.93499657 -0.7358704 -2.280593718  2.3174164    F-test
#> 14 -2.10176188 0.14800529 -3.3146998 -1.679329037 -0.4224328    F-test
#> 15  1.69088074 0.60143044  2.4735444  0.534749040  1.1561317    F-test
#> 16  1.64757702 0.34569943 -1.4990178  1.381753017  0.2658240    F-test
#> 17  4.12931366 0.62633143 -1.4166369  2.886491094  1.2428226    F-test
#> 18  1.77468919 0.91547932 -0.4728891 -0.474779344  2.2494685    F-test
#> 19  0.87289533 0.57854717  0.9555486 -0.203570042  1.0764654    F-test
#> 20  2.31223174 0.93091577 -2.5662708  0.009022389  2.3032094    F-test
#> 21  1.26798773 0.30085960 -1.7159187  1.158270145  0.1097176    F-test
#> 22  0.76239712 0.02907391 -1.5574160  1.598880357 -0.8364832    F-test
#> 23  2.01494048 0.93499657 -0.7358704 -0.302475876  2.3174164    F-test
#> 24 -2.51575667 0.14800529 -3.3146998 -2.093323828 -0.4224328    F-test
#> 25  2.61115603 0.60143044  2.4735444  1.455024324  1.1561317    F-test
#> 26 -1.27312886 0.34569943 -1.4990178 -1.538952867  0.2658240    F-test
#> 27  0.14201394 0.62633143 -1.4166369 -1.100808631  1.2428226    F-test
#> 28  0.36847100 0.91547932 -0.4728891 -1.880997526  2.2494685    F-test
#> 29  1.79013431 0.57854717  0.9555486  0.713668943  1.0764654    F-test
#> 30  4.29392432 0.93091577 -2.5662708  1.990714961  2.3032094    F-test
#>     statistic    p_value k  null
#> 1  1.85324654 0.21560117 2 FALSE
#> 2  1.85324654 0.21560117 2 FALSE
#> 3  1.85324654 0.21560117 2 FALSE
#> 4  1.85324654 0.21560117 2 FALSE
#> 5  1.85324654 0.21560117 2 FALSE
#> 6  1.85324654 0.21560117 2 FALSE
#> 7  1.85324654 0.21560117 2 FALSE
#> 8  1.85324654 0.21560117 2 FALSE
#> 9  1.85324654 0.21560117 2 FALSE
#> 10 1.85324654 0.21560117 2 FALSE
#> 11 7.89901280 0.02612648 1  TRUE
#> 12 7.89901280 0.02612648 1  TRUE
#> 13 7.89901280 0.02612648 1  TRUE
#> 14 7.89901280 0.02612648 1  TRUE
#> 15 7.89901280 0.02612648 1  TRUE
#> 16 7.89901280 0.02612648 1  TRUE
#> 17 7.89901280 0.02612648 1  TRUE
#> 18 7.89901280 0.02612648 1  TRUE
#> 19 7.89901280 0.02612648 1  TRUE
#> 20 7.89901280 0.02612648 1  TRUE
#> 21 0.04153412 0.84430859 3  TRUE
#> 22 0.04153412 0.84430859 3  TRUE
#> 23 0.04153412 0.84430859 3  TRUE
#> 24 0.04153412 0.84430859 3  TRUE
#> 25 0.04153412 0.84430859 3  TRUE
#> 26 0.04153412 0.84430859 3  TRUE
#> 27 0.04153412 0.84430859 3  TRUE
#> 28 0.04153412 0.84430859 3  TRUE
#> 29 0.04153412 0.84430859 3  TRUE
#> 30 0.04153412 0.84430859 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e          x       .resid  .fitted test_name  statistic   p_value
#> 1   3.508418  1 0.82605337 -1.926434599 5.434852    F-test 1.03162788 0.3435901
#> 2   4.163803  2 0.68903458 -1.837496560 6.001300    F-test 1.03162788 0.3435901
#> 3   4.442420  3 0.33211763 -3.034405865 7.476826    F-test 1.03162788 0.3435901
#> 4   5.210454  4 0.17856724 -2.901162789 8.111616    F-test 1.03162788 0.3435901
#> 5   6.617034  5 0.43114648 -0.450397950 7.067432    F-test 1.03162788 0.3435901
#> 6   8.552450  6 0.84255355  3.185811097 5.366639    F-test 1.03162788 0.3435901
#> 7   8.187411  7 0.16137063  0.004702503 8.182709    F-test 1.03162788 0.3435901
#> 8   9.839141  8 0.54361936  3.236682396 6.602459    F-test 1.03162788 0.3435901
#> 9  10.070076  9 0.06575303  1.492076376 8.578000    F-test 1.03162788 0.3435901
#> 10 11.015867 10 0.01562306  2.230625391 8.785242    F-test 1.03162788 0.3435901
#> 11  6.811157  1 0.82605337  1.376305129 5.434852    F-test 0.00714281 0.9350132
#> 12  5.162366  2 0.68903458 -0.838933599 6.001300    F-test 0.00714281 0.9350132
#> 13  5.578102  3 0.33211763 -1.898724086 7.476826    F-test 0.00714281 0.9350132
#> 14 13.560591  4 0.17856724  5.448974365 8.111616    F-test 0.00714281 0.9350132
#> 15  8.468635  5 0.43114648  1.401203768 7.067432    F-test 0.00714281 0.9350132
#> 16  4.902834  6 0.84255355 -0.463805420 5.366639    F-test 0.00714281 0.9350132
#> 17  5.938170  7 0.16137063 -2.244538202 8.182709    F-test 0.00714281 0.9350132
#> 18  5.149772  8 0.54361936 -1.452687207 6.602459    F-test 0.00714281 0.9350132
#> 19  9.736123  9 0.06575303  1.158123022 8.578000    F-test 0.00714281 0.9350132
#> 20  6.299324 10 0.01562306 -2.485917770 8.785242    F-test 0.00714281 0.9350132
#> 21  6.147802  1 0.82605337  0.712949506 5.434852    F-test 1.48674053 0.2622101
#> 22  4.876695  2 0.68903458 -1.124604585 6.001300    F-test 1.48674053 0.2622101
#> 23 11.092708  3 0.33211763  3.615882817 7.476826    F-test 1.48674053 0.2622101
#> 24  8.683319  4 0.17856724  0.571702924 8.111616    F-test 1.48674053 0.2622101
#> 25  3.121500  5 0.43114648 -3.945931958 7.067432    F-test 1.48674053 0.2622101
#> 26  8.221632  6 0.84255355  2.854993535 5.366639    F-test 1.48674053 0.2622101
#> 27  9.018867  7 0.16137063  0.836158203 8.182709    F-test 1.48674053 0.2622101
#> 28  2.983157  8 0.54361936 -3.619302257 6.602459    F-test 1.48674053 0.2622101
#> 29  8.773870  9 0.06575303  0.195869977 8.578000    F-test 1.48674053 0.2622101
#> 30  8.687524 10 0.01562306 -0.097718162 8.785242    F-test 1.48674053 0.2622101
#>    k  null
#> 1  1 FALSE
#> 2  1 FALSE
#> 3  1 FALSE
#> 4  1 FALSE
#> 5  1 FALSE
#> 6  1 FALSE
#> 7  1 FALSE
#> 8  1 FALSE
#> 9  1 FALSE
#> 10 1 FALSE
#> 11 2  TRUE
#> 12 2  TRUE
#> 13 2  TRUE
#> 14 2  TRUE
#> 15 2  TRUE
#> 16 2  TRUE
#> 17 2  TRUE
#> 18 2  TRUE
#> 19 2  TRUE
#> 20 2  TRUE
#> 21 3  TRUE
#> 22 3  TRUE
#> 23 3  TRUE
#> 24 3  TRUE
#> 25 3  TRUE
#> 26 3  TRUE
#> 27 3  TRUE
#> 28 3  TRUE
#> 29 3  TRUE
#> 30 3  TRUE