Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic.

Value

A data frame.

Examples

# Instantiate
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>             y          x            e       .resid  .fitted test_name
#> 1   1.9249476 0.76725533 -0.430988445 -0.873754108 2.798702    F-test
#> 2   2.0852221 0.79365558 -0.338322628 -0.782111129 2.867333    F-test
#> 3   3.5135030 0.93103977  0.715628137  0.289018284 3.224485    F-test
#> 4   0.1203557 0.07812263 -0.963870064 -0.886839154 1.007195    F-test
#> 5   5.0994314 0.69677176  2.917168790  2.483962624 2.615469    F-test
#> 6   1.6752950 0.61869833 -0.326190932 -0.737209854 2.412505    F-test
#> 7   2.1522678 0.16787863  0.956205980  0.911738395 1.240529    F-test
#> 8   3.1056570 0.85623219  0.516291210  0.075646142 3.030011    F-test
#> 9   2.2876769 0.58250399  0.365862025 -0.030735185 2.318412    F-test
#> 10  2.5698771 0.85222483 -0.008634934 -0.449716014 3.019593    F-test
#> 11  3.7406749 0.76725533 -0.430988445  0.941973192 2.798702    F-test
#> 12  0.8526322 0.79365558 -0.338322628 -2.014701003 2.867333    F-test
#> 13  4.0056352 0.93103977  0.715628137  0.781150472 3.224485    F-test
#> 14  2.2154586 0.07812263 -0.963870064  1.208263723 1.007195    F-test
#> 15  3.5524564 0.69677176  2.917168790  0.936987593 2.615469    F-test
#> 16  1.4618430 0.61869833 -0.326190932 -0.950661889 2.412505    F-test
#> 17  0.1655248 0.16787863  0.956205980 -1.075004602 1.240529    F-test
#> 18  3.2070039 0.85623219  0.516291210  0.176993105 3.030011    F-test
#> 19  2.2329126 0.58250399  0.365862025 -0.085499522 2.318412    F-test
#> 20  3.1000920 0.85222483 -0.008634934  0.080498931 3.019593    F-test
#> 21  3.2363847 0.76725533 -0.430988445  0.437682945 2.798702    F-test
#> 22  3.9369368 0.79365558 -0.338322628  1.069603507 2.867333    F-test
#> 23  3.4388798 0.93103977  0.715628137  0.214395119 3.224485    F-test
#> 24 -0.1931027 0.07812263 -0.963870064 -1.200297545 1.007195    F-test
#> 25  0.5512502 0.69677176  2.917168790 -2.064218643 2.615469    F-test
#> 26  2.2982043 0.61869833 -0.326190932 -0.114300561 2.412505    F-test
#> 27  2.9297300 0.16787863  0.956205980  1.689200538 1.240529    F-test
#> 28  2.8562632 0.85623219  0.516291210 -0.173747628 3.030011    F-test
#> 29  2.4559692 0.58250399  0.365862025  0.137557104 2.318412    F-test
#> 30  3.0237182 0.85222483 -0.008634934  0.004125164 3.019593    F-test
#>     statistic   p_value k  null
#> 1  0.18123589 0.6830965 3 FALSE
#> 2  0.18123589 0.6830965 3 FALSE
#> 3  0.18123589 0.6830965 3 FALSE
#> 4  0.18123589 0.6830965 3 FALSE
#> 5  0.18123589 0.6830965 3 FALSE
#> 6  0.18123589 0.6830965 3 FALSE
#> 7  0.18123589 0.6830965 3 FALSE
#> 8  0.18123589 0.6830965 3 FALSE
#> 9  0.18123589 0.6830965 3 FALSE
#> 10 0.18123589 0.6830965 3 FALSE
#> 11 0.80307707 0.3999421 1  TRUE
#> 12 0.80307707 0.3999421 1  TRUE
#> 13 0.80307707 0.3999421 1  TRUE
#> 14 0.80307707 0.3999421 1  TRUE
#> 15 0.80307707 0.3999421 1  TRUE
#> 16 0.80307707 0.3999421 1  TRUE
#> 17 0.80307707 0.3999421 1  TRUE
#> 18 0.80307707 0.3999421 1  TRUE
#> 19 0.80307707 0.3999421 1  TRUE
#> 20 0.80307707 0.3999421 1  TRUE
#> 21 0.02302314 0.8836774 2  TRUE
#> 22 0.02302314 0.8836774 2  TRUE
#> 23 0.02302314 0.8836774 2  TRUE
#> 24 0.02302314 0.8836774 2  TRUE
#> 25 0.02302314 0.8836774 2  TRUE
#> 26 0.02302314 0.8836774 2  TRUE
#> 27 0.02302314 0.8836774 2  TRUE
#> 28 0.02302314 0.8836774 2  TRUE
#> 29 0.02302314 0.8836774 2  TRUE
#> 30 0.02302314 0.8836774 2  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>            y  e         x      .resid  .fitted test_name statistic    p_value k
#> 1   2.112949  1 0.1024523 -2.87330867 4.986257    F-test 2.2219497 0.17968149 1
#> 2   3.153171  2 0.1349574 -2.00906486 5.162236    F-test 2.2219497 0.17968149 1
#> 3   5.562918  3 0.8464466 -3.45122050 9.014139    F-test 2.2219497 0.17968149 1
#> 4   6.453081  4 0.8050213 -2.33678754 8.789868    F-test 2.2219497 0.17968149 1
#> 5   6.950377  5 0.5956172 -0.70580644 7.656183    F-test 2.2219497 0.17968149 1
#> 6   7.645675  6 0.4464011  0.79732658 6.848349    F-test 2.2219497 0.17968149 1
#> 7   9.974446  7 0.9914577  0.17523714 9.799209    F-test 2.2219497 0.17968149 1
#> 8   9.771362  8 0.5106247  2.57531632 7.196046    F-test 2.2219497 0.17968149 1
#> 9  11.433941  9 0.7976675  2.68388521 8.750056    F-test 2.2219497 0.17968149 1
#> 10 11.473247 10 0.3504392  5.14442276 6.328824    F-test 2.2219497 0.17968149 1
#> 11  8.634383  1 0.1024523  3.64812549 4.986257    F-test 4.3304994 0.07597455 2
#> 12  4.511254  2 0.1349574 -0.65098154 5.162236    F-test 4.3304994 0.07597455 2
#> 13  9.464218  3 0.8464466  0.45007942 9.014139    F-test 4.3304994 0.07597455 2
#> 14 11.707306  4 0.8050213  2.91743815 8.789868    F-test 4.3304994 0.07597455 2
#> 15  3.245063  5 0.5956172 -4.41111997 7.656183    F-test 4.3304994 0.07597455 2
#> 16  7.681936  6 0.4464011  0.83358781 6.848349    F-test 4.3304994 0.07597455 2
#> 17 11.725377  7 0.9914577  1.92616753 9.799209    F-test 4.3304994 0.07597455 2
#> 18  2.292634  8 0.5106247 -4.90341193 7.196046    F-test 4.3304994 0.07597455 2
#> 19  8.229557  9 0.7976675 -0.52049913 8.750056    F-test 4.3304994 0.07597455 2
#> 20  7.039438 10 0.3504392  0.71061419 6.328824    F-test 4.3304994 0.07597455 2
#> 21  5.083401  1 0.1024523  0.09714327 4.986257    F-test 0.0137285 0.91001734 3
#> 22  6.180789  2 0.1349574  1.01855326 5.162236    F-test 0.0137285 0.91001734 3
#> 23  9.219716  3 0.8464466  0.20557725 9.014139    F-test 0.0137285 0.91001734 3
#> 24 13.742042  4 0.8050213  4.95217371 8.789868    F-test 0.0137285 0.91001734 3
#> 25  9.603979  5 0.5956172  1.94779552 7.656183    F-test 0.0137285 0.91001734 3
#> 26  1.265167  6 0.4464011 -5.58318149 6.848349    F-test 0.0137285 0.91001734 3
#> 27  8.391411  7 0.9914577 -1.40779824 9.799209    F-test 0.0137285 0.91001734 3
#> 28  7.496942  8 0.5106247  0.30089624 7.196046    F-test 0.0137285 0.91001734 3
#> 29  6.073290  9 0.7976675 -2.67676619 8.750056    F-test 0.0137285 0.91001734 3
#> 30  7.474431 10 0.3504392  1.14560667 6.328824    F-test 0.0137285 0.91001734 3
#>     null
#> 1  FALSE
#> 2  FALSE
#> 3  FALSE
#> 4  FALSE
#> 5  FALSE
#> 6  FALSE
#> 7  FALSE
#> 8  FALSE
#> 9  FALSE
#> 10 FALSE
#> 11  TRUE
#> 12  TRUE
#> 13  TRUE
#> 14  TRUE
#> 15  TRUE
#> 16  TRUE
#> 17  TRUE
#> 18  TRUE
#> 19  TRUE
#> 20  TRUE
#> 21  TRUE
#> 22  TRUE
#> 23  TRUE
#> 24  TRUE
#> 25  TRUE
#> 26  TRUE
#> 27  TRUE
#> 28  TRUE
#> 29  TRUE
#> 30  TRUE