Generating random values from the expression of y, and forms a lineup

This function generates random values from the expression of y, and keeps all the right hand side information in a data frame.

Arguments

n

Integer. Number of observations.

k

Integer. Number of plots in the lineup. Default is k = 20.

pos

Integer. Position of the true data plot. Default is pos = NULL, which means the position is random.

computed

List. Default is NULL. If it is provided, random variables or random closed form expression will use the values from the list, which makes the expression potentially deterministic. In this function, only the actual data plot will use these computed values.

Value

A data frame.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)
#>              y          x          e      .resid   .fitted test_name statistic
#> 1   2.68060394 0.36969843  1.1742286  1.19283877 1.4877652    F-test 0.9785859
#> 2   3.31504130 0.92095060  0.5459407  0.12353830 3.1915030    F-test 0.9785859
#> 3   0.32149852 0.25573015 -0.9996295 -0.81402843 1.1355270    F-test 0.9785859
#> 4   0.74677702 0.03556033 -0.2900478  0.29172204 0.4550550    F-test 0.9785859
#> 5   1.65018516 0.15611184  0.4697024  0.82254540 0.8276398    F-test 0.9785859
#> 6   0.66005460 0.36800498 -0.8433781 -0.82247668 1.4825313    F-test 0.9785859
#> 7   1.63808183 0.66048648 -0.4586470 -0.74841285 2.3864947    F-test 0.9785859
#> 8   3.13154089 0.93340177  0.3269003 -0.09844454 3.2299854    F-test 0.9785859
#> 9   0.50585691 0.19889838 -0.7326020 -0.45402187 0.9598788    F-test 0.9785859
#> 10  3.86369826 0.97448445  0.9395939  0.50673986 3.3569584    F-test 0.9785859
#> 11  0.68537853 0.36969843  1.1742286 -0.80238664 1.4877652    F-test 5.5428881
#> 12  2.91668178 0.92095060  0.5459407 -0.27482123 3.1915030    F-test 5.5428881
#> 13 -0.15992896 0.25573015 -0.9996295 -1.29545592 1.1355270    F-test 5.5428881
#> 14  1.25727107 0.03556033 -0.2900478  0.80221609 0.4550550    F-test 5.5428881
#> 15  1.62733416 0.15611184  0.4697024  0.79969439 0.8276398    F-test 5.5428881
#> 16  1.68517034 0.36800498 -0.8433781  0.20263906 1.4825313    F-test 5.5428881
#> 17  1.87045932 0.66048648 -0.4586470 -0.51603536 2.3864947    F-test 5.5428881
#> 18  3.89455663 0.93340177  0.3269003  0.66457120 3.2299854    F-test 5.5428881
#> 19  1.00534917 0.19889838 -0.7326020  0.04547040 0.9598788    F-test 5.5428881
#> 20  3.73106641 0.97448445  0.9395939  0.37410801 3.3569584    F-test 5.5428881
#> 21  0.57348750 0.36969843  1.1742286 -0.91427766 1.4877652    F-test 2.9812253
#> 22  3.31078923 0.92095060  0.5459407  0.11928623 3.1915030    F-test 2.9812253
#> 23  1.80057525 0.25573015 -0.9996295  0.66504829 1.1355270    F-test 2.9812253
#> 24  0.81638008 0.03556033 -0.2900478  0.36132511 0.4550550    F-test 2.9812253
#> 25  1.51876014 0.15611184  0.4697024  0.69112038 0.8276398    F-test 2.9812253
#> 26 -0.09182046 0.36800498 -0.8433781 -1.57435174 1.4825313    F-test 2.9812253
#> 27  2.51694262 0.66048648 -0.4586470  0.13044793 2.3864947    F-test 2.9812253
#> 28  3.64012385 0.93340177  0.3269003  0.41013841 3.2299854    F-test 2.9812253
#> 29  1.03814392 0.19889838 -0.7326020  0.07826515 0.9598788    F-test 2.9812253
#> 30  3.38995631 0.97448445  0.9395939  0.03299790 3.3569584    F-test 2.9812253
#>       p_value k  null
#> 1  0.35550162 1 FALSE
#> 2  0.35550162 1 FALSE
#> 3  0.35550162 1 FALSE
#> 4  0.35550162 1 FALSE
#> 5  0.35550162 1 FALSE
#> 6  0.35550162 1 FALSE
#> 7  0.35550162 1 FALSE
#> 8  0.35550162 1 FALSE
#> 9  0.35550162 1 FALSE
#> 10 0.35550162 1 FALSE
#> 11 0.05076271 2  TRUE
#> 12 0.05076271 2  TRUE
#> 13 0.05076271 2  TRUE
#> 14 0.05076271 2  TRUE
#> 15 0.05076271 2  TRUE
#> 16 0.05076271 2  TRUE
#> 17 0.05076271 2  TRUE
#> 18 0.05076271 2  TRUE
#> 19 0.05076271 2  TRUE
#> 20 0.05076271 2  TRUE
#> 21 0.12787769 3  TRUE
#> 22 0.12787769 3  TRUE
#> 23 0.12787769 3  TRUE
#> 24 0.12787769 3  TRUE
#> 25 0.12787769 3  TRUE
#> 26 0.12787769 3  TRUE
#> 27 0.12787769 3  TRUE
#> 28 0.12787769 3  TRUE
#> 29 0.12787769 3  TRUE
#> 30 0.12787769 3  TRUE

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
#>             y  e            x      .resid  .fitted test_name statistic
#> 1   2.8680445  1 0.5573762537 -4.70421469 7.572259    F-test  0.308427
#> 2   3.5449789  2 0.3916158732 -3.44700281 6.991982    F-test  0.308427
#> 3   4.0166796  3 0.0164102563 -1.66181939 5.678499    F-test  0.308427
#> 4   5.0004839  4 0.0004836891 -0.62226086 5.622745    F-test  0.308427
#> 5   7.9984221  5 0.9994739308 -1.12148902 9.119911    F-test  0.308427
#> 6   8.8179641  6 0.9380417725 -0.08689132 8.904855    F-test  0.308427
#> 7   8.8633099  7 0.5551350075  1.29889658 7.564413    F-test  0.308427
#> 8  10.2143039  8 0.7100842546  2.10745959 8.106844    F-test  0.308427
#> 9  10.3799727  9 0.2937081754  3.73073658 6.649236    F-test  0.308427
#> 10 11.5947051 10 0.4190783838  4.50658535 7.088120    F-test  0.308427
#> 11  7.3090976  1 0.5573762537 -0.26316168 7.572259    F-test  2.267943
#> 12 10.6410894  2 0.3916158732  3.64910770 6.991982    F-test  2.267943
#> 13 -0.9861649  3 0.0164102563 -6.66466389 5.678499    F-test  2.267943
#> 14  7.7321194  4 0.0004836891  2.10937465 5.622745    F-test  2.267943
#> 15  8.8201971  5 0.9994739308 -0.29971401 9.119911    F-test  2.267943
#> 16  6.0233237  6 0.9380417725 -2.88153180 8.904855    F-test  2.267943
#> 17  7.1733064  7 0.5551350075 -0.39110694 7.564413    F-test  2.267943
#> 18  9.2075685  8 0.7100842546  1.10072417 8.106844    F-test  2.267943
#> 19  8.7375205  9 0.2937081754  2.08828441 6.649236    F-test  2.267943
#> 20  8.6408071 10 0.4190783838  1.55268740 7.088120    F-test  2.267943
#> 21  4.8915565  1 0.5573762537 -2.68070278 7.572259    F-test  2.148826
#> 22  1.6050355  2 0.3916158732 -5.38694617 6.991982    F-test  2.148826
#> 23  7.7482160  3 0.0164102563  2.06971702 5.678499    F-test  2.148826
#> 24  8.5347656  4 0.0004836891  2.91202086 5.622745    F-test  2.148826
#> 25  9.5186049  5 0.9994739308  0.39869377 9.119911    F-test  2.148826
#> 26 10.5378471  6 0.9380417725  1.63299167 8.904855    F-test  2.148826
#> 27 10.4290737  7 0.5551350075  2.86466044 7.564413    F-test  2.148826
#> 28  9.1395430  8 0.7100842546  1.03269869 8.106844    F-test  2.148826
#> 29  2.6130173  9 0.2937081754 -4.03621878 6.649236    F-test  2.148826
#> 30  8.2812050 10 0.4190783838  1.19308527 7.088120    F-test  2.148826
#>      p_value k  null
#> 1  0.5959408 1 FALSE
#> 2  0.5959408 1 FALSE
#> 3  0.5959408 1 FALSE
#> 4  0.5959408 1 FALSE
#> 5  0.5959408 1 FALSE
#> 6  0.5959408 1 FALSE
#> 7  0.5959408 1 FALSE
#> 8  0.5959408 1 FALSE
#> 9  0.5959408 1 FALSE
#> 10 0.5959408 1 FALSE
#> 11 0.1757970 2  TRUE
#> 12 0.1757970 2  TRUE
#> 13 0.1757970 2  TRUE
#> 14 0.1757970 2  TRUE
#> 15 0.1757970 2  TRUE
#> 16 0.1757970 2  TRUE
#> 17 0.1757970 2  TRUE
#> 18 0.1757970 2  TRUE
#> 19 0.1757970 2  TRUE
#> 20 0.1757970 2  TRUE
#> 21 0.1861055 3  TRUE
#> 22 0.1861055 3  TRUE
#> 23 0.1861055 3  TRUE
#> 24 0.1861055 3  TRUE
#> 25 0.1861055 3  TRUE
#> 26 0.1861055 3  TRUE
#> 27 0.1861055 3  TRUE
#> 28 0.1861055 3  TRUE
#> 29 0.1861055 3  TRUE
#> 30 0.1861055 3  TRUE