Generate null residuals from a null model

This function uses the data and the null model to generate null residuals. The null residuals are generated by first regressing random noises on the original regressors, then multiply the obtained residuals by original RSS divided by the current RSS. The results are the null residuals.

Arguments

dat

Data frame. A data frame that used to fit the model.

mod

Linear Model. The null model.

test

Boolean. Whether or not to use test on the newly generated data.

Value

A data frame with updated y, .resid, and potentially updated test result.

Examples

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

dat <- test$gen(10)
test$null_resid(dat, test$fit(dat), test = TRUE)
#>             y          x          e      .resid test_name statistic    p_value
#> 1  -0.2922812 0.42488680 -0.9548376 -1.38241455    F-test   5.61281 0.04966923
#> 2   0.9060785 0.59644345 -0.6856668 -0.14107374    F-test   5.61281 0.04966923
#> 3   0.1361617 0.16278370 -0.2936651 -1.01963806    F-test   5.61281 0.04966923
#> 4   2.3670261 0.11029014 -0.7859837  1.19807479    F-test   5.61281 0.04966923
#> 5   1.1546019 0.70637704 -1.8413646  0.13499202    F-test   5.61281 0.04966923
#> 6   0.3935047 0.08080247  1.0169591 -0.78283431    F-test   5.61281 0.04966923
#> 7   0.2734278 0.27545435  1.9702283 -0.85414386    F-test   5.61281 0.04966923
#> 8   1.0892173 0.67458660 -0.2986945  0.06164277    F-test   5.61281 0.04966923
#> 9   2.1394801 0.67282937 -1.8102191  1.11146535    F-test   5.61281 0.04966923
#> 10  2.8526833 0.07116442 -1.1416138  1.67392959    F-test   5.61281 0.04966923
#>    null
#> 1  TRUE
#> 2  TRUE
#> 3  TRUE
#> 4  TRUE
#> 5  TRUE
#> 6  TRUE
#> 7  TRUE
#> 8  TRUE
#> 9  TRUE
#> 10 TRUE