Calculate p-value for a visual test

This function calculate the p-value for a visual test. The null distribution is simulated by using function sim_dist(). The p-value is the probability of the number of detections greater or equal to the observed value.

calc_p_value(
  n_detect,
  n_eval,
  n_sel,
  n_plot = 20,
  n_sim = 50000,
  cache_env = NULL,
  seed = NULL
)

Arguments

n_detect

Integer. Observed value of number of detections.

n_eval

Integer. Number of evaluations.

n_sel

Integer. A vector of the number of selections.

n_plot

Integer. Number of plots in the lineup.

n_sim

Integer. Number of simulations draws.

cache_env

Environment. A provided environment for caching.

seed

Integer. set.seed() will be run at the beginning of the function if seed is provided.

Value

A numeric value representing the p-value.

Details

It is encouraged to provide a cache environment to boost up the performance when this function needs to be reused. The cache environment will remember the result corresponding to the combinations of n_eval and n_sim.

Examples

calc_p_value(2, 3, c(1,1,2))
#> [1] 0.01534
calc_p_value(1, 1, c(1))
#> [1] 0.05068