Calculate p-value for all combinations of evaluations of a lineup.

This function calculate p-value for all combinations of evaluations given the desired number of evaluations. For a lineup that being evaluated by \(K\) subjects, all combinations of \(K\) evaluations, taken \(X\) at a time will be generated, where \(X\) is the desired number of evaluations.

calc_p_value_comb(
  detected,
  n_eval,
  n_sel,
  n_plot = 20,
  n_sim = 50000,
  cache_env = NULL
)

Arguments

detected

Boolean. A vector of Boolean values indicating whether the lineup has been detected by the subjects.

n_eval

Integer. Desired number of evaluations.

n_sel

Integer. A vector of the number of selections.

n_plot

Integer. Number of plots in the lineup.

n_sim

Integer. Number of simulation draws.

cache_env

Environment. A provided environment for caching.

Value

A vector of p-value with the combination matrix as the attribute.

Details

It is encouraged to provide a cache environment to boost up the performance when this function needs to be reused. The cache environment will remember the result corresponding to the combinations of n_eval and n_sim.

See also

Examples

calc_p_value_comb(c(TRUE, FALSE, TRUE), 2, c(1, 1, 2))
#> [1] 0.09574 0.00660 0.14480
#> attr(,"combinations")
#>      [,1] [,2] [,3]
#> [1,]    1    1    2
#> [2,]    2    3    3
calc_p_value_comb(c(TRUE, FALSE), 1, c(1, 1))
#> [1] 0.05098 1.00000
#> attr(,"combinations")
#>      [,1] [,2]
#> [1,]    1    2